### 2024-06-27 #6936 @editorialbot
[PRE REVIEW]: ExaDEM: a HPC application based on exaNBody targeting scalable DEM simulations with complex particle shapes
**Submitting author:** @rprat-pro (<a href="http://orcid.org/0009-0002-3808-5401">Raphaël Prat</a>)
**Repository:** https://github.com/Collab4exaNBody/exaDEM
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.0.0
