+++ doc/paper/joss/paper.bib
@@ -12,6 +12,7 @@
   title={Simulation of liquids and solids},
   author={Ciccotti, Giovanni and Frenkel, Daan and Mc Donald, Ian R},
   year={1987},
+  doi = {},
   publisher={Elsevier Science Pub. Co. Inc., New York, NY}
 }
 
@@ -32,6 +33,7 @@
   volume={271},
   pages={108171},
   year={2022},
+  doi = {10.1016/j.cpc.2021.108171},
   publisher={Elsevier}
 }
 
@@ -43,6 +45,7 @@
   number={1},
   pages={1--19},
   year={1995},
+  doi = {10.2172/10176421},
   publisher={Elsevier}
 }
 
@@ -55,6 +58,7 @@
   number={1},
   pages={98},
   year={1967},
+  doi={10.1103/physrev.159.98 },
   publisher={APS}
 }
 
@@ -66,6 +70,7 @@
   number={2-3},
   pages={140--152},
   year={2012},
+  doi={10.1504/pcfd.2012.047457},
   publisher={Inderscience Publishers}
 }
 
@@ -75,6 +80,7 @@
   booktitle={Supercomputing Frontiers: 4th Asian Conference, SCFA 2018, Singapore, March 26-29, 2018, Proceedings 4},
   pages={159--178},
   year={2018},
+  doi={10.1007/978-3-319-69953-0_10},
   organization={Springer}
 }
 
@@ -85,6 +91,7 @@
   volume={319},
   pages={318--336},
   year={2018},
+  doi={10.1016/j.amc.2017.03.037},
   publisher={Elsevier}
 }
 
@@ -94,6 +101,7 @@
   booktitle={Euro-Par 2014: Parallel Processing Workshops: Euro-Par 2014 International Workshops, Porto, Portugal, August 25-26, 2014, Revised Selected Papers, Part II 20},
   pages={121--132},
   year={2014},
+  doi={10.1007/978-3-319-14313-2_11},
   organization={Springer}
 }
 
@@ -102,6 +110,7 @@
   author={Prat, Rapha{\"e}l and Colombet, Laurent and Namyst, Raymond},
   booktitle={Proceedings of the 47th International Conference on Parallel Processing},
   pages={1--10},
+  doi={10.1145/3225058.3225085},
   year={2018}
 }
 
@@ -112,6 +121,7 @@
   volume={253},
   pages={107177},
   year={2020},
+  doi={10.1016/j.cpc.2020.107177},
   publisher={Elsevier}
 }
 
@@ -122,6 +132,7 @@
   volume={249},
   pages={107129},
   year={2020},
+  doi = {10.1016/j.cpc.2019.107129},
   publisher={Elsevier}
 }
 
@@ -129,6 +140,7 @@
   title={Yade documentation},
   author={Smilauer, Vaclav and Angelidakis, Vasileios and Catalano, Emanuele and Caulk, Robert and Chareyre, Bruno and Chevremont, William and Dorofeenko, Sergei and Duriez, Jerome and Dyck, Nolan and Elias, Jan and others},
   journal={arXiv preprint arXiv:2301.00611},
+  doi = {10.48550/arXiv.2301.00611},
   year={2023}
 }
 
@@ -138,7 +150,6 @@
   editor = {Richefeu, Vincent and Villard, Pascal},
   year = {2016},
   publisher = {{Elsevier}},
-  doi = {10.1016/B978-1-78548-076-8},
   isbn = {978-1-78548-076-8}
 }
 
