+++ doc/paper/joss/paper.bib
@@ -150,6 +150,7 @@
   editor = {Richefeu, Vincent and Villard, Pascal},
   year = {2016},
   publisher = {{Elsevier}},
+  doi = {10.1016/c2015-0-01294-1},
   isbn = {978-1-78548-076-8}
 }
 
