### 2024-06-27 #6936 @danielskatz
That's fine - not all papers have DOIs. As you can see above, we've also moved this to another track, so another track editor will take over.
