### 2024-06-28 b28c0b9
update classifier, it works with spheres and polyhedra
