### 2024-06-28 fada508
rename hooke sphere operators
