+++ src/driver/drive_ball_wall_radius.cpp
@@ -1,244 +1,0 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-//#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
-
-#include <mpi.h>
-#include <exanb/core/operator.h>
-#include <exanb/core/operator_slot.h>
-#include <exanb/core/operator_factory.h>
-#include <exanb/core/make_grid_variant_operator.h>
-#include <exanb/core/parallel_grid_algorithm.h>
-#include <exanb/core/grid.h>
-#include <exanb/core/domain.h>
-#include <exanb/compute/compute_cell_particles.h>
-#include <memory>
-#include <exaDEM/common_compute_kernels.h>
-#include <exaDEM/compute_hooke_force.h>
-//#include <exaDEM/ball_wall.h>
-
-namespace exaDEM
-{
-  using namespace exanb;
-  struct BallWallFunctor
-  {
-    Vec3d m_ball_center ;
-    double m_ball_radius;
-    double m_sigma;
-    double m_dt;
-    double m_kt;
-    double m_kn;
-    double m_kr;
-    double m_mu;
-    double m_dampRate;
-
-    ONIKA_HOST_DEVICE_FUNC inline void operator() (
-        const double a_rx, const double a_ry, const double a_rz,
-        const double a_vx, const double a_vy, const double a_vz,
-        const Vec3d& a_vrot, 
-        const double a_particle_radius,
-        double& a_fx, double& a_fy, double& a_fz, 
-        const double a_mass,
-        Vec3d& a_mom,
-        Vec3d& a_ft, 
-        double& ball_f) const
-    {
-      const Vec3d m_ball_angular_velocity = {0.0 , 0.0 , 0.0};
-      const Vec3d m_ball_velocity = {0.0,0.0,0.0}; // TODO : do not touch
-
-      Vec3d pos   = Vec3d{a_rx, a_ry, a_rz};
-      Vec3d vel   = Vec3d{a_vx, a_vy, a_vz};
-      // === direction
-      const auto dir  = pos - m_ball_center;
-
-      // === interpenetration
-      const double dn = m_ball_radius - ( norm(dir) + a_particle_radius );
-      if(dn > 0.0)
-      {
-        a_ft = {0.0,0.0,0.0};
-        return;
-      }
-
-      // === figure out the contact position 
-      const auto dir_norm   = dir / norm(dir) ;
-      const auto contact  = m_ball_center + dir_norm * m_ball_radius ; // compute contact position between the particle and the ball
-
-      // === compute damp
-      const double meff = 1;//mass; // mass ball >>>>> mass i
-      const double damp = exaDEM::compute_damp(m_dampRate, m_kn, meff);
-
-      // === relative velocity  
-      const auto total_vel = exaDEM::compute_relative_velocity(contact,
-          m_ball_center, m_ball_velocity , m_ball_angular_velocity,
-          pos     , vel               , a_vrot);
-
[…]
