+++ src/interaction/include/exaDEM/interaction/classifier.hpp
@@ -6,121 +6,193 @@
 
 namespace exaDEM
 {
-	struct Classifier
-	{
-		static constexpr int types = 13;
-		template <typename T> using VectorT =  onika::memory::CudaMMVector<T>;
-		std::vector<VectorT<exaDEM::Interaction>> waves;
+  /**
+   * @brief Classifier for managing interactions categorized into different types.
+   *
+   * The Classifier struct manages interactions categorized into different types (up to 13 types).
+   * It provides functionalities to store interactions in CUDA memory-managed vectors (`VectorT`).
+   */
+  struct Classifier
+  {
+    static constexpr int types = 13;
+    template <typename T> using VectorT =  onika::memory::CudaMMVector<T>;
+    std::vector<VectorT<exaDEM::Interaction>> waves; ///< Storage for interactions categorized by type.
 
-		Classifier() { waves.resize(types); }
+    /**
+     * @brief Default constructor.
+     *
+     * Initializes the waves vector to hold interactions for each type.
+     */
+    Classifier() { waves.resize(types); }
 
-		void initialize() { waves.resize(types); }
+    /**
+     * @brief Initializes the waves vector to hold interactions for each type.
+     */
+    void initialize() { waves.resize(types); }
 
-		void reset_waves()
-		{
-			for(auto& wave : waves)
-			{
-				wave.clear();
-			}
-		}
 
-		VectorT<exaDEM::Interaction>& get_wave(size_t id) {return waves[id];}
+    /**
+     * @brief Clears all stored interactions in the waves vector.
+     */
+    void reset_waves()
+    {
+      for(auto& wave : waves)
+      {
+        wave.clear();
+      }
+    }
 
-		std::pair<exaDEM::Interaction*, size_t> get_info(size_t id) 
-		{
-			const unsigned int  data_size = onika::cuda::vector_size( waves[id]);
-			exaDEM::Interaction* const data_ptr = onika::cuda::vector_data( waves[id] );
-			return {data_ptr, data_size};
-		}
+    /**
+     * @brief Retrieves the CUDA memory-managed vector of interactions for a specific type.
+     *
+     * @param id Type identifier for the interaction wave.
+     * @return Reference to the CUDA memory-managed vector storing interactions of the specified type.
+     */
+    VectorT<exaDEM::Interaction>& get_wave(size_t id) {return waves[id];}
 
-		size_t number_of_waves() {return waves.size();}
+    /**
+     * @brief Retrieves the pointer and size of the data stored in the CUDA memory-managed vector for a specific type.
+     *
+     * @param id Type identifier for the interaction wave.
+     * @return Pair containing the pointer to the interaction data and the size of the data.
+     */
+    std::pair<exaDEM::Interaction*, size_t> get_info(size_t id) 
+    {
+      const unsigned int  data_size = onika::cuda::vector_size( waves[id]);
+      exaDEM::Interaction* const data_ptr = onika::cuda::vector_data( waves[id] );
+      return {data_ptr, data_size};
+    }
 
-		void classify(GridCellParticleInteraction& ges)
-		{
-			reset_waves();
-			// first loop to figure out the number of interactions per wave
-			std::array<int, types> sizes;
-			std::array<int, types> shifts;
+    /**
+     * @brief Returns the number of interaction types managed by the classifier.
+     *
+     * @return Number of interaction types.
+     */
+    size_t number_of_waves() {return waves.size();}
 
-			for(int  w = 0 ; w < types ; w++) {sizes[w] = 0 ; shifts[w]=0;}
+    /**
+     * @brief Classifies interactions into categorized waves based on their types.
+     *
[…]
