+++ src/interaction/include/exaDEM/interaction/classifier_for_all.hpp
@@ -8,6 +8,9 @@ namespace exaDEM
 {
 	using namespace onika::parallel;
 
+	/**
+	 * @brief Namespace for utilities related to tuple manipulation.
+	 */
 	namespace tuple_helper
 	{
 		template <size_t... Is>
@@ -20,27 +23,60 @@ namespace exaDEM
 			struct gen_seq<0, Is...> : index<Is...> {};
 	}
 
-
+	/**
+	 * @brief Wrapper for applying a kernel function to elements of an array in parallel.
+	 *
+	 * This struct provides a mechanism to apply a kernel function to each element of
+	 * an array in parallel. It stores a pointer to the array (`ptr`),
+	 * the kernel function (`kernel`), and a tuple of parameters (`params`) to be passed
+	 * to the kernel function along with each element.
+	 *
+	 * @tparam T Type of elements in the array.
+	 * @tparam K Type of the kernel function.
+	 * @tparam Args Types of additional parameters passed to the kernel function.
+	 */
 	template<typename T, typename K, typename... Args>
 		struct WrapperForAll
 		{
-			T* const ptr;
-      uint64_t size;
-			const K kernel;
-			std::tuple<Args...> params;
+			T* const ptr;              /**< Pointer to the array of elements. */
+			const K kernel;            /**< Kernel function to be applied. */
+			std::tuple<Args...> params; /**< Tuple of parameters to be passed to the kernel function. */
 
-			WrapperForAll(T* p, uint64_t s, K& k, Args... args) 
-				: ptr(p), size(s), 
+
+			/**
+			 * @brief Constructor to initialize the WrapperForAll struct.
+			 *
+			 * @param p Pointer to the array of elements.
+			 * @param k Kernel function to be applied.
+			 * @param args Additional parameters passed to the kernel function.
+			 */
+			WrapperForAll(T* p, K& k, Args... args) 
+				: ptr(p),
 				kernel(k), 
 				params(std::tuple<Args...>(args...)) 
 			{} 
 
-			template <size_t... Is>
+
+			/**
+			 * @brief Helper function to apply the kernel function to a single element.
+			 *
+			 * @tparam Is Index sequence for unpacking the parameter tuple.
+			 * @param item Reference to the element from the array.
+			 * @param indexes Index sequence to unpack the parameter tuple.
+			 */
+			templat
+				template <size_t... Is>
 				ONIKA_HOST_DEVICE_FUNC inline void apply(T& item, tuple_helper::index<Is...> indexes) const
 				{
 					kernel(item, std::get<Is>(params)...);
 				}
 
+
+			/**
+			 * @brief Functor operator to apply the kernel function to each element in the array.
+			 *
+			 * @param i Index of the element in the array.
+			 */
 			ONIKA_HOST_DEVICE_FUNC inline void operator()(uint64_t i) const 
 			{
 				T& item = ptr[i];
@@ -51,13 +87,13 @@ namespace exaDEM
 
 namespace onika
 {
-  namespace parallel
+	namespace parallel
 	{
-  	template<typename T, typename K, typename... Args> struct ParallelForFunctorTraits < exaDEM::WrapperForAll <T,K,Args... > >
-  	{
-	  	static inline constexpr bool CudaCompatible = true;
-	  };
-  }
+		template<typename T, typename K, typename... Args> struct ParallelForFunctorTraits < exaDEM::WrapperForAll <T,K,Args... > >
+		{
+			static inline constexpr bool CudaCompatible = true;
+		};
+	}
 }
 
 namespace exaDEM
