### 2024-06-28 c3b0ec8
update example with new operator names
