+++ src/meshset/include/exaDEM/hooke_stl_meshes.h
@@ -1,219 +1,0 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-#pragma once
-#include <exanb/core/basic_types.h>
-#include <exaDEM/common_compute_kernels.h>
-#include <exaDEM/compute_hooke_force.h>
-#include <exaDEM/face.h>
-#include <exaDEM/stl_mesh.h>
-
-namespace exaDEM
-{
-	using exanb::Vec3d; 
-
-	/**
-	 * @brief Functor for applying Hooke's law to multiple Faces within a grid.
-	 *
-	 * The `ApplyHookeSTLMeshesFunctor` struct represents a functor used to apply Hooke's law to particles interacting
-	 * with multiple STL meshes within a grid. It is designed to be used as an operator in simulations. The functor takes
-	 * various parameters and updates forces and torques acting on the particles.
-	 *
-	 * @tparam GridT The type of grid.
-	 */
-	struct ApplyHookeSTLMeshesFunctor
-	{
-		std::vector<exaDEM::stl_mesh>& meshes; /**< A collection of STL meshes. */
-		double m_dt; /**< Time step. */
-		double m_kt; /**< Tangential spring constant. */
-		double m_kn; /**< Normal spring constant. */
-		double m_kr; /**< Rotational spring constant. */
-		double m_mu; /**< Friction coefficient. */
-		double m_dampRate; /**< Damping rate. */
-
-		/**
-		 * @brief Operator for applying Hooke's law.
-		 *
-		 * This operator applies Hooke's law to particles interacting with multiple STL meshes within a grid. It updates
-		 * forces and torques acting on the particles based on the specified parameters and properties.
-		 *
-		 * @param cell_idx The cell index.
-		 * @param a_rx The x-coordinate of the particle's position.
-		 * @param a_ry The y-coordinate of the particle's position.
-		 * @param a_rz The z-coordinate of the particle's position.
-		 * @param a_vx The x-component of the particle's velocity.
-		 * @param a_vy The y-component of the particle's velocity.
-		 * @param a_vz The z-component of the particle's velocity.
-		 * @param a_vrot The rotational velocity of the particle.
-		 * @param a_particle_radius The radius of the particle.
-		 * @param a_fx Reference to store the x-component of the calculated force.
-		 * @param a_fy Reference to store the y-component of the calculated force.
-		 * @param a_fz Reference to store the z-component of the calculated force.
-		 * @param a_mass The mass of the particle.
-		 * @param a_mom Reference to store the angular momentum.
-		 * @param a_ft Reference to store the torque.
-		 */
-		ONIKA_HOST_DEVICE_FUNC inline void operator () (
-				const size_t cell_idx,
[…]
