+++ src/interaction/include/exaDEM/interaction/classifier_for_all.hpp
@@ -6,104 +6,103 @@
 
 namespace exaDEM
 {
-	using namespace onika::parallel;
+  using namespace onika::parallel;
 
-	/**
-	 * @brief Namespace for utilities related to tuple manipulation.
-	 */
-	namespace tuple_helper
-	{
-		template <size_t... Is>
-			struct index {};
+  /**
+   * @brief Namespace for utilities related to tuple manipulation.
+   */
+  namespace tuple_helper
+  {
+    template <size_t... Is>
+      struct index {};
 
-		template <size_t N, size_t... Is>
-			struct gen_seq : gen_seq<N - 1, N - 1, Is...> {};
+    template <size_t N, size_t... Is>
+      struct gen_seq : gen_seq<N - 1, N - 1, Is...> {};
 
-		template <size_t... Is>
-			struct gen_seq<0, Is...> : index<Is...> {};
-	}
+    template <size_t... Is>
+      struct gen_seq<0, Is...> : index<Is...> {};
+  }
 
-	/**
-	 * @brief Wrapper for applying a kernel function to elements of an array in parallel.
-	 *
-	 * This struct provides a mechanism to apply a kernel function to each element of
-	 * an array in parallel. It stores a pointer to the array (`ptr`),
-	 * the kernel function (`kernel`), and a tuple of parameters (`params`) to be passed
-	 * to the kernel function along with each element.
-	 *
-	 * @tparam T Type of elements in the array.
-	 * @tparam K Type of the kernel function.
-	 * @tparam Args Types of additional parameters passed to the kernel function.
-	 */
-	template<typename T, typename K, typename... Args>
-		struct WrapperForAll
-		{
-			T* const ptr;              /**< Pointer to the array of elements. */
-			const K kernel;            /**< Kernel function to be applied. */
-			std::tuple<Args...> params; /**< Tuple of parameters to be passed to the kernel function. */
-
-
-			/**
-			 * @brief Constructor to initialize the WrapperForAll struct.
-			 *
-			 * @param p Pointer to the array of elements.
-			 * @param k Kernel function to be applied.
-			 * @param args Additional parameters passed to the kernel function.
-			 */
-			WrapperForAll(T* p, K& k, Args... args) 
-				: ptr(p),
-				kernel(k), 
-				params(std::tuple<Args...>(args...)) 
-			{} 
+  /**
+   * @brief Wrapper for applying a kernel function to elements of an array in parallel.
+   *
+   * This struct provides a mechanism to apply a kernel function to each element of
+   * an array in parallel. It stores a pointer to the array (`ptr`),
+   * the kernel function (`kernel`), and a tuple of parameters (`params`) to be passed
+   * to the kernel function along with each element.
+   *
+   * @tparam T Type of elements in the array.
+   * @tparam K Type of the kernel function.
+   * @tparam Args Types of additional parameters passed to the kernel function.
+   */
+  template<typename T, typename K, typename... Args>
+    struct WrapperForAll
+    {
+      T* const ptr;              /**< Pointer to the array of elements. */
+      const K kernel;            /**< Kernel function to be applied. */
+      std::tuple<Args...> params; /**< Tuple of parameters to be passed to the kernel function. */
 
 
-			/**
-			 * @brief Helper function to apply the kernel function to a single element.
-			 *
-			 * @tparam Is Index sequence for unpacking the parameter tuple.
-			 * @param item Reference to the element from the array.
-			 * @param indexes Index sequence to unpack the parameter tuple.
-			 */
-			templat
-				template <size_t... Is>
[…]
