+++ src/CMakeLists.txt
@@ -26,7 +26,6 @@ add_subdirectory(friction)
 add_subdirectory(forcefield)
 add_subdirectory(io)
 add_subdirectory(driver)
-add_subdirectory(meshset)
 add_subdirectory(interaction)
 add_subdirectory(sphere)
 add_subdirectory(polyhedron)
