### 2024-06-28 145c55b
remove old friction data structures
