+++ src/friction/migrate_cell_particles_friction.cpp
@@ -1,158 +1,0 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-#pragma xstamp_grid_variant
-
-#include <exanb/core/operator.h>
-#include <exanb/core/operator_slot.h>
-#include <exanb/core/operator_factory.h>
-#include <exanb/core/log.h>
-#include <exanb/core/basic_types_stream.h>
-#include <exanb/core/grid.h>
-#include <exanb/core/domain.h>
-#include <exanb/core/make_grid_variant_operator.h>
-#include <exanb/grid_cell_particles/grid_cell_values.h>
-
-#include <exanb/mpi/migrate_cell_particles.h>
-#include <exaDEM/neighbor_friction.h>
-
-#include <mpi.h>
-#include <vector>
-#include <cstring>
-
-
-namespace exaDEM
-{
-	using namespace exanb;
-
-	// operator interface
-	template<class GridT, class CellValueMergeOperatorT = exanb::UpdateValueAdd >
-		class MigrateCellParticlesFriction : public OperatorNode
-	{
-		using CellParticles = typename GridT::CellParticles;
-		using ParticleTuple = typename CellParticles::TupleValueType; // decltype( GridT().cells()[0][0] );
-		using ParticleBuffer = std::vector<ParticleTuple>;
-		//using BufListIt = typename std::list<ParticleBuffer>::iterator;
-		//using GridCellValueType = typename GridCellValues::GridCellValueType;
-		using MergeOp = CellValueMergeOperatorT;
-
-		// -----------------------------------------------
-		// Operator slots
-		// -----------------------------------------------
-		ADD_SLOT( MPI_Comm  , mpi        , INPUT , MPI_COMM_WORLD );
-		ADD_SLOT( long      , mpi_tag    , INPUT , 0 );
-		ADD_SLOT( MergeOp   , merge_func , INPUT, MergeOp{} );
-
-		ADD_SLOT( GridBlock , lb_block   , INPUT , REQUIRED );
-		ADD_SLOT( double    , ghost_dist , INPUT , REQUIRED );
-		ADD_SLOT( Domain    , domain     , INPUT , REQUIRED );
-		ADD_SLOT( GridT     , grid       , INPUT_OUTPUT);
-
-		// optional particles "outside the box", not in the grid. (possibly due to move_particles operator)
-		// it is in/out because elements of this array may be swapped (order changed), though it will contain the same element after operation
-		ADD_SLOT( ParticleBuffer , otb_particles, INPUT_OUTPUT , DocString{"Particles outside of local processor's grid"} );
-
-		// MPI buffer size for exchanges
-		ADD_SLOT( long , buffer_size, INPUT , 100000 , DocString{"Performance tuning parameter. Size of send/receive buffers in number of particles"} );
-
-		// threshold, in number of particles, above which a task is generated to copy cell content to send buffer
[…]
