+++ src/CMakeLists.txt
@@ -22,7 +22,6 @@
 add_subdirectory(grid)
 add_subdirectory(field)
 add_subdirectory(numericalscheme)
-add_subdirectory(friction)
 add_subdirectory(forcefield)
 add_subdirectory(io)
 add_subdirectory(driver)
