+++ src/driver/CMakeLists.txt
@@ -14,5 +14,5 @@
 # kind, either express or implied.  see the license for the
 # specific language governing permissions and limitations
 # under the license.
-set(exadem_driver_LINK_LIBRARIES exanbIO exanbDefBox exanbParticleNeighbors exadem_numerical_scheme exadem_friction exadem_force_field exadem_io exadem_interaction)
+set(exadem_driver_LINK_LIBRARIES exanbIO exanbDefBox exanbParticleNeighbors exadem_numerical_scheme exadem_force_field exadem_io exadem_interaction)
 xstamp_add_plugin(exadem_driver ${CMAKE_CURRENT_SOURCE_DIR})
