### 2024-06-28 060a7b8
update stl mesh examples with spheres
