+++ example/spheres/mesh-stl/CMakeLists.txt
@@ -1,5 +1,5 @@
 
-set(ExampleName "QuadraticForce")
+set(ExampleName "STLMesh")
 set(FileName "mesh_stl.msp")
 
 file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/particles.xyz DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
