### 2024-06-28 81e10e7
Integration first dev carlo gpu (#35)
* update data structure for gpu + add gpu parallelization with classifier for spherical particles

* refactoring sym/no sym stuff for the contact law

* update hooke force with a classifier (block parallel for)

* fix issue in classifier sphere gpu (// exec context)

* remove size from wrapper

* fix the computation of the cylinder center

* move for all function using the interaction classifier in the interaction directory

* add classifier for polyhedra with gpu

* update classifier, it works with spheres and polyhedra

* rename hooke sphere operators

*
[…]
