+++ src/interaction/include/exaDEM/type/OBB.hpp
@@ -28,277 +28,288 @@
 /// @brief Oriented Bounding Box
 class OBB
 {
-public:
-	vec3r center;  //< Center
-	vec3r e[3];    //< 3 directions (normalized vectors)
-	vec3r extent;  //< 3 extents (in the the 3 directions)
-	
-	// Constructors
-	OBB(): center(), extent(0.0, 0.0, 0.0) {
-		e[0].set(1.0, 0.0, 0.0);
-		e[1].set(0.0, 1.0, 0.0);
-		e[2].set(0.0, 0.0, 1.0);
-	}
+	public:
+		vec3r center;  //< Center
+		vec3r e1;    //< 3 directions (normalized vectors)
+		vec3r e2;    //< 3 directions (normalized vectors)
+		vec3r e3;    //< 3 directions (normalized vectors)
+		vec3r extent;  //< 3 extents (in the the 3 directions)
 
-	OBB(const OBB &obb): center(obb.center), extent(obb.extent) {
-		e[0] = obb.e[0];
-		e[1] = obb.e[1];
-		e[2] = obb.e[2];
-	}
-  
-	OBB & operator = (const OBB &obb) {
-		center = obb.center;
-		e[0] = obb.e[0];
-		e[1] = obb.e[1];
-		e[2] = obb.e[2];
-    extent = obb.extent;
-		return (*this);
-	}
+		// Constructors
+		ONIKA_HOST_DEVICE_FUNC
+			OBB(): center(), extent(0.0, 0.0, 0.0) {
+				e1.set(1.0, 0.0, 0.0);
+				e2.set(0.0, 1.0, 0.0);
+				e3.set(0.0, 0.0, 1.0);
+			}
 
-	void enlarge(double more) {
-		extent.x += more;
-		extent.y += more;
-		extent.z += more;
-	}
+		ONIKA_HOST_DEVICE_FUNC
+			OBB(const OBB &obb): center(obb.center), extent(obb.extent) {
+				e1 = obb.e1;
+				e2 = obb.e2;
+				e3 = obb.e3;
+			}
 
-	void translate(const vec3r & v) {
-		center += v;
-	}
+		ONIKA_HOST_DEVICE_FUNC
+			OBB & operator = (const OBB &obb) {
+				center = obb.center;
+				e1 = obb.e1;
+				e2 = obb.e2;
+				e3 = obb.e3;
+				extent = obb.extent;
+				return (*this);
+			}
 
-	void rotate(const quat & Q) {
-		e[0] = Q * e[0];
-		e[1] = Q * e[1];
-		e[2] = Q * e[2];
-		center = Q * center;
-	}
+		ONIKA_HOST_DEVICE_FUNC
+			void enlarge(double more) {
+				extent.x += more;
+				extent.y += more;
+				extent.z += more;
+			}
 
-	// see page 101 of the book 'Real-Time Collision Detection' (Christer Ericson)
-	bool intersect(const OBB& obb, double tol = FLT_EPSILON) const {
-		double  ra, rb;
-		mat9r R, AbsR;
+		ONIKA_HOST_DEVICE_FUNC
+			void translate(const vec3r & v) {
+				center += v;
+			}
 
-		// Compute first terms of rotation matrix expressing obb frame in this OBB coordinate frame
-		// (other terms will be computed later)
-		R.xx = e[0] * obb.e[0];
-		R.xy = e[0] * obb.e[1];
-		R.xz = e[0] * obb.e[2];
+		ONIKA_HOST_DEVICE_FUNC
+			void rotate(const quat & Q) {
[…]

[…]
