### 2024-06-30 0f9a7a6
update the exaDEM version in the spack repository (v1.0.1)
