+++ .github/workflows/cmake.yml
@@ -37,16 +37,16 @@ jobs:
       working-directory: ${{github.workspace}}/build
       run: make UpdatePluginDataBase && ctest --test-dir example -C --rerun-failed --output-on-failure ${{env.BUILD_TYPE}}
 
-#    - name: Generate a code coverage report
-#      uses: threeal/gcovr-action@v1.0.0
-#      with:
-#        root: ${{github.workspace}}
+    - name: Generate a code coverage report
+      uses: threeal/gcovr-action@v1.0.0
+      with:
+        root: ${{github.workspace}}
 
-#    - name: Generate a code coverage report
-#      run: gcovr --html-details coverage.html
+    - name: Generate a code coverage report
+      run: gcovr --html-details coverage.html
       
-#    - name: Archive code coverage results
-#      uses: actions/upload-artifact@v4
-#      with:
-#        name: code-coverage-report
-#        path: coverage.html        
+    - name: Archive code coverage results
+      uses: actions/upload-artifact@v4
+      with:
+        name: code-coverage-report
+        path: coverage.html        
