### 2024-07-03 9a9f793
Update paper.md
typo
