+++ doc/paper/joss/paper.md
@@ -63,7 +63,7 @@ Several `DEM` software packages have emerged in recent years, offering `HPC` cap
 
 # Implementation Details
 
-`ExaDEM` leverages the `exaNBody` data structures (grid, cells, fields) as well as key parallel algorithms (domain decomposition, particles migration, numerical schemes) while proposing `DEM`-specific mechanism. `ExaDEM` achieves a `MPI` parallelization where the simulation domain is decomposed into subdomains using spatial domain decomposition and the Recursive Coordinate Bisection (`RCB`) partitioning method to evenly distribute the workload among `MPI` processes. Particle information stored within these cells, and a subdomain corresponds to a grid of cells. The use of cells aims to apply the state-of-the-art linked cells method to expedite neighbor searches with a complexity of O(N), where N represetns the number of particles. Additionally, the Verlet lists method maintains larger neighbor lists over timesteps as long as a particle has a displacement shorter than half of the Verlet radius. Regarding the data layout, it is decomposed into two levels. The first level consists of cells (`SOA`), each composed of fields (`Array`) containing particle data. The second level is associated with the grid of cells (`AOSOA`), corresponding to a subdomain. The `DEM` grid includes the following fields: type, position, velocity, acceleration, radius, angular velocity, orientation. The `AOSAO` data structure facilitates data movement between `MPI` processes while maintaining good data locality, ensuring that particles in the same cell or neighboring cells interact with a good memory hit ratio. Moreover, the use of `SOA` storage (cell layout) enhances the use of `SIMD` instructions.
+`ExaDEM` leverages the `exaNBody` data structures (grid, cells, fields) as well as key parallel algorithms (domain decomposition, particles migration, numerical schemes) while proposing `DEM`-specific mechanism. `ExaDEM` achieves a `MPI` parallelization where the simulation domain is decomposed into subdomains using spatial domain decomposition and the Recursive Coordinate Bisection (`RCB`) partitioning method to evenly distribute the workload among `MPI` processes. Particle information stored within these cells, and a subdomain corresponds to a grid of cells. The use of cells aims to apply the state-of-the-art linked cells method to expedite neighbor searches with a complexity of O(N), where N represents the number of particles. Additionally, the Verlet lists method maintains larger neighbor lists over timesteps as long as a particle has a displacement shorter than half of the Verlet radius. Regarding the data layout, it is decomposed into two levels. The first level consists of cells (`SOA`), each composed of fields (`Array`) containing particle data. The second level is associated with the grid of cells (`AOSOA`), corresponding to a subdomain. The `DEM` grid includes the following fields: type, position, velocity, acceleration, radius, angular velocity, orientation. The `AOSAO` data structure facilitates data movement between `MPI` processes while maintaining good data locality, ensuring that particles in the same cell or neighboring cells interact with a good memory hit ratio. Moreover, the use of `SOA` storage (cell layout) enhances the use of `SIMD` instructions.
 
 Regarding the intra-`MPI` parallelization, we distinguish two main differences based on the type of particle, i.e. sphere or polyhedron: 
 
