+++ .github/workflows/cmake.yml
@@ -21,10 +21,11 @@ jobs:
     - uses: actions/setup-node@v4
     - name: Get Yaml cpp
       run: sudo apt install libyaml-cpp-dev
+    - name: Install
+      run: sudo apt-get update && sudo apt-get install lcov
     - uses: mpi4py/setup-mpi@v1
     - name: Get ExaNBody
       run: git clone https://github.com/Collab4exaNBody/exaNBody.git
-
       
     - name: Configure CMake
       run: cmake -DCMAKE_CXX_FLAGS="--coverage -g" -B ${{github.workspace}}/build -DCMAKE_BUILD_TYPE=${{env.BUILD_TYPE}} -DexaNBody_DIR=${{github.workspace}}/exaNBody
@@ -37,13 +38,13 @@ jobs:
       working-directory: ${{github.workspace}}/build
       run: make UpdatePluginDataBase && ctest --test-dir example -C --rerun-failed --output-on-failure ${{env.BUILD_TYPE}}
 
-    - name: Generate a code coverage report
+    - name: Generate a code coverage repor part1
       uses: threeal/gcovr-action@v1.0.0
       with:
         root: ${{github.workspace}}
 
-    - name: Generate a code coverage report
-      run: gcovr --gcov-ignore-errors=no_working_dir_found --html-details coverage.html
+    - name: Generate a code coverage report part2
+      run: gcovr --gcov-ignore-errors=no_working_dir_found --merge-mode-functions=MERGE_MODE --html-details coverage.html
       
     - name: Archive code coverage results
       uses: actions/upload-artifact@v4
