### 2024-07-05 d8b9bd1
Liste de voisins GPU
