+++ src/grid/chunk_neighbors_contact.cpp
@@ -44,6 +44,7 @@ under the License.
 //#include <exaDEM/interactions_PP.h>
 //#include <exaDEM/neighbor_friction.h>
 #include <../forcefield/include/exaDEM/interactions_PP.h>
+#include <exanb/particle_neighbors/neighbor_filter_func.h>
 
 namespace exaDEM
 {
@@ -108,24 +109,200 @@ namespace exaDEM
   	}
   }
   
-  /*__global__ void cellGPU(int* cell,
-  			int* cell_size,
-  			onika::memory::CudaMMVector<int> cell2,
-  			onika::memory::CudaMMVector<int> cell_size2,
-  			int size)
+  
+  template< class GridT > ONIKA_HOST_DEVICE_FUNC bool nbh_filter_GPU(GridT* cells,
+  						double rcut_inc,
+  						double d2,
+  						double rcut2,
+  						size_t cell_a,
+  						size_t p_a,
+  						size_t cell_b,
+  						size_t p_b)
+  {
+      const double r_a = cells[cell_a][field::radius][p_a];
+      const double r_b = cells[cell_b][field::radius][p_b];
+      const double rmax = r_a + r_b + rcut_inc;
+      const double rmax2 = rmax * rmax;
+      return d2 > 0.0 && d2 < rmax2 && d2 < rcut2 ;
+  }
+  
+  
+ template< class GridT > __global__ void neighbor_GPU(GridT* cells, 
+ 							int* cells2,
+ 							int* cells_size,
+ 							int* cells_start,
+ 							int* cells_tot,
+ 							Mat3d xform,
+ 							const double dist_lab,
+ 							double rcut_inc,
+ 							int* cells_b,
+ 							int* cells_b_size,
+ 							int* result,
+ 							int* result_start
+  							)
+  {
+  	int cell = cells2[blockIdx.x];
+  	int size = cells_size[blockIdx.x];
+  	int start = cells_start[blockIdx.x];
+  	int tot = cells_tot[blockIdx.x];
+  	int res_start = result_start[blockIdx.x];
+  	
+  	if(threadIdx.x < size)
+  	{
+  		int p_a = threadIdx.x;
+  		
+  		for(int i = start; i < start + tot; i++)
+  		{
+  			int cell_b = cells_b[i];
+  			
+  			for(int p_b = 0; p_b < cells_b_size[i]; p_b++)
+  			{
+  				double rcut2 = dist_lab * dist_lab;
+  				
+  			 	const Vec3d dr = { cells[cell][field::rx][p_a] - cells[cell_b][field::rx][p_b] , cells[cell][field::ry][p_a] - cells[cell_b][field::ry][p_b] , cells[cell][field::rz][p_a] - cells[cell_b][field::ry][p_b] };
+                    		double d2 = norm2(xform * dr );
+  				if(nbh_filter_GPU(cells, rcut_inc, d2, rcut2, cell, p_a, cell_b, p_b)) { /*result[res_start + threadIdx.x * p_b] = 1;*/}
+  				else{ /*result[res_start + threadIdx.x * p_b] = 0;*/}
+  				
+  				result_start[blockIdx.x]++;
+  				
+  			}
+  		}
+  	}
+  	
+  }
+  
+  template< class GridT > __global__ void kernel_GPU(GridT* cells,
+  							int cell,
+  							int size,
+  							int* result,
+  							int* cellb,
+  							int* pb,
+  							const double dist_lab,
+  							Mat3d xform,
+  							double rcut_inc)
   {
   	int idx = threadIdx.x + blockIdx.x * blockDim.x;
   	if(idx < size)
   	{
-  		auto cell3 = onika::cuda::vector_data(cell2);
[…]
