+++ .github/workflows/cmake.yml
@@ -45,8 +45,6 @@ jobs:
 
     - name: Generate a code coverage report part2
       run: gcovr --exclude-directories='.*/exanb*' --exclude-directories='.*/onika*' --html-details coverage.html
-      with:
-        root: ${{github.workspace}}/build
 
     - name: Archive code coverage results
       uses: actions/upload-artifact@v4
