+++ .github/workflows/cmake.yml
@@ -42,9 +42,11 @@ jobs:
       uses: threeal/gcovr-action@v1.0.0
       with:
         root: ${{github.workspace}}/build
+        excludes: |
+          exaNBody/src/*
 
     - name: Generate a code coverage report part2
-      run: gcovr --exclude-directories='.*/exanb*' --exclude-directories='.*/onika*' --html-details coverage.html
+      run: gcovr --exclude-directories='.*/exanb*' --exclude-directories='.*/onika*' --exclude-directories='exaNBody' --html-details coverage.html
 
     - name: Archive code coverage results
       uses: actions/upload-artifact@v4
