### 2024-07-07 4a7c066
Create Spack.yml
