### 2024-07-07 347f392
Update Spack.yml
