+++ .github/workflows/Spack.yml
@@ -13,4 +13,17 @@ jobs:
 
     steps:
     - uses: actions/checkout@v4
+    
+    - name: Set up Spack
+      uses: spack/setup-spack@v2
+      with:
+        ref: develop      # Spack version (examples: develop, releases/v0.21)
+        buildcache: true  # Configure oci://ghcr.io/spack/github-actions-buildcache
+        color: true       # Force color output (SPACK_COLOR=always)
+        path: spack       # Where to clone Spack
+        
+    - name: Install ExaDEM With Spack
+      run: |
+        spack repo add spack_repo
+        spack install exadem
 
