### 2024-07-07 c6ff841
Create Spack.yml (#39)
* Create Spack.yml

* Update Spack.yml

* Update package.py

* Rename Spack.yml to spack.yml
