### 2024-07-07 26dc6cd
Update spack.yml
