+++ .github/workflows/spack.yml
@@ -1,4 +1,4 @@
-name: C/C++ CI
+name: Check Spack Installation
 
 on:
   push:
