+++ src/grid/chunk_neighbors_contact.cpp
@@ -125,83 +125,7 @@ namespace exaDEM
       const double rmax2 = rmax * rmax;
       return d2 > 0.0 && d2 < rmax2 && d2 < rcut2 ;
   }
-  
-  
- template< class GridT > __global__ void neighbor_GPU(GridT* cells, 
- 							int* cells2,
- 							int* cells_size,
- 							int* cells_start,
- 							int* cells_tot,
- 							Mat3d xform,
- 							const double dist_lab,
- 							double rcut_inc,
- 							int* cells_b,
- 							int* cells_b_size,
- 							int* result,
- 							int* result_start
-  							)
-  {
-  	int cell = cells2[blockIdx.x];
-  	int size = cells_size[blockIdx.x];
-  	int start = cells_start[blockIdx.x];
-  	int tot = cells_tot[blockIdx.x];
-  	int res_start = result_start[blockIdx.x];
-  	
-  	if(threadIdx.x < size)
-  	{
-  		int p_a = threadIdx.x;
-  		
-  		for(int i = start; i < start + tot; i++)
-  		{
-  			int cell_b = cells_b[i];
-  			
-  			for(int p_b = 0; p_b < cells_b_size[i]; p_b++)
-  			{
-  				double rcut2 = dist_lab * dist_lab;
-  				
-  			 	const Vec3d dr = { cells[cell][field::rx][p_a] - cells[cell_b][field::rx][p_b] , cells[cell][field::ry][p_a] - cells[cell_b][field::ry][p_b] , cells[cell][field::rz][p_a] - cells[cell_b][field::ry][p_b] };
-                    		double d2 = norm2(xform * dr );
-  				if(nbh_filter_GPU(cells, rcut_inc, d2, rcut2, cell, p_a, cell_b, p_b)) { /*result[res_start + threadIdx.x * p_b] = 1;*/}
-  				else{ /*result[res_start + threadIdx.x * p_b] = 0;*/}
-  				
-  				result_start[blockIdx.x]++;
-  				
-  			}
-  		}
-  	}
-  	
-  }
-  
-  template< class GridT > __global__ void kernel_GPU(GridT* cells,
-  							int cell,
-  							int size,
-  							int* result,
-  							int* cellb,
-  							int* pb,
-  							const double dist_lab,
-  							Mat3d xform,
-  							double rcut_inc)
-  {
-  	int idx = threadIdx.x + blockIdx.x * blockDim.x;
-  	if(idx < size)
-  	{
-  		int p_a = (int)(threadIdx.x/size);
-  		int nbh = threadIdx.x - (p_a * size);
-  		
-  		int cell_b = cellb[nbh];
-  		int p_b = pb[nbh];
-  		
-  		double rcut2 = dist_lab * dist_lab;
-  		
-  		const Vec3d dr = { cells[cell][field::rx][p_a] - cells[cell_b][field::rx][p_b] , cells[cell][field::ry][p_a] - cells[cell_b][field::ry][p_b] , cells[cell][field::rz][p_a] - cells[cell_b][field::ry][p_b] };
-                double d2 = norm2(xform * dr );
-  		
-  		if(nbh_filter_GPU(cells, rcut_inc, d2, rcut2, cell, p_a, cell_b, p_b)) { result[idx] = 1;}
-  		else{ result[idx] = 1;}
-  		
-  	}
-  }
- 
+
   
   template< class GridT > __global__ void kernel_GPU2(GridT* cells,
   							int size,
@@ -234,9 +158,18 @@ namespace exaDEM
   	int start = cells_start[blockIdx.x];
   	int div = cells_256_div[blockIdx.x];
   	int interactions = 0;
+  	int total_interactions = 0;
+  	int total_interactions2 = 0;
+  	
+  	//__shared__ int shared_threads[256];
[…]
