### 2024-07-12 b252dc3
Update cmake.yml (#37)
* Update cmake.yml
