### 2024-07-13 af3c412
add the random feature in the set_quaternion operator
