### 2024-07-15 f9d883a
41 add a random orientation feature (#42)
* some updates to run generator.msp

* add the random feature in the set_quaternion operator

* remove debug output

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
