### 2024-07-19 1f8740d
fix compilation issues on GPU with quaternions
