### 2024-07-22 b0a17a0
remove old hooke force version for polyhedron
