+++ src/polyhedron/include/exaDEM/hooke_polyhedron.h
@@ -169,103 +169,6 @@ namespace exaDEM
 
       };
 
-      /**
-       * @brief Operator function for performing interactions between particles.
-       *
-       * This function performs interactions between particles.
-       * It uses precomputed shapes (`shps`) and synchronization (`locker`) for thread safety.
-       *
-       * @tparam TMPLC Type of the cells or particles container.
-       * @param item Reference to the Interaction object representing the interaction details.
-       * @param cells Reference to the cells or particles container.
-       * @param hkp Reference to the HookeParams object containing interaction parameters.
-       * @param shps Reference to the shapes object providing shape information for interactions.
-       * @param dt Simulation dt increment.
-       * @param locker Reference to mutexes object for synchronization.
-       */
-      template<typename TMPLC>
-        ONIKA_HOST_DEVICE_FUNC 
-        inline void operator()(
-            Interaction& item, 
-            TMPLC& cells, 
-            const HookeParams& hkp,
-            const shapes& shps, 
-            const double dt, 
-            mutexes& locker) const
-        {
-          // === cell
-          auto& cell_i =  cells[item.cell_i];
-          auto& cell_j =  cells[item.cell_j];
-
-          // === positions
-          const Vec3d ri = get_r(cell_i, item.p_i);
-          const Vec3d rj = get_r(cell_j, item.p_j);
-
-          // === vrot
-          const Vec3d& vrot_i = cell_i[field::vrot][item.p_i];
-          const Vec3d& vrot_j = cell_j[field::vrot][item.p_j];
-
-          // === type
-          const auto& type_i = cell_i[field::type][item.p_i];
-          const auto& type_j = cell_j[field::type][item.p_j];
-
-          // === vertex array
-          const auto& vertices_i =  cell_i[field::vertices][item.p_i];
-          const auto& vertices_j =  cell_j[field::vertices][item.p_j];
-
-          // === shapes
-          const shape* shp_i = shps[type_i];
-          const shape* shp_j = shps[type_j];
-
-          auto [contact, dn, n, contact_position] = detect(item.type, vertices_i, item.sub_i, shp_i, vertices_j, item.sub_j, shp_j);
-          if(contact)
-          {
-            const Vec3d vi = get_v(cell_i, item.p_i);
-            const Vec3d vj = get_v(cell_j, item.p_j);
-            const auto& m_i = cell_i[field::mass][item.p_i];
-            const auto& m_j = cell_j[field::mass][item.p_j];
-
-            // temporary vec3d to store forces.
-            Vec3d f = {0,0,0};
-            const double meff = compute_effective_mass(m_i, m_j);
-
-            hooke_force_core(dn, n, dt, hkp.m_kn, hkp.m_kt, hkp.m_kr,
-                hkp.m_mu, hkp.m_damp_rate, meff,
-                item.friction, contact_position,
-                ri, vi, f, item.moment, vrot_i,  // particle 1
-                rj, vj, vrot_j // particle nbh
-                );
-
-
-            // === update particle informations
-            // ==== Particle i
-            locker.lock(item.cell_i, item.p_i);
-
-            auto& mom_i = cell_i[field::mom][item.p_i];
-            mom_i += compute_moments(contact_position, ri, f, item.moment);
-            cell_i[field::fx][item.p_i] += f.x;
-            cell_i[field::fy][item.p_i] += f.y;
-            cell_i[field::fz][item.p_i] += f.z;
-
-            locker.unlock(item.cell_i, item.p_i);
-
-            // ==== Particle j
[…]
