### 2024-07-22 2dd13f1
fix issue in debug mode && remove old hooke force with mutexes && add buffers for analysis asked by lhassan in the hooke force operator
