+++ src/polyhedron/include/exaDEM/hooke_polyhedron.h
@@ -169,296 +169,315 @@ namespace exaDEM
 
       };
 
-			/**
-			 * @brief Operator function for performing interactions between particles (polyhedra).
-			 *
-			 * @tparam TMPLC Type of the cells or particles container.
-			 * @param item Reference to the Interaction object representing the interaction details.
-			 * @param cells Pointer to the cells or particles container.
-			 * @param hkp Reference to the HookeParams object containing interaction parameters.
-			 * @param shps Pointer to the shapes array providing shape information for interactions.
-			 * @param dt Time increment for the simulation step.
-			 */
-			template<typename TMPLC>
-				ONIKA_HOST_DEVICE_FUNC inline void operator()(
-						Interaction& item, 
-						TMPLC* const cells, 
-						const HookeParams& hkp, 
-						const shape* const shps, 
-						const double dt) const
-				{
-					// === cell
-					auto& cell_i =  cells[item.cell_i];
-					auto& cell_j =  cells[item.cell_j];
+      /**
+       * @brief Operator function for performing interactions between particles (polyhedra).
+       *
+       * @tparam TMPLC Type of the cells or particles container.
+       * @param item Reference to the Interaction object representing the interaction details.
+       * @param cells Pointer to the cells or particles container.
+       * @param hkp Reference to the HookeParams object containing interaction parameters.
+       * @param shps Pointer to the shapes array providing shape information for interactions.
+       * @param dt Time increment for the simulation step.
+       */
+      template<typename TMPLC>
+        ONIKA_HOST_DEVICE_FUNC inline std::tuple<Vec3d, Vec3d, Vec3d> operator()(
+            Interaction& item, 
+            TMPLC* const cells, 
+            const HookeParams& hkp, 
+            const shape* const shps, 
+            const double dt) const
+        {
+          // === cell
+          auto& cell_i =  cells[item.cell_i];
+          auto& cell_j =  cells[item.cell_j];
 
-					// === positions
-					const Vec3d ri = get_r(cell_i, item.p_i);
-					const Vec3d rj = get_r(cell_j, item.p_j);
+          // === positions
+          const Vec3d ri = get_r(cell_i, item.p_i);
+          const Vec3d rj = get_r(cell_j, item.p_j);
 
-					// === vrot
-					const Vec3d& vrot_i = cell_i[field::vrot][item.p_i];
-					const Vec3d& vrot_j = cell_j[field::vrot][item.p_j];
+          // === vrot
+          const Vec3d& vrot_i = cell_i[field::vrot][item.p_i];
+          const Vec3d& vrot_j = cell_j[field::vrot][item.p_j];
 
-					// === type
-					const auto& type_i = cell_i[field::type][item.p_i];
-					const auto& type_j = cell_j[field::type][item.p_j];
+          // === type
+          const auto& type_i = cell_i[field::type][item.p_i];
+          const auto& type_j = cell_j[field::type][item.p_j];
 
-					// === vertex array
-					const auto& vertices_i =  cell_i[field::vertices][item.p_i];
-					const auto& vertices_j =  cell_j[field::vertices][item.p_j];
+          // === vertex array
+          const auto& vertices_i =  cell_i[field::vertices][item.p_i];
+          const auto& vertices_j =  cell_j[field::vertices][item.p_j];
 
-					// === shapes
-					const shape& shp_i = shps[type_i];
-					const shape& shp_j = shps[type_j];
+          // === shapes
+          const shape& shp_i = shps[type_i];
+          const shape& shp_j = shps[type_j];
 
-					auto [contact, dn, n, contact_position] = detect(item.type, vertices_i, item.sub_i, &shp_i, vertices_j, item.sub_j, &shp_j);
[…]
