### 2024-07-23 226137b
add an example with interaction output files
