### 2024-07-24 b505b23
clean and update path for interaction analyses
