+++ example/CMakeLists.txt
@@ -1,4 +1,5 @@
 enable_testing()
 
+add_subdirectory(commands)
 add_subdirectory(spheres)
 add_subdirectory(polyhedra)
