### 2024-07-29 63ab83e
add example using write_xyz operator and a post processing script
