### 2024-07-29 c91086e
factorize cmake code in example
