+++ example/polyhedra/analyses/CMakeLists.txt
@@ -1,10 +1,3 @@
-
-set(ExampleName "InteractionOutput")
-set(FileName "interaction_minimal.msp")
-
 file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/shapes.shp DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
-
-add_test(Example_ExaDEM_${ExampleName} ${CMAKE_BINARY_DIR}/exaDEM ${CMAKE_CURRENT_SOURCE_DIR}/${FileName} --omp_num_threads 1)
-add_test(Example_ExaDEM_${ExampleName}_omp ${CMAKE_BINARY_DIR}/exaDEM ${CMAKE_CURRENT_SOURCE_DIR}/${FileName} --omp_num_threads 2)
-add_test(Example_ExaDEM_${ExampleName}_mpi mpirun -n 2 ${CMAKE_BINARY_DIR}/exaDEM ${CMAKE_CURRENT_SOURCE_DIR}/${FileName} --omp_num_threads 1)
-
+add_omp_mpi_test("InteractionOutput" "interaction_minimal.msp")
+add_omp_mpi_test("WritePosVelOutput" "write_pos_vel_minimal.msp")
