### 2024-07-30 49d7827
add the dn vector and rename co to cp
