### 2024-07-30 3f2bf31
43 add post processing interaction output files (#46)
* remove old hooke force version for polyhedron

* fix issue in debug mode && remove old hooke force with mutexes && add buffers for analysis asked by lhassan in the hooke force operator

* It works in serial [interaction output file]

* clean classifier analyses

* add a post_processing script

* add an example with interaction output files

* clean and update path for interaction analyses

* fix issue with mpi

* add the dn vector and rename co to cp

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
