+++ src/polyhedron/include/exaDEM/hooke_polyhedron.h
@@ -137,7 +137,7 @@ namespace exaDEM
          * was detected, the distance between the polyhedrons, norm and contact point.
          *
          * @param type Type of interaction to detect:
-         *             - 0: Vertex-Vertex interaction
+         *             - 0; Vertex-Vertex interaction
          *             - 1: Vertex-Edge interaction
          *             - 2: Vertex-Face interaction
          *             - 3: Edge-Edge interaction
@@ -170,28 +170,22 @@ namespace exaDEM
       };
 
       /**
-       * @brief Operator function for performing interactions between particles.
-       *
-       * This function performs interactions between particles.
-       * It uses precomputed shapes (`shps`) and synchronization (`locker`) for thread safety.
+       * @brief Operator function for performing interactions between particles (polyhedra).
        *
        * @tparam TMPLC Type of the cells or particles container.
        * @param item Reference to the Interaction object representing the interaction details.
-       * @param cells Reference to the cells or particles container.
+       * @param cells Pointer to the cells or particles container.
        * @param hkp Reference to the HookeParams object containing interaction parameters.
-       * @param shps Reference to the shapes object providing shape information for interactions.
-       * @param dt Simulation dt increment.
-       * @param locker Reference to mutexes object for synchronization.
+       * @param shps Pointer to the shapes array providing shape information for interactions.
+       * @param dt Time increment for the simulation step.
        */
       template<typename TMPLC>
-        ONIKA_HOST_DEVICE_FUNC 
-        inline void operator()(
+        ONIKA_HOST_DEVICE_FUNC inline std::tuple<double, Vec3d, Vec3d, Vec3d> operator()(
             Interaction& item, 
-            TMPLC& cells, 
-            const HookeParams& hkp,
-            const shapes& shps, 
-            const double dt, 
-            mutexes& locker) const
+            TMPLC* const cells, 
+            const HookeParams& hkp, 
+            const shape* const shps, 
+            const double dt) const
         {
           // === cell
           auto& cell_i =  cells[item.cell_i];
@@ -214,10 +208,13 @@ namespace exaDEM
           const auto& vertices_j =  cell_j[field::vertices][item.p_j];
 
           // === shapes
-          const shape* shp_i = shps[type_i];
-          const shape* shp_j = shps[type_j];
+          const shape& shp_i = shps[type_i];
+          const shape& shp_j = shps[type_j];
 
-          auto [contact, dn, n, contact_position] = detect(item.type, vertices_i, item.sub_i, shp_i, vertices_j, item.sub_j, shp_j);
+          auto [contact, dn, n, contact_position] = detect(item.type, vertices_i, item.sub_i, &shp_i, vertices_j, item.sub_j, &shp_j);
+          // temporary vec3d to store forces.
+          Vec3d f = {0,0,0};
+          Vec3d fn = {0,0,0};
           if(contact)
           {
             const Vec3d vi = get_v(cell_i, item.p_i);
@@ -225,8 +222,6 @@ namespace exaDEM
             const auto& m_i = cell_i[field::mass][item.p_i];
             const auto& m_j = cell_j[field::mass][item.p_j];
 
-            // temporary vec3d to store forces.
-            Vec3d f = {0,0,0};
             const double meff = compute_effective_mass(m_i, m_j);
 
             hooke_force_core(dn, n, dt, hkp.m_kn, hkp.m_kt, hkp.m_kr,
@@ -236,473 +231,256 @@ namespace exaDEM
[…]
