### 2024-07-30 b76606f
47 add output containing positions and velocities (#48)
* remove old hooke force version for polyhedron

* fix issue in debug mode && remove old hooke force with mutexes && add buffers for analysis asked by lhassan in the hooke force operator

* It works in serial [interaction output file]

* clean classifier analyses

* add a post_processing script

* add an example with interaction output files

* clean and update path for interaction analyses

* update the example Interaction analyses

* update coverage

* update spack repo

* add example using write_xyz operator and a post processing scrip
[…]
