+++ src/sphere/include/exaDEM/hooke_sphere.h
@@ -419,7 +419,6 @@ namespace exaDEM
 						item.reset();
             dn = 0;
 					}
-
 					return {dn, contact_position, fn, item.friction};
 				}
 		};
