+++ example/spheres/rigid-surface/CMakeLists.txt
@@ -1,9 +1,2 @@
-
-set(ExampleName "RigidSurface")
-set(FileName "rigid_surface.msp")
-
 file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/input_file_rigid_surface.xyz DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
-
-add_test(Example_ExaDEM_${ExampleName} ${CMAKE_BINARY_DIR}/exaDEM ${CMAKE_CURRENT_SOURCE_DIR}/${FileName} --omp_num_threads 1)
-add_test(Example_ExaDEM_${ExampleName}_omp ${CMAKE_BINARY_DIR}/exaDEM ${CMAKE_CURRENT_SOURCE_DIR}/${FileName} --omp_num_threads 2)
-add_test(Example_ExaDEM_${ExampleName}_mpi mpirun -n 2 ${CMAKE_BINARY_DIR}/exaDEM ${CMAKE_CURRENT_SOURCE_DIR}/${FileName} --omp_num_threads 1)
+add_omp_mpi_test("RigidSurface" "rigid_surface.msp")
