+++ CMakeLists.txt
@@ -23,7 +23,7 @@ project(exaDEM LANGUAGES C CXX)
 # ===========================
 # === CMake customization ===
 # ===========================
-set_property(CACHE CMAKE_BUILD_TYPE PROPERTY STRINGS Release RelWithDebInfo Debug)
+set_property(CACHE CMAKE_BUILD_TYPE PROPERTY STRINGS Release Coverage RelWithDebInfo Debug)
 if (NOT CMAKE_BUILD_TYPE)
   set(CMAKE_BUILD_TYPE "Release" CACHE STRING "" FORCE)
 endif()
