+++ example/polyhedra/CMakeLists.txt
@@ -1,6 +1,6 @@
 enable_testing()
 
-list(APPEND DLIST generator rotating_drum balls stl_mesh)
+list(APPEND DLIST generator rotating_drum balls stl_mesh analyses)
 
 foreach(DirExample IN LISTS DLIST)
   add_subdirectory(${DirExample})
