### 2024-07-31 9fa5701
add stress tensor operator (not GPU)
