+++ src/polyhedron/hooke_polyhedron.cpp
@@ -58,6 +58,7 @@ namespace exaDEM
     ADD_SLOT( shapes      , shapes_collection , INPUT_OUTPUT , DocString{"Collection of shapes"});
     // analysis
     ADD_SLOT( long        , timestep          , INPUT , REQUIRED );
+    ADD_SLOT( bool        , save_interactions , INPUT , false           , DocString{"Store interactions into the classifier"});
     ADD_SLOT( long        , analysis_interaction_dump_frequency , INPUT , REQUIRED, DocString{"Write an interaction dump file"});
 
 
@@ -74,7 +75,7 @@ namespace exaDEM
 
       /** Analysis */
       const long frequency = *analysis_interaction_dump_frequency;
-      bool write_interactions = ( frequency > 0 && (*timestep) % frequency == 0);
+      bool write_interactions = *save_interactions || ( frequency > 0 && (*timestep) % frequency == 0);
 
       /** Get driver and particles data */
       Drivers* drvs =  drivers.get_pointer();
