### 2024-08-04 7c4500b
add compute/write avg stress oeprators + adapt other operators + adapt config files + add test (not checked)
