### 2024-08-05 27514da
update stress tensor dev
