### 2024-08-13 fb4af68
49 compute global stress (#51)
* add stress tensor operator (not GPU)

* add compute/write avg stress oeprators + adapt other operators + adapt config files + add test (not checked)

* update stress tensor dev

* add test with avg stress + script

* update write_avg_stress test

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
