### 2024-08-13 0d0e7e7
propagate new compute_cell_particles optimizations
