### 2024-08-13 45eff22
propagate optimization to contact and classifier operators
