### 2024-08-13 642882b
propagate optimization from the reduce_cell_particles operator (exanbody)
