### 2024-08-13 c47cb0d
52 free empty cells computation (#53)
* propagate new compute_cell_particles optimizations

* propagate optimization to contact and classifier operators

* propagate optimization from the reduce_cell_particles operator (exanbody)

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
