+++ src/io/CMakeLists.txt
@@ -14,5 +14,5 @@
 # kind, either express or implied.  see the license for the
 # specific language governing permissions and limitations
 # under the license.
-set(exadem_io_LINK_LIBRARIES exanbIO exanbDefBox exanbParticleNeighbors exanbExtraStorage exadem_force_field exadem_interaction)
+set(exadem_io_LINK_LIBRARIES exanbIO exanbDefBox exanbParticleNeighbors exanbExtraStorage exadem_grid exadem_force_field exadem_interaction)
 xstamp_add_plugin(exadem_io ${CMAKE_CURRENT_SOURCE_DIR})

[…]
