### 2024-08-14 739ab21
add new logs: avg active ineractions, avg interactions, |dn| and rename thoughout
