### 2024-08-15 a688234
add new logs + update interaction_summary
