### 2024-08-15 1df825c
add save logs in an output file named by default logs
