+++ src/io/print_simulation_state.cpp
@@ -63,15 +63,7 @@ namespace exaDEM
 
 		inline void execute () override final
 		{
-			double conv_temperature = 1.e4 * legacy_constant::atomicMass / legacy_constant::boltzmann ;
-			//double conv_energy = 1.e4 * legacy_constant::atomicMass / legacy_constant::elementaryCharge;
-			//static const std::string header = "     Step     Time (ps)     Particles   Mv/Ext/Imb.  Tot. E. (eV/part)  Kin. E. (eV/part)  Pot. E. (eV/part)  Temperature   Pressure     sMises     Volume       Mass";
-			static const std::string header = "     Step     Time (ps)     Particles  Mv/Ext/Imb.       |dn|  avg. act. I     avg. I  Temperature     Volume       Mass  Throughput";
-
-			if( *internal_units )
-			{
-				conv_temperature = 1.0;
-			}
+			static const std::string header = "     Step     Time          Particles  Mv/Ext/Imb.       |dn|  avg. act. I     avg. I   avg. Ec   avg. Erot     Volume       Mass  Throughput";
 
 			bool lb_flag = (*lb_counter) > 0 ;
 			long move_count = *move_counter ;
@@ -136,7 +128,7 @@ namespace exaDEM
       double avg_act_I = double(active_interactions) / sim_info.particle_count();
       double avg_I = double(total_interactions) / sim_info.particle_count();
 
-			std::string line = format_string("%9ld % .6e %13ld  %c %c %8s %.3e    %9.3f  %9.3f  % 11.3f % .3e % .3e % .4e",
+			std::string line = format_string("%9ld % .6e %13ld  %c %c %8s %.3e    %9.3f  %9.3f % .3e % .3e % .3e % .3e % .4e",
 					*timestep, // %9ld
 					*physical_time, // %.6e
 					sim_info.particle_count(), // %13ld 
@@ -144,7 +136,8 @@ namespace exaDEM
           std::abs(sim_info.dn()),
           avg_act_I,
           avg_I,
-					sim_info.temperature_scal() / sim_info.particle_count() * conv_temperature, //%11.3f
+					sim_info.kinetic_energy_scal() / sim_info.particle_count(),
+					sim_info.rotation_energy_scal() / sim_info.particle_count(),
 					sim_info.volume(),
 					sim_info.mass(),
 					throughput) ;
@@ -163,8 +156,6 @@ namespace exaDEM
 
 				if ( rank == 0 )
 				{
-          lout << "ici " << full_path << std::endl;
-          lout << line << std::endl;
 					fs::create_directory(*dir_name);
 				  std::fstream file(full_path, std::ios::out | std::ios::in | std::ios::app); 
           if(*print_header) file << header << std::endl;
