### 2024-08-22 e3469e3
factorize output file within the io_config operator
