### 2024-08-26 a1bb56a
45 update simulation logs (#54)
* add new logs: avg active ineractions, avg interactions, |dn| and rename thoughout

* add new logs + update interaction_summary

* add save logs in an output file named by default logs

* add erot and ec

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
