+++ src/io/simulation_state.cpp
@@ -37,13 +37,15 @@ under the License.
 #include <exaDEM/simulation_state.h>
 #include <exaDEM/dem_simulation_state.h>
 #include <exaDEM/cell_list_wrapper.hpp>
+#include <exaDEM/interaction/classifier.hpp>
+#include <exaDEM/itools/itools.hpp>
 
 // ================== Thermodynamic state compute operator ======================
 namespace exaDEM
 {
 	using namespace exanb;
 
-	template<class GridT ,	class = AssertGridHasFields< GridT, field::_vx, field::_vy, field::_vz, field::_mass >> struct SimulationStateNode : public OperatorNode
+	template<class GridT ,	class = AssertGridHasFields< GridT, field::_vx, field::_vy, field::_vz, field::_vrot, field::_mass >> struct SimulationStateNode : public OperatorNode
 	{
 		// compile time constant indicating if grid has type field
 
@@ -53,46 +55,55 @@ namespace exaDEM
     ADD_SLOT( CellListWrapper    , cell_list , INPUT   , DocString{"list of non empty cells within the current grid"});
 		ADD_SLOT( SimulationState    , simulation_state    , OUTPUT );
 
-		static constexpr FieldSet<field::_vx ,field::_vy ,field::_vz, field::_mass> reduce_field_set {};
-		static constexpr FieldSet<field::_vx> reduce_vx_field_set {};
-		static constexpr FieldSet<field::_vy> reduce_vy_field_set {};
-		static constexpr FieldSet<field::_vz> reduce_vz_field_set {};
-		static constexpr FieldSet<field::_mass> reduce_mass_field_set {};
+    // DEM data
+    ADD_SLOT( Classifier         , ic                  , INPUT , DocString{"Interaction lists classified according to their types"} );
+    ADD_SLOT( bool               , symetric            , INPUT , REQUIRED , DocString{"Use of symetric feature (contact law)"});
+
+		static constexpr FieldSet<field::_vx ,field::_vy ,field::_vz, field::_vrot, field::_mass> reduce_field_set {};
 		inline void execute () override final
 		{
 			MPI_Comm comm = *mpi;
 			SimulationState& sim_info = *simulation_state;
 
-			Vec3d momentum;  // constructs itself with 0s
 			Vec3d kinetic_energy;  // constructs itself with 0s
+			Vec3d rotation_energy;  // constructs itself with 0s
 			double mass = 0.;
-			unsigned long long int total_particles = 0;
+			uint64_t total_particles = 0;
 
       auto [cell_ptr, cell_size] = cell_list->info();
-			exaDEM::simulation_state_variables sim {}; //kinetic_energy, momentum, mass, potential_energy, total_particles};
+			exaDEM::simulation_state_variables sim {}; //kinetic_energy, rotation_energy, mass, potential_energy, total_particles};
 			ReduceSimulationStateFunctor func = {};
 			reduce_cell_particles( *grid , false , func , sim, reduce_field_set , parallel_execution_context() , {} , cell_ptr, cell_size );
 
+      // get interaction informations
+      const Classifier& classifier = *ic;
+     	uint64_t active_interactions = itools::get_act_interaction_size(*grid, classifier, *symetric);
+     	uint64_t total_interactions  = itools::get_tot_interaction_size(*grid, classifier, *symetric);
+      double dn                    = itools::get_min_dn(classifier);
+
 			// reduce partial sums and share the result
 			{
[…]
