+++ src/grid/chunk2.cpp
@@ -1,0 +1,791 @@
+/*
+Licensed to the Apache Software Foundation (ASF) under one
+or more contributor license agreements.  See the NOTICE file
+distributed with this work for additional information
+regarding copyright ownership.  The ASF licenses this file
+to you under the Apache License, Version 2.0 (the
+"License"); you may not use this file except in compliance
+with the License.  You may obtain a copy of the License at
+
+  http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing,
+software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+KIND, either express or implied.  See the License for the
+specific language governing permissions and limitations
+under the License.
+*/
+//#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
+#include <exanb/core/operator.h>
+#include <exanb/core/operator_slot.h>
+#include <exanb/core/operator_factory.h>
+#include <exanb/core/make_grid_variant_operator.h>
+#include <exanb/amr/amr_grid.h>
+#include <exanb/core/grid.h>
+#include <exanb/core/parallel_grid_algorithm.h>
+#include <exanb/core/particle_id_codec.h>
+#include <exanb/amr/amr_grid_algorithm.h>
+#include <exanb/core/particle_type_pair.h>
+
+#include <onika/cuda/cuda_context.h>
+#include <onika/memory/allocator.h>
+
+#include <exanb/particle_neighbors/chunk_neighbors.h>
+#include <exanb/particle_neighbors/chunk_neighbors_config.h>
+#include <exanb/particle_neighbors/chunk_neighbors_scratch.h>
+#include <exanb/particle_neighbors/chunk_neighbors_host_write_accessor.h>
+
+#include <exanb/core/domain.h>
+#include <exanb/core/xform.h>
+
+#include <exanb/particle_neighbors/chunk_neighbors_execute_2.h>
+
+//#include <exaDEM/interactions_PP.h>
+//#include <exaDEM/neighbor_friction.h>
+#include <../forcefield/include/exaDEM/interactions_PP.h>
+#include <exanb/particle_neighbors/neighbor_filter_func.h>
+
+namespace exaDEM
+{
+  using namespace exanb;
+   
+  template<class CellsT>
+  struct ContactNeighborFilterFunc
+  {
+    CellsT cells;
+    const double rcut_inc = 0.0;
+    inline bool operator () (double d2, double rcut2,size_t cell_a,size_t p_a,size_t cell_b,size_t p_b) const
+    {
+      assert( cell_a!=cell_b || p_a!=p_b );
+      const double r_a = cells[cell_a][field::radius][p_a];
+      const double r_b = cells[cell_b][field::radius][p_b];
+      const double rmax = r_a + r_b + rcut_inc;
+      const double rmax2 = rmax * rmax;
+      return d2 > 0.0 && d2 < rmax2 && d2 < rcut2 ;
+    }
+  };
+  
+  __global__ void setGPU(int* pa, 
+  			int* cella,
+  			int* pb,
+  			int* cellb,
+  			double* ftx,
+  			double* fty,
+  			double* ftz,
+  			onika::memory::CudaMMVector<int> pa2, 
+  			onika::memory::CudaMMVector<int> cella2,
[…]
