+++ doc/paper/joss/paper.bib
@@ -8,6 +8,19 @@
   organization={Springer}
 }
 
+@article{JOSIEN2024109354,
+title = {Parallel and bias-free RSA algorithm for maximal Poisson-sphere sampling},
+journal = {Computer Physics Communications},
+volume = {305},
+pages = {109354},
+year = {2024},
+issn = {0010-4655},
+doi = {https://doi.org/10.1016/j.cpc.2024.109354},
+url = {https://www.sciencedirect.com/science/article/pii/S0010465524002777},
+author = {Marc Josien and Raphaël Prat},
+keywords = {RSA, HPC, MPI},
+}
+
 @article{ciccotti1987simulation,
   title={Simulation of liquids and solids},
   author={Ciccotti, Giovanni and Frenkel, Daan and Mc Donald, Ian R},
