+++ src/driver/add_surface.cpp
@@ -43,6 +43,7 @@ namespace exaDEM
       ADD_SLOT( int     , id        , INPUT , REQUIRED           , DocString{"Driver index"});
       ADD_SLOT( double  , offset    , INPUT , 0.0                , DocString{"Offset from the origin (0,0,0) of the rigid surface"} );
       ADD_SLOT( double  , velocity  , INPUT , 0.0                , DocString{"Surface velocity"});
+      ADD_SLOT( Vec3d   , vrot      , INPUT , Vec3d{0.0,0.0,0.0} , DocString{"Angular velocity of the rigid surface, default is 0 m.s-"});
       ADD_SLOT( Vec3d   , normal    , INPUT , Vec3d{0.0,0.0,1.0} , DocString{"Normal vector of the rigid surface"});
 
       public:
@@ -56,8 +57,7 @@ namespace exaDEM
 
       inline void execute () override final
       {
-        constexpr Vec3d null= { 0.0, 0.0, 0.0 };
-        exaDEM::Surface driver = {*offset, *normal, null, *velocity, null}; // 
+        exaDEM::Surface driver = {*offset, *normal, null, *velocity, *vrot}; // 
         driver.initialize();
         drivers->add_driver(*id, driver);
       }
