+++ README.md
@@ -10,160 +10,10 @@ Additionally, ExaDEM offers compatibility with MPI+GPUs, using the CUDA programm
 ## Documentation
 
 Documentation is available here: 
-- Github: [ExaDEM Documentation](https://github.com/Collab4exaNBody/doc_exaDEM.git)
+
 - Website: [ExaDEM Website](https://collab4exanbody.github.io/exaDEM/)
-
-Build the documentation:
-
-```
-git clone https://github.com/Collab4exaNBody/doc_exaDEM.git
-cd doc_exaDEM/
-git submodule init
-git submodule update
-cd doc_exaNBody && git pull && cd ..
-make html
-firefox build/html/index.html 
-```
-
-## Installation
-
-### Minimal Requirements
-
-To proceed with the installation, your system must meet the minimum prerequisites. The first step involves the installation of exaNBody:
-
-```
-git clone https://github.com/Collab4exaNBody/exaNBody.git
-export exaNBody_DIR=${PWD}/exaNBody 
-```
-
-The next step involves the installation of yaml-cpp, which can be achieved using either the spack package manager or cmake:
-
-```
-spack install yaml-cpp@0.6.3
-spack load yaml-cpp@0.6.3
-```
-
-Variant: 
-
-```
-apt install libyaml-cpp-dev
-```
-
-### Optional Dependencies
-
-Before proceeding further, you have the option to consider the following dependencies:
-
-- Cuda
-- MPI
-- HIP
-
-### Buidling ExaDEM With CMAKE
-
-To install `ExaDEM`, follow these steps:
-
-Set the `exaNBody_DIR` environment variable to the installation path. Clone the ExaDEM repository using the command:
-
-```
-git clone https://github.com/Collab4exaNBody/exaDEM.git
-```
-
-
-Create a directory named build-exaDEM and navigate into it:
-
-```	
-mkdir build-exaDEM && cd build-exaDEM
-```
-
-Run CMake to configure the ExaDEM build, specifying that CUDA support should be turned off:
-
-```		
-cmake ../exaDEM -DXNB_BUILD_CUDA=OFF
-```
-
-Build ExaDEM using the make command with a specified number of parallel jobs (e.g., -j 4 for 4 parallel jobs):
-
-```	
-make -j 4
-```
-
-Build Plugins
-
-```	
-make UpdatePluginDataBase
-```
-
-This command will display all plugins and related operators. Example: 
-
-```	
- + exadem_force_fieldPlugin
-   operator    cylinder_wall
-   operator    gravity_force
-   operator    hooke_force
-   operator    rigid_surface
- + exadem_ioPlugin
-   operator    print_simulation_state
-   operator    read_xyz
-   operator    read_dump_particles
-```
-
-### Building ExaDEM With SPACK
-
-Installation with spack is preferable for people who don't want to develop in `ExaDEM`. Only stable versions are added when you install `ExaDEM` with `Spack`. Note: `ExaDEM` main will never be directly accessible via this installation method.
-
-#### Installing Spack
-
-```
-git clone https://github.com/spack/spack.git
-export SPACK_ROOT=$PWD/spack
-source ${SPACK_ROOT}/share/spack/setup-env.sh
-```
-
-#### Installing ExaDEM
-
-First get the spack repository in `ExaDEM` directory and it to spack. It contains two packages: `exanbody` and `exadem`.
-
-```
-git clone https://github.com/Collab4exaNBody/exaDEM.git
-cd exaDEM
-spack repo add spack_repo
-```
-
-Current variante(s):
-  
-- +cuda: Add GPU support (disabled by default)
-
-Second install `ExaDEM` (this command will install cmake, yaml-cpp and exanbody).
-
-```
-spack install exadem
-```
-
[…]
