### 2024-09-06 8f9473d
Resolution of errors during the GPU compilation and execution
