+++ src/interaction/include/exaDEM/shape/shape.hpp
@@ -189,7 +189,8 @@ namespace exaDEM
 		ONIKA_HOST_DEVICE_FUNC
 			const std::pair<int*, int> get_face(const int i)
 			{
-				assert(i < m_faces[0]);
+				//assert(i < m_faces[0]);
+				assert(i < onika::cuda::vector_data(m_faces)[0]);
 				int * ptr = this->get_faces() + 1;
 				for(int it = i ; it > 0 ; it --)
 				{
@@ -202,7 +203,8 @@ namespace exaDEM
 		ONIKA_HOST_DEVICE_FUNC
 		const std::pair<int*, int> get_face(const int i) const
 		{
-			assert(i < m_faces[0]);
+			//assert(i < m_faces[0]);
+			assert(i < onika::cuda::vector_data(m_faces)[0]);
 			int * ptr = this->get_faces() + 1;
 			for(int it = i ; it > 0 ; it --)
 			{
