+++ src/driver/init_drivers.cpp
@@ -29,7 +29,7 @@ under the License.
 #include <memory>
 #include <exaDEM/driver_base.h>
 #include <exaDEM/drivers.h>
-#include <exaDEM/cylinder.h>
+#include <exaDEM/surface.h>
 
 namespace exaDEM
 {
@@ -37,44 +37,31 @@ namespace exaDEM
   using namespace exanb;
 
   template<typename GridT>
-    class AddCylinder : public OperatorNode
+    class InitDrivers : public OperatorNode
     {
-      static constexpr Vec3d default_axis = { 1.0, 0.0, 1.0 };
-      static constexpr Vec3d null= { 0.0, 0.0, 0.0 };
-
-
-      ADD_SLOT( Drivers , drivers    , INPUT_OUTPUT, REQUIRED    , DocString{"List of Drivers"});
-      ADD_SLOT( int     , id         , INPUT       , REQUIRED , DocString{"Driver index"});
-      ADD_SLOT( Vec3d   , center     , INPUT       , REQUIRED , DocString{"Center of the cylinder"});
-      ADD_SLOT( Vec3d   , axis       , INPUT       , default_axis , DocString{"Define the plan of the cylinder"});
-      ADD_SLOT( Vec3d   , angular_velocity, INPUT  , null     , DocString{"Angular velocity of the cylinder, default is 0 m.s-"});
-      ADD_SLOT( Vec3d   , velocity   , INPUT       , null     , DocString{"Cylinder velocity, could be used in 'expert mode'"});
-      ADD_SLOT( double  , radius     , INPUT       , REQUIRED , DocString{"Radius of the cylinder, positive and should be superior to the biggest sphere radius in the cylinder"});
+      ADD_SLOT( Drivers , drivers , OUTPUT , DocString{"List of Drivers"});
 
       public:
 
       inline std::string documentation() const override final
       {
         return R"EOF(
-        This operator add a cylinder to the drivers list.
+        This operator creates a slot for drivers.
         )EOF";
       }
 
       inline void execute () override final
       {
-				// proj center over axis
-				Vec3d c = (*center) * (*axis);
-        exaDEM::Cylinder driver = {*radius, *axis, c, *velocity, *angular_velocity};
-        drivers->add_driver(*id, driver);
+        // do nothing
       }
     };
 
-  template<class GridT> using AddCylinderTmpl = AddCylinder<GridT>;
+  template<class GridT> using InitDriversTmpl = InitDrivers<GridT>;
 
   // === register factories ===  
   CONSTRUCTOR_FUNCTION
   {
-    OperatorNodeFactory::instance()->register_factory( "add_cylinder", make_grid_variant_operator< AddCylinderTmpl > );
+    OperatorNodeFactory::instance()->register_factory( "init_drivers", make_grid_variant_operator< InitDriversTmpl > );
   }
 }
 
