+++ src/io/io_config.cpp
@@ -33,6 +33,12 @@ namespace exaDEM
     ADD_SLOT( std::string , interaction_basename   , INPUT_OUTPUT , REQUIRED , DocString{"Write an Output file containing interactions."   } );
 
 		public:
+
+    inline std::string documentation() const override final
+    {
+      return R"EOF(This operator defines the tree structure of output files.)EOF";
+    }
+
 		inline bool is_sink() const override final { return true; }
 
 		inline void execute () override final
