### 2024-09-08 043dcc1
55 add a configuration operator for outputs (#56)
* add new logs: avg active ineractions, avg interactions, |dn| and rename thoughout

* add new logs + update interaction_summary

* add save logs in an output file named by default logs

* add erot and ec

* factorize output file within the io_config operator

* update io_conf operator + add a paraview file name operator + update configuration files

* disable to write avg stress tensor at the first step

* add an angular velocty slot for add surface and add ball

* .

* update io config

---------

Co-authored-by: rprat-pro <rp269144@pleiades1
[…]
