### 2024-09-09 45c58ef
initialize (surface) is not a gpu function
