### 2024-09-09 9d6f43b
fix warning into classify.h
