+++ src/sphere/contact_force_sphere.cpp
@@ -27,8 +27,8 @@ under the License.
 
 #include <memory>
 
-#include <exaDEM/hooke_force_parameters.h>
-#include <exaDEM/compute_hooke_force.h>
+#include <exaDEM/contact_force_parameters.h>
+#include <exaDEM/compute_contact_force.h>
 #include <exaDEM/interaction/interaction.hpp>
 #include <exaDEM/interaction/grid_cell_interaction.hpp>
 #include <exaDEM/interaction/classifier.hpp>
@@ -37,38 +37,44 @@ under the License.
 #include <exaDEM/shape/shapes.hpp>
 #include <exaDEM/shape/shape_detection.hpp>
 #include <exaDEM/shape/shape_detection_driver.hpp>
+#include <exaDEM/mutexes.h>
 #include <exaDEM/drivers.h>
-#include <exaDEM/hooke_polyhedron.h>
+#include <exaDEM/contact_sphere.h>
 
 namespace exaDEM
 {
   using namespace exanb;
-  using namespace polyhedron;
+  using namespace sphere;
 
   template<typename GridT , class = AssertGridHasFields< GridT, field::_radius >>
-    class ComputeHookeClassifierPolyhedronGPU : public OperatorNode
+    class ComputeContactClassifierSphereGPU : public OperatorNode
   {
+    // attributes processed during computation
+    using ComputeFields = FieldSet< field::_vrot, field::_arot >;
+    static constexpr ComputeFields compute_field_set {};
+
     ADD_SLOT( GridT       , grid              , INPUT_OUTPUT , REQUIRED );
-    ADD_SLOT( HookeParams , config            , INPUT , REQUIRED ); // can be re-used for to dump contact network
-    ADD_SLOT( HookeParams , config_driver     , INPUT , OPTIONAL ); // can be re-used for to dump contact network
-    ADD_SLOT( double      , dt                , INPUT , REQUIRED );
+    ADD_SLOT( ContactParams , config            , INPUT        , REQUIRED , DocString{"Contact parameters for sphere interactions"}); // can be re-used for to dump contact network
+    ADD_SLOT( ContactParams , config_driver     , INPUT        , OPTIONAL , DocString{"Contact parameters for drivers, optional"}); // can be re-used for to dump contact network
+    ADD_SLOT( double      , dt                , INPUT        , REQUIRED , DocString{"Time step value"});
     ADD_SLOT( bool        , symetric          , INPUT_OUTPUT , REQUIRED , DocString{"Activate the use of symetric feature (contact law)"});
-    ADD_SLOT( Drivers     , drivers           , INPUT , DocString{"List of Drivers {Cylinder, Surface, Ball, Mesh}"});
+    ADD_SLOT( Drivers     , drivers           , INPUT        , REQUIRED ,DocString{"List of Drivers {Cylinder, Surface, Ball, Mesh}"});
     ADD_SLOT( Classifier  , ic                , INPUT_OUTPUT , DocString{"Interaction lists classified according to their types"} );
-    ADD_SLOT( shapes      , shapes_collection , INPUT_OUTPUT , DocString{"Collection of shapes"});
-    // analyses
+    // analysis
     ADD_SLOT( long        , timestep          , INPUT , REQUIRED );
+    ADD_SLOT( bool        , save_interactions , INPUT , false           , DocString{"Store interactions into the classifier"});
     ADD_SLOT( long        , analysis_interaction_dump_frequency  , INPUT , REQUIRED , DocString{"Write an interaction dump file"});
     ADD_SLOT( long        , analysis_dump_stress_tensor_frequency, INPUT , REQUIRED , DocString{"Compute avg Stress Tensor."});
     ADD_SLOT( long        , simulation_log_frequency             , INPUT , REQUIRED , DocString{"Log frequency."});
[…]
