### 2024-09-10 512d6ed
fix a driver issue on GPU
