+++ src/sphere/hooke_sphere.cpp
@@ -51,6 +51,7 @@ namespace exaDEM
   {
     // attributes processed during computation
     using ComputeFields = FieldSet< field::_vrot, field::_arot >;
+    using driver_t = std::variant<exaDEM::Cylinder, exaDEM::Surface, exaDEM::Ball, exaDEM::Stl_mesh, exaDEM::UndefinedDriver>;
     static constexpr ComputeFields compute_field_set {};
 
     ADD_SLOT( GridT       , grid              , INPUT_OUTPUT , REQUIRED );
@@ -94,7 +95,7 @@ namespace exaDEM
       bool store_interactions = write_interactions || compute_stress_tensor || need_interactions_for_log_frequency;
 
       /** Get Driver */
-      Drivers* drvs =  drivers.get_pointer();
+      driver_t* drvs =  drivers->data();
       auto* cells = grid->cells();
       const HookeParams hkp = *config;
       HookeParams hkp_drvs{};
