+++ src/sphere/include/exaDEM/hooke_sphere.h
@@ -15,7 +15,7 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
- */
+*/
 #pragma once
 
 #include <exaDEM/shape/shape.hpp>
@@ -82,345 +82,347 @@ namespace exaDEM
       struct hooke_law
       {
 
-        /**
-         * @brief Retrieves the position vector of a particle.
-         *
-         * This function returns the position vector of a particle within a cell.
-         *
-         * @tparam Cell The type representing a cell in the simulation.
-         * @param cell Reference to the cell containing the particle.
-         * @param p_id The ID of the particle whose position is to be retrieved.
-         * @return The position vector of the particle.
-         */
-        template<typename Cell>
-          ONIKA_HOST_DEVICE_FUNC inline const Vec3d get_r(Cell& cell, const int p_id) const
+      /**
+       * @brief Retrieves the position vector of a particle.
+       *
+       * This function returns the position vector of a particle within a cell.
+       *
+       * @tparam Cell The type representing a cell in the simulation.
+       * @param cell Reference to the cell containing the particle.
+       * @param p_id The ID of the particle whose position is to be retrieved.
+       * @return The position vector of the particle.
+       */
+      template<typename Cell>
+        ONIKA_HOST_DEVICE_FUNC inline const Vec3d get_r(Cell& cell, const int p_id) const
+        {
+          const Vec3d res = {
+            cell[field::rx][p_id],
+            cell[field::ry][p_id],
+            cell[field::rz][p_id]};
+          return res;
+        };
+
+
+      /**
+       * @brief Retrieves the velocity vector of a particle.
+       *
+       * This function returns the velocity vector of a particle within a cell.
+       *
+       * @tparam Cell The type representing a cell in the simulation.
+       * @param cell Reference to the cell containing the particle.
+       * @param p_id The ID of the particle whose velocity is to be retrieved.
+       * @return The velocity vector of the particle.
+       */
+      template<typename Cell>
+        ONIKA_HOST_DEVICE_FUNC inline const Vec3d get_v(Cell& cell, const int p_id) const
+        {
+          const Vec3d res = {
+            cell[field::vx][p_id],
+            cell[field::vy][p_id],
+            cell[field::vz][p_id]};
+          return res;
+        };
+
+      /**
+       * @brief Handles particle interactions using Hooke's law parameters.
+       *
+       * This function applies Hooke's law to calculate forces between interacting particles
+       * within the given cells. It updates the interaction item based on the specified
+       * parameters and uses mutexes for thread safety.
+       *
+       * @tparam TMPC The type representing the collection of cells in the simulation.
+       * @param item Reference to the interaction item to be updated.
+       * @param cells Pointer to the collection of cells containing the particles.
+       * @param hkp Reference to the Hooke's law parameters.
+       * @param time Increment simulation time.
+       */
+      template<typename TMPC>
+        ONIKA_HOST_DEVICE_FUNC 
+        inline std::tuple<double, Vec3d, Vec3d, Vec3d> operator()(
+            Interaction& item, 
+            TMPC* cells, 
+            const HookeParams& hkp, 
+            const double time) const
+        {
+          // === cell
+          auto& cell_i =  cells[item.cell_i];
[…]
