+++ src/interaction/include/exaDEM/interaction/classifier.hpp
@@ -3,6 +3,7 @@
 #include <exaDEM/interaction/interaction.hpp>
 #include <exaDEM/interaction/grid_cell_interaction.hpp>
 #include <exaDEM/itools/buffer.hpp>
+#include <exaDEM/interaction/interactionSOA.hpp>
 
 namespace exaDEM
 {
@@ -24,18 +25,23 @@ namespace exaDEM
   struct InteractionWrapper
   {
     const int m_type;
-    exaDEM::Interaction* const m_data;
-    ONIKA_HOST_DEVICE_FUNC inline exaDEM::Interaction& operator()(const uint64_t idx) const
+    //exaDEM::Interaction* const m_data;
+    InteractionSOA& m_data;
+    /*ONIKA_HOST_DEVICE_FUNC inline exaDEM::Interaction& operator()(const uint64_t idx) const
     {
       return m_data[idx];
+    }*/
+    ONIKA_HOST_DEVICE_FUNC inline exaDEM::Interaction operator()(const uint64_t idx) const
+    {
+    	return m_data[idx];
     }
-    ONIKA_HOST_DEVICE_FUNC inline uint8_t type() { return m_type; }
+    /*ONIKA_HOST_DEVICE_FUNC inline uint8_t type() { return m_type; }
     ONIKA_HOST_DEVICE_FUNC inline size_t pi      (const uint64_t idx) { return m_data[idx].p_i; }
     ONIKA_HOST_DEVICE_FUNC inline size_t pj      (const uint64_t idx) { return m_data[idx].p_j; }
     ONIKA_HOST_DEVICE_FUNC inline size_t celli   (const uint64_t idx) { return m_data[idx].cell_i; }
     ONIKA_HOST_DEVICE_FUNC inline size_t cellj   (const uint64_t idx) { return m_data[idx].cell_j; }
     ONIKA_HOST_DEVICE_FUNC inline Vec3d& moment  (const uint64_t idx) { return m_data[idx].moment; }
-    ONIKA_HOST_DEVICE_FUNC inline Vec3d& friction(const uint64_t idx) { return m_data[idx].friction; }
+    ONIKA_HOST_DEVICE_FUNC inline Vec3d& friction(const uint64_t idx) { return m_data[idx].friction; }*/
   };
 
 
@@ -49,7 +55,8 @@ namespace exaDEM
   {
     static constexpr int types = 13;
     template <typename T> using VectorT =  onika::memory::CudaMMVector<T>;
-    std::vector< VectorT< exaDEM::Interaction >> waves; ///< Storage for interactions categorized by type.
+    //std::vector< VectorT< exaDEM::Interaction >> waves; ///< Storage for interactions categorized by type.
+    VectorT<InteractionSOA> waves;
     std::vector< itools::interaction_buffers > buffers; ///< Storage for analysis. Empty if there is no analysis
 
     /**
@@ -81,8 +88,10 @@ namespace exaDEM
      * @param id Type identifier for the interaction wave.
      * @return Reference to the CUDA memory-managed vector storing interactions of the specified type.
      */
-    VectorT<exaDEM::Interaction>& get_wave(size_t id) {return waves[id];}
-    const VectorT<exaDEM::Interaction>& get_wave(size_t id) const {return waves[id];}
+    //VectorT<exaDEM::Interaction>& get_wave(size_t id) {return waves[id];}
+    InteractionSOA& get_wave(size_t id) {return waves[id];}
+    //const VectorT<exaDEM::Interaction>& get_wave(size_t id) const {return waves[id];}
+    const InteractionSOA& get_wave(size_t id) const {return waves[id];}
 
     /**
      * @brief Retrieves the pointer and size of the data stored in the CUDA memory-managed vector for a specific type.
@@ -90,17 +99,29 @@ namespace exaDEM
[…]
