### 2024-09-13 da8658c
add reduction on gpu (classifier)
