+++ src/interaction/include/exaDEM/itools/counter.hpp
@@ -1,7 +1,171 @@
 #pragma once
 
+//#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
+
 #include<exaDEM/interaction/classifier.hpp>
 #include<exaDEM/itools/buffer.hpp>
+
+#include <onika/cuda/cuda.h>
+#include <onika/cuda/cuda_error.h>
+#include <onika/cuda/device_storage.h>
+#include <onika/soatl/field_id.h>
+#include <exanb/core/grid_particle_field_accessor.h>
+
+#include <exanb/core/parallel_grid_algorithm.h>
+#include <onika/parallel/parallel_execution_context.h>
+#include <onika/parallel/block_parallel_for.h>
+#include <onika/parallel/parallel_for.h>
+#include <exanb/compute/reduce_cell_particles.h>
+
+// mini macro here
+//#ifdef ONIKA_CUDA_VERSION
+#   if defined(__CUDA_ARCH__) || defined(__HIP_DEVICE_COMPILE__)
+
+__device__ double atomicMin_double(double* address, double val)
+{
+    unsigned long long int* address_as_ull = (unsigned long long int*) address;
+    unsigned long long int old = *address_as_ull, assumed;
+    do {
+        assumed = old;
+        old = atomicCAS(address_as_ull, assumed,
+            __double_as_longlong(fmin(val, __longlong_as_double(assumed))));
+    } while (assumed != old);
+    return __longlong_as_double(old);
+}
+
+#   define EXADEM_CU_ATOMIC_MIN(x,a,...) atomicMin_double( &x , static_cast<std::remove_reference_t<decltype(x)> >(a) )
+
+#else
+#   define EXADEM_CU_ATOMIC_MIN(x,a,...) ::onika::capture_atomic_min( x , static_cast<std::remove_reference_t<decltype(x)> >(a) )
+#endif
+
+
+
+namespace exaDEM
+{
+	namespace itools /* interaction tools */
+	{
+		using namespace onika::parallel;
+    template <typename T> using VectorT =  onika::memory::CudaMMVector<T>;
+
+		template<typename T, class FuncT, class ResultT>
+			struct ReduceTFunctor
+			{
+				const T* const m_data;
+				const FuncT m_func;
+				ResultT* m_reduced_val ;
+
+				ONIKA_HOST_DEVICE_FUNC inline void operator () ( uint64_t i ) const
+				{
+					ResultT local_val = ResultT();
+					m_func( local_val, i, m_data,  reduce_thread_local_t{} );
+
+					ONIKA_CU_BLOCK_SHARED onika::cuda::UnitializedPlaceHolder<ResultT> team_val_place_holder;
+					ResultT& team_val = team_val_place_holder.get_ref();
+
+					if( ONIKA_CU_THREAD_IDX == 0 ) { team_val = local_val; }
+					ONIKA_CU_BLOCK_SYNC();
+
+					if( ONIKA_CU_THREAD_IDX != 0 )
+					{
+						m_func( team_val, local_val , reduce_thread_block_t{} );
+					}
+					ONIKA_CU_BLOCK_SYNC();
+
+					if( ONIKA_CU_THREAD_IDX == 0 )
+					{
+						m_func( *m_reduced_val, team_val , reduce_global_t{} );
+					}
+				}
+			};
+
+		struct IOSimInteractionResult
+		{
+			unsigned long long int n_act_interaction = 0;
+			unsigned long long int n_tot_interaction = 0;
+			double min_dn = 0;
+      void update(IOSimInteractionResult& in)
+      {
+        n_act_interaction += in.n_act_interaction;
+        n_tot_interaction += in.n_tot_interaction;
[…]
