### 2024-09-15 c60ce73
use GPU with reduc_data
