### 2024-09-15 b251524
change name hooke to contact (#58)
Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
