+++ src/interaction/unclassify_interactions.cpp
@@ -28,8 +28,8 @@ under the License.
 
 #include <memory>
 
-#include <exaDEM/hooke_force_parameters.h>
-#include <exaDEM/compute_hooke_force.h>
+#include <exaDEM/contact_force_parameters.h>
+#include <exaDEM/compute_contact_force.h>
 #include <exaDEM/interaction/interaction.hpp>
 #include <exaDEM/interaction/grid_cell_interaction.hpp>
 #include <exaDEM/interaction/classifier.hpp>
