+++ src/driver/include/exaDEM/ball.h
@@ -45,11 +45,11 @@ namespace exaDEM
      */
     void print()
     {
-      std::cout << "Driver Type: Ball"  << std::endl;
-      std::cout << "Radius: " << radius << std::endl;
-      std::cout << "Center: " << center << std::endl;
-      std::cout << "Vel   : " << vel    << std::endl;
-      std::cout << "AngVel: " << vrot   << std::endl;
+      lout << "Driver Type: Ball"  << std::endl;
+      lout << "Radius: " << radius << std::endl;
+      lout << "Center: " << center << std::endl;
+      lout << "Vel   : " << vel    << std::endl;
+      lout << "AngVel: " << vrot   << std::endl;
     }
 
     /**
