### 2024-09-16 f56103b
fix performance issue (reduce)
