+++ src/interaction/include/exaDEM/itools/counter.hpp
@@ -23,14 +23,14 @@
 
 __device__ double atomicMin_double(double* address, double val)
 {
-    unsigned long long int* address_as_ull = (unsigned long long int*) address;
-    unsigned long long int old = *address_as_ull, assumed;
-    do {
-        assumed = old;
-        old = atomicCAS(address_as_ull, assumed,
-            __double_as_longlong(fmin(val, __longlong_as_double(assumed))));
-    } while (assumed != old);
-    return __longlong_as_double(old);
+  unsigned long long int* address_as_ull = (unsigned long long int*) address;
+  unsigned long long int old = *address_as_ull, assumed;
+  do {
+    assumed = old;
+    old = atomicCAS(address_as_ull, assumed,
+	__double_as_longlong(fmin(val, __longlong_as_double(assumed))));
+  } while (assumed != old);
+  return __longlong_as_double(old);
 }
 
 #   define EXADEM_CU_ATOMIC_MIN(x,a,...) atomicMin_double( &x , static_cast<std::remove_reference_t<decltype(x)> >(a) )
@@ -43,241 +43,242 @@ __device__ double atomicMin_double(double* address, double val)
 
 namespace exaDEM
 {
-	namespace itools /* interaction tools */
-	{
-		using namespace onika::parallel;
+  namespace itools /* interaction tools */
+  {
+    using namespace onika::parallel;
     template <typename T> using VectorT =  onika::memory::CudaMMVector<T>;
 
-		template<typename T, class FuncT, class ResultT>
-			struct ReduceTFunctor
-			{
-				const T* const m_data;
-				const FuncT m_func;
-				ResultT* m_reduced_val ;
+    template<typename T, class FuncT, class ResultT>
+      struct ReduceTFunctor
+      {
+	const T* const m_data;
+	const FuncT m_func;
+	ResultT* m_reduced_val ;
 
-				ONIKA_HOST_DEVICE_FUNC inline void operator () ( uint64_t i ) const
-				{
-					ResultT local_val = ResultT();
-					m_func( local_val, i, m_data,  reduce_thread_local_t{} );
+	ONIKA_HOST_DEVICE_FUNC inline void operator () ( uint64_t i ) const
+	{
+	  ResultT local_val = ResultT();
+	  m_func( local_val, i, m_data,  reduce_thread_local_t{} );
 
-					ONIKA_CU_BLOCK_SHARED onika::cuda::UnitializedPlaceHolder<ResultT> team_val_place_holder;
-					ResultT& team_val = team_val_place_holder.get_ref();
+	  ONIKA_CU_BLOCK_SHARED onika::cuda::UnitializedPlaceHolder<ResultT> team_val_place_holder;
+	  ResultT& team_val = team_val_place_holder.get_ref();
 
-					if( ONIKA_CU_THREAD_IDX == 0 ) { team_val = local_val; }
-					ONIKA_CU_BLOCK_SYNC();
+	  if( ONIKA_CU_THREAD_IDX == 0 ) { team_val = local_val; }
+	  ONIKA_CU_BLOCK_SYNC();
 
-					if( ONIKA_CU_THREAD_IDX != 0 )
-					{
-						m_func( team_val, local_val , reduce_thread_block_t{} );
-					}
-					ONIKA_CU_BLOCK_SYNC();
+	  if( ONIKA_CU_THREAD_IDX != 0 )
+	  {
+	    m_func( team_val, local_val , reduce_thread_block_t{} );
+	  }
+	  ONIKA_CU_BLOCK_SYNC();
 
-					if( ONIKA_CU_THREAD_IDX == 0 )
-					{
-						m_func( *m_reduced_val, team_val , reduce_global_t{} );
-					}
-				}
-			};
+	  if( ONIKA_CU_THREAD_IDX == 0 )
+	  {
[…]
