### 2024-09-17 f371507
add dem_rcut_max operator
