### 2024-09-17 7aef364
add doxy -> itools counter.hpp
