+++ src/interaction/include/exaDEM/itools/counter.hpp
@@ -23,14 +23,14 @@
 
 __device__ double atomicMin_double(double* address, double val)
 {
-  unsigned long long int* address_as_ull = (unsigned long long int*) address;
-  unsigned long long int old = *address_as_ull, assumed;
-  do {
-    assumed = old;
-    old = atomicCAS(address_as_ull, assumed,
-	__double_as_longlong(fmin(val, __longlong_as_double(assumed))));
-  } while (assumed != old);
-  return __longlong_as_double(old);
+	unsigned long long int* address_as_ull = (unsigned long long int*) address;
+	unsigned long long int old = *address_as_ull, assumed;
+	do {
+		assumed = old;
+		old = atomicCAS(address_as_ull, assumed,
+				__double_as_longlong(fmin(val, __longlong_as_double(assumed))));
+	} while (assumed != old);
+	return __longlong_as_double(old);
 }
 
 #   define EXADEM_CU_ATOMIC_MIN(x,a,...) atomicMin_double( &x , static_cast<std::remove_reference_t<decltype(x)> >(a) )
@@ -43,242 +43,196 @@ __device__ double atomicMin_double(double* address, double val)
 
 namespace exaDEM
 {
-  namespace itools /* interaction tools */
-  {
-    using namespace onika::parallel;
-    template <typename T> using VectorT =  onika::memory::CudaMMVector<T>;
 
-    template<typename T, class FuncT, class ResultT>
-      struct ReduceTFunctor
-      {
-	const T* const m_data;
-	const FuncT m_func;
-	ResultT* m_reduced_val ;
-
-	ONIKA_HOST_DEVICE_FUNC inline void operator () ( uint64_t i ) const
+  /**
+   * @namespace itools
+   * Contains interaction tools for parallel processing in the exaDEM library.
+   */
+	namespace itools /* interaction tools */
 	{
-	  ResultT local_val = ResultT();
-	  m_func( local_val, i, m_data,  reduce_thread_local_t{} );
+		using namespace onika::parallel;
+		template <typename T> using VectorT =  onika::memory::CudaMMVector<T>;
 
-	  ONIKA_CU_BLOCK_SHARED onika::cuda::UnitializedPlaceHolder<ResultT> team_val_place_holder;
-	  ResultT& team_val = team_val_place_holder.get_ref();
 
-	  if( ONIKA_CU_THREAD_IDX == 0 ) { team_val = local_val; }
-	  ONIKA_CU_BLOCK_SYNC();
+    /**
+     * @struct ReduceTFunctor
+     * @brief Functor for reducing data in a parallel execution.
+     * 
+     * @tparam T The data type.
+     * @tparam FuncT The function type to perform the reduction.
+     * @tparam ResultT The result type that stores the reduction result.
+     */
+		template<typename T, class FuncT, class ResultT>
+			struct ReduceTFunctor
+			{
+				const T* const m_data;    /**< Pointer to the data to be reduced. */
+				const FuncT m_func;       /**< Functor that defines how reduction is performed. */
+				ResultT* m_reduced_val ;  /**< Pointer to the result of the reduction. */
 
-	  if( ONIKA_CU_THREAD_IDX != 0 )
-	  {
-	    m_func( team_val, local_val , reduce_thread_block_t{} );
-	  }
-	  ONIKA_CU_BLOCK_SYNC();
+      /**
+       * @brief Operator to perform the reduction.
+       * @param i The index of the data to reduce.
+       */
+				ONIKA_HOST_DEVICE_FUNC inline void operator () ( uint64_t i ) const
+				{
+					ResultT local_val = ResultT();
+					m_func( local_val, i, m_data,  reduce_thread_local_t{} );
[…]
