+++ src/interaction/include/exaDEM/interaction/classifier_for_all.hpp
@@ -8,114 +8,142 @@ namespace exaDEM
 {
   using namespace onika::parallel;
 
-  /**
-   * @brief Namespace for utilities related to tuple manipulation.
-   */
-  namespace tuple_helper
-  {
-    template <size_t... Is>
-      struct index {};
+	/**
+	 * @brief Namespace for utilities related to tuple manipulation.
+	 */
+	namespace tuple_helper
+	{
+		template <size_t... Is>
+			struct index {};
 
-    template <size_t N, size_t... Is>
-      struct gen_seq : gen_seq<N - 1, N - 1, Is...> {};
+		template <size_t N, size_t... Is>
+			struct gen_seq : gen_seq<N - 1, N - 1, Is...> {};
 
-    template <size_t... Is>
-      struct gen_seq<0, Is...> : index<Is...> {};
-  }
+		template <size_t... Is>
+			struct gen_seq<0, Is...> : index<Is...> {};
+	}
 
-  struct AnalysisDataPackerNull
-  {
-    template<typename... Args>
-      ONIKA_HOST_DEVICE_FUNC inline void operator() (Args&&...  args) const { /* do nothing */ }
-  };
+	struct AnalysisDataPackerNull
+	{
+		template<typename... Args>
+			ONIKA_HOST_DEVICE_FUNC inline void operator() (Args&&...  args) const { /* do nothing */ }
+	};
 
-  struct AnalysisDataPacker
-  {
-    double* dnp;
-    Vec3d*  cpp;
-    Vec3d*  fnp;
-    Vec3d*  ftp;
+	/**
+	 * @struct AnalysisDataPacker
+	 * @brief Packs analysis data for interactions, including overlap (dn), contact point, and forces.
+	 * 
+	 * This structure stores interaction-related data such as overlap (dn), contact point, 
+	 * normal force (fn), and tangential force (ft) into their respective buffers.
+	 */
+	struct AnalysisDataPacker
+	{
+		double* dnp; /**< Pointer to the buffer storing overlap (dn) values between particles. */
+		Vec3d*  cpp; /**< Pointer to the buffer storing contact point positions. */
+		Vec3d*  fnp; /**< Pointer to the buffer storing normal force vectors (fn). */
+		Vec3d*  ftp; /**< Pointer to the buffer storing tangential force vectors (ft). */
 
-    AnalysisDataPacker(Classifier& ic, int type)
-    {
-      auto [_dnp, _cpp, _fnp, _ftp] = ic.buffer_p(type);
-      dnp = _dnp;
-      cpp = _cpp;
-      fnp = _fnp;
-      ftp = _ftp;
-    }
+		/**
+		 * @brief Constructor that initializes the data packer with buffers from a classifier.
+		 * 
+		 * @param ic The classifier used to access the buffers for a given interaction type.
+		 * @param type The interaction type identifier to retrieve the appropriate buffers.
+		 * 
+		 * The constructor retrieves the buffers for overlap (dn), contact points, normal forces,
+		 * and tangential forces from the classifier based on the interaction type.
+		 */
+		AnalysisDataPacker(Classifier& ic, int type)
+		{
+			auto [_dnp, _cpp, _fnp, _ftp] = ic.buffer_p(type);
+			dnp = _dnp;
+			cpp = _cpp;
+			fnp = _fnp;
+			ftp = _ftp;
+		}
 
-    ONIKA_HOST_DEVICE_FUNC inline void operator() (const uint64_t idx, const double dn, const Vec3d& contact, const Vec3d& fn, const Vec3d& ft) const
-    { 
-      dnp[idx] = dn;
-      cpp[idx] = contact;
-      fnp[idx] = fn;
-      ftp[idx] = ft;
-    }
-  };
+		/**
[…]
