### 2024-09-18 318ddc1
create branch v1.0.2 for xsp-release-3.6
