### 2024-09-20 7b96bbc
update examples with the lastest version of exanbody
