### 2024-09-25 b11cf6a
update stl operators + move
