### 2024-09-25 99ee44e
InteractionSOA and InteractionAOS
