+++ src/interaction/include/exaDEM/interaction/interactionSOA.hpp
@@ -27,48 +27,38 @@ under the License.
 namespace exaDEM
 {
 	/**
-	 * @brief Structure representing an interaction in a Discrete Element Method (DEM) simulation.
+	 * @brief Structure representing the Structure of Arrays data structure for the interactions in a Discrete Element Method (DEM) simulation.
 	 */
 	struct InteractionSOA
 	{
 		template <typename T> using VectorT =  onika::memory::CudaMMVector<T>;
 		
-		VectorT<double> ft_x;
-		VectorT<double> ft_y;
-		VectorT<double> ft_z;
+		VectorT<double> ft_x;  /**< List of the x coordinate for the friction.  */
+		VectorT<double> ft_y;  /**< List of the y coordinate for the friction.  */
+		VectorT<double> ft_z;  /**< List of the z coordinate for the friction.  */
 		
-		VectorT<double> mom_x;
-		VectorT<double> mom_y;
-		VectorT<double> mom_z;
+		VectorT<double> mom_x;  /**< List of the x coordinate for the moment.  */
+		VectorT<double> mom_y;  /**< List of the y coordinate for the moment.  */
+		VectorT<double> mom_z;  /**< List of the z coordinate for the moment.  */
 		
-		VectorT<size_t> id_i;
-		VectorT<size_t> id_j;
+		VectorT<uint64_t> id_i;  /**< List of the ids of the first particle involved in the interaction.  */
+		VectorT<uint64_t> id_j;  /**< List of the ids of the second particle involved in the interaction.  */
 		
-		VectorT<size_t> cell_i;
-		VectorT<size_t> cell_j;
+		VectorT<uint32_t> cell_i;  /**< List of the indexes of the cell for the first particle involved in the interaction.  */
+		VectorT<uint32_t> cell_j;  /**< List of the indexes of the cell for the second particle involved in the interaction.  */
 		
-		VectorT<int> p_i;
-		VectorT<int> p_j;
+		VectorT<uint16_t> p_i;  /**< List of the indexes of the particle within its cell for the first particle involved in the interaction. */
+		VectorT<uint16_t> p_j;  /**< List of the indexes of the particle within its cell for the second particle involved in the interaction.  */
 		
-		VectorT<size_t> sub_i;
-		VectorT<size_t> sub_j;
+		VectorT<uint16_t> sub_i;  /**< List of the sub-particle indexes for the first particle involved in the interaction.  */
+		VectorT<uint16_t> sub_j;  /**< List of the sub-particle indexes for the first particle involved in the interaction.  */
 		
 		uint16_t type;          /**< Type of the interaction (e.g., contact type). */
 
+
 		/**
-		 * @brief Resets the Interaction structure by setting friction and moment vectors to zero.
-		 */
-		ONIKA_HOST_DEVICE_FUNC void reset( size_t id )
-		{
-			onika::cuda::vector_data(ft_x)[id] = 0;
-			onika::cuda::vector_data(ft_y)[id] = 0;
-			onika::cuda::vector_data(ft_z)[id] = 0;
-			
-			onika::cuda::vector_data(mom_x)[id] = 0;
-			onika::cuda::vector_data(mom_y)[id] = 0;
-			onika::cuda::vector_data(mom_z)[id] = 0;
-		}
-		
+		 *@briefs CLears all the lists.
+		 */		
 		void clear()
 		{
 			ft_x.clear();
[…]
