+++ src/sphere/contact_force_sphere.cpp
@@ -30,6 +30,8 @@ under the License.
 #include <exaDEM/contact_force_parameters.h>
 #include <exaDEM/compute_contact_force.h>
 #include <exaDEM/interaction/interaction.hpp>
+#include <exaDEM/interaction/interactionSOA.hpp>
+#include <exaDEM/interaction/interactionAOS.hpp>
 #include <exaDEM/interaction/grid_cell_interaction.hpp>
 #include <exaDEM/interaction/classifier.hpp>
 #include <exaDEM/interaction/classifier_for_all.hpp>
@@ -60,7 +62,7 @@ namespace exaDEM
     ADD_SLOT( double      , dt                , INPUT        , REQUIRED , DocString{"Time step value"});
     ADD_SLOT( bool        , symetric          , INPUT_OUTPUT , REQUIRED , DocString{"Activate the use of symetric feature (contact law)"});
     ADD_SLOT( Drivers     , drivers           , INPUT        , REQUIRED ,DocString{"List of Drivers {Cylinder, Surface, Ball, Mesh}"});
-    ADD_SLOT( Classifier  , ic                , INPUT_OUTPUT , DocString{"Interaction lists classified according to their types"} );
+    ADD_SLOT( Classifier<InteractionAOS>  , ic                , INPUT_OUTPUT , DocString{"Interaction lists classified according to their types"} );
     // analysis
     ADD_SLOT( long        , timestep          , INPUT , REQUIRED );
     ADD_SLOT( bool        , save_interactions , INPUT , false           , DocString{"Store interactions into the classifier"});
@@ -130,6 +132,9 @@ namespace exaDEM
       {
         run_contact_law(parallel_execution_context(), type, classifier, stlm, store_interactions, cells, drvs, hkp_drvs, time);  
       }
+      
+      //printf("EXECUTE\n");
+      //getchar();
 
       if(write_interactions)
       {
