### 2024-09-27 bd4fd78
V1.0.2-tmp (#61)
* update lattice examples

* update examples with the lastest version of exanbody

* update exaDEM to prepare for the integration of moving drivers

* update stl operators + move

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
