+++ src/interaction/include/exaDEM/interaction/classifier.hpp
@@ -211,6 +211,7 @@ namespace exaDEM
 					}
 				}
 			}
+			reset_waves();
 		}
 	};
 }
