+++ src/io/simulation_state.cpp
@@ -88,7 +88,6 @@ namespace exaDEM
 					}
 				}
 			} // synchronize 
-
 			for(int i = 0 ; i < types ; i++)
 			{ 
 				res.update(results[i]);

[…]
