### 2024-09-27 c3df2e7
Revert "New data structures for the interactions"
