+++ src/interaction/include/exaDEM/interaction/interactionSOA.hpp
@@ -1,163 +1,0 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-#pragma once
-
-#include <iostream>
-//#include <ostream>
-#include <exanb/core/basic_types.h>
-#include <exanb/core/basic_types_stream.h>
-#include <exaDEM/interaction/interaction.hpp>
-
-namespace exaDEM
-{
-	/**
-	 * @brief Structure representing the Structure of Arrays data structure for the interactions in a Discrete Element Method (DEM) simulation.
-	 */
-	struct InteractionSOA
-	{
-		template <typename T> using VectorT =  onika::memory::CudaMMVector<T>;
-		
-		VectorT<double> ft_x;  /**< List of the x coordinate for the friction.  */
-		VectorT<double> ft_y;  /**< List of the y coordinate for the friction.  */
-		VectorT<double> ft_z;  /**< List of the z coordinate for the friction.  */
-		
-		VectorT<double> mom_x;  /**< List of the x coordinate for the moment.  */
-		VectorT<double> mom_y;  /**< List of the y coordinate for the moment.  */
-		VectorT<double> mom_z;  /**< List of the z coordinate for the moment.  */
-		
-		VectorT<uint64_t> id_i;  /**< List of the ids of the first particle involved in the interaction.  */
-		VectorT<uint64_t> id_j;  /**< List of the ids of the second particle involved in the interaction.  */
-		
-		VectorT<uint32_t> cell_i;  /**< List of the indexes of the cell for the first particle involved in the interaction.  */
-		VectorT<uint32_t> cell_j;  /**< List of the indexes of the cell for the second particle involved in the interaction.  */
-		
-		VectorT<uint16_t> p_i;  /**< List of the indexes of the particle within its cell for the first particle involved in the interaction. */
-		VectorT<uint16_t> p_j;  /**< List of the indexes of the particle within its cell for the second particle involved in the interaction.  */
-		
-		VectorT<uint16_t> sub_i;  /**< List of the sub-particle indexes for the first particle involved in the interaction.  */
-		VectorT<uint16_t> sub_j;  /**< List of the sub-particle indexes for the first particle involved in the interaction.  */
-		
-		uint16_t type;          /**< Type of the interaction (e.g., contact type). */
-
-
-		/**
-		 *@briefs CLears all the lists.
-		 */		
-		void clear()
-		{
-			ft_x.clear();
-			ft_y.clear();
-			ft_z.clear();
-			
-			mom_x.clear();
-			mom_y.clear();
-			mom_z.clear();
-			
-			id_i.clear();
-			id_j.clear();
-			
[…]
