### 2024-09-28 0fb4357
add push v to r and av to quat functions for drivers
