### 2024-09-29 84c3c5d
remove debug logs
