+++ src/driver/update_grid_stl_mesh_v2.cpp
@@ -156,7 +156,7 @@ namespace exaDEM
 									}
 								}
 					}
-          mesh.grid_indexes_summary(); //for debug
+          //mesh.grid_indexes_summary(); //for debug
 				}
 			}
 		}
