+++ src/io/simulation_state.cpp
@@ -80,8 +80,7 @@ namespace exaDEM
 					if( i < 4 && symetric ) coef *= 2;
 
 					IOSimInteractionFunctor func = {dnp, coef};
-
-					if ( size > 0 && dnp != nullptr ) // skip it if forces has not been computed
+					if ( size > 0 && dnp != nullptr ) // skip it if forces have not been computed
 					{
 						//pexw[i] = exaDEM::itools::reduce_data<exaDEM::Interaction, IOSimInteractionFunctor, IOSimInteractionResult>(parallel_execution_context(), ptr, func, size, results[i]);
 						reduce_data<exaDEM::Interaction, IOSimInteractionFunctor, IOSimInteractionResult>(parallel_execution_context(), ptr, func, size, results[i]);
@@ -118,13 +117,14 @@ namespace exaDEM
       uint64_t active_interactions, total_interactions;
       double dn;
 			{
-				double tmpDouble[8] = {
+				double tmpDouble[7] = {
 					sim.rotation_energy.x, sim.rotation_energy.y, sim.rotation_energy.z,
 					sim.kinetic_energy.x, sim.kinetic_energy.y, sim.kinetic_energy.z,
-					sim.mass , red.min_dn };
+					sim.mass };
         uint64_t tmpUInt64T[3] = {sim.n_particles, red.n_act_interaction, red.n_tot_interaction};
-				MPI_Allreduce(MPI_IN_PLACE, tmpDouble, 8, MPI_DOUBLE, MPI_SUM, comm);
-				MPI_Allreduce(MPI_IN_PLACE, tmpUInt64T, 3, MPI_UINT64_T, MPI_SUM, comm);
+				MPI_Allreduce(MPI_IN_PLACE, tmpDouble  , 8, MPI_DOUBLE, MPI_SUM, comm);
+				MPI_Allreduce(MPI_IN_PLACE, &red.min_dn, 1, MPI_DOUBLE, MPI_MAX, comm);
+				MPI_Allreduce(MPI_IN_PLACE, tmpUInt64T , 3, MPI_UINT64_T, MPI_SUM, comm);
 
 				rotation_energy.x = tmpDouble[0];
 				rotation_energy.y = tmpDouble[1];
@@ -133,7 +133,7 @@ namespace exaDEM
 				kinetic_energy.y  = tmpDouble[4];
 				kinetic_energy.z  = tmpDouble[5];
 				mass              = tmpDouble[6];
-				dn                = tmpDouble[7];
+				dn                = red.min_dn;
 
 				total_particles     = tmpUInt64T[0];
 				active_interactions = tmpUInt64T[1];
