### 2024-09-30 e4c7258
update movable wall test case
