### 2024-09-30 f64bca1
clean stl mesh driver
