+++ src/driver/include/exaDEM/driver_stl_mesh.h
@@ -1,217 +1,0 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-#pragma once
-#include <exanb/core/basic_types.h>
-#include <exaDEM/driver_base.h>
-#include <exaDEM/shape/shape.hpp>
-#include <exaDEM/interaction/interaction.hpp>
-
-namespace exaDEM
-{
-  using namespace exanb;
-
-  template<typename T> using vector_t = std::vector<T>;
-
-  /**
-   * @brief Struct representing a list of elements( vertex, edge, or face).
-   */
-  struct list_of_elements
-  {
-    vector_t<int> vertices;  /**< List of vertex indices. */
-    vector_t<int> edges;     /**< List of edge indices. */
-    vector_t<int> faces;     /**< List of face indices. */
-  };
-
-  /**
-   * @brief Struct representing a STL mesh in the exaDEM simulation.
-   */
-  struct Stl_mesh
-  {
-    exanb::Vec3d center;            /**< Center position of the STL mesh. */
-    exanb::Vec3d vel;               /**< Velocity of the STL mesh. */
-    exanb::Vec3d vrot;              /**< Angular velocity of the STL mesh. */
-    exanb::Quaternion quat;         /**< Quaternion of the STL mesh. */
-    shape shp;                      /**< Shape of the STL mesh. */
-  //  vector_t<Vec3d> verticies;      /**< Collection of vertices (computed from shp, quat and center). */
-    vector_t<list_of_elements> grid_indexes; /**< Grid indices of the STL mesh. */
-
-    /**
-     * @brief Get the type of the driver (in this case, STL_MESH).
-     * @return The type of the driver.
-     */
-    constexpr DRIVER_TYPE get_type() {return DRIVER_TYPE::STL_MESH;}
-
-    /**
-     * @brief Print information about the STL mesh.
-     */
-    void print()
-    {
-      lout << "Driver Type: MESH STL" << std::endl;
-      lout << "center : " << center   << std::endl;
-      lout << "Vel    : " << vel << std::endl;
-      lout << "AngVel : " << vrot << std::endl;
-      lout << "Quat   : " << quat.w << " " << quat.x << " " << quat.y << " " << quat.z  << std::endl;
-      lout << "Number of faces    : " << shp.get_number_of_faces() << std::endl;
-      lout << "Number of edges    : " << shp.get_number_of_edges() << std::endl;
-      lout << "Number of vertices : " << shp.get_number_of_vertices() << std::endl;
-    }
-
-    /**
-     * @brief Print information about the STL mesh.
-     */
-    ONIKA_HOST_DEVICE_FUNC inline void initialize ()
-    {
-      // checks
-    }
-
-    /**
-     * @brief return driver velocity
-     */
-    ONIKA_HOST_DEVICE_FUNC inline Vec3d& get_vel()
[…]
