### 2024-10-01 a12f8fe
Résolution du problème avec les logs
