+++ src/interaction/include/exaDEM/interaction/interactionSOA.hpp
@@ -158,6 +158,30 @@ namespace exaDEM
 			onika::cuda::vector_data(mom_z)[id] = item.moment.z;
 		}
 		
+		exaDEM::Interaction* interactions_list()
+		{
+			std::vector<exaDEM::Interaction> interactions;
+			
+			for(int i = 0; i < ft_x.size(); i++)
+			{
+				interactions.push_back({{ft_x[i], ft_y[i], ft_z[i]}, {mom_x[i], mom_y[i], mom_z[i]}, id_i[i], id_j[i], cell_i[i], cell_j[i], p_i[i], p_j[i], sub_i[i], sub_j[i], type});
+			}
+			
+			return interactions.data();
+		}
+		
+		const exaDEM::Interaction* interactions_list() const
+		{
+			std::vector<exaDEM::Interaction> interactions;
+			
+			for(int i = 0; i < ft_x.size(); i++)
+			{
+				interactions.push_back({{ft_x[i], ft_y[i], ft_z[i]}, {mom_x[i], mom_y[i], mom_z[i]}, id_i[i], id_j[i], cell_i[i], cell_j[i], p_i[i], p_j[i], sub_i[i], sub_j[i], type});
+			}
+			
+			return interactions.data();
+		}
+		
 
 	};
 }
