### 2024-10-03 0fa06ae
Problems GPU data
