### 2024-10-05 3f2799a
Port the set_type operator on gpu
