### 2024-10-05 f74af6b
fix set_quaternion operator with cuda
