### 2024-10-07 eab2508
log issue reoslution
