+++ src/io/simulation_state.cpp
@@ -62,7 +62,7 @@ namespace exaDEM
 		static constexpr FieldSet<field::_vx ,field::_vy ,field::_vz, field::_vrot, field::_mass> reduce_field_set {};
 		
 		template< typename T >
-		inline IOSimInteractionResult reduce_sim_io(const Classifier<T>& classifier, bool symetric)
+		inline IOSimInteractionResult reduce_sim_io(Classifier<T>& classifier, bool symetric)
 		{
 			IOSimInteractionResult res;
 			VectorT<IOSimInteractionResult> results;
@@ -79,13 +79,17 @@ namespace exaDEM
 
 					int coef = 1;
 					if( i < 4 && symetric ) coef *= 2;
+					
+					InteractionWrapper<T> data(ptr);
 
-					IOSimInteractionFunctor func = {dnp, coef};
+					//IOSimInteractionFunctor func = {dnp, coef};
+					IOSimInteractionFunctorWrapper func = {dnp, coef};
 
 					if ( size > 0 && dnp != nullptr ) // skip it if forces has not been computed
 					{
 						//pexw[i] = exaDEM::itools::reduce_data<exaDEM::Interaction, IOSimInteractionFunctor, IOSimInteractionResult>(parallel_execution_context(), ptr, func, size, results[i]);
 						//reduce_data<exaDEM::Interaction, IOSimInteractionFunctor, IOSimInteractionResult>(parallel_execution_context(), ptr, func, size, results[i]);
+						reduce_data_wrapper<T, IOSimInteractionFunctorWrapper, IOSimInteractionResult>(parallel_execution_context(), data, func, size, results[i]);
 					}
 				}
 			} // synchronize 
@@ -112,7 +116,7 @@ namespace exaDEM
 			reduce_cell_particles( *grid , false , func , sim, reduce_field_set , parallel_execution_context() , {} , cell_ptr, cell_size );
 
       // get interaction informations
-      const Classifier<InteractionSOA>& classifier = *ic;
+      Classifier<InteractionSOA>& classifier = *ic;
       exaDEM::itools::IOSimInteractionResult red = reduce_sim_io(classifier, *symetric );
 
 			// reduce partial sums and share the result
