+++ src/driver/add_cylinder.cpp
@@ -16,17 +16,9 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-//#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
 #include "exanb/core/operator.h"
 #include "exanb/core/operator_slot.h"
 #include "exanb/core/operator_factory.h"
-#include "exanb/core/make_grid_variant_operator.h"
-#include "exanb/core/parallel_grid_algorithm.h"
-#include "exanb/core/grid.h"
-#include "exanb/core/domain.h"
-#include "exanb/compute/compute_cell_particles.h"
-#include <mpi.h>
-#include <memory>
 #include <exaDEM/driver_base.h>
 #include <exaDEM/drivers.h>
 #include <exaDEM/cylinder.h>
@@ -36,45 +28,42 @@ namespace exaDEM
 
   using namespace exanb;
 
-  template<typename GridT>
-    class AddCylinder : public OperatorNode
-    {
-      static constexpr Vec3d default_axis = { 1.0, 0.0, 1.0 };
-      static constexpr Vec3d null= { 0.0, 0.0, 0.0 };
+	class AddCylinder : public OperatorNode
+	{
+		static constexpr Vec3d default_axis = { 1.0, 0.0, 1.0 };
+		static constexpr Vec3d null= { 0.0, 0.0, 0.0 };
 
 
-      ADD_SLOT( Drivers , drivers    , INPUT_OUTPUT, REQUIRED    , DocString{"List of Drivers"});
-      ADD_SLOT( int     , id         , INPUT       , REQUIRED , DocString{"Driver index"});
-      ADD_SLOT( Vec3d   , center     , INPUT       , REQUIRED , DocString{"Center of the cylinder"});
-      ADD_SLOT( Vec3d   , axis       , INPUT       , default_axis , DocString{"Define the plan of the cylinder"});
-      ADD_SLOT( Vec3d   , angular_velocity, INPUT  , null     , DocString{"Angular velocity of the cylinder, default is 0 m.s-"});
-      ADD_SLOT( Vec3d   , velocity   , INPUT       , null     , DocString{"Cylinder velocity, could be used in 'expert mode'"});
-      ADD_SLOT( double  , radius     , INPUT       , REQUIRED , DocString{"Radius of the cylinder, positive and should be superior to the biggest sphere radius in the cylinder"});
+		ADD_SLOT( Drivers , drivers    , INPUT_OUTPUT, REQUIRED    , DocString{"List of Drivers"});
+		ADD_SLOT( int     , id         , INPUT       , REQUIRED , DocString{"Driver index"});
+		ADD_SLOT( Vec3d   , center     , INPUT       , REQUIRED , DocString{"Center of the cylinder"});
+		ADD_SLOT( Vec3d   , axis       , INPUT       , default_axis , DocString{"Define the plan of the cylinder"});
+		ADD_SLOT( Vec3d   , angular_velocity, INPUT  , null     , DocString{"Angular velocity of the cylinder, default is 0 m.s-"});
+		ADD_SLOT( Vec3d   , velocity   , INPUT       , null     , DocString{"Cylinder velocity, could be used in 'expert mode'"});
+		ADD_SLOT( double  , radius     , INPUT       , REQUIRED , DocString{"Radius of the cylinder, positive and should be superior to the biggest sphere radius in the cylinder"});
 
-      public:
+		public:
 
-      inline std::string documentation() const override final
-      {
-        return R"EOF(
+		inline std::string documentation() const override final
+		{
+			return R"EOF(
         This operator add a cylinder to the drivers list.
         )EOF";
-      }
+		}
 
[…]
