+++ src/driver/add_stl_mesh.cpp
@@ -16,15 +16,9 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-//#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
 #include "exanb/core/operator.h"
 #include "exanb/core/operator_slot.h"
 #include "exanb/core/operator_factory.h"
-#include "exanb/core/make_grid_variant_operator.h"
-#include "exanb/core/parallel_grid_algorithm.h"
-#include "exanb/core/grid.h"
-#include "exanb/core/domain.h"
-#include "exanb/compute/compute_cell_particles.h"
 #include <mpi.h>
 #include <memory>
 #include <exaDEM/driver_base.h>
@@ -38,7 +32,6 @@ namespace exaDEM
 
 	using namespace exanb;
 
-	template<typename GridT>
 		class AddSTLMesh : public OperatorNode
 	{
 		static constexpr Vec3d null= { 0.0, 0.0, 0.0 };
@@ -84,12 +77,10 @@ namespace exaDEM
 		}
 	};
 
-	template<class GridT> using AddSTLMeshTmpl = AddSTLMesh<GridT>;
-
 	// === register factories ===  
 	CONSTRUCTOR_FUNCTION
 	{
-		OperatorNodeFactory::instance()->register_factory( "add_stl_mesh", make_grid_variant_operator< AddSTLMeshTmpl > );
+		OperatorNodeFactory::instance()->register_factory( "add_stl_mesh", make_simple_operator< AddSTLMesh > );
 	}
 }
 
