+++ src/driver/add_surface.cpp
@@ -16,17 +16,9 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-//#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
 #include "exanb/core/operator.h"
 #include "exanb/core/operator_slot.h"
 #include "exanb/core/operator_factory.h"
-#include "exanb/core/make_grid_variant_operator.h"
-#include "exanb/core/parallel_grid_algorithm.h"
-#include "exanb/core/grid.h"
-#include "exanb/core/domain.h"
-#include "exanb/compute/compute_cell_particles.h"
-#include <mpi.h>
-#include <memory>
 #include <exaDEM/driver_base.h>
 #include <exaDEM/drivers.h>
 #include <exaDEM/surface.h>
@@ -36,7 +28,6 @@ namespace exaDEM
 
   using namespace exanb;
 
-  template<typename GridT>
     class AddSurface : public OperatorNode
     {
       ADD_SLOT( Drivers , drivers  , INPUT_OUTPUT , REQUIRED    , DocString{"List of Drivers"});
@@ -64,12 +55,11 @@ namespace exaDEM
       }
     };
 
-  template<class GridT> using AddSurfaceTmpl = AddSurface<GridT>;
 
   // === register factories ===  
   CONSTRUCTOR_FUNCTION
   {
-    OperatorNodeFactory::instance()->register_factory( "add_surface", make_grid_variant_operator< AddSurfaceTmpl > );
+    OperatorNodeFactory::instance()->register_factory( "add_surface", make_simple_operator< AddSurface > );
   }
 }
 

[…]
