+++ src/interaction/dem_cost_model.cpp
@@ -20,6 +20,7 @@ under the License.
 #include <exanb/core/operator_slot.h>
 #include <exanb/core/operator_factory.h>
 #include <exanb/core/make_grid_variant_operator.h>
+#include <exanb/core/parallel_grid_algorithm.h>
 #include <exanb/core/log.h>
 #include <exanb/core/grid.h>
 #include <exanb/fields.h>
