### 2024-10-07 5fb2f8e
remove unused functions
