### 2024-10-07 b1c1b9f
Interactions soa (#63)
* Structure SOA
* InteractionSOA and InteractionAOS
* Corrections Pull Resquest
* Résolution du problème avec les logs
* Problems GPU data
* log issue reoslution
* remove unused functions

---------

Co-authored-by: DONCECCHI Carlo-Elia <doncecchicarloelia@gmail.com>
Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
