+++ src/interaction/stats_interactions.cpp
@@ -55,8 +55,12 @@ namespace exaDEM
 		{
 			auto & cells = ges->m_data;
 
-			int nvv(0), nve(0), nvf(0), nee(0); // interaction counters
+      // v vertex, e edge, f face, c cylinder, s surface, b ball, S stl
+
+			int nvv(0), nve(0), nvf(0), nee(0); // interaction counters [particles]
 			int an(0), anvv(0), anve(0), anvf(0), anee(0); // active interaction counters
+			int nvc(0), nvs(0), nvb(0), nSvv(0), nSve(0), nSvf(0), nSee(0), nSev(0), nSfv(0); // interaction counters [drivers]
+			int anvc(0), anvs(0), anvb(0), anSvv(0), anSve(0), anSvf(0), anSee(0), anSev(0), anSfv(0); // interaction counters [drivers]
 
 			const exanb::Vec3d null = {0.,0.,0.};
 
@@ -75,14 +79,40 @@ namespace exaDEM
 			{
 				for( auto& item : cells[i].m_data)
 				{
-					if( item.type == 0 ) incr_interaction_counters(item, nvv, anvv, an); 
-					if( item.type == 1 ) incr_interaction_counters(item, nve, anve, an); 
-					if( item.type == 2 ) incr_interaction_counters(item, nvf, anvf, an); 
-					if( item.type == 3 ) incr_interaction_counters(item, nee, anee, an); 
+					// particles
+					if( item.type == 0 )  incr_interaction_counters(item, nvv, anvv, an); 
+					if( item.type == 1 )  incr_interaction_counters(item, nve, anve, an); 
+					if( item.type == 2 )  incr_interaction_counters(item, nvf, anvf, an); 
+					if( item.type == 3 )  incr_interaction_counters(item, nee, anee, an); 
+					// drivers
+					// cylinder
+					if( item.type == 4 )  incr_interaction_counters(item, nvc, anvc, an); 
+					//surface
+					if( item.type == 5 )  incr_interaction_counters(item, nvs, anvs, an); 
+					// ball
+					if( item.type == 6 )  incr_interaction_counters(item, nvb, anvb, an); 
+					//stl
+					if( item.type == 7 )  incr_interaction_counters(item, nSvv, anSvv, an); 
+					if( item.type == 8 )  incr_interaction_counters(item, nSve, anSve, an); 
+					if( item.type == 9 )  incr_interaction_counters(item, nSvf, anSvf, an); 
+					if( item.type == 10 ) incr_interaction_counters(item, nSee, anSee, an); 
+					if( item.type == 11 ) incr_interaction_counters(item, nSev, anSev, an); 
+					if( item.type == 12 ) incr_interaction_counters(item, nSfv, anSfv, an); 
 				}
 			}
 
-			std::vector<int> val = {nvv, nve, nvf, nee, an, anvv, anve, anvf, anee};
+			std::vector<int> val = {
+				//particle
+				nvv, nve, nvf, nee, 
+				// driver
+				nvc, nvs, nvb, nSvv, nSve, nSvf, nSee, nSev, nSfv,
+				// total
+				an, 
+				// particle
+				anvv, anve, anvf, anee,
+				// driver
+				anvc, anvs, anvb, anSvv, anSve, anSvf, anSee, anSev, anSfv
+			};
 
 			int rank;
 			MPI_Comm_rank(*mpi, &rank);
@@ -93,15 +123,24 @@ namespace exaDEM
 				MPI_Reduce(val.data(), val.data(), val.size(), MPI_INT, MPI_SUM, 0, *mpi);
 
 			int idx = 0 ;
[…]
