### 2024-10-07 38eae4f
V1.0.2 (#64)
* update stl operators + move

* add push v to r and av to quat functions for drivers

* remove debug logs

* update logs with stl drivers

* rename some operators

* update movable wall test case

* use visit instead if

* clean stl mesh driver

* udpate actions/download-artifact@v4 in draft-pdf

* fix using visit in push av to quat

* add test case cylinder_stl

* Port the set_type operator on gpu

* add backup for drivers + driver displ over. remove spec verlet .msp

* fix set_quaternion operator with cuda

* clean includes and make_simple_operators for add driver operators

* 
[…]
