### 2024-10-09 479b7a2
add write_shape_file operator
