+++ src/interaction/include/exaDEM/shape/shapesSOA.hpp
@@ -1,0 +1,624 @@
+/*
+Licensed to the Apache Software Foundation (ASF) under one
+or more contributor license agreements.  See the NOTICE file
+distributed with this work for additional information
+regarding copyright ownership.  The ASF licenses this file
+to you under the Apache License, Version 2.0 (the
+"License"); you may not use this file except in compliance
+with the License.  You may obtain a copy of the License at
+
+  http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing,
+software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+KIND, either express or implied.  See the License for the
+specific language governing permissions and limitations
+under the License.
+*/
+#pragma once
+
+#include<exaDEM/shape/shape.hpp>
+
+namespace exaDEM
+{
+	// data collection of shapes
+	struct shapesSOA
+	{
+		template <typename T> using VectorT =  onika::memory::CudaMMVector<T>; 
+		
+		int num_shapes = 0; //Number of shapes
+		
+		//VectorT <exanb::Vec3d> m_vertices; ///<
+		VectorT<int> m_vertices_start;
+		VectorT<double> m_vertices_x;
+		VectorT<double> m_vertices_y;
+		VectorT<double> m_vertices_z;
+		VectorT<int> m_vertices_end;
+		  
+		//exanb::Vec3d m_inertia_on_mass;
+		
+		VectorT<double> m_inertia_on_mass_x;
+		VectorT<double> m_inertia_on_mass_y;
+		VectorT<double> m_inertia_on_mass_z;
+		
+		VectorT<double> obb_center_x;
+		VectorT<double> obb_center_y;
+		VectorT<double> obb_center_z;
+		VectorT<double> obb_e1_x;
+		VectorT<double> obb_e1_y;
+		VectorT<double> obb_e1_z;
+		VectorT<double> obb_e2_x;
+		VectorT<double> obb_e2_y;
+		VectorT<double> obb_e2_z;
+		VectorT<double> obb_e3_x;
+		VectorT<double> obb_e3_y;
+		VectorT<double> obb_e3_z;
+		VectorT<double> obb_extent_x;
+		VectorT<double> obb_extent_y;
+		VectorT<double> obb_extent_z;
+		
+		VectorT<double> m_radius;
+		VectorT<double> m_volume;
+		
+		VectorT<std::string> m_name;
+
+
+		//OBB		
+		//vec3r center;  //< Center
+		//vec3r e1;    //< 3 directions (normalized vectors)
+		//vec3r e2;    //< 3 directions (normalized vectors)
+		//vec3r e3;    //< 3 directions (normalized vectors)
+		//vec3r extent;  //< 3 extents (in the the 3 directions)
+		
+		//VectorT <OBB> m_obb_vertices; ///< only used for stl meshes
+		VectorT<int> m_obb_vertices_start;
+		//center
+		VectorT<double> m_obb_vertices_center_x;
+		VectorT<double> m_obb_vertices_center_y;
+		VectorT<double> m_obb_vertices_center_z;
+		//e1
+		VectorT<double> m_obb_vertices_e1_x;
+		VectorT<double> m_obb_vertices_e1_y;
[…]
