+++ src/driver/dump_drivers.cpp
@@ -1,0 +1,81 @@
+/*
+Licensed to the Apache Software Foundation (ASF) under one
+or more contributor license agreements.  See the NOTICE file
+distributed with this work for additional information
+regarding copyright ownership.  The ASF licenses this file
+to you under the Apache License, Version 2.0 (the
+"License"); you may not use this file except in compliance
+with the License.  You may obtain a copy of the License at
+
+  http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing,
+software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+KIND, either express or implied.  See the License for the
+specific language governing permissions and limitations
+under the License.
+*/
+//#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
+#include "exanb/core/operator.h"
+#include "exanb/core/operator_slot.h"
+#include "exanb/core/operator_factory.h"
+#include "exanb/core/parallel_grid_algorithm.h"
+#include <mpi.h>
+#include <memory>
+#include <exaDEM/stl_mesh.h>
+#include <exaDEM/drivers.h>
+#include <exanb/core/string_utils.h>
+
+namespace exaDEM
+{
+  using namespace exanb;
+  class DumpDriver : public OperatorNode
+  {
+    static constexpr Vec3d null= { 0.0, 0.0, 0.0 };
+
+    ADD_SLOT( Drivers     , drivers  , INPUT_OUTPUT, REQUIRED , DocString{"List of Drivers"});
+    ADD_SLOT( long        , timestep , INPUT                  , DocString{"Iteration number"});
+    ADD_SLOT( std::string , dir_name , INPUT       , REQUIRED , DocString{"Main output directory."} );
+
+    public:
+
+    inline std::string documentation() const override final
+    {
+      return R"EOF( This operator outputs driver information. )EOF";
+    }
+
+    inline void execute () override final
+    {
+      auto& drvs = *drivers;
+      size_t n_drivers = drivers->get_size();
+
+      if( n_drivers == 0 ) return;
+
+      std::string path = *dir_name + "/CheckpointFiles/";
+      int id = 0;
+      std::stringstream data_stream; 
+      std::string filename = path + "driver_%010d.msp";
+      filename = format_string(filename, *timestep);
+      data_stream << "setup_drivers:" << std::endl;
+
+      for(size_t i = 0 ; i < n_drivers ; i++)
+      {
+        if ( drvs.type(i) == DRIVER_TYPE::STL_MESH)
+        {
+          exaDEM::Stl_mesh& mesh = std::get<exaDEM::Stl_mesh> (drvs.data(i));
+          mesh.dump_driver(id++, path, data_stream);
+        }
+      }
+      std::ofstream file(filename.c_str());
+      file << data_stream.rdbuf();
+    }
+  };
+
+  // === register factories ===  
+  CONSTRUCTOR_FUNCTION
+  {
+    OperatorNodeFactory::instance()->register_factory( "dump_driver", make_simple_operator< DumpDriver > );
+  }
+}
+
