### 2024-10-11 a6c4e95
fix performance issue with reduce_data in simulation_state
