+++ src/interaction/include/exaDEM/itools/counter.hpp
@@ -39,8 +39,6 @@ __device__ double atomicMin_double(double* address, double val)
 #   define EXADEM_CU_ATOMIC_MIN(x,a,...) ::onika::capture_atomic_min( x , static_cast<std::remove_reference_t<decltype(x)> >(a) )
 #endif
 
-
-
 namespace exaDEM
 {
 
@@ -53,7 +51,6 @@ namespace exaDEM
 		using namespace onika::parallel;
 		template <typename T> using VectorT =  onika::memory::CudaMMVector<T>;
 
-
     /**
      * @struct ReduceTFunctor
      * @brief Functor for reducing data in a parallel execution.
@@ -64,41 +61,6 @@ namespace exaDEM
      */
 		template<typename T, class FuncT, class ResultT>
 			struct ReduceTFunctor
-			{
-				const T* const m_data;    /**< Pointer to the data to be reduced. */
-				const FuncT m_func;       /**< Functor that defines how reduction is performed. */
-				ResultT* m_reduced_val ;  /**< Pointer to the result of the reduction. */
-
-      /**
-       * @brief Operator to perform the reduction.
-       * @param i The index of the data to reduce.
-       */
-				ONIKA_HOST_DEVICE_FUNC inline void operator () ( uint64_t i ) const
-				{
-					ResultT local_val = ResultT();
-					m_func( local_val, i, m_data,  reduce_thread_local_t{} );
-
-					ONIKA_CU_BLOCK_SHARED onika::cuda::UnitializedPlaceHolder<ResultT> team_val_place_holder;
-					ResultT& team_val = team_val_place_holder.get_ref();
-
-					if( ONIKA_CU_THREAD_IDX == 0 ) { team_val = local_val; }
-					ONIKA_CU_BLOCK_SYNC();
-
-					if( ONIKA_CU_THREAD_IDX != 0 )
-					{
-						m_func( team_val, local_val , reduce_thread_block_t{} );
-					}
-					ONIKA_CU_BLOCK_SYNC();
-
-					if( ONIKA_CU_THREAD_IDX == 0 )
-					{
-						m_func( *m_reduced_val, team_val , reduce_global_t{} );
-					}
-				}
-			};
-
-		template<typename T, class FuncT, class ResultT>
-			struct ReduceTFunctorWrapper
 			{
 				InteractionWrapper<T> m_data;    /**< Pointer to the data to be reduced. */
 				const FuncT m_func;       /**< Functor that defines how reduction is performed. */
@@ -132,20 +94,21 @@ namespace exaDEM
 				}
 			};
 
-    /**
-     * @struct IOSimInteractionResult
-     * @brief Stores the results.
-     */
+
+		/**
+		 * @struct IOSimInteractionResult
+		 * @brief Stores the results.
+		 */
 		struct IOSimInteractionResult
 		{
 			unsigned long long int n_act_interaction = 0; /**< Number of active interactions. */
 			unsigned long long int n_tot_interaction = 0; /**< Total number of interactions. */
 			double min_dn = 0; /**< Minimum distance between particles during interactions. Or Max overlapping between between two particles. */
 
-      /**
-       * @brief Updates the current interaction result with data from another result.
-       * @param in The input interaction result to merge.
-       */
+			/**
+			 * @brief Updates the current interaction result with data from another result.
+			 * @param in The input interaction result to merge.
+			 */
 			void update(IOSimInteractionResult& in)
 			{
 				n_act_interaction += in.n_act_interaction;
@@ -154,54 +117,52 @@ namespace exaDEM
 			}   
 		};
 
-    /**
-     * @struct IOSimInteractionFunctor
-     * @brief Functor that reduces interaction information (act interactions, tot interaction and max(|dn|) <-> min(dn).
-     */
 		struct IOSimInteractionFunctor
 		{
[…]
