### 2024-10-11 b792fe4
fix performance issue with reduce_data in simulation_state
