+++ src/interaction/include/exaDEM/shape/shape_printer.hpp
@@ -23,83 +23,86 @@ under the License.
 
 namespace exaDEM
 {
-	void build_buffer (const exanb::Vec3d& pos, const shape* shp, const exanb::Quaternion& orient,  size_t& polygon_offset_in_stream,
-			size_t& n_vertices, 	size_t& n_polygon, 			std::stringstream& buff_position, std::stringstream& buff_faces, std::stringstream& buff_offset)
-	{
-		auto writer_v = [] (const exanb::Vec3d& v, std::stringstream& out, const exanb::Vec3d& p, const exanb::Quaternion& q)
-		{
-			exanb::Vec3d new_pos = p + q * v;
-			out << " " << new_pos.x << " " << new_pos.y << " " << new_pos.z;
-		};
+  void build_buffer(const exanb::Vec3d &pos, const shape *shp, const exanb::Quaternion &orient, size_t &polygon_offset_in_stream, size_t &n_vertices, size_t &n_polygon, std::stringstream &buff_position, std::stringstream &buff_faces, std::stringstream &buff_offset)
+  {
+    auto writer_v = [](const exanb::Vec3d &v, std::stringstream &out, const exanb::Vec3d &p, const exanb::Quaternion &q)
+    {
+      exanb::Vec3d new_pos = p + q * v;
+      out << " " << new_pos.x << " " << new_pos.y << " " << new_pos.z;
+    };
 
-		shp->for_all_vertices(writer_v, buff_position, pos, orient);
+    shp->for_all_vertices(writer_v, buff_position, pos, orient);
 
-		size_t n_faces = shp->get_number_of_faces();
-		n_polygon += n_faces;
+    size_t n_faces = shp->get_number_of_faces();
+    n_polygon += n_faces;
 
-		// faces 
-		auto writer_f = [] (const size_t size, const int * data, std::stringstream& sface, std::stringstream& soffset, size_t& offset, size_t point_off)
-		{
-			soffset << offset + size << " ";
-			offset += size;
-			for (size_t it = 0 ; it < size ; it++) sface << " " << data[it] + point_off;
-		};
-		shp->for_all_faces(writer_f, buff_faces, buff_offset, polygon_offset_in_stream, n_vertices);
-		n_vertices += shp->get_number_of_vertices();
-	}
+    // faces
+    auto writer_f = [](const size_t size, const int *data, std::stringstream &sface, std::stringstream &soffset, size_t &offset, size_t point_off)
+    {
+      soffset << offset + size << " ";
+      offset += size;
+      for (size_t it = 0; it < size; it++)
+        sface << " " << data[it] + point_off;
+    };
+    shp->for_all_faces(writer_f, buff_faces, buff_offset, polygon_offset_in_stream, n_vertices);
+    n_vertices += shp->get_number_of_vertices();
+  }
 
-	void write_vtp (std::string name, size_t n_vertices, size_t n_polygons, 
-			std::stringstream& buff_vertices, std::stringstream& buff_faces, std::stringstream& buff_offsets )
-	{
-		name = name + ".vtp";
-		std::ofstream outFile(name);
-		if (!outFile) {
-			std::cerr << "Erreur : impossible de créer le fichier de sortie suivant: " << name << std::endl;
-			return;
-		}
+  void write_vtp(std::string name, size_t n_vertices, size_t n_polygons, std::stringstream &buff_vertices, std::stringstream &buff_faces, std::stringstream &buff_offsets)
+  {
+    name = name + ".vtp";
+    std::ofstream outFile(name);
+    if (!outFile)
+    {
[…]
