+++ src/interaction/include/exaDEM/type/OBB.hpp
@@ -1,241 +1,252 @@
 // Copyright (C) OBB <vincent.richefeu@3sr-grenoble.fr>
-// 
+//
 // This file is part of mbox.
-// 
+//
 // OBB can not be copied and/or distributed without the express
 // permission of the authors.
 // It is coded for academic purposes.
 //
-// Note 
-// Without a license, the code is copyrighted by default. 
-// People can read the code, but they have no legal right to use it. 
+// Note
+// Without a license, the code is copyrighted by default.
+// People can read the code, but they have no legal right to use it.
 // To use the code, you must contact the author directly and ask permission.
 
 #ifndef OBB_HPP_D3958314
-#define OBB_HPP_D3958314
-/// @file
-/// @brief Oriented Bounding Box
-/// @author Vincent Richefeu <Vincent.Richefeu@3sr-grenoble.fr>,
-/// Lab 3SR, Grenoble University
+# define OBB_HPP_D3958314;
+  /// @file
+  /// @brief Oriented Bounding Box
+  /// @author Vincent Richefeu <Vincent.Richefeu@3sr-grenoble.fr>,
+  /// Lab 3SR, Grenoble University
 
-#include <cfloat>
+# include <cfloat>
 
-#include "vec3.hpp"
-#include "quat.hpp"
-#include "mat9.hpp"
-#include <exanb/core/basic_types.h>
+# include "vec3.hpp"
+# include "quat.hpp"
+# include "mat9.hpp"
+# include <exanb/core/basic_types.h>
 
-
-/// @ingroup Bounding_Volumes
-/// @brief Oriented Bounding Box
-class OBB
-{
-	public:
-		vec3r center;  //< Center
-		vec3r e1;    //< 3 directions (normalized vectors)
-		vec3r e2;    //< 3 directions (normalized vectors)
-		vec3r e3;    //< 3 directions (normalized vectors)
-		vec3r extent;  //< 3 extents (in the the 3 directions)
-
-		// Constructors
-		ONIKA_HOST_DEVICE_FUNC
-			OBB(): center(), extent(0.0, 0.0, 0.0) {
-				e1.set(1.0, 0.0, 0.0);
-				e2.set(0.0, 1.0, 0.0);
-				e3.set(0.0, 0.0, 1.0);
-			}
-
-		ONIKA_HOST_DEVICE_FUNC
-			OBB(const OBB &obb): center(obb.center), extent(obb.extent) {
-				e1 = obb.e1;
-				e2 = obb.e2;
-				e3 = obb.e3;
-			}
-
-		ONIKA_HOST_DEVICE_FUNC
-			OBB & operator = (const OBB &obb) {
-				center = obb.center;
-				e1 = obb.e1;
-				e2 = obb.e2;
-				e3 = obb.e3;
-				extent = obb.extent;
-				return (*this);
-			}
-
-		ONIKA_HOST_DEVICE_FUNC
-			void enlarge(double more) {
-				extent.x += more;
-				extent.y += more;
-				extent.z += more;
-			}
-
-		ONIKA_HOST_DEVICE_FUNC
-			void translate(const vec3r & v) {
-				center += v;
-			}
-
-		ONIKA_HOST_DEVICE_FUNC
-			void rotate(const quat & Q) {
-				e1 = Q * e1;
-				e2 = Q * e2;
-				e3 = Q * e3;
-				center = Q * center;
-			}
-
[…]
