+++ src/driver/include/exaDEM/dump_driver.h
@@ -9,15 +9,15 @@ namespace exaDEM
   struct driver_io_writer
   {
     std::string filename;
-    inline  operator() (Drivers& drvs)
+    inline operator()(Drivers &drvs)
     {
       std::stringstream buffer;
-			size_t size = drvs.size();
-      buffer << size  << std::endl;
-      for(int i = 0; i < size ; i++)
+      size_t size = drvs.size();
+      buffer << size << std::endl;
+      for (int i = 0; i < size; i++)
       {
-        //drvs[i].write(bufffer);
+        // drvs[i].write(bufffer);
       }
     }
   };
-}
+} // namespace exaDEM
