+++ src/interaction/include/exaDEM/type/mat9.hpp
@@ -1,391 +1,388 @@
-#ifndef MAT9_HPP
-# define MAT9_HPP;
+#pragma once
 
-  /// @file
-  /// @brief 3 by 3 matrix
-  /// @author Vincent Richefeu <Vincent.Richefeu@3sr-grenoble.fr>,
-  /// Lab 3SR, Grenoble University
+/// @file
+/// @brief 3 by 3 matrix
+/// @author Vincent Richefeu <Vincent.Richefeu@3sr-grenoble.fr>,
+/// Lab 3SR, Grenoble University
 
-# include "vec3.hpp"
+#include "vec3.hpp"
 
-  /// Matrix 3x3
-  template <typename T> class mat9
+/// Matrix 3x3
+template <typename T> class mat9
+{
+public:
+  T xx, xy, xz;
+  T yx, yy, yz;
+  T zx, zy, zz;
+
+  // static const mat9 zero;
+  // static const mat9 unit;
+
+  mat9() : xx(0), xy(0), xz(0), yx(0), yy(0), yz(0), zx(0), zy(0), zz(0) {}
+  mat9(const T XX, const T XY, const T XZ, const T YX, const T YY, const T YZ, const T ZX, const T ZY, const T ZZ) : xx(XX), xy(XY), xz(XZ), yx(YX), yy(YY), yz(YZ), zx(ZX), zy(ZY), zz(ZZ) {}
+  /*explicit*/ mat9(const T val) : xx(val), xy(val), xz(val), yx(val), yy(val), yz(val), zx(val), zy(val), zz(val) {}
+
+  mat9(const vec3<T> &col1, const vec3<T> &col2, const vec3<T> &col3) { set(col1, col2, col3); }
+
+  /*explicit*/ mat9(const T M[])
   {
-  public:
-    T xx, xy, xz;
-    T yx, yy, yz;
-    T zx, zy, zz;
+    xx = M[0];
+    xy = M[1];
+    xz = M[2];
+    yx = M[3];
+    yy = M[4];
+    yz = M[5];
+    zx = M[6];
+    zy = M[7];
+    zz = M[8];
+  }
 
-    // static const mat9 zero;
-    // static const mat9 unit;
+  mat9(const mat9 &M) : xx(M.xx), xy(M.xy), xz(M.xz), yx(M.yx), yy(M.yy), yz(M.yz), zx(M.zx), zy(M.zy), zz(M.zz) {}
 
-    mat9() : xx(0), xy(0), xz(0), yx(0), yy(0), yz(0), zx(0), zy(0), zz(0) {}
-    mat9(const T XX, const T XY, const T XZ, const T YX, const T YY, const T YZ, const T ZX, const T ZY, const T ZZ) : xx(XX), xy(XY), xz(XZ), yx(YX), yy(YY), yz(YZ), zx(ZX), zy(ZY), zz(ZZ) {}
-    /*explicit*/ mat9(const T val) : xx(val), xy(val), xz(val), yx(val), yy(val), yz(val), zx(val), zy(val), zz(val) {}
+  mat9 &operator=(const mat9 &M)
+  {
+    xx = M.xx;
+    xy = M.xy;
+    xz = M.xz;
+    yx = M.yx;
+    yy = M.yy;
+    yz = M.yz;
+    zx = M.zx;
+    zy = M.zy;
+    zz = M.zz;
+    return (*this);
+  }
 
-    mat9(const vec3<T> &col1, const vec3<T> &col2, const vec3<T> &col3) { set(col1, col2, col3); }
+  // Constants
+  static mat9 zero() { return mat9(0, 0, 0, 0, 0, 0, 0, 0, 0); }
[…]
