### 2024-10-18 9a0e252
fix compilation error into src/field/dem_rcut_max.cpp
