+++ src/interaction/stats_interactions.cpp
@@ -35,123 +35,123 @@ namespace exaDEM
 {
   using namespace exanb;
 
-  template<typename GridT
-    , class = AssertGridHasFields< GridT >
-    >
-  class StatsInteractions : public OperatorNode
+  template <typename GridT, class = AssertGridHasFields<GridT>> class StatsInteractions : public OperatorNode
   {
-		ADD_SLOT( MPI_Comm                    , mpi  , INPUT        , MPI_COMM_WORLD );
-		ADD_SLOT( GridT                       , grid , INPUT_OUTPUT , REQUIRED );
-		ADD_SLOT( GridCellParticleInteraction , ges  , INPUT_OUTPUT , DocString{"Interaction list"} );
+    ADD_SLOT(MPI_Comm, mpi, INPUT, MPI_COMM_WORLD);
+    ADD_SLOT(GridT, grid, INPUT_OUTPUT, REQUIRED);
+    ADD_SLOT(GridCellParticleInteraction, ges, INPUT_OUTPUT, DocString{"Interaction list"});
 
-		public:
+  public:
+    inline std::string documentation() const override final { return R"EOF( This operator displays DEM simulation data for a given frequency.)EOF"; }
 
-		inline std::string documentation() const override final
-		{
-			return R"EOF( This operator displays DEM simulation data for a given frequency.)EOF";
-		}
-
-		inline void execute () override final
-		{
-			auto & cells = ges->m_data;
+    inline void execute() override final
+    {
+      auto &cells = ges->m_data;
 
       // v vertex, e edge, f face, c cylinder, s surface, b ball, S stl
 
-			int nvv(0), nve(0), nvf(0), nee(0); // interaction counters [particles]
-			int an(0), anvv(0), anve(0), anvf(0), anee(0); // active interaction counters
-			int nvc(0), nvs(0), nvb(0), nSvv(0), nSve(0), nSvf(0), nSee(0), nSev(0), nSfv(0); // interaction counters [drivers]
-			int anvc(0), anvs(0), anvb(0), anSvv(0), anSve(0), anSvf(0), anSee(0), anSev(0), anSfv(0); // interaction counters [drivers]
+      int nvv(0), nve(0), nvf(0), nee(0);                                                        // interaction counters [particles]
+      int an(0), anvv(0), anve(0), anvf(0), anee(0);                                             // active interaction counters
+      int nvc(0), nvs(0), nvb(0), nSvv(0), nSve(0), nSvf(0), nSee(0), nSev(0), nSfv(0);          // interaction counters [drivers]
+      int anvc(0), anvs(0), anvb(0), anSvv(0), anSve(0), anSvf(0), anSee(0), anSev(0), anSfv(0); // interaction counters [drivers]
 
-			const exanb::Vec3d null = {0.,0.,0.};
+      const exanb::Vec3d null = {0., 0., 0.};
 
-			auto incr_interaction_counters = [null] (const Interaction& I, int& count, int& active_count, int& active_global_count) -> void
-			{
-				count ++;
-				if( I.friction != null ) 
-				{
-					active_count ++;
-					active_global_count ++;
-				}
-			};
+      auto incr_interaction_counters = [null](const Interaction &I, int &count, int &active_count, int &active_global_count) -> void
+      {
+        count++;
+        if (I.friction != null)
+        {
+          active_count++;
+          active_global_count++;
+        }
+      };
 
-#pragma omp parallel for reduction(+:nvv, nve, nvf, nee, an, anvv, anve, anvf, anee)
-			for( size_t i = 0 ; i < cells.size() ; i++ )
-			{
-				for( auto& item : cells[i].m_data)
-				{
[…]
