+++ src/driver/add_stl_mesh.cpp
@@ -16,9 +16,11 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-#include "exanb/core/operator.h"
-#include "exanb/core/operator_slot.h"
-#include "exanb/core/operator_factory.h"
+#include <exanb/core/operator.h>
+#include <exanb/core/operator_slot.h>
+#include <exanb/core/operator_factory.h>
+#include <exanb/core/quaternion.h>
+#include <exanb/core/quaternion_yaml.h>
 #include <mpi.h>
 #include <memory>
 #include <exaDEM/driver_base.h>
@@ -30,57 +32,80 @@ under the License.
 namespace exaDEM
 {
 
-	using namespace exanb;
+  using namespace exanb;
 
-		class AddSTLMesh : public OperatorNode
-	{
-		static constexpr Vec3d null= { 0.0, 0.0, 0.0 };
-		static constexpr Quaternion default_quat= { 1.0 , 0.0, 0.0, 0.0 };
+  class AddSTLMesh : public OperatorNode
+  {
+    static constexpr Vec3d null = {0.0, 0.0, 0.0};
+    static constexpr Quaternion default_quat = {1.0, 0.0, 0.0, 0.0};
 
-    ADD_SLOT( Drivers     , drivers         , INPUT_OUTPUT, REQUIRED , DocString{"List of Drivers"});
-		ADD_SLOT( int         , id              , INPUT       , REQUIRED , DocString{"Driver index"});
-		ADD_SLOT( std::string , filename        , INPUT       , REQUIRED , DocString{"Input filename"});
-		ADD_SLOT( Vec3d       , center          , INPUT       , null     , DocString{"Defined but not used"});
-		ADD_SLOT( Vec3d       , angular_velocity, INPUT       , null     , DocString{"Defined but not used"});
-		ADD_SLOT( Vec3d       , velocity        , INPUT       , null     , DocString{"Defined but not used"});
-		ADD_SLOT( Quaternion  , orientation     , INPUT       , default_quat , DocString{"Defined but not used"});
-    ADD_SLOT( double      , minskowski      , INPUT       , REQUIRED , DocString{"Minskowski radius value"} );
-    ADD_SLOT( double      , rcut_inc        , INPUT       , DocString{"value added to the search distance to update neighbor list less frequently. in physical space"} );
+    ADD_SLOT(Drivers, drivers, INPUT_OUTPUT, REQUIRED, DocString{"List of Drivers"});
+    ADD_SLOT(int, id, INPUT, REQUIRED, DocString{"Driver index"});
+    ADD_SLOT(std::string, filename, INPUT, REQUIRED, DocString{"Input filename"});
+    ADD_SLOT(Vec3d, center, INPUT, null, DocString{"Defined but not used"});
+    ADD_SLOT(Vec3d, angular_velocity, INPUT, null, DocString{"Defined but not used"});
+    ADD_SLOT(Vec3d, velocity, INPUT, null, DocString{"Defined but not used"});
+    ADD_SLOT(Quaternion, orientation, INPUT, default_quat, DocString{"Defined but not used"});
+    ADD_SLOT(double, minskowski, INPUT, REQUIRED, DocString{"Minskowski radius value"});
+    ADD_SLOT(double, rcut_inc, INPUT, DocString{"value added to the search distance to update neighbor list less frequently. in physical space"});
 
-		public:
-
-		inline std::string documentation() const override final
-		{
-			return R"EOF(
+  public:
+    inline std::string documentation() const override final
+    {
+      return R"EOF(
         This operator add a stl mesh to the drivers list.
         )EOF";
-		}
+    }
 
-		inline void execute () override final
-		{
-			stl_mesh_reader reader;
-			reader(*filename);
[…]
