### 2024-10-23 1e7a2c6
Préparation de l'extrait
