+++ src/sphere/contact_force_sphere.cpp
@@ -47,6 +47,145 @@ namespace exaDEM
 {
   using namespace exanb;
   using namespace sphere;
+ 
+  
+  __device__ void kernelContact(Interaction2& item,
+  				double* rx,
+  				double* ry,
+  				double* rz,
+  				double* rad,
+  				double* vrotx,
+  				double* vroty,
+  				double* vrotz,
+  				double* vx,
+  				double* vy,
+  				double* vz,
+  				double* mass,
+  				double* momx,
+  				double* momy,
+  				double* momz,
+  				double* fx,
+  				double* fy,
+  				double* fz,
+  				const double time,
+  				const ContactParams hkp)
+  {
+  	auto& idi = item.id_i;
+  	auto& idj = item.id_j;
+  		
+  	auto& rx_i = rx[idi];
+  	auto& ry_i = ry[idi];
+  	auto& rz_i = rz[idi];
+  		
+  	auto& rx_j = rx[idj];
+  	auto& ry_j = ry[idj];
+  	auto& rz_j = rz[idj];
+  		
+  	Vec3d ri = {rx_i, ry_i, rz_i};
+  	Vec3d rj = {rx_j, ry_j, rz_j};
+  		
+  	auto& rad_i = rad[idi];
+  	auto& rad_j = rad[idj];
+  		
+  	auto& vrotx_i = vrotx[idi];
+  	auto& vroty_i = vroty[idi];
+  	auto& vrotz_i = vrotz[idi];
+  		
+  	auto& vrotx_j = vrotx[idj];
+  	auto& vroty_j = vroty[idj];
+  	auto& vrotz_j = vrotz[idj];
+  		
+  	Vec3d vrot_i = {vrotx_i, vroty_i, vrotz_i};
+  	Vec3d vrot_j = {vrotx_j, vroty_j, vrotz_j};
+  		
+  	auto [contact, dn, n, contact_position] = detection_vertex_vertex_core(ri, rad_i, rj, rad_j);	
+  	Vec3d fn = {0, 0, 0};
+  		
+  	if (contact)
+  	{
+  		const Vec3d vi = {vx[idi], vy[idi], vz[idi]};
+  		const Vec3d vj = {vx[idj], vy[idj], vz[idj]};
+  		const auto& m_i = mass[idi];
+  		const auto& m_j = mass[idj];
+  			
+  		Vec3d f = {0, 0, 0};
+  		const double meff = compute_effective_mass(m_i, m_j);
+  			
+  		contact_force_core(dn, n, time, hkp.m_kn, hkp.m_kt, hkp.m_kr, hkp.m_mu, hkp.m_damp_rate, meff, item.friction, contact_position, ri, vi, f, item.moment, vrot_i, rj, vj, vrot_j);
+  			
+  		fn = f - item.friction;
+  			
+  		Vec3d mom_i = {momx[idi], momy[idi], momz[idi]};
+  		Vec3d mom_i_res = compute_moments(contact_position, ri, f, item.moment);
+  		atomicAdd(&momx[idi], mom_i_res.x);
+  		atomicAdd(&momy[idi], mom_i_res.y);
+  		atomicAdd(&momz[idi], mom_i_res.z);
+  		atomicAdd(&fx[idi], f.x);
+  		atomicAdd(&fy[idi], f.y);
+  		atomicAdd(&fz[idi], f.z);
+  			
+  		Vec3d mom_j = {momx[idj], momy[idj], momz[idj]};
+  		Vec3d mom_j_res = compute_moments(contact_position, rj, -f, -item.moment);
+  		atomicAdd(&momx[idj], mom_j_res.x);
[…]
