+++ src/interaction/stress_tensor.cpp
@@ -52,7 +52,7 @@ namespace exaDEM
     ADD_SLOT(MPI_Comm, mpi, INPUT, MPI_COMM_WORLD);
     ADD_SLOT(GridT, grid, INPUT_OUTPUT, REQUIRED);
     ADD_SLOT(GridCellParticleInteraction, ges, INPUT, DocString{"Interaction list"});
-    ADD_SLOT(Classifier<InteractionSOA>, ic, INPUT_OUTPUT, DocString{"Interaction lists classified according to their types"});
+    ADD_SLOT(Classifier<InteractionAOS>, ic, INPUT_OUTPUT, DocString{"Interaction lists classified according to their types"});
     ADD_SLOT(double, volume, INPUT, REQUIRED, DocString{"Volume of the domain simulation. >0 "});
     ADD_SLOT(Mat3d, stress_tensor, OUTPUT, DocString{"Write an Output file containing stress tensors."});
 
@@ -68,7 +68,7 @@ namespace exaDEM
 
       // get slot data
       auto cells = grid->cells();
-      Classifier<InteractionSOA> &cf = *ic;
+      Classifier<InteractionAOS> &cf = *ic;
       exanb::Mat3d &stress = *stress_tensor;
       stress = exanb::make_zero_matrix();
 
