+++ src/polyhedron/contact_force_polyhedron.cpp
@@ -48,12 +48,35 @@ namespace exaDEM
   using namespace exanb;
   using namespace polyhedron;
   
+  ONIKA_HOST_DEVICE_FUNC
+  inline std::tuple<bool, double, Vec3d, Vec3d> detection_precompute(const uint16_t type, const VertexArray &pi, const int i, const shape *shpi, const VertexArray &pj, const int j, const shape *shpj)
+  {
+#       define __params__     pi, i, shpi, pj, j, shpj
+  	if(type==0)
+  	{
+  		return detection_vertex_vertex_precompute(__params__);
+  	}
+  	else if(type==1)
+  	{
+  		return detection_vertex_edge_precompute(__params__);
+  	}
+  	else if(type==2)
+  	{
+  		return detection_vertex_face_precompute(__params__);
+  	}
+  	else if(type==3)
+  	{
+  		return detection_edge_edge_precompute(__params__);
+  	}
+  	
+  	return std::tuple<bool, double, Vec3d, Vec3d>();
+  }
   
-  __device__ void kernelContactUN(Interaction2& item,
+  
+  __device__ void kernelContact1(Interaction2& item,
   					double* rx,
   					double* ry,
   					double* rz,
-  					double* rad,
   					double* vrotx,
   					double* vroty,
   					double* vrotz,
@@ -69,10 +92,691 @@ namespace exaDEM
   					double* fz,
   					uint32_t* type,
   					VertexArray* vertices,
+  					const shape *const shps,
   					const double time,
   					const ContactParams hkp)
   {
+  	auto& idi = item.id_i;
+  	auto& idj = item.id_j;
+  	
+  	auto& rx_i = rx[idi];
+  	auto& ry_i = ry[idi];
+  	auto& rz_i = rz[idi];
+  	
+  	auto& rx_j = rx[idj];
+  	auto& ry_j = ry[idj];
+  	auto& rz_j = rz[idj];
+  	
+  	Vec3d ri = {rx_i, ry_i, rz_i};
+  	Vec3d rj = {rx_j, ry_j, rz_j};
+  	
+  	auto& vrotx_i = vrotx[idi];
+  	auto& vroty_i = vroty[idi];
+  	auto& vrotz_i = vrotz[idi];
+  	
+  	auto& vrotx_j = vrotx[idj];
+  	auto& vroty_j = vroty[idj];
+  	auto& vrotz_j = vrotz[idj];
+  	
+  	Vec3d vrot_i = {vrotx_i, vroty_i, vrotz_i};
+  	Vec3d vrot_j = {vrotx_j, vroty_j, vrotz_j};
+  	
+  	auto& type_i = type[idi];
+  	auto& type_j = type[idj];
+  	
+  	auto& vertices_i = vertices[idi];
+  	auto& vertices_j = vertices[idj];
+  	
+  	const shape& shp_i = shps[type_i];
+  	const shape& shp_j = shps[type_j];
+  	
+  	auto [contact, dn, n, contact_position] = detection_vertex_vertex_precompute(vertices_i, item.sub_i, &shp_i, vertices_j, item.sub_j, &shp_j);
+  	
+  	Vec3d f = {0, 0, 0};
+  	Vec3d fn = {0, 0, 0};
+  	if(contact)
+  	{
+  		printf("CONTACT\n");
+  		const Vec3d vi = {vx[idi], vy[idi], vz[idi]};
+  		const Vec3d vj = {vx[idj], vy[idj], vz[idj]};
+  		const auto& m_i = mass[idi];
+  		const auto& m_j = mass[idj];
+  		
+  		const double meff = compute_effective_mass(m_i, m_j);
+  		
[…]
