### 2024-10-26 47b22d0
fix performance issues on GPU
