+++ src/interaction/include/exaDEM/network.hpp
@@ -26,6 +26,7 @@ under the License.
 #include <vector>
 #include <tuple>
 #include <algorithm>
+#include <exaDEM/type/contact.hpp>
 
 namespace exaDEM
 {
@@ -51,7 +52,7 @@ namespace exaDEM
     ContactParams params;                 ///< Parameters for Contact operations.
     double time;                          ///< Incrementation time value
 
-    using signature = std::tuple<bool, double, exanb::Vec3d, exanb::Vec3d> (*)(const onika::oarray_t<exanb::Vec3d, EXADEM_MAX_VERTICES> &, int, const exaDEM::shape *, const onika::oarray_t<exanb::Vec3d, EXADEM_MAX_VERTICES> &, int, const exaDEM::shape *);
+    using signature = contact (*)(const onika::oarray_t<exanb::Vec3d, EXADEM_MAX_VERTICES> &, int, const exaDEM::shape *, const onika::oarray_t<exanb::Vec3d, EXADEM_MAX_VERTICES> &, int, const exaDEM::shape *);
 
     /**
      * @brief Array of function pointers for precomputed detection signatures.
