### 2024-10-26 5c598e6
improve classifier::classify, twice faster
