### 2024-10-30 #6936 @jedbrown
@vsangelidakis :wave: Hi, would you be available to review this submission for JOSS?
