+++ src/interaction/include/exaDEM/shape/shape_reader.hpp
@@ -206,8 +206,7 @@ namespace exaDEM
     std::string first;
     for (std::string line; getline(input, line);)
     {
-      input >> first;
-      if (first == "<")
+      if (line == "<")
       {
         first = ""; // reset key
         shape shp = read_shp(input, big_shape);
