### 2024-10-30 1f52a01
remove relative path of driver shape names
