### 2024-10-31 d625884
fix GPU compilation with stl_mesh::initialize
