### 2024-11-01 #6936 @vsangelidakis
> @vsangelidakis 👋 Hi, would you be available to review this submission for JOSS?

Hi Jed 👋 I would be happy to review this submission. It aligns well with my current activities, so I should be able to assess and advise. This shouldn't take me more than a month (hopefully much less). Do let me know for anything.
