### 2024-11-05 f6c2d23
Polyhedron kernels
