+++ src/forcefield/include/exaDEM/compute_contact_force.h
@@ -132,6 +132,65 @@ namespace exaDEM
 
     // === sum forces
     f_i = fn + ft;
+    
+    //printf("FORCE(%f,%f,%f)\n", f_i.x,f_i.y,f_i.z);
+
+    // === update moments
+    mom_i += kr * (vrot_j - vrot_i) * dt;
+
+    ///*
+    // test
+    Vec3d branch = contact_position - pos_i;
+    double r = (exanb::dot(branch, vrot_i)) / (exanb::dot(vrot_i, vrot_i));
+    branch -= r * vrot_i;
+
+    constexpr double mur = 0;
+    double threshold_mom = std::abs(mur * exanb::norm(branch) * exanb::norm(fn)); // even without fabs, the value should
+                                                                                  // be positive
+    double mom_square = exanb::dot(mom_i, mom_i);
+    if (mom_square > 0.0 && mom_square > threshold_mom * threshold_mom)
+      mom_i = mom_i * (threshold_mom / sqrt(mom_square));
+    //*/
+  }
+  
+  ONIKA_HOST_DEVICE_FUNC inline void contact_force_core3(const double dn,
+                                                        const Vec3d &n, // -normal
+                                                        const double dt, const double kn, const double kt, const double kr, const double mu, const double dampRate, const double meff,
+                                                        Vec3d &ft, // tangential force between particle i and j
+                                                        const Vec3d &contact_position,
+                                                        const Vec3d &pos_i,  // positions i
+                                                        const Vec3d &vel_i,  // positions i
+                                                        Vec3d f_i,          // forces i
+                                                        Vec3d &mom_i,        // moments i
+                                                        const Vec3d &vrot_i, // angular velocities i
+                                                        const Vec3d &pos_j,  // positions j
+                                                        const Vec3d &vel_j,  // positions j
+                                                        const Vec3d &vrot_j  // angular velocities j
+  )
+  {
+    const double damp = compute_damp(dampRate, kn, meff);
+
+    // === Relative velocity (j relative to i)
+    auto vel = compute_relative_velocity(contact_position, pos_i, vel_i, vrot_i, pos_j, vel_j, vrot_j);
+
+    // compute relative velocity
+    const double vn = exanb::dot(vel, n);
+
+    // === Normal force (elatic contact + viscous damping)
+    const Vec3d fn = compute_normal_force(kn, damp, dn, vn, n); // fc ==> cohesive force
+
+    // === Tangential force (friction)
+    ft += exaDEM::compute_tangential_force(kt, dt, vn, n, vel);
+    // ft	 	+= exaDEM::compute_tangential_force(kt, dt, vn, n, vel);
+
+    // fit tangential force
+    auto threshold_ft = exaDEM::compute_threshold_ft(mu, kn, dn);
+    exaDEM::fit_tangential_force(threshold_ft, ft);
+
+    // === sum forces
+    f_i = fn + ft;
+    
+    //printf("FORCE(%f,%f,%f)\n", fn.x,fn.y,fn.z);
 
     // === update moments
     mom_i += kr * (vrot_j - vrot_i) * dt;
