### 2024-11-06 cce4bb6
Position's shift in the shape reader
