+++ src/polyhedron/contact_force_polyhedron.cpp
@@ -47,483 +47,544 @@ namespace exaDEM
 {
   using namespace exanb;
   using namespace polyhedron;
-  
-  
-  
-  __device__ double compute_damp_gpu(const double a_damp_rate, const double a_kn, const double a_meff)
-  {
-    const double ret = a_damp_rate * 2.0 * sqrt(a_kn * a_meff);
-    return ret;
-  }
 
-  __device__ Vec3d compute_relative_velocity_gpu(const Vec3d &contact_position, const Vec3d &pos_i, const Vec3d &vel_i, const Vec3d &vrot_i, const Vec3d &pos_j, const Vec3d &vel_j, const Vec3d &vrot_j)
-  {
-    const auto contribution_i = vel_i - exanb::cross(contact_position - pos_i, vrot_i);
-    const auto contribution_j = vel_j - exanb::cross(contact_position - pos_j, vrot_j);
-    const auto ret = contribution_j - contribution_i;
-    return ret;
-  }
-  
- __device__ Vec3d compute_normal_force_gpu(const double a_kn, const double a_damp, const double a_dn, const double a_vn, const Vec3d &a_n)
-  {
-    const double fne = -a_kn * a_dn;  // elastic contact
-    const double fnv = a_damp * a_vn; // viscous damping
-    const double fn = fnv + fne;
-    const auto ret = fn * a_n;
-    return ret;
-  }
-  
-  __device__ Vec3d compute_tangential_force_gpu(const double a_kt, const double a_dt, const double a_vn, const Vec3d &a_n, const Vec3d &a_vel)
-  {
-    const Vec3d vt = a_vel - a_vn * a_n;
-    const Vec3d ft = a_kt * (a_dt * vt);
-    return ft;
-  }
-  
-  __device__ double compute_threshold_ft_gpu(const double a_mu, const double a_kn, const double a_dn)
-  {
-    // === recompute fne
-    const double fne = a_kn * a_dn; //(remove -)
-    const double threshold_ft = std::fabs(a_mu * fne);
-    return threshold_ft;
-  }
-
-  __device__ void fit_tangential_force_gpu(const double threshold_ft, Vec3d &a_ft)
-  {
-    double ft_square = exanb::dot(a_ft, a_ft);
-    if (ft_square > 0.0 && ft_square > threshold_ft * threshold_ft)
-      a_ft *= (threshold_ft / sqrt(ft_square));
-  }
-
-  
-  __device__ void contact_force_core_gpu(const double dn,
-                                                        const Vec3d &n, // -normal
-                                                        const double dt, const double kn, const double kt, const double kr, const double mu, const double dampRate, const double meff,
-                                                        Vec3d &ft, // tangential force between particle i and j
-                                                        const Vec3d &contact_position,
-                                                        const Vec3d &pos_i,  // positions i
-                                                        const Vec3d &vel_i,  // positions i
-                                                        Vec3d f_i,          // forces i
-                                                        Vec3d &mom_i,        // moments i
-                                                        const Vec3d &vrot_i, // angular velocities i
-                                                        const Vec3d &pos_j,  // positions j
[…]
