### 2024-11-06 0b90d3d
remove switch used into the contact law
