+++ src/polyhedron/include/exaDEM/contact_polyhedron.h
@@ -55,65 +55,13 @@ namespace exaDEM
     }
 
     /**
-     * @struct polyhedron_detector
-     * @brief Structure for detecting interactions between polyhedrons in exaDEM.
-     *
-     * This structure provides methods to detect interactions between polyhedrons
-     * based on the specified type using precomputed data.
-     */
-    struct polyhedron_detector
-    {
-      /**
-       * @brief Default constructor for polyhedron_detector struct.
-       */
-      polyhedron_detector() {}
-
-      /**
-       * @brief Detects interactions between polyhedrons based on the specified type.
-       *
-       * This function detects interactions between polyhedrons based on the specified
-       * type using precomputed data. It returns a tuple indicating whether an interaction
-       * was detected, the distance between the polyhedrons, norm and contact point.
-       *
-       * @param type Type of interaction to detect:
-       *             - 0; Vertex-Vertex interaction
-       *             - 1: Vertex-Edge interaction
-       *             - 2: Vertex-Face interaction
-       *             - 3: Edge-Edge interaction
-       * @param pi Vertex array of the first polyhedron.
-       * @param i Index of the vertex or edge in the first polyhedron.
-       * @param shpi Shape data of the first polyhedron.
-       * @param pj Vertex array of the second polyhedron.
-       * @param j Index of the vertex or edge in the second polyhedron.
-       * @param shpj Shape data of the second polyhedron.
-       * @return contact()
-       */
-      ONIKA_HOST_DEVICE_FUNC inline contact operator()(const uint16_t type, const VertexArray &pi, const int i, const shape *shpi, const VertexArray &pj, const int j, const shape *shpj) const
-      {
-#       define __params__     pi, i, shpi, pj, j, shpj
-        assert(type >= uint16_t(0) && type <= uint16_t(4));
-        switch (type)
-        {
-        case 0:
-          return exaDEM::detection_vertex_vertex_precompute(__params__);
-        case 1:
-          return exaDEM::detection_vertex_edge_precompute(__params__);
-        case 2:
-          return exaDEM::detection_vertex_face_precompute(__params__);
-        case 3:
-          return exaDEM::detection_edge_edge_precompute(__params__);
-        }
-#undef __params__
-        return contact();
-      }
-    };
-
-    /**
      * @struct contact_law
      * @brief Structure defining contact law interactions for particles (polyhedra).
      */
+    template<int interaction_type>
     struct contact_law
     {
+      detect<interaction_type> detection;
       /**
        * @brief Default constructor for contact_law struct.
        */
@@ -191,7 +139,7 @@ namespace exaDEM
         const shape &shp_i = shps[type_i];
         const shape &shp_j = shps[type_j];
 
-        auto [contact, dn, n, contact_position] = detect(item.type, vertices_i, item.sub_i, &shp_i, vertices_j, item.sub_j, &shp_j);
+        auto [contact, dn, n, contact_position] = detection(vertices_i, item.sub_i, &shp_i, vertices_j, item.sub_j, &shp_j);
         // temporary vec3d to store forces.
         Vec3d f = {0, 0, 0};
         Vec3d fn = {0, 0, 0};
@@ -233,7 +181,6 @@ namespace exaDEM
 
         return {dn, contact_position, fn, item.friction};
       }
-      const polyhedron_detector detect;
     };
 
     /**
@@ -310,76 +257,13 @@ namespace exaDEM
     };
 
     /**
-     * @brief Functor for detecting interactions between two objects in an STL mesh.
-     *
-     * This functor detects interactions between two objects represented by vertices,
-     * edges, or faces in an STL mesh. It returns a tuple indicating whether an interaction
[…]

[…]
