### 2024-11-06 c0e57c5
remove switch used into the contact law (#72)
Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
