### 2024-11-06 af686db
update spack receipt
