+++ README.md
@@ -10,7 +10,7 @@ Additionally, ExaDEM offers compatibility with MPI+GPUs, using the CUDA programm
 
 Documentation is available here: 
 
-- Website: [ExaDEM Website](https://collab4exanbody.github.io/exaDEM/)
+- Website: [ExaDEM Website](https://collab4exanbody.github.io/doc_exaDEM/)
 - Github: [ExaDEM Documentation](https://github.com/Collab4exaNBody/doc_exaDEM.git)
 
 ## Authors and acknowledgment
