+++ .github/workflows/sphinx-docs.yml
@@ -1,33 +1,0 @@
-name: Deploy Sphinx documentation to Pages
-
-on:
-  push:
-    branches: [main] # branch to trigger deployment
-
-jobs:
-  pages:
-    runs-on: ubuntu-20.04
-    environment:
-      name: github-pages
-      url: ${{ steps.deployment.outputs.page_url }}
-    permissions:
-      pages: write
-      id-token: write
-    steps:
-    
-    - name: test
-      uses: actions/checkout@v1
-
-    - name: Pull & update submodules 
-      run: |
-        git clone https://github.com/Collab4exaNBody/doc_exaDEM.git
-        cd doc_exaDEM
-        git submodule init
-        git submodule update
-        
-    - id: deployment
-      uses: sphinx-notes/pages@v3
-      with:
-        checkout: false
-        documentation_path: doc_exaDEM/docs/source
-        requirements_path: doc_exaDEM/docs/source/requirements.txt
