### 2024-11-06 b7a4ff4
Create codeql.yml
