### 2024-11-06 46efdfa
Update codeql.yml
