### 2024-11-06 d2d9d3d
Update codeql.yml
