### 2024-11-07 179c9ca
Update codacy-analysis.yaml
