+++ .github/workflows/codacy-analysis.yaml
@@ -10,6 +10,8 @@ jobs:
   codacy-security-scan:
     name: Codacy Security Scan
     runs-on: ubuntu-latest
+    permissions:
+      security-events: write    
     steps:
       - name: Checkout code
         uses: actions/checkout@main
