### 2024-11-07 f2fdaba
Update patch_library_path.sh (#74)
