+++ scripts/patch_library_path.sh
@@ -1,6 +1,6 @@
 if [ $LD_LIBRARY_PATH ]
 then
-  HAS_PATH=`echo "${LD_LIBRARY_PATH}" | tr ':' '\n' | grep -c "${PLUGIN_PATH}"`
+  HAS_PATH=`echo "$(LD_LIBRARY_PATH)" | tr ':' '\n' | grep -c "${PLUGIN_PATH}"`
   # echo "HAS_PATH=${HAS_PATH}"
   if [ "${HAS_PATH}" == "0" ]
   then
