### 2024-11-07 8923683
Update patch_library_path.sh
