### 2024-11-07 ccf5f9a
Update patch_library_path.sh
