### 2024-11-07 a6f0468
Update patch_library_path.sh
