+++ scripts/patch_library_path.sh
@@ -1,12 +1,12 @@
 #!/bin/sh
-if [ $(LD_LIBRARY_PATH) ]
+if [ ${LD_LIBRARY_PATH} ]
 then
-  HAS_PATH=`$(LD_LIBRARY_PATH) | tr ':' '\n' | grep -c $(PLUGIN_PATH)`
+  HAS_PATH=$(${LD_LIBRARY_PATH} | tr ':' '\n' | grep -c ${PLUGIN_PATH})
   # echo "HAS_PATH=${HAS_PATH}"
   if [ "${HAS_PATH}" == "0" ]
   then
     echo "Append $(PLUGIN_PATH) to LD_LIBRARY_PATH"
-    export LD_LIBRARY_PATH=$(PLUGIN_PATH):$(LD_LIBRARY_PATH)
+    export LD_LIBRARY_PATH=${PLUGIN_PATH}:${LD_LIBRARY_PATH}
   fi
 else
   echo "Set LD_LIBRARY_PATH to ${PLUGIN_PATH}"
