### 2024-11-07 6e66a50
fix bash script
