+++ scripts/script-clang-format.sh
@@ -1,8 +1,9 @@
+#!/bin/sh
 list=$(find . \( -name "*.h" -o -name "*.cpp" -o -name "*.hpp" -o -name "*.cxx" \))
 
-for files in $list; do
+for files in i"$list"; do
     echo "$files"
-    ../exaNBody/scripts/code-format.sh $files
+    ../exaNBody/scripts/code-format.sh "$files"
 done
 
 
