### 2024-11-07 bc73852
fix bash script
