+++ scripts/script-clang-format.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 list=$(find . \( -name "*.h" -o -name "*.cpp" -o -name "*.hpp" -o -name "*.cxx" \))
 
-for files in i"$list"; do
+for files in "$list"; do
     echo "$files"
     ../exaNBody/scripts/code-format.sh "$files"
 done
