### 2024-11-07 6e0088f
fix script bash
