+++ scripts/script-clang-format.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
-list=$(find . \( -name "*.h" -o -name "*.cpp" -o -name "*.hpp" -o -name "*.cxx" \))
+cmd_list=$(find . \( -name "*.h" -o -name "*.cpp" -o -name "*.hpp" -o -name "*.cxx" \))
 
-for files in "$list"; do
+for files in ${cmd_list}; do
     echo "$files"
     ../exaNBody/scripts/code-format.sh "$files"
 done
