+++ .github/workflows/cpplint.yml
@@ -7,4 +7,4 @@ jobs:
     - uses: actions/checkout@v1
     - uses: actions/setup-python@v1
     - run: pip install cpplint
-    - run: cpplint --recursive src/
+    - run: cpplint --recursive --extensions=cpp,cxx,hpp,h,hxx --linelength=1000  src/
