+++ src/driver/print_drivers.cpp
@@ -33,14 +33,14 @@ namespace exaDEM
     ADD_SLOT(Drivers, drivers, INPUT, REQUIRED, DocString{"List of Drivers"});
 
   public:
-    inline std::string documentation() const override final
+    inline std::string documentation() const final
     {
       return R"EOF(
         This operator prints drivers.
         )EOF";
     }
 
-    inline void execute() override final
+    inline void execute() final
     {
       auto &drvs = *drivers;
       lout << std::endl;
