### 2024-11-07 da8d8bd
remove error with using namespace from cpplint outputs
