+++ .github/workflows/cpplint.yml
@@ -1,10 +1,0 @@
-name: cpplint
-on: [push, pull_request]
-jobs:
-  cpplint:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v1
-    - uses: actions/setup-python@v1
-    - run: pip install cpplint
-    - run: cpplint --recursive --extensions=cpp,cxx,hpp,h,hxx --linelength=1000  src/
