+++ src/io/write_paraview_obb_particles.cpp
@@ -40,44 +40,52 @@ under the License.
 #include <mpi.h>
 #include <exaDEM/shape/shapes.hpp>
 #include <exaDEM/shape/shape_printer.hpp>
+#include <exanb/core/string_utils.h>
 
 namespace exaDEM
 {
   using namespace exanb;
-  template <class GridT, class = AssertGridHasFields<GridT>> class WriteParaviewPolyhedraOperator : public OperatorNode
+
+
+  template <class GridT, class = AssertGridHasFields<GridT>> class WriteParaviewOBBParticlesOperator : public OperatorNode
   {
     using ComputeFields = FieldSet<field::_rx, field::_ry, field::_rz, field::_type, field::_orient>;
     static constexpr ComputeFields compute_field_set{};
     ADD_SLOT(MPI_Comm, mpi, INPUT, MPI_COMM_WORLD);
     ADD_SLOT(GridT, grid, INPUT_OUTPUT);
     ADD_SLOT(Domain, domain, INPUT, REQUIRED);
-    ADD_SLOT(std::string, basename, INPUT, REQUIRED, DocString{"Output filename"});
-    ADD_SLOT(std::string, basedir, INPUT, "polyhedra_paraview", DocString{"Output directory, default is polyhedra_paraview"});
+    ADD_SLOT(std::string, basename, INPUT, "obb", DocString{"Output filename"});
+    ADD_SLOT(std::string, dir_name, INPUT_OUTPUT, REQUIRED, DocString{"Main output directory."});
     ADD_SLOT(long, timestep, INPUT, DocString{"Iteration number"});
     ADD_SLOT(shapes, shapes_collection, INPUT_OUTPUT, DocString{"Collection of shapes"});
 
   public:
-    inline std::string documentation() const override final
+    inline std::string documentation() const final
     {
-      return R"EOF( This operator initialize shapes data structure from a shape input file.
+      return R"EOF( This operator dumps obb into a paraview output file.
     	    			)EOF";
     }
 
-    inline void execute() override final
+    inline void execute() final
     {
       // mpi stuff
       int rank, size;
       MPI_Comm_rank(*mpi, &rank);
       MPI_Comm_size(*mpi, &size);
 
-      std::string directory = (*basedir) + "/" + (*basename) + "_" + std::to_string(*timestep);
-      std::string filename = directory + "/" + (*basename) + "_" + std::to_string(*timestep) + "_" + std::to_string(rank);
+      std::string ts = "%010d"; 
+      std::string rk = "%06d"; 
+
+      std::string directory = (*dir_name) + "/ParaviewOutputFiles/" + (*basename) + "_" + ts;
+      directory = format_string(directory, *timestep);
+      std::string filename = directory + "/" + rk + ".vtp";
+      filename  = format_string(filename,  rank);
 
       // prepro
       if (rank == 0)
       {
         namespace fs = std::filesystem;
-        fs::create_directory(*basedir);
+        fs::create_directory(*dir_name);
         fs::create_directory(directory);
       }
 
@@ -87,12 +95,8 @@ namespace exaDEM
       const auto cells = grid->cells();
       const size_t n_cells = grid->number_of_cells();
 
-      size_t count_vertex(0), count_face(0), polygon_offset_in_stream(0);
-      std::stringstream buff_vertices; // store vertices
-      std::stringstream buff_faces;    // store faces
-      std::stringstream buff_offsets;  // store face offsets
+      par_obb_helper buffers;
 
-#     define __PARAMS__ count_vertex, count_face, buff_vertices, buff_faces, buff_offsets
       // fill string buffers
[…]
