### 2024-11-09 3584eb7
add paraview operator to display OBBs (#76)
Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
