+++ src/interaction/include/exaDEM/shape/shape_printer.hpp
@@ -10,7 +10,7 @@ with the License.  You may obtain a copy of the License at
   http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
+software DISTributed under the License is distributed on an
 "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
@@ -19,11 +19,23 @@ under the License.
 #pragma once
 
 #include <exaDEM/shape/shape.hpp>
+#include <exanb/core/string_utils.h>
 #include <iostream>
 
 namespace exaDEM
 {
-  void build_buffer(const exanb::Vec3d &pos, const shape *shp, const exanb::Quaternion &orient, size_t &polygon_offset_in_stream, size_t &n_vertices, size_t &n_polygon, std::stringstream &buff_position, std::stringstream &buff_faces, std::stringstream &buff_offset)
+  struct par_obb_helper
+  {
+    int count = 0;  // use to count the number of corners
+    int off = 0;    // use for lines
+    std::stringstream corners;
+    std::stringstream lines;
+    std::stringstream offsets;
+    std::stringstream ids;
+    std::stringstream types;
+  };
+
+  inline void build_buffer_polyhedron(const exanb::Vec3d &pos, const shape *shp, const exanb::Quaternion &orient, size_t &polygon_offset_in_stream, size_t &n_vertices, size_t &n_polygon, std::stringstream &buff_position, std::stringstream &buff_faces, std::stringstream &buff_offset)
   {
     auto writer_v = [](const exanb::Vec3d &v, std::stringstream &out, const exanb::Vec3d &p, const exanb::Quaternion &q)
     {
@@ -48,7 +60,88 @@ namespace exaDEM
     n_vertices += shp->get_number_of_vertices();
   }
 
-  void write_vtp(std::string name, size_t n_vertices, size_t n_polygons, std::stringstream &buff_vertices, std::stringstream &buff_faces, std::stringstream &buff_offsets)
+
+  inline void build_buffer_obb(const exanb::Vec3d &pos, const uint64_t id, const uint16_t type, const shape *shp, const exanb::Quaternion &orient, par_obb_helper& buffers)
+  {
+    // add [1] corner [2] lines [3] id [4] type
+
+    double homothety = 1;
+    vec3r p = conv_to_vec3r(pos);
+    OBB obbi = shp->obb;
+    quat Q = conv_to_quat(orient);
+    obbi.rotate(Q);
+    obbi.extent *= homothety;
+    obbi.center *= homothety;
+    obbi.center += p;
+    vec3r corner;
+
+
+    auto add_corner = [&buffers] (vec3r& c) { buffers.corners << " " << c[0] << " " << c[1] << " " << c[2]; };
+    auto add_offset = [&buffers] (int incr) { buffers.offsets << " " << buffers.off; buffers.off += incr; };
+
+    // [1] Add Corners
+
+    vec3r e0 = obbi.e1; // 
+    vec3r e1 = obbi.e2; // 
+    vec3r e2 = obbi.e3; // 
+
+    corner = obbi.center - obbi.extent[0] * e0 - obbi.extent[1] * e1 - obbi.extent[2] * e2;
+    add_corner(corner);
+    corner += 2.0 * obbi.extent[0] * e0;
+    add_corner(corner);
+    corner += 2.0 * obbi.extent[1] * e1;
+    add_corner(corner);
+    corner -= 2.0 * obbi.extent[0] * e0;
+    add_corner(corner);
+
[…]
