### 2024-11-10 a4743dc
add components when using write_paraview_polyhedra
