### 2024-11-10 5820a9f
add ctest tests with write paraview and avg stress
