### 2024-11-11 ee8ef96
create set_fields.cpp, not working
