### 2024-11-18 a881805
obbs of stl meshes are not rebuilt if the didn't move
