+++ CMakeLists.txt
@@ -78,7 +78,7 @@ endforeach()
 # ==============================
 # === Application version    ===
 # ==============================
-set(EXASDEM_VERSION "exaDEM-1.1.0") 
+set(EXASDEM_VERSION "exaDEM-1.0.2") 
 message(STATUS "Project version : ${EXASDEM_VERSION}")
 
 # ==================================
