### 2024-11-23 272b034
add filter functions instead of using obb. The file containing detection functions is splitted into two: core and api.
