+++ src/interaction/include/exaDEM/shape/shape_detection_core.hpp
@@ -1,0 +1,641 @@
+/*
+Licensed to the Apache Software Foundation (ASF) under one
+or more contributor license agreements.  See the NOTICE file
+distributed with this work for additional information
+regarding copyright ownership.  The ASF licenses this file
+to you under the Apache License, Version 2.0 (the
+"License"); you may not use this file except in compliance
+with the License.  You may obtain a copy of the License at
+
+  http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing,
+software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+KIND, either express or implied.  See the License for the
+specific language governing permissions and limitations
+under the License.
+*/
+#pragma once
+
+#include <exanb/core/basic_types.h>
+#include <exanb/core/basic_types_operators.h>
+#include <exaDEM/shape/shape.hpp>
+#include <math.h>
+#include <exaDEM/shape/shape_prepro.hpp>
+#include <exaDEM/type/contact.hpp>
+
+namespace exaDEM
+{
+  using namespace exanb;
+  using VertexArray = ::onika::oarray_t<::exanb::Vec3d, EXADEM_MAX_VERTICES>;
+
+  /**
+   * @brief Normalizes a 3D vector in-place.
+   *
+   * @param in The 3D vector to be normalized.
+   *
+   * @note If the input vector has a length of zero, the behavior is undefined.
+   * @note The input vector is modified in-place, and the normalized vector is also returned.
+   * @note It is recommended to ensure that the input vector is non-zero before calling this function.
+   */
+  ONIKA_HOST_DEVICE_FUNC inline void normalize(Vec3d &in)
+  {
+    const double norm = exanb::norm(in);
+    in = in / norm;
+  }
+
+  // This function returns : if there is a contact, interpenetration value, normal vector, and the contact position
+  ONIKA_HOST_DEVICE_FUNC inline contact detection_vertex_vertex_core(const Vec3d &pi, double ri, const Vec3d &pj, double rj)
+  {
+    // sphero-polyhedron
+    double R = ri + rj;
+
+    // === compute distance
+    const Vec3d dist = pi - pj;
+
+    // === compute norm
+    const double dist_norm = exanb::norm(dist);
+
+    // === inv norm
+    const double inv_dist_norm = 1.0 / dist_norm;
+
+    // === compute overlap in dn
+    const double dn = dist_norm - R;
+
+    if (dn > 0)
+    {
+      return contact();
+    }
+    else
+    {
+      // === normal vector
+      const Vec3d n = dist * inv_dist_norm;
+
+      // === compute contact position
+      const Vec3d contact_position = pi - n * (ri + 0.5 * dn);
+
+      return {true, dn, n, contact_position};
+    }
+  }
+
+  /**
+   * @brief Filters vertex-vertex interactions based on a specified Verlet radius.
+   * @param rVerlet The Verlet radius used for filtering interactions.
+   * @param vi The position of the first vertex.
+   * @param ri The radius of the first vertex.
+   * @param vj The position of the second vertex.
+   * @param rj The radius of the second vertex.
[…]
