+++ src/interaction/include/exaDEM/shape/shape_detection_api.hpp
@@ -1,0 +1,405 @@
+/*
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing,
+software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+KIND, either express or implied.  See the License for the
+specific language governing permissions and limitations
+under the License.
+ */
+#pragma once
+
+#include <exanb/core/basic_types.h>
+#include <exanb/core/basic_types_operators.h>
+#include <exaDEM/shape/shape.hpp>
+#include <math.h>
+#include <exaDEM/shape/shape_prepro.hpp>
+#include <exaDEM/type/contact.hpp>
+
+namespace exaDEM
+{
+  using namespace exanb;
+  using VertexArray = ::onika::oarray_t<::exanb::Vec3d, EXADEM_MAX_VERTICES>;
+
+  ONIKA_HOST_DEVICE_FUNC inline contact detection_vertex_vertex(
+      const Vec3d &pi, 
+      const int i, 
+      const shape *shpi, 
+      const exanb::Quaternion &oi, 
+      const Vec3d &pj, const int j, 
+      const shape *shpj, 
+      const exanb::Quaternion &oj)
+  {
+
+    // === compute vertex position
+    Vec3d vi = shpi->get_vertex(i, pi, oi);
+    Vec3d vj = shpj->get_vertex(j, pj, oj);
+    return detection_vertex_vertex_core(vi, shpi->m_radius, vj, shpj->m_radius);
+  }
+
+  ONIKA_HOST_DEVICE_FUNC inline contact detection_vertex_vertex(
+      const Vec3d &pi, 
+      const double radius, 
+      const Vec3d &pj, 
+      const int j, 
+      const shape *shpj, 
+      const exanb::Quaternion &oj)
+  {
+
+    // === compute vertex position
+    Vec3d vj = shpj->get_vertex(j, pj, oj);
+    return detection_vertex_vertex_core(pi, radius, vj, shpj->m_radius);
+  }
+
+  ONIKA_HOST_DEVICE_FUNC inline contact detection_vertex_vertex_precompute(
+      const VertexArray &vai, 
+      const int i, 
+      const shape *shpi, 
+      const VertexArray &vaj, 
+      const int j, 
+      const shape *shpj)
+  {
+    // === get vertex position
+    const Vec3d &vi = vai[i];
+    const Vec3d &vj = vaj[j];
+    return detection_vertex_vertex_core(vi, shpi->m_radius, vj, shpj->m_radius);
+  }
+
+  /**
+   * @brief Filters vertex-vertex interactions based on a specified Verlet radius.
+   * @param rVerlet The Verlet radius used for filtering interactions.
+   * @param pi The position vector of the first vertex.
+   * @param i The index of the first vertex.
+   * @param shpi The shape associated with the first vertex.
+   * @param oi The orientation of the first vertex.
+   * @param pj The position vector of the second vertex.
+   * @param j The index of the second vertex.
+   * @param shpj The shape associated with the second vertex.
[…]

[…]
