### 2024-11-23 bd4af04
remove obb flag into nbh_polyhedron
