### 2024-11-23 e993548
remove useless sloth in contact_force_polyhedron.cpp
