### 2024-11-23 b8fd7c1
now, the operator that stored the contact network uses data stored within the classifier
