### 2024-11-23 fb390f2
add filter functions instead of using obb. The file containing detect… (#82)
* add filter functions instead of using obb. The file containing detection functions is splitted into two: core and api.

* remove obb flag into nbh_polyhedron

* remove useless operators in /rotating-drum/rotating-drum.msp

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
