+++ src/interaction/include/exaDEM/shape/shape_detection.hpp
@@ -18,675 +18,6 @@ under the License.
 */
 #pragma once
 
-#include <exanb/core/basic_types.h>
-#include <exanb/core/basic_types_operators.h>
-#include <exaDEM/shape/shape.hpp>
-#include <math.h>
-#include <exaDEM/shape/shape_prepro.hpp>
-#include <exaDEM/type/contact.hpp>
-
-namespace exaDEM
-{
-  using namespace exanb;
-  using VertexArray = ::onika::oarray_t<::exanb::Vec3d, EXADEM_MAX_VERTICES>;
-
-  /**
-   * @brief Normalizes a 3D vector in-place.
-   *
-   * @param in The 3D vector to be normalized.
-   *
-   * @note If the input vector has a length of zero, the behavior is undefined.
-   * @note The input vector is modified in-place, and the normalized vector is also returned.
-   * @note It is recommended to ensure that the input vector is non-zero before calling this function.
-   */
-  ONIKA_HOST_DEVICE_FUNC inline void normalize(Vec3d &in)
-  {
-    const double norm = exanb::norm(in);
-    in = in / norm;
-  }
-
-  // This function returns : if there is a contact, interpenetration value, normal vector, and the contact position
-  ONIKA_HOST_DEVICE_FUNC inline contact detection_vertex_vertex_core(const Vec3d &pi, double ri, const Vec3d &pj, double rj)
-  {
-    // sphero-polyhedron
-    double R = ri + rj;
-
-    // === compute distance
-    const Vec3d dist = pi - pj;
-
-    // === compute norm
-    const double dist_norm = exanb::norm(dist);
-
-    // === inv norm
-    const double inv_dist_norm = 1.0 / dist_norm;
-
-    // === compute overlap in dn
-    const double dn = dist_norm - R;
-
-    if (dn > 0)
-    {
-      return contact();
-    }
-    else
-    {
-      // === normal vector
-      const Vec3d n = dist * inv_dist_norm;
-
-      // === compute contact position
-      const Vec3d contact_position = pi - n * (ri + 0.5 * dn);
-
-      return {true, dn, n, contact_position};
-    }
-  }
-
-  ONIKA_HOST_DEVICE_FUNC inline contact detection_vertex_vertex(const Vec3d &pi, const int i, const shape *shpi, const exanb::Quaternion &oi, const Vec3d &pj, const int j, const shape *shpj, const exanb::Quaternion &oj)
-  {
-
-    // === compute vertex position
-    Vec3d vi = shpi->get_vertex(i, pi, oi);
-    Vec3d vj = shpj->get_vertex(j, pj, oj);
-    return detection_vertex_vertex_core(vi, shpi->m_radius, vj, shpj->m_radius);
-  }
-
-  ONIKA_HOST_DEVICE_FUNC inline contact detection_vertex_vertex(const Vec3d &pi, const double radius, const Vec3d &pj, const int j, const shape *shpj, const exanb::Quaternion &oj)
-  {
-
-    // === compute vertex position
-    Vec3d vj = shpj->get_vertex(j, pj, oj);
-    return detection_vertex_vertex_core(pi, radius, vj, shpj->m_radius);
-  }
-
-  ONIKA_HOST_DEVICE_FUNC inline contact detection_vertex_vertex_precompute(const VertexArray &vai, const int i, const shape *shpi, const VertexArray &vaj, const int j, const shape *shpj)
-  {
-    // === get vertex position
-    const Vec3d &vi = vai[i];
-    const Vec3d &vj = vaj[j];
-    return detection_vertex_vertex_core(vi, shpi->m_radius, vj, shpj->m_radius);
-  }
-
-  /**
-   * @brief Filters vertex-vertex interactions based on a specified Verlet radius.
[…]
