+++ src/polyhedron/nbh_polyhedron.cpp
@@ -1,13 +1,13 @@
 /*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
 
-  http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing,
 software distributed under the License is distributed on an
@@ -15,7 +15,7 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
-*/
+ */
 #include <memory>
 #include <exanb/core/operator.h>
 #include <exanb/core/operator_slot.h>
@@ -56,7 +56,7 @@ namespace exaDEM
     ADD_SLOT(Drivers, drivers, INPUT, DocString{"List of Drivers"});
     ADD_SLOT(CellListWrapper, cell_list, INPUT, DocString{"list of non empty cells within the current grid"});
 
-  public:
+    public:
     inline std::string documentation() const override final
     {
       return R"EOF(
@@ -234,9 +234,6 @@ namespace exaDEM
         return;
       }
 
-      // use OBB for vertex/edge and vertex/faces
-      constexpr bool skip_obb = false;
-
       auto [cell_ptr, cell_size] = cell_list->info();
 
 #     pragma omp parallel
@@ -344,171 +341,165 @@ namespace exaDEM
 
           // Second, we add interactions between two polyhedra.
           apply_cell_particle_neighbors(*grid, *chunk_neighbors, cell_a, loc_a, std::false_type() /* not symetric */,
-                                        [&g, cells, &info_particles, cell_a, &item, &shps, rVerlet, id_a, rx_a, ry_a, rz_a, t_a, orient_a, vertices_a, &add_contact](int p_a, size_t cell_b, unsigned int p_b, size_t p_nbh_index)
-                                        {
-                                          // default value of the interaction studied (A or i -> B or j)
-                                          const uint64_t id_nbh = cells[cell_b][field::id][p_b];
-                                          if (id_a[p_a] >= id_nbh)
-                                          {
-                                            if (!g.is_ghost_cell(cell_b))
-                                              return;
-                                          }
+              [&g, cells, &info_particles, cell_a, &item, &shps, rVerlet, id_a, rx_a, ry_a, rz_a, t_a, orient_a, vertices_a, &add_contact](int p_a, size_t cell_b, unsigned int p_b, size_t p_nbh_index)
+              {
+              // default value of the interaction studied (A or i -> B or j)
+              const uint64_t id_nbh = cells[cell_b][field::id][p_b];
+              if (id_a[p_a] >= id_nbh)
+              {
+              if (!g.is_ghost_cell(cell_b))
+              return;
+              }
 
-                                          // Get particle pointers for the particle b.
-                                          const uint32_t type_nbh = cells[cell_b][field::type][p_b];
-                                          const Quaternion orient_nbh = cells[cell_b][field::orient][p_b];
-                                          const double rx_nbh = cells[cell_b][field::rx][p_b];
[…]
