### 2024-11-23 d3130c7
79 update dump contact network (#84)
* remove useless sloth in contact_force_polyhedron.cpp

* now, the operator that stored the contact network uses data stored within the classifier

* fix generator.msp [tutorial]

* update tutorial

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
