+++ src/interaction/include/exaDEM/network.hpp
@@ -46,187 +46,49 @@ namespace exaDEM
     using StorageType = ForceType;                ///< Type for storing forces.
     using KeyType = CoupleType;                   ///< Type used as keys in data storage.
 
-    // Members
-    typename GridT::CellParticles *cells; ///< Pointer to the cells of the grid.
-    shapes &shps;                         ///< Reference to the list of shapes.
-    ContactParams params;                 ///< Parameters for Contact operations.
-    double time;                          ///< Incrementation time value
-
-    using signature = contact (*)(const onika::oarray_t<exanb::Vec3d, EXADEM_MAX_VERTICES> &, int, const exaDEM::shape *, const onika::oarray_t<exanb::Vec3d, EXADEM_MAX_VERTICES> &, int, const exaDEM::shape *);
-
-    /**
-     * @brief Array of function pointers for precomputed detection signatures.
-     *
-     * This static constexpr array contains function pointers for precomputed detection.
-     * Each element of the array corresponds to a specific type of interaction detection:
-     * - detection[0]: Vertex-Vertex interaction detection.
-     * - detection[1]: Vertex-Edge interaction detection.
-     * - detection[2]: Vertex-Face interaction detection.
-     * - detection[3]: Edge-Edge interaction detection.
-     */
-    static constexpr signature detection[4] = {exaDEM::detection_vertex_vertex_precompute, exaDEM::detection_vertex_edge_precompute, exaDEM::detection_vertex_face_precompute, exaDEM::detection_edge_edge_precompute};
+    GridT& grid; ///< Pointer to the cells of the grid.
+    std::stringstream pos; // store particle positions
+    std::stringstream connect; // store interaction connections
+    std::stringstream val; // store force values
 
     // TODO optimize if later with another data storage
     std::map<KeyType, StorageType> network; ///< Stores network data, with keys defined by KeyType and values by StorageType.
 
-    /**
-     * @brief Constructor for NetworkFunctor.
-     *
-     * Initializes a NetworkFunctor object with the provided grid, shapes, configuration parameters, and time.
-     *
-     * @param g Reference to the grid.
-     * @param s Reference to the shapes.
-     * @param config Reference to the configuration parameters.
-     * @param t Time parameter.
-     */
-    NetworkFunctor(GridT &g, shapes &s, ContactParams &config, double t) : cells(g.cells()), shps(s), params(config), time(t) {}
+    NetworkFunctor(GridT &g) : grid(g) {}
 
-    /**
-     * @brief Gets the position vector of a particle in a specified cell.
-     *
-     * Retrieves the position vector of the particle with the position in the specified cell.
-     *
-     * @param cell_id The ID of the cell containing the particle.
-     * @param p_id The position of the particle in the cell_id.
-     * @return The position vector of the particle.
-     */
-    const Vec3d get_position(const int cell_id, const int p_id)
-    {
-      const Vec3d res = {cells[cell_id][field::rx][p_id], cells[cell_id][field::ry][p_id], cells[cell_id][field::rz][p_id]};
-      return res;
-    };
-
-    /**
-     * @brief Gets the velocity vector of a particle in a specified cell.
-     *
-     * Retrieves the velocity vector of the particle with the position in the specified cell.
-     *
-     * @param cell_id The ID of the cell containing the particle.
-     * @param p_id The position of the particle in the cell_id.
-     * @return The velocity vector of the particle.
-     */
-    const Vec3d get_velocity(const int cell_id, const int p_id)
-    {
[…]
