+++ src/grid/chunk_neighbors_contact.cpp
@@ -16,7 +16,7 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-
+//#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
 #include <exanb/core/operator.h>
 #include <exanb/core/operator_slot.h>
 #include <exanb/core/operator_factory.h>
@@ -39,6 +39,8 @@ under the License.
 #include <exanb/core/domain.h>
 #include <exanb/core/xform.h>
 
+#include <exaDEM/interaction/interaction.hpp>
+
 #include <exanb/particle_neighbors/chunk_neighbors_execute.h>
 
 namespace exaDEM
@@ -59,6 +61,141 @@ namespace exaDEM
       return d2 > 0.0 && d2 < rmax2 && d2 < rcut2;
     }
   };
+  
+    template< class GridT > ONIKA_HOST_DEVICE_FUNC bool nbh_filter_GPU(GridT* cells,
+  						double rcut_inc,
+  						double d2,
+  						double rcut2,
+  						size_t cell_a,
+  						size_t p_a,
+  						size_t cell_b,
+  						size_t p_b)
+  {
+      const double r_a = cells[cell_a][field::radius][p_a];
+      const double r_b = cells[cell_b][field::radius][p_b];
+      const double rmax = r_a + r_b + rcut_inc;
+      const double rmax2 = rmax * rmax;
+      return d2 > 0.0 && d2 < rmax2 && d2 < rcut2 ;
+  }
+  
+  template< class GridT > __global__ void kernelUN(GridT* cells,
+  							int* cell_id,
+  							int* nb_total,
+  							int* nb_particles_nbh,
+  							int* incr_cell,
+  							int* incr_cell2,
+  							int* cell_nb_nbh,
+  							int* start_cell,
+  							int* cell_div,
+  							int* cells_nbh_ids,
+  							int* cells_nbh_particles,
+  							const double dist_lab,
+  							Mat3d xform,
+  							double rcut_inc,
+  							exaDEM::Interaction* interactions,
+  							int* nb_nbh2,
+  							int* interactions_blocks )
+  {
+  	
+  	int cell = cell_id[blockIdx.x];
+  	int total_size = nb_total[blockIdx.x];
+  	int total = nb_particles_nbh[blockIdx.x];
+  	int incr = incr_cell[blockIdx.x];
+  	int incr4 = incr_cell2[blockIdx.x];
+  	int nb_nbh = cell_nb_nbh[blockIdx.x];
+  	int start = start_cell[blockIdx.x];
+  	int div = cell_div[blockIdx.x];
+  	
+  	//__shared__ int shared_threads[256];
+  	__shared__ int shared_threads[256];
+
+	shared_threads[threadIdx.x] = -1;
+	__syncthreads();
+	
+	int nbh_pa;
+  	
+  	for(int i = 0; i < div; i++)
+  	{
+  		//int idx = 256*i + threadIdx.x
+  		int idx = 256*i + threadIdx.x;
+  		if(idx < total_size)
+  		{
+  			int p_a = (int)(idx / total);
+  			int idx2 = idx - (p_a * total);
+  			int cell_b;
+  			int p_b;
+  			bool b2 = true;
+  			int incr2 = 0;
+  			int incr3 = 0;
+  			
+  			for(int j = start; j < start + nb_nbh; j++)
+  			{
+  				incr2+= cells_nbh_particles[j];
+  				if(b2 && idx2 < incr2)
+  				{
[…]
