### 2024-11-30 8d2ef47
rework contacts -> stl / particles (slow)
