### 2024-11-30 15ba13c
use OBB to speedup the detection particle/stl mesh
