### 2024-11-30 db5ee09
add operator compute_driver_vertices.cpp
