### 2024-11-30 e628a8a
perf for stl meshes
