+++ src/interaction/include/exaDEM/shape/shape_detection_api.hpp
@@ -1,11 +1,11 @@
 /*
-   Licensed to the Apache Software Foundation (ASF) under one
-   or more contributor license agreements.  See the NOTICE file
-   distributed with this work for additional information
-   regarding copyright ownership.  The ASF licenses this file
-   to you under the Apache License, Version 2.0 (the
-   "License"); you may not use this file except in compliance
-   with the License.  You may obtain a copy of the License at
+	 Licensed to the Apache Software Foundation (ASF) under one
+	 or more contributor license agreements.  See the NOTICE file
+	 distributed with this work for additional information
+	 regarding copyright ownership.  The ASF licenses this file
+	 to you under the Apache License, Version 2.0 (the
+	 "License"); you may not use this file except in compliance
+	 with the License.  You may obtain a copy of the License at
 
 http://www.apache.org/licenses/LICENSE-2.0
 
@@ -27,401 +27,438 @@ under the License.
 
 namespace exaDEM
 {
-  using namespace exanb;
-  using VertexArray = ::onika::oarray_t<::exanb::Vec3d, EXADEM_MAX_VERTICES>;
+	using namespace exanb;
+	using VertexArray = ::onika::oarray_t<::exanb::Vec3d, EXADEM_MAX_VERTICES>;
 
-  ONIKA_HOST_DEVICE_FUNC inline contact detection_vertex_vertex(
-      const Vec3d &pi, 
-      const int i, 
-      const shape *shpi, 
-      const exanb::Quaternion &oi, 
-      const Vec3d &pj, const int j, 
-      const shape *shpj, 
-      const exanb::Quaternion &oj)
-  {
+  /*********/
+  /* Utils */
+  /*********/
+  inline const Vec3d* get_ptr(const Vec3d* ptr) { return ptr; }
+  inline const Vec3d* get_ptr(const VertexArray& vec) { return vec.data(); }
 
-    // === compute vertex position
-    Vec3d vi = shpi->get_vertex(i, pi, oi);
-    Vec3d vj = shpj->get_vertex(j, pj, oj);
-    return detection_vertex_vertex_core(vi, shpi->m_radius, vj, shpj->m_radius);
-  }
+	/********************/
+	/* Vertex - Vertex  */
+	/********************/
 
-  ONIKA_HOST_DEVICE_FUNC inline contact detection_vertex_vertex(
-      const Vec3d &pi, 
-      const double radius, 
-      const Vec3d &pj, 
-      const int j, 
-      const shape *shpj, 
-      const exanb::Quaternion &oj)
-  {
+	ONIKA_HOST_DEVICE_FUNC inline bool filter_vertex_vertex(
+			const double rVerlet, 
+			const Vec3d &pi, 
+			const double radius, 
+			const Vec3d &pj, 
+			const int j, 
+			const shape *shpj, 
+			const exanb::Quaternion &oj)
+	{
+		// === compute vertex position
+		Vec3d vj = shpj->get_vertex(j, pj, oj);
+		return filter_vertex_vertex(rVerlet, pi, radius, vj, shpj->m_radius);
+	}
 
-    // === compute vertex position
-    Vec3d vj = shpj->get_vertex(j, pj, oj);
-    return detection_vertex_vertex_core(pi, radius, vj, shpj->m_radius);
-  }
+	ONIKA_HOST_DEVICE_FUNC 
+		inline bool filter_vertex_vertex(
+				const double rVerlet, 
+				const Vec3d &pi, 
+				const int i, 
+				const shape *shpi, 
+				const exanb::Quaternion &oi, 
[…]
