+++ src/driver/include/exaDEM/stl_mesh.h
@@ -77,169 +77,177 @@ namespace exaDEM
 
     /**
      * @brief Print information about the STL mesh.
-     */
-    inline void initialize()
-    {
-      // checks
-      if( shp.get_number_of_faces() == 0 && shp.get_number_of_edges() == 0 && shp.get_number_of_vertices() == 0)
+		 */
+		inline void initialize()
+		{
+			// checks
+			if( shp.get_number_of_faces() == 0 
+					&& shp.get_number_of_edges() == 0 
+					&& shp.get_number_of_vertices() == 0)
+			{
+				lout << "Your shape is not correctly defined, no vertex, no edge, and no face" << std::endl;
+				std::abort();
+			}
+
+			// resize and initialize vertices
+			vertices.resize(shp.get_number_of_vertices());  
+#pragma omp parallel for schedule(static)
+      for(int i = 0; i < shp.get_number_of_vertices() ; i++)
       {
-        lout << "Your shape is not correctly defined, no vertex, no edge, and no face" << std::endl;
-        std::abort();
+        this->update_vertex(i);
       }
 
-      // resize vertices
-      vertices.resize(shp.get_number_of_vertices());  
+			// remove relative paths
+			std::filesystem::path full_name = this->shp.m_name;
+			this->shp.m_name = full_name.filename();
+		}
 
-      // remove relative paths
-      std::filesystem::path full_name = this->shp.m_name;
-      this->shp.m_name = full_name.filename();
-    }
+		ONIKA_HOST_DEVICE_FUNC 
+			inline void update_vertex(int i)
+			{
+				vertices[i] = shp.get_vertex(i, this->center, this->quat);
+			}
 
-    ONIKA_HOST_DEVICE_FUNC inline void update_vertex(int i)
-    {
-      vertices[i] = shp.get_vertex(i, this->center, this->quat);
-    }
+		/**
+		 * @brief return driver velocity
+		 */
+		ONIKA_HOST_DEVICE_FUNC inline Vec3d &get_vel() { return vel; }
 
-    /**
-     * @brief return driver velocity
-     */
-    ONIKA_HOST_DEVICE_FUNC inline Vec3d &get_vel() { return vel; }
+		/**
+		 * @brief return driver velocity
+		 */
+		ONIKA_HOST_DEVICE_FUNC inline exanb::Quaternion &get_quat() { return quat; }
 
-    /**
-     * @brief return driver velocity
-     */
-    ONIKA_HOST_DEVICE_FUNC inline exanb::Quaternion &get_quat() { return quat; }
+		ONIKA_HOST_DEVICE_FUNC inline bool stationary()
+		{
+			constexpr exanb::Vec3d null = {0,0,0};
+			return (this->vel == null && this->vrot == null);
+		}
 
-    ONIKA_HOST_DEVICE_FUNC inline bool stationary()
-    {
-      constexpr exanb::Vec3d null = {0,0,0};
-      return (this->vel == null && this->vrot == null);
-    }
+		void dump_driver(int id, std::string path, std::stringstream &stream)
+		{
+			std::string filename = path + this->shp.m_name + ".shp";
+			stream << "  - add_stl_mesh:" << std::endl;
+			stream << "     id: " << id << std::endl;
+			stream << "     center: [" << this->center << "]" << std::endl;
+			stream << "     filename: " << filename << std::endl;
+			stream << "     minskowski: " << this->shp.m_radius << std::endl;
+			stream << "     velocity: [" << this->vel << "]" << std::endl;
+			stream << "     angular_velocity: [" << this->vrot << "]" << std::endl;
[…]
