### 2024-12-01 aadff36
it was already divided by two
