### 2024-12-01 bc8ad32
typo include/exaDEM/stl_mesh.h
