+++ src/driver/include/exaDEM/stl_mesh.h
@@ -1,11 +1,11 @@
 /*
-         Licensed to the Apache Software Foundation (ASF) under one
-         or more contributor license agreements.  See the NOTICE file
-         distributed with this work for additional information
-         regarding copyright ownership.  The ASF licenses this file
-         to you under the Apache License, Version 2.0 (the
-         "License"); you may not use this file except in compliance
-         with the License.  You may obtain a copy of the License at
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
 
 http://www.apache.org/licenses/LICENSE-2.0
 
@@ -77,177 +77,177 @@ namespace exaDEM
 
     /**
      * @brief Print information about the STL mesh.
-		 */
-		inline void initialize()
-		{
-			// checks
-			if( shp.get_number_of_faces() == 0 
-					&& shp.get_number_of_edges() == 0 
-					&& shp.get_number_of_vertices() == 0)
-			{
-				lout << "Your shape is not correctly defined, no vertex, no edge, and no face" << std::endl;
-				std::abort();
-			}
+     */
+    inline void initialize()
+    {
+      // checks
+      if( shp.get_number_of_faces() == 0 
+          && shp.get_number_of_edges() == 0 
+          && shp.get_number_of_vertices() == 0)
+      {
+        lout << "Your shape is not correctly defined, no vertex, no edge, and no face" << std::endl;
+        std::abort();
+      }
 
-			// resize and initialize vertices
-			vertices.resize(shp.get_number_of_vertices());  
+      // resize and initialize vertices
+      vertices.resize(shp.get_number_of_vertices());  
 #pragma omp parallel for schedule(static)
       for(int i = 0; i < shp.get_number_of_vertices() ; i++)
       {
         this->update_vertex(i);
       }
 
-			// remove relative paths
-			std::filesystem::path full_name = this->shp.m_name;
-			this->shp.m_name = full_name.filename();
-		}
+      // remove relative paths
+      std::filesystem::path full_name = this->shp.m_name;
+      this->shp.m_name = full_name.filename();
+    }
 
-		ONIKA_HOST_DEVICE_FUNC 
-			inline void update_vertex(int i)
-			{
-				vertices[i] = shp.get_vertex(i, this->center, this->quat);
-			}
+    ONIKA_HOST_DEVICE_FUNC 
+      inline void update_vertex(int i)
+      {
+        vertices[i] = shp.get_vertex(i, this->center, this->quat);
+      }
 
-		/**
-		 * @brief return driver velocity
-		 */
-		ONIKA_HOST_DEVICE_FUNC inline Vec3d &get_vel() { return vel; }
+    /**
+     * @brief return driver velocity
+     */
+    ONIKA_HOST_DEVICE_FUNC inline Vec3d &get_vel() { return vel; }
 
-		/**
-		 * @brief return driver velocity
-		 */
-		ONIKA_HOST_DEVICE_FUNC inline exanb::Quaternion &get_quat() { return quat; }
+    /**
+     * @brief return driver velocity
+     */
+    ONIKA_HOST_DEVICE_FUNC inline exanb::Quaternion &get_quat() { return quat; }
 
[…]
