+++ src/interaction/include/exaDEM/shape/shape_detection_core.hpp
@@ -1,13 +1,13 @@
 /*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
 
-  http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing,
 software distributed under the License is distributed on an
@@ -15,7 +15,7 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
-*/
+ */
 #pragma once
 
 #include <exanb/core/basic_types.h>
@@ -99,7 +99,7 @@ namespace exaDEM
       return {true, dn, n, contact_position};
     }
   }
-  
+
   ONIKA_HOST_DEVICE_FUNC inline bool filter_vertex_edge_core(const double rVerlet, const Vec3d &vi, const double ri, const Vec3d &vf, const Vec3d &vs, const double rj)
   {
     // === compute distances
@@ -114,7 +114,7 @@ namespace exaDEM
 
     // === r max
     const double Rmax = ri + rj + rVerlet;
-  
+
     return (exanb::dot(dist, dist) < (Rmax * Rmax) );
   }
 
@@ -164,495 +164,495 @@ namespace exaDEM
       const Vec3d contact_position = vi - n * (ri + 0.5 * dn);
 
       return {true, dn, n, contact_position};
-		}
-	}
+    }
+  }
 
 
-	ONIKA_HOST_DEVICE_FUNC 
-		inline contact detection_vertex_face(
-				const Vec3d &pi, 
-				const double radius, 
-				const Vec3d& pj, 
-				const int j, 
-				const shape *shpj, 
-				const exanb::Quaternion &oj)
-		{
-			double ri = radius;
-			double rj = shpj->m_radius;
+  ONIKA_HOST_DEVICE_FUNC 
+    inline contact detection_vertex_face(
+        const Vec3d &pi, 
+        const double radius, 
+        const Vec3d& pj, 
+        const int j, 
+        const shape *shpj, 
+        const exanb::Quaternion &oj)
+    {
+      double ri = radius;
+      double rj = shpj->m_radius;
 
-			// === compute vertices
-			const Vec3d vi = pi;
-			auto [data, nf] = shpj->get_face(j);
-			assert(nf >= 3);
-			Vec3d va = shpj->get_vertex(data[0], pj, oj);
-			Vec3d vb = shpj->get_vertex(data[1], pj, oj);
[…]
