+++ src/polyhedron/nbh_polyhedron.cpp
@@ -64,46 +64,46 @@ namespace exaDEM
                 )EOF";
     }
 
-		template <typename Func> 
-			void add_driver_interaction(
-					Stl_mesh &mesh, 
-					size_t cell_a, 
-					Func &add_contact, 
-					Interaction &item, 
-					const size_t n_particles, 
-					const double rVerlet, 
-					const uint32_t *__restrict__ type, 
-					const uint64_t *__restrict__ id, 
-					const double *__restrict__ rx, 
-					const double *__restrict__ ry, 
-					const double *__restrict__ rz, 
-					const VertexArray *__restrict__ vertices, 
-					const exanb::Quaternion *__restrict__ orient, 
-					shapes &shps)
-			{
+    template <typename Func> 
+      void add_driver_interaction(
+          Stl_mesh &mesh, 
+          size_t cell_a, 
+          Func &add_contact, 
+          Interaction &item, 
+          const size_t n_particles, 
+          const double rVerlet, 
+          const uint32_t *__restrict__ type, 
+          const uint64_t *__restrict__ id, 
+          const double *__restrict__ rx, 
+          const double *__restrict__ ry, 
+          const double *__restrict__ rz, 
+          const VertexArray *__restrict__ vertices, 
+          const exanb::Quaternion *__restrict__ orient, 
+          shapes &shps)
+      {
 #define __particle__ vertices_i, i, shpi
 #define __driver__ mesh.vertices.data(), idx, &mesh.shp
-				assert(cell_a < mesh.grid_indexes.size());
-				auto &list = mesh.grid_indexes[cell_a];
-				const size_t stl_nv = list.vertices.size();
-				const size_t stl_ne = list.edges.size();
-				const size_t stl_nf = list.faces.size();
+        assert(cell_a < mesh.grid_indexes.size());
+        auto &list = mesh.grid_indexes[cell_a];
+        const size_t stl_nv = list.vertices.size();
+        const size_t stl_ne = list.edges.size();
+        const size_t stl_nf = list.faces.size();
 
-				if (stl_nv == 0 && stl_ne == 0 && stl_nf == 0)
-					return;
+        if (stl_nv == 0 && stl_ne == 0 && stl_nf == 0)
+          return;
 
-				for (size_t p = 0; p < n_particles; p++)
-				{
-					Vec3d r = {rx[p], ry[p], rz[p]}; // position
+        for (size_t p = 0; p < n_particles; p++)
+        {
+          Vec3d r = {rx[p], ry[p], rz[p]}; // position
           auto& vertices_i = vertices[p];
           const Quaternion& orient_i = orient[p];
-					item.p_i = p;
-					item.id_i = id[p];
-					auto ti = type[p];
-					const shape *shpi = shps[ti];
-					const size_t nv = shpi->get_number_of_vertices();
-					const size_t ne = shpi->get_number_of_edges();
-					const size_t nf = shpi->get_number_of_faces();
+          item.p_i = p;
+          item.id_i = id[p];
+          auto ti = type[p];
+          const shape *shpi = shps[ti];
+          const size_t nv = shpi->get_number_of_vertices();
+          const size_t ne = shpi->get_number_of_edges();
+          const size_t nf = shpi->get_number_of_faces();
 
           // Get OBB from stl mesh
           auto &stl_shp = mesh.shp;
@@ -121,431 +121,430 @@ namespace exaDEM
           // Note:
           // loop i = particle p
           // loop j = stl mesh
-					for (size_t i = 0; i < nv; i++)
-					{
-						vec3r v = conv_to_vec3r(vertices[p][i]);
+          for (size_t i = 0; i < nv; i++)
+          {
+            vec3r v = conv_to_vec3r(vertices[p][i]);
             OBB obb_v_i;
             obb_v_i.center = v; 
[…]
