### 2024-12-04 4e73b5d
Fix contact stl (#85)
* rework contacts -> stl / particles (slow)

* use OBB to speedup the detection particle/stl mesh

* add operator compute_driver_vertices.cpp

* add tutorial example (blade)

* perf for stl meshes

* add the shape file for the blade tutorial

* wrong path into the blade tutorial

* update tuto blade

* fix exec_ctx

* it was already divided by two

* typo include/exaDEM/stl_mesh.h

* typo shape

* type

* Update cmake.yml

* wrong driver name -> cylinder

* fix densities and radii

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
