+++ src/polyhedron/include/exaDEM/contact_polyhedron.h
@@ -38,7 +38,7 @@ namespace exaDEM
      * @param add Value to add atomically to the variable.
      */
     ONIKA_HOST_DEVICE_FUNC
-    inline void lockAndAdd(double &val, double add) { ONIKA_CU_ATOMIC_ADD(val, add); }
+      inline void lockAndAdd(double &val, double add) { ONIKA_CU_ATOMIC_ADD(val, add); }
 
     /**
      * @brief Atomically adds components of a Vec3d to another Vec3d variable.
@@ -47,141 +47,141 @@ namespace exaDEM
      * @param add Rvalue reference to the source Vec3d whose components are to be added.
      */
     ONIKA_HOST_DEVICE_FUNC
-    inline void lockAndAdd(Vec3d &val, Vec3d &&add)
-    {
-      ONIKA_CU_ATOMIC_ADD(val.x, add.x);
-      ONIKA_CU_ATOMIC_ADD(val.y, add.y);
-      ONIKA_CU_ATOMIC_ADD(val.z, add.z);
-    }
+      inline void lockAndAdd(Vec3d &val, Vec3d &&add)
+      {
+        ONIKA_CU_ATOMIC_ADD(val.x, add.x);
+        ONIKA_CU_ATOMIC_ADD(val.y, add.y);
+        ONIKA_CU_ATOMIC_ADD(val.z, add.z);
+      }
 
     /**
      * @struct contact_law
      * @brief Structure defining contact law interactions for particles (polyhedra).
      */
     template<int interaction_type>
-    struct contact_law
-    {
-      detect<interaction_type> detection;
-      /**
-       * @brief Default constructor for contact_law struct.
-       */
-      contact_law() {}
-
-      /**
-       * @brief Retrieves the position vector of a particle from a cell.
-       *
-       * This function retrieves the position vector of a particle identified
-       * by `p_id` from the given cell using field indices `field::rx`, `field::ry`,
-       * and `field::rz`.
-       *
-       * @tparam TMPLC Type of the cell.
-       * @param cell Reference to the cell containing particle data.
-       * @param p_id Index of the particle.
-       * @return Vec3d Position vector of the particle.
-       */
-      template <typename TMPLC> ONIKA_HOST_DEVICE_FUNC inline const Vec3d get_r(TMPLC &cell, const int p_id) const
+      struct contact_law
       {
-        const Vec3d res = {cell[field::rx][p_id], cell[field::ry][p_id], cell[field::rz][p_id]};
-        return res;
-      };
+        detect<interaction_type> detection;
+        /**
+         * @brief Default constructor for contact_law struct.
+         */
+        contact_law() {}
 
-      /**
-       * @brief Retrieves the velocity vector of a particle from a cell.
-       *
-       * This function retrieves the velocity vector of a particle identified
-       * by `p_id` from the given cell using field indices `field::vx`, `field::vy`,
-       * and `field::vz`.
-       *
-       * @tparam TMPLC Type of the cell.
-       * @param cell Reference to the cell containing particle data.
-       * @param p_id Index of the particle.
-       * @return Vec3d Velocity vector of the particle.
-       */
-      template <typename TMPLC> ONIKA_HOST_DEVICE_FUNC inline const Vec3d get_v(TMPLC &cell, const int p_id) const
-      {
-        const Vec3d res = {cell[field::vx][p_id], cell[field::vy][p_id], cell[field::vz][p_id]};
-        return res;
-      };
-
-      /**
-       * @brief Operator function for performing interactions between particles (polyhedra).
-       *
-       * @tparam TMPLC Type of the cells or particles container.
-       * @param item Reference to the Interaction object representing the interaction details.
-       * @param cells Pointer to the cells or particles container.
-       * @param hkp Reference to the ContactParams object containing interaction parameters.
[…]
