+++ src/polyhedron/nbh_polyhedron.cpp
@@ -64,148 +64,184 @@ namespace exaDEM
                 )EOF";
     }
 
-    template <typename Func> void add_driver_interaction(Stl_mesh &mesh, size_t cell_a, Func &add_contact, Interaction &item, const size_t n_particles, const double rVerlet, const uint32_t *__restrict__ type, const uint64_t *__restrict__ id, const double *__restrict__ rx, const double *__restrict__ ry, const double *__restrict__ rz, const VertexArray *__restrict__ vertices, const exanb::Quaternion *__restrict__ orient, shapes &shps)
-    {
-      assert(cell_a < mesh.grid_indexes.size());
-      auto &list = mesh.grid_indexes[cell_a];
-      const size_t stl_nv = list.vertices.size();
-      const size_t stl_ne = list.edges.size();
-      const size_t stl_nf = list.faces.size();
-      auto &stl_shp = mesh.shp;
-      OBB *__restrict__ stl_obb_vertices = onika::cuda::vector_data(stl_shp.m_obb_vertices);
-      OBB *__restrict__ stl_obb_edges = onika::cuda::vector_data(stl_shp.m_obb_edges);
-      OBB *__restrict__ stl_obb_faces = onika::cuda::vector_data(stl_shp.m_obb_faces);
-
-      if (stl_nv == 0 && stl_ne == 0 && stl_nf == 0)
-        return;
-
-      for (size_t p = 0; p < n_particles; p++)
+    template <typename Func> 
+      void add_driver_interaction(
+          Stl_mesh &mesh, 
+          size_t cell_a, 
+          Func &add_contact, 
+          Interaction &item, 
+          const size_t n_particles, 
+          const double rVerlet, 
+          const uint32_t *__restrict__ type, 
+          const uint64_t *__restrict__ id, 
+          const double *__restrict__ rx, 
+          const double *__restrict__ ry, 
+          const double *__restrict__ rz, 
+          const VertexArray *__restrict__ vertices, 
+          const exanb::Quaternion *__restrict__ orient, 
+          shapes &shps)
       {
-        Vec3d r = {rx[p], ry[p], rz[p]}; // position
-        item.p_i = p;
-        item.id_i = id[p];
-        auto ti = type[p];
-        const shape *shpi = shps[ti];
-        const size_t nv = shpi->get_number_of_vertices();
-        const size_t ne = shpi->get_number_of_edges();
-        const size_t nf = shpi->get_number_of_faces();
+#define __particle__ vertices_i, i, shpi
+#define __driver__ mesh.vertices.data(), idx, &mesh.shp
+        assert(cell_a < mesh.grid_indexes.size());
+        auto &list = mesh.grid_indexes[cell_a];
+        const size_t stl_nv = list.vertices.size();
+        const size_t stl_ne = list.edges.size();
+        const size_t stl_nf = list.faces.size();
 
-        for (size_t i = 0; i < nv; i++)
+        if (stl_nv == 0 && stl_ne == 0 && stl_nf == 0)
+          return;
+
+        for (size_t p = 0; p < n_particles; p++)
         {
-          vec3r v = conv_to_vec3r(vertices[p][i]);
-          // vertex - vertex
-          item.type = 7;
-          item.sub_i = i;
-          for (size_t j = 0; j < stl_nv; j++)
-          {
-            size_t idx = list.vertices[j];
-            OBB obb = stl_obb_vertices[idx];
-            obb.enlarge(rVerlet);
-            // obb.enlarge( shpi->m_radius );
-            if (obb.intersect(v))
-            {
-              add_contact(p, item, i, idx);
-            }
-          }
-          // vertex - edge
-          item.type = 8;
[…]
