+++ src/interaction/include/exaDEM/type/OBB.hpp
@@ -71,6 +71,15 @@ public:
     extent.z += more;
   }
 
+  void rescale(const double scale)
+  {
+    e1 = scale * e1;
+    e2 = scale * e2;
+    e3 = scale * e3;
+    extent = scale * extent;
+    center = scale * center;
+  }
+
   ONIKA_HOST_DEVICE_FUNC
   void translate(const vec3r &v) { center += v; }
 
