### 2024-12-04 3cf170d
fix compilation issue on GPU (stl mesh)
