### 2024-12-04 2475003
stl meshes rework on GPU now
