+++ src/driver/driver_displ_over.cpp
@@ -18,143 +18,177 @@ under the License.
  */
 //#pragma xstamp_cuda_enable
 
+#include <mpi.h>
 #include <exanb/core/operator.h>
 #include <exanb/core/operator_slot.h>
 #include <exanb/core/operator_factory.h>
 #include <exanb/core/basic_types.h>
-#include <mpi.h>
-#include <exaDEM/shape/shapes.hpp>
 #include <exaDEM/drivers.h>
+#include <exaDEM/reduce_stl_mesh.hpp>
 
 namespace exaDEM
 {
+	using namespace exanb;
+	template <typename T> using VectorT = onika::memory::CudaMMVector<T>;
 
-  struct driver_displ_over
-  {
-    double r2; // square distance
-    MPI_Comm &comm;
-    int operator()(exaDEM::UndefinedDriver &a, exaDEM::UndefinedDriver &b) { return 0; }
-    int operator()(exaDEM::Cylinder &a, exaDEM::Cylinder &b) { return 0; } // WARNING should not move
-    int operator()(exaDEM::Surface &a, exaDEM::Surface &b)
-    {
-      Vec3d d = a.center_proj - b.center_proj;
-      if (exanb::dot(d, d) >= r2)
-        return 1;
-      else
-        return 0;
-    }
-    int operator()(exaDEM::Ball &a, exaDEM::Ball &b)
-    {
-      Vec3d d = a.center - b.center;
-      if (exanb::dot(d, d) >= r2)
-        return 1;
-      else
-        return 0;
-    }
-    int operator()(exaDEM::Stl_mesh &a, exaDEM::Stl_mesh &b)
-    {
-      if ((a.center == b.center) && (a.quat == b.quat))
-        return 0;
-      // check each vertex
-      auto check_dist = [](Vec3d &v1, Vec3d &v2, double r2) -> bool
-      {
-        Vec3d d = v1 - v2;
-        if (exanb::dot(d, d) >= r2)
-          return true;
-        return false;
-      };
-      auto &shp_a = a.shp;
-      size_t size = shp_a.get_number_of_vertices();
-      int sum = 0;
+  template<typename Operator> struct driver_displ_over
+	{
+    Operator* op;
+		double r2; // square distance
+		MPI_Comm &comm;
+    int * storage;
+		/* Ignored */ 	
+		int operator()(exaDEM::UndefinedDriver &a, exaDEM::UndefinedDriver &b) { return 0; }
+		int operator()(exaDEM::Cylinder &a, exaDEM::Cylinder &b) { return 0; } // WARNING should not move
 
-      size_t start(0), end(size);
-      // optimization for big shapes
-      if (size > 100)
-      {
-        int mpi_rank, mpi_size;
-        MPI_Comm_rank(comm, &mpi_rank);
-        MPI_Comm_size(comm, &mpi_size);
-        start = (size * mpi_rank) / mpi_size;
-        end = (size * (mpi_rank + 1)) / mpi_size;
-      }
+		/* Active */
+		int operator()(exaDEM::Surface &a, exaDEM::Surface &b)
+		{
+			Vec3d d = a.center_proj - b.center_proj;
+			if (exanb::dot(d, d) >= r2)
+				return 1;
+			else
+				return 0;
+		}
+
+		int operator()(exaDEM::Ball &a, exaDEM::Ball &b)
+		{
+			Vec3d d = a.center - b.center;
+			if (exanb::dot(d, d) >= r2)
+				return 1;
+			else
+				return 0;
+		}
+
+		int operator()(exaDEM::Stl_mesh &a, exaDEM::Stl_mesh &b)
+		{
[…]
