+++ src/driver/driver_displ_over.cpp
@@ -28,167 +28,166 @@ under the License.
 
 namespace exaDEM
 {
-	using namespace exanb;
-	template <typename T> using VectorT = onika::memory::CudaMMVector<T>;
+  using namespace exanb;
+  template <typename T> using VectorT = onika::memory::CudaMMVector<T>;
 
   template<typename Operator> struct driver_displ_over
-	{
+  {
     Operator* op;
-		double r2; // square distance
-		MPI_Comm &comm;
+    double r2; // square distance
+    MPI_Comm &comm;
     int * storage;
-		/* Ignored */ 	
-		int operator()(exaDEM::UndefinedDriver &a, exaDEM::UndefinedDriver &b) { return 0; }
-		int operator()(exaDEM::Cylinder &a, exaDEM::Cylinder &b) { return 0; } // WARNING should not move
+    /* Ignored */   
+    int operator()(exaDEM::UndefinedDriver &a, exaDEM::UndefinedDriver &b) { return 0; }
+    int operator()(exaDEM::Cylinder &a, exaDEM::Cylinder &b) { return 0; } // WARNING should not move
 
-		/* Active */
-		int operator()(exaDEM::Surface &a, exaDEM::Surface &b)
-		{
-			Vec3d d = a.center_proj - b.center_proj;
-			if (exanb::dot(d, d) >= r2)
-				return 1;
-			else
-				return 0;
-		}
+    /* Active */
+    int operator()(exaDEM::Surface &a, exaDEM::Surface &b)
+    {
+      Vec3d d = a.center_proj - b.center_proj;
+      if (exanb::dot(d, d) >= r2)
+        return 1;
+      else
+        return 0;
+    }
 
-		int operator()(exaDEM::Ball &a, exaDEM::Ball &b)
-		{
-			Vec3d d = a.center - b.center;
-			if (exanb::dot(d, d) >= r2)
-				return 1;
-			else
-				return 0;
-		}
+    int operator()(exaDEM::Ball &a, exaDEM::Ball &b)
+    {
+      Vec3d d = a.center - b.center;
+      if (exanb::dot(d, d) >= r2)
+        return 1;
+      else
+        return 0;
+    }
 
-		int operator()(exaDEM::Stl_mesh &a, exaDEM::Stl_mesh &b)
-		{
-			if ((a.center == b.center) && (a.quat == b.quat))
-				return 0;
+    int operator()(exaDEM::Stl_mesh &a, exaDEM::Stl_mesh &b)
+    {
+      if ((a.center == b.center) && (a.quat == b.quat))
+        return 0;
 
 #ifdef ONIKA_CUDA_VERSION
-			size_t size = a.shp.get_number_of_vertices();
+      size_t size = a.shp.get_number_of_vertices();
       const Vec3d * __restrict__ ptr_shp_vertices = onika::cuda::vector_data(a.shp.m_vertices);
-			STLVertexDisplacementFunctor SVDFunc = {r2, ptr_shp_vertices, a.center, a.quat, b.center, b.quat};
-			ReduceMaxSTLVertexDisplacementFunctor func = {SVDFunc, storage};
+      STLVertexDisplacementFunctor SVDFunc = {r2, ptr_shp_vertices, a.center, a.quat, b.center, b.quat};
+      ReduceMaxSTLVertexDisplacementFunctor func = {SVDFunc, storage};
 
-			ParallelForOptions opts;
-			opts.omp_scheduling = OMP_SCHED_STATIC;
-			parallel_for(size, func, op->parallel_execution_context(), opts);   
-			return *storage;
+      ParallelForOptions opts;
+      opts.omp_scheduling = OMP_SCHED_STATIC;
[…]
