### 2024-12-04 8c056a9
add educe_stl_mesh.hpp
