### 2024-12-04 2d667a9
add educe_stl_mesh.hpp
