### 2024-12-05 f28f55b
improve get faces with an offset array
