+++ src/interaction/include/exaDEM/shape/shape_prepro.hpp
@@ -1,13 +1,13 @@
 /*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
 
-  http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing,
 software distributed under the License is distributed on an
@@ -15,7 +15,7 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
-*/
+ */
 #pragma once
 
 #include <exaDEM/shape/shape.hpp>
@@ -25,98 +25,99 @@ namespace exaDEM
 {
   using namespace exanb;
 
-  inline OBB build_OBB(const std::vector<vec3r> &vec, double radius)
-  {
-    OBB obb;
-    vec3r mu;
-    mat9r C;
-    for (size_t i = 0; i < vec.size(); i++)
+  template<typename Vec>
+    inline OBB build_OBB(const Vec &vec, double radius)
     {
-      mu += vec[i];
+      OBB obb;
+      vec3r mu;
+      mat9r C;
+      for (size_t i = 0; i < vec.size(); i++)
+      {
+        mu += vec[i];
+      }
+      mu /= (double)vec.size();
+
+      // loop over the points again to build the
+      // covariance matrix.  Note that we only have
+      // to build terms for the upper trianglular
+      // portion since the matrix is symmetric
+      double cxx = 0.0, cxy = 0.0, cxz = 0.0, cyy = 0.0, cyz = 0.0, czz = 0.0;
+      for (size_t i = 0; i < vec.size(); i++)
+      {
+        vec3r p = vec[i];
+        cxx += p.x * p.x - mu.x * mu.x;
+        cxy += p.x * p.y - mu.x * mu.y;
+        cxz += p.x * p.z - mu.x * mu.z;
+        cyy += p.y * p.y - mu.y * mu.y;
+        cyz += p.y * p.z - mu.y * mu.z;
+        czz += p.z * p.z - mu.z * mu.z;
+      }
+
+      // now build the covariance matrix
+      C.xx = cxx;
+      C.xy = cxy;
+      C.xz = cxz;
+      C.yx = cxy;
+      C.yy = cyy;
+      C.yz = cyz;
+      C.zx = cxz;
+      C.zy = cyz;
+      C.zz = czz;
+
+      // ==== set the OBB parameters from the covariance matrix
+      // extract the eigenvalues and eigenvectors from C
+      mat9r eigvec;
+      vec3r eigval;
+      C.sym_eigen(eigvec, eigval);
+
[…]
