### 2024-12-05 b6cd319
Fix STL Meshes on GPU (#87)
* stl meshes rework on GPU now

* tab driver_displ_over.cpp

* add educe_stl_mesh.hpp

* add educe_stl_mesh.hpp

* improve get faces with an offset array

* Update config_move_particles.msp

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
