+++ src/interaction/include/exaDEM/shape/shape.hpp
@@ -1,13 +1,13 @@
 /*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
 
-  http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing,
 software distributed under the License is distributed on an
@@ -15,7 +15,7 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
-*/
+ */
 #pragma once
 #include <fstream>
 #include <vector>
@@ -59,157 +59,177 @@ namespace exaDEM
     OBB obb;
     VectorT<int> m_edges; ///<
     VectorT<int> m_faces; ///<
+    VectorT<int> m_offset_faces; ///<
     double m_radius;      ///< use for detection
     double m_volume;      ///< use for detection
     std::string m_name = "undefined";
 
     // functions in shape_prepro.hpp
-    /*		ONIKA_HOST_DEVICE_FUNC inline void pre_compute_obb_vertices();
-                    ONIKA_HOST_DEVICE_FUNC inline void pre_compute_obb_edges();
-                    ONIKA_HOST_DEVICE_FUNC inline void pre_compute_obb_faces();
-        ONIKA_HOST_DEVICE_FUNC inline void increase_obb ( const double value );
-    */
-    inline void pre_compute_obb_vertices(const exanb::Vec3d &particle_center, const exanb::Quaternion &particle_quat);
+    /*    ONIKA_HOST_DEVICE_FUNC inline void pre_compute_obb_vertices();
+          ONIKA_HOST_DEVICE_FUNC inline void pre_compute_obb_edges();
+          ONIKA_HOST_DEVICE_FUNC inline void pre_compute_obb_faces();
+          ONIKA_HOST_DEVICE_FUNC inline void increase_obb ( const double value );
+     */
     inline void pre_compute_obb_edges(const exanb::Vec3d &particle_center, const exanb::Quaternion &particle_quat);
     inline void pre_compute_obb_faces(const exanb::Vec3d &particle_center, const exanb::Quaternion &particle_quat);
+    inline void pre_compute_obb_vertices(const exanb::Vec3d * scratch);
+    inline void pre_compute_obb_edges(const exanb::Vec3d * scratch);
+    inline void pre_compute_obb_faces(const exanb::Vec3d * scratch);
     inline void increase_obb(const double value);
+    void compute_prepro_obb(exanb::Vec3d * scratch, const exanb::Vec3d &particle_center, const exanb::Quaternion &particle_quat);
 
     ONIKA_HOST_DEVICE_FUNC
-    inline double get_volume() const
[…]

[…]
