### 2024-12-06 1b6274e
Friction et oment
