+++ src/interaction/classify_interactions.cpp
@@ -44,6 +44,7 @@ namespace exaDEM
     ADD_SLOT(GridCellParticleInteraction, ges, INPUT, DocString{"Interaction list"});
     ADD_SLOT(Classifier<InteractionSOA>, ic, INPUT_OUTPUT, DocString{"Interaction lists classified according to their types"});
     ADD_SLOT(CellListWrapper, cell_list, INPUT, DocString{"list of non empty cells within the current grid"});
+    ADD_SLOT(InteractionSOA, interaction_neighbors, INPUT_OUTPUT);
 
   public:
     inline std::string documentation() const override final
@@ -61,7 +62,7 @@ namespace exaDEM
       auto [cell_ptr, cell_size] = cell_list->info();
       if (!ic.has_value())
         ic->initialize();
-      ic->classify(*ges, cell_ptr, cell_size);
+      ic->classify(*ges, cell_ptr, cell_size/*, *interaction_neighbors*/);
       //ic->prefetch_memory_on_gpu(); // GPU only
     }
   };
