+++ src/grid/chunk_neighbors_contact.cpp
@@ -41,7 +41,13 @@ under the License.
 
 #include <exaDEM/interaction/interaction.hpp>
 
+#include <exaDEM/interaction/interactionSOA.hpp>
+
 #include <exanb/particle_neighbors/chunk_neighbors_execute.h>
+
+#include <cub/cub.cuh>
+
+#define THREADS_PER_BLOCK 256
 
 namespace exaDEM
 {
@@ -80,121 +86,144 @@ namespace exaDEM
   
   template< class GridT > __global__ void kernelUN(GridT* cells,
   							int* cell_id,
-  							int* nb_total,
-  							int* nb_particles_nbh,
-  							int* incr_cell,
-  							int* incr_cell2,
-  							int* cell_nb_nbh,
-  							int* start_cell,
-  							int* cell_div,
-  							int* cells_nbh_ids,
-  							int* cells_nbh_particles,
+  							int* nb_particles,
+  							int* cell_neighbors_ids,
+  							int* cell_neighbors_size,
+  							int* cell_start,
+  							int* cell_end,
+  							int* cell_particle_start,
   							const double dist_lab,
   							Mat3d xform,
   							double rcut_inc,
-  							exaDEM::Interaction* interactions,
-  							int* nb_nbh2,
-  							int* interactions_blocks )
+  							uint64_t* id_i,
+  							uint64_t* id_j,
+  							uint32_t* cell_i,
+  							uint32_t* cell_j,
+  							uint16_t* p_i,
+  							uint16_t* p_j,
+  							int* nb_nbh,
+  							int* id,
+  							int* total_interactions )
   {
   	
   	int cell = cell_id[blockIdx.x];
-  	int total_size = nb_total[blockIdx.x];
-  	int total = nb_particles_nbh[blockIdx.x];
-  	int incr = incr_cell[blockIdx.x];
-  	int incr4 = incr_cell2[blockIdx.x];
-  	int nb_nbh = cell_nb_nbh[blockIdx.x];
-  	int start = start_cell[blockIdx.x];
-  	int div = cell_div[blockIdx.x];
-  	
-  	//__shared__ int shared_threads[256];
-  	__shared__ int shared_threads[256];
-
-	shared_threads[threadIdx.x] = -1;
-	__syncthreads();
+	int nb = nb_particles[blockIdx.x];
+	int start = cell_start[blockIdx.x];
+	int end = cell_end[blockIdx.x];
+	int incr = cell_particle_start[blockIdx.x];
 	
-	int nbh_pa;
-  	
-  	for(int i = 0; i < div; i++)
-  	{
-  		//int idx = 256*i + threadIdx.x
-  		int idx = 256*i + threadIdx.x;
-  		if(idx < total_size)
-  		{
-  			int p_a = (int)(idx / total);
-  			int idx2 = idx - (p_a * total);
-  			int cell_b;
-  			int p_b;
-  			bool b2 = true;
-  			int incr2 = 0;
-  			int incr3 = 0;
-  			
-  			for(int j = start; j < start + nb_nbh; j++)
-  			{
-  				incr2+= cells_nbh_particles[j];
-  				if(b2 && idx2 < incr2)
-  				{
-  					cell_b = cells_nbh_ids[j];
-  					p_b = idx2 - incr3;
-  					b2 = false;
-  				}
-  				
-  				incr3+= cells_nbh_particles[j];
-  			
-  			}
-  			
-  			double rcut2 = dist_lab * dist_lab;
+	if(threadIdx.x < nb)
+	{
[…]
