### 2024-12-07 3a2c855
Update spack.yml
