### 2024-12-07 8d3cb13
Update spack.yml
