### 2024-12-10 #7484 @slamont1
Minor clarification needed in the installation documentation. Issue created: https://github.com/Collab4exaNBody/doc_exaDEM/issues/10
