### 2024-12-10 Collab4exaNBody/exaDEM#91 @slamont1
Add community guidelines
Please supply a visible link within your documentation expressing guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
