### 2024-12-10 #7484 @slamont1
Missing community guidelines. Issue created: https://github.com/Collab4exaNBody/exaDEM/issues/91
