### 2024-12-11 be6fae5
add CONTRIBUTING.md
