+++ README.md
@@ -13,10 +13,40 @@ Documentation is available here:
 - Website: [ExaDEM Website](https://collab4exanbody.github.io/doc_exaDEM/)
 - Github: [ExaDEM Documentation](https://github.com/Collab4exaNBody/doc_exaDEM.git)
 
+- Main Sections:
+  - [Overview](https://collab4exanbody.github.io/doc_exaDEM/project_exaDEM/Introduction.html)
+  - [Installation](https://collab4exanbody.github.io/doc_exaDEM/project_exaDEM/Installation.html)
+  - [ExaNBody] (https://collab4exanbody.github.io/doc_exaDEM/project_exaNBody/index.html)
+  - [Spheres] (https://collab4exanbody.github.io/doc_exaDEM/project_exaDEM/Spheres.html) / [Polyhedra] (https://collab4exanbody.github.io/doc_exaDEM/project_exaDEM/Polyhedra.html)
+  - [Force field](https://collab4exanbody.github.io/doc_exaDEM/project_exaDEM/ForceField.html)
+  - [I/O](https://collab4exanbody.github.io/doc_exaDEM/project_exaDEM/Analysis.html)
+  - [Examples](https://collab4exanbody.github.io/doc_exaDEM/project_exaDEM/Test_cases.html)
+  - [Tutorials](https://collab4exanbody.github.io/doc_exaDEM/project_exaDEM/Tutorial.html)
+
+## Community Guidelines
+
+- For any bug, please create an issue and add the label “bug”. We welcome all feedback to make exaDEM as robust as possible.
+- If you would like to participate and add functionality to `exaDEM`, you can find instructions for coding style, tests and pull request process in `CONTRIBUTING.md`.
+- If you have any support-related / collaboration questions, please contact the team at `raphael.prat@cea.fr`. If you are a `CEA` member, please request access to the group : "exaNBody & Co. (exaStamp, exaDEM, exaSPH)", an external access can also be provided. 
+
 ## Authors and acknowledgment
+
+### Main developers
+
+- Raphaël Prat (raphael.prat@cea.fr)
+- Thierry Carrard
+- Carlo Elia Doncecchi
+
+### Other Developers
+
+- Paul Lafourcade
+- Lhassan Amarsid
+- Vincent Richefeu
+
+### Acknowledgment
 
 `ExaDEM` is part of the `PLEIADES` platform which has been developped in collaboration with the French nuclear industry - mainly `CEA`, `EDF`, and `Framatome` - for simulation of fuel1 elements.
 
 ## License
 
-See `LICENSE.txt` 
+See `LICENSE.txt`
