### 2024-12-11 64e072e
Faster on GPU (#89)
