### 2024-12-11 f9e9ab5
91 add community guidelines (#93)
* add CONTRIBUTING.md

* update README.md

* add Support.md

* rename doc to docs

* fix CI [draft-pdf]

* update developers

* Update CONTRIBUTING.md

* Update README.md

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
