### 2024-12-11 #7484 @rprat-pro
Thank you @slamont1 for your remarks, I've completed the documentation for installing yaml-cpp and I've completed the community guidelines in the `README.md` and `CONTRIBUTE.md` files.
