### 2024-12-11 Collab4exaNBody/exaDEM#94 @slamont1
Add clarification on tests and examples
It isn't clear where the examples from the documentation are located within the codebase. Specifically, the last section 3.3 for simulations that can run on a laptop, which I would imagine many people would try to run first.

I notice in the repository there are folders titled 'example,' 'tests,' and 'tutorial,' but as a new user I'm not sure where I should go first to test functionality versus usage, etc. I see some references to these titles within the documentation, but the language sometimes isn't consistent (for instance, the documentation on 'test cases' refers to simulations located within the 'examples' folder instead of the 'tests' folder).

I think including a README inside of each of these folders would be useful to outline what the expected usage for these simulations are.
