### 2024-12-11 #7484 @slamont1
Hi @rprat-pro, I'm running through some of the provided examples to test functionality and I think some clarification is needed on their expected usage. While the documentation provides screenshots of initial/final conditions (which is great), it's not always easy to match the simulation in the documentation to its location in the actual codebase.

I think this is easily solved with some extra READMEs to point the user in the right direction. Please refer to the issue I created here: https://github.com/Collab4exaNBody/exaDEM/issues/94

I also can't seem to find the simulations in Test Cases 3.3 titled 'Show-cases' - are they in a separate example folder?
