### 2024-12-11 3b07a1e
update sphere test cases + add README.md in example
