+++ src/CMakeLists.txt
@@ -28,7 +28,7 @@ add_subdirectory(driver)
 add_subdirectory(interaction)
 add_subdirectory(sphere)
 add_subdirectory(polyhedron)
-add_subdirectory(analysis)
+#add_subdirectory(analysis)
 
 if(USE_RSA)
 	add_subdirectory(rsa)
