### 2024-12-12 27fcded
update documentation and test cases
