+++ example/README.md
@@ -1,3 +1,3 @@
 # Examples
-
-Here you'll find a very short list of examples, divided into two types: spheres and polyhedra.
+- Here you'll find a very short list of examples, divided into two types: spheres and polyhedra.
+- Two tutorials are also available into the folder `tutorial`.
