### 2024-12-12 Collab4exaNBody/exaDEM#94 @rprat-pro
Hi @slamont1

- I removed the tests folder for clarity (which contained old files) and migrated the tutorials to example.

- `README.md` files have been added to the folders in example with very minimalist descriptions (to minimize documentation duplication and maintenance).

- The website documentation has been modified to add paths to improve test case identification.

I hope this will improve the experience of new users.
