### 2024-12-12 c98bc63
94 add clarification on tests and examples (#95)
* update sphere test cases + add README.md in example

* add README.md for examples

* sphere examples -> ok

* update documentation and test cases

* move tutorial to example

* Update README.md

* Create README.md

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
