### 2024-12-12 #7484 @rprat-pro
Hi @slamont1 , I've added the path to the first "show case" in the documentation. I'll need a few days/weeks to double-check that my input files work correctly for this blade -> https://github.com/Collab4exaNBody/exaDEM/issues/96 .

The documentation (website) has been updated (Test cases) with file paths.

Thank you for your remarks.
