### 2024-12-12 c9fb58e
add analysis directory
