### 2024-12-13 #7484 @slamont1
@jedbrown, I've finished my review. The software is functional, well-documented, and reflects a significant amount of work from the authors. I believe it meets all of the relevant criteria for publication in JOSS.
