+++ src/analysis/counter.cpp
@@ -22,27 +22,33 @@ under the License.
 #include <exanb/core/operator_slot.h>
 #include <exanb/core/operator_factory.h>
 #include <exanb/core/make_grid_variant_operator.h>
-#include <exanb/core/parallel_grid_algorithm.h>
+#include <exanb/compute/reduce_cell_particles.h>
+#include <exanb/grid_cell_particles/particle_region.h>
 #include <exanb/core/grid.h>
+
 #include <memory>
 #include <exaDEM/cell_list_wrapper.hpp>
 #include <exaDEM/AnalysisManager.hpp>
+#include <exaDEM/counter.hpp>
+
 
 namespace exaDEM
 {
   using namespace exanb;
 
-  template <typename GridT, class = AssertGridHasFields<GridT>> class CounterAnalysis : public OperatorNode
+  template <typename GridT, class = AssertGridHasFields<GridT, field::_rx, field::_ry, field::_rz, field::_type>> class CounterAnalysis : public OperatorNode
   {
 
+    static constexpr FieldSet<field::_rx,field::_ry,field::_rz,field::_type> reduce_field_set{};
+    ADD_SLOT(MPI_Comm, mpi, INPUT, MPI_COMM_WORLD);
     ADD_SLOT(GridT, grid, INPUT_OUTPUT);
     ADD_SLOT(CellListWrapper, cell_list, INPUT_OUTPUT, DocString{"list of non empty cells within the current grid"});
     ADD_SLOT(double, dt, INPUT, REQUIRED);
     ADD_SLOT(long, timestep, INPUT, REQUIRED, DocString{"Iteration number"});
     ADD_SLOT(ParticleRegions, particle_regions, INPUT, OPTIONAL);
     ADD_SLOT(ParticleRegionCSG, region, INPUT, OPTIONAL);
-    ADD_SLOT(std::string, dir_name, INPUT, REQUIRED, DocString{"Output directory"});
-    ADD_SLOT(std::string, name, INPUT, "counter_undefined.txt", DocString{"Filename. Default is: counter_undefined.txt"});
+    ADD_SLOT(std::string, dir_name, INPUT, REQUIRED, DocString{"Output directory, usually defined into io_config."});
+    ADD_SLOT(std::string, name, INPUT, "ParticleCounter.txt", DocString{"Filename. Default is: ParticleCounter.txt"});
     ADD_SLOT(std::vector<int>, types, INPUT, REQUIRED , DocString{"List of types"});
 
 
@@ -61,39 +67,49 @@ namespace exaDEM
       manager.set_filename (*name);
       double t = (*dt) * (*timestep);
       manager.add_element("Time", t, "%3f");
+      auto& list_of_types = (*types);
 
-      // now, fill the radius field
-      if (region.has_value())
-      {
-        ParticleRegionCSGShallowCopy prcsg = *region;
-        UpdateRadiusPolyhedronFunctor func = {prcsg, onika::cuda::vector_data(r)};
-        if (!particle_regions.has_value())
-        {
-          fatal_error() << "Region is defined, but particle_regions has no value" << std::endl;
-        }
+      if( list_of_types.size() == 0 ) lout << "[Analysis/counter] types is empty, this operator is skipped" << std::endl;
 
-        if (region->m_nb_operands == 0)
-        {
-          ldbg << "rebuild CSG from expr " << region->m_user_expr << std::endl;
-          region->build_from_expression_string(particle_regions->data(), particle_regions->size());
-        }
-        compute_cell_particles(*grid, false, func, compute_region_field_set, parallel_execution_context());
-      }
-      else
-      {
-        ParticleRegionCSGShallowCopy prcsg;
-        UpdateRadiusPolyhedronFunctor func = {prcsg, onika::cuda::vector_data(r)};
[…]
