### 2024-12-14 9851978
fix cmakelist for anaylises
