### 2024-12-14 cf99282
add particle counter example
