### 2024-12-16 91c5394
update analysis test cases + add particle_counter example
