### 2024-12-16 172e83d
90 add analysis count particles into a region (#97)
* update sphere test cases + add README.md in example

* add analysis directory

* add counter, it works correctly

* add analyses freq into the exaDEM execution task graph.

* fix cmakelist for anaylises

* add particle counter example

* remove old command in lattice (repeats) used into example files

* remove old command in lattice (repeats) used into example files

* fix test cases

* update analysis test cases + add particle_counter example

* Add file particle_counter_minimal.msp

---------

Co-authored-by: rprat-pro <rp269144@pleiades1
[…]
