+++ example/polyhedra/analyses/CMakeLists.txt
@@ -1,4 +1,5 @@
-file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/shapes.shp DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
+file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/shapes.shp ${CMAKE_CURRENT_SOURCE_DIR}/box_avg_stress.stl DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
+add_omp_mpi_test("ParticleCount" "particle_counter_minimal.msp")
 add_omp_mpi_test("InteractionOutput" "interaction_minimal.msp")
 add_omp_mpi_test("WritePosVelOutput" "write_pos_vel_minimal.msp")
 add_omp_mpi_test("WriteAvgStress" "write_avg_stress_minimal.msp")
