+++ docs/paper/joss/paper.md
@@ -47,30 +47,30 @@ bibliography: paper.bib
 
 # Summary 
 
-`ExaDEM` is a Discrete Element Method (`DEM`) code developed using the `exaNBody` framework [@carrard2023exanbody] at the French atomic commission (`CEA`). This software provides `DEM` modeling capabilities to enable the modeling of mechanical interactions occuring at contact between spheres and polyhedra particles while offering performance optimizations on modern `HPC` platforms. A notable aspect of `ExaDEM` stands in its hybrid parallelization approach, combining the use of `MPI` and Threads (OpenMP). Additionally, `ExaDEM` offers portability to clusters of `GPU`s, using the `CUDA` programming model and `MPI` for `DEM` simulations with spherical particles. Developed in `C++17`, `ExaDEM` is designed to offer `HPC` capabilities to the scientific community. 
+`ExaDEM` is a Discrete Element Method (`DEM`) code developed using the `exaNBody` framework [@carrard2023exanbody] at the French atomic commission (`CEA`). This software provides `DEM` modeling capabilities to enable the modeling of mechanical interactions occurring at contact between spheres and polyhedra particles while offering performance optimizations on modern `HPC` platforms. A notable aspect of `ExaDEM` stands in its hybrid parallelization approach, combining the use of `MPI` and Threads (OpenMP). Additionally, `ExaDEM` offers portability to clusters of `GPU`s, using the `CUDA` programming model and `MPI` for `DEM` simulations with spherical particles. Developed in `C++17`, `ExaDEM` is designed to offer `HPC` capabilities to the scientific community. 
 
 # Statement of need
 
-A comprehesive description of the behavior of granular media remains a significant challenge for the scientific community. `DEM` simulations enhance our knowledge by examining phenomena that are otherwise unreachable or too costly to study through experiments. To accurately reproduce these phenomena originated at particle scale, we must simulate a representative number of particles, ranging from thousands to billions. While a single processor can easily handle simulations involving thousands of particles, simulating millions requires `HPC` resources. These simulations are constrained either by memory footprint or excessively long execution time. The `DEM` method is inherently parallel, with various techniques avaiable, such as spatial domain decomposition [@plimpton1995fast], thread parallelization over cells and vectorization. In this paper, we introduce the code `ExaDEM`, designed for large-scale `DEM` simulation on `HPC` platforms. This code leverages numerous features of the `exaNBody` framework, including data structure management, `MPI`+X parallelization and add-on modules (optimized `IO`, Paraview export).
+A comprehensive description of the behavior of granular media remains a significant challenge for the scientific community. `DEM` simulations enhance our knowledge by examining phenomena that are otherwise unreachable or too costly to study through experiments. To accurately reproduce these phenomena originated at particle scale, we must simulate a representative number of particles, ranging from thousands to billions. While a single processor can easily handle simulations involving thousands of particles, simulating millions requires `HPC` resources. These simulations are constrained either by memory footprint or excessively long execution time. The `DEM` method is inherently parallel, with various techniques available, such as spatial domain decomposition [@plimpton1995fast], thread parallelization over cells and vectorization. In this paper, we introduce the code `ExaDEM`, designed for large-scale `DEM` simulation on `HPC` platforms. This code leverages numerous features of the `exaNBody` framework, including data structure management, `MPI`+X parallelization and add-on modules (optimized `IO`, Paraview export).
 
 # DEM Background
 
[…]
