### 2024-12-17 cad410a
Other gammar checks (joss paper)
