### 2024-12-17 #7484 @rprat-pro
> Hi @rprat-pro - Cool code! I went over the manuscript, and everything is clear. I note a few typos in the manuscript, provided per line number, hoping to improve presentation of the work:
>
> 13: "occuring" → "occurring" 21: "comprehesive" → "comprehensive" 28: "avaiable" → "available" 39: “moment" → “momentum” 61: “In-situ” → “in-situ” 70: “DEM specific mechanism” → “a DEM specific mechanism” or “DEM specific mechanisms” 75: "represetns" → "represents" 96: “exaNBody” → “the exaNBody“ 98: "theses" → "these" 98: "independents" → "independent" 104: "to to" → "to" 132: "developped" → "developed" 132: “Rocakble” → “Rockable”
>
> I will follow up with comments soon, as I go over the docs and test runs.

Thank you @vsangelidakis , I've just corrected the typos mentioned (and others). Don't hesitate to let me know of any compilation problems =).
