+++ src/interaction/unclassify_interactions.cpp
@@ -16,6 +16,9 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
+
+//#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
+
 #include <exanb/core/operator.h>
 #include <exanb/core/operator_slot.h>
 #include <exanb/core/operator_factory.h>
@@ -37,10 +40,140 @@ under the License.
 #include <exaDEM/shape/shape_detection_driver.hpp>
 #include <exaDEM/mutexes.h>
 #include <exaDEM/drivers.h>
+#include <cub/cub.cuh>
 
 namespace exaDEM
 {
   using namespace exanb;
+  
+  __global__ void kernelUN( double* ft_x,
+  			double* ft_y,
+  			double* ft_z,
+  			double* mom_x,
+  			double* mom_y,
+  			double* mom_z,
+  			size_t size,
+  			int* blocks,
+  			int* total)
+  {
+  	int idx = threadIdx.x + blockIdx.x * blockDim.x;
+  	
+  	if(idx < size)
+  	{
+  		if(ft_x[idx]!=0 || ft_y[idx]!=0 || ft_z[idx]!=0 || mom_x[idx]!=0 || mom_y[idx]!=0 || mom_z[idx]!=0)
+  		{
+  			atomicAdd(&blocks[blockIdx.x], 1);
+  			atomicAdd(&total[0], 1);
+  		}
+  	}
+  }
+  
+  __global__ void kernelDEUX( uint64_t* id_i,
+  				uint64_t* id_j,
+  				uint64_t* id_i_res,
+  				uint64_t* id_j_res,
+  				double* ft_x,
+  				double* ft_y,
+  				double* ft_z,
+  				double* ft_x_res,
+  				double* ft_y_res,
+  				double* ft_z_res,
+  				double* mom_x,
+  				double* mom_y,
+  				double* mom_z,
+  				double* mom_x_res,
+  				double* mom_y_res,
+  				double* mom_z_res,
+  				int* blocks_incr,
+  				int* indices,
+  				int* particles,
+  				int size)
+  {
+  	int idx = threadIdx.x + blockIdx.x * blockDim.x;
+  	
+  	int incr = blocks_incr[blockIdx.x];
+  	
+  	__shared__ int s[256];
+  	
+  	s[threadIdx.x] = 0;
+  	
+  	__syncthreads();
+  	
+  	
+  	if(idx < size)
+  	{
+  		if(ft_x[idx]!=0 || ft_y[idx]!=0 || ft_z[idx]!=0 || mom_x[idx]!=0 || mom_y[idx]!=0 || mom_z[idx]!=0)
+  		{
+  			s[threadIdx.x] = 1;
+  		}
+  		
+  		__syncthreads();
+  		
+  		int index = 0;
+  		
+  		if(s[threadIdx.x] == 1)
+  		{
+  			for(int i = 0; i < threadIdx.x; i++)
+  			{
+  				if( s[i] == 1 ) index++;
+  			}
+  			
+  			id_i_res[index + incr] = id_i[threadIdx.x];
+  			id_j_res[index + incr] = id_j[threadIdx.x];
+  			ft_x_res[index + incr] = ft_x[threadIdx.x];
+  			ft_y_res[index + incr] = ft_y[threadIdx.x];
+  			ft_z_res[index + incr] = ft_z[threadIdx.x];
+  			mom_x_res[index + incr] = mom_x[threadIdx.x];
+  			mom_y_res[index + incr] = mom_y[threadIdx.x];
+  			mom_z_res[index + incr] = mom_z[threadIdx.x];
+  			
[…]
