+++ src/interaction/classify_interactions.cpp
@@ -16,6 +16,9 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
+
+//#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
+
 #include <exanb/core/operator.h>
 #include <exanb/core/operator_slot.h>
 #include <exanb/core/operator_factory.h>
@@ -33,6 +36,56 @@ under the License.
 namespace exaDEM
 {
   using namespace exanb;
+  
+  __global__ void kernel( double* ft_x,
+  			double* ft_y,
+  			double* ft_z,
+  			double* mom_x,
+  			double* mom_y,
+  			double* mom_z,
+  			uint64_t* id_i,
+  			uint64_t* id_j,
+  			double* ft_x_old,
+  			double* ft_y_old,
+  			double* ft_z_old,
+  			double* mom_x_old,
+  			double* mom_y_old,
+  			double* mom_z_old,
+  			uint64_t* id_i_old,
+  			uint64_t* id_j_old,
+  			int* indices,
+  			int* particles,
+  			int* particles_incr,
+  			size_t size)
+  {
+  	int idx = threadIdx.x + blockIdx.x * blockDim.x;
+  	
+  	if( idx < size )
+  	{
+  		int id = id_i[idx];
+  		
+  		if(particles[id] > 0)
+  		{
+  			int p = particles_incr[id];
+  			
+  			int id2 = id_j[idx];
+  			
+  			for(int i = 0; i < particles[id]; i++)
+  			{
+  				if( id2 == id_j_old[indices[p + i]])
+  				{
+  					ft_x[idx] = ft_x_old[indices[p + i]];
+  					ft_y[idx] = ft_y_old[indices[p + i]]; 
+  					ft_z[idx] = ft_z_old[indices[p + i]];
+  					
+  					mom_x[idx] = mom_x_old[indices[p + i]];  
+  					mom_y[idx] = mom_y_old[indices[p + i]];  
+  					mom_z[idx] = mom_z_old[indices[p + i]];   
+  				}
+  			}
+  		}
+  	}
+  }
 
   template <typename GridT, class = AssertGridHasFields<GridT, field::_radius>> class ClassifyInteractions : public OperatorNode
   {
@@ -44,6 +97,7 @@ namespace exaDEM
     ADD_SLOT(GridCellParticleInteraction, ges, INPUT, DocString{"Interaction list"});
     ADD_SLOT(Classifier<InteractionSOA>, ic, INPUT_OUTPUT, DocString{"Interaction lists classified according to their types"});
     ADD_SLOT(CellListWrapper, cell_list, INPUT, DocString{"list of non empty cells within the current grid"});
+    ADD_SLOT(OldClassifier, ic_old, INPUT_OUTPUT);
 
   public:
     inline std::string documentation() const override final
@@ -54,6 +108,9 @@ namespace exaDEM
 
     inline void execute() override final
     {
+    
+      //printf("CLASSIFY\n");
+    
       if (grid->number_of_cells() == 0)
       {
         return;
@@ -63,6 +120,27 @@ namespace exaDEM
         ic->initialize();
       ic->classify(*ges, cell_ptr, cell_size);
       //ic->prefetch_memory_on_gpu(); // GPU only
+      
+      auto [data, size] = ic->get_info(0);
+      
+      InteractionWrapper<InteractionSOA> interactions(data);
+      
+      int blockSize = 256;
+      int numBlocks = ( size + blockSize - 1 ) / blockSize;
+      
+      auto& old = *ic_old;
+      
[…]
