### 2024-12-19 Collab4exaNBody/exaDEM#100 @vsangelidakis
JOSS: Review on exaDEM docs
Hello, I share some suggestions to consider regarding your docs as part of my JOSS review. My comments look extensive but they are really small things that should not take more than a few minutes to implement.

1. Some optional suggestions, not requirements within this review, hoping they might help (@rprat-pro feel free to tick them when you have read them or optionally implemented them):

- [x] In CONTRIBUTING.md, you state _“Please do not use the issue tracker for personal support requests (contact directly the authors [raphael.prat@cea.fr](mailto:raphael.prat@cea.fr)”_.  As the code grows, you might start receiving an unsustainably high amount of emails with questions of how to use the different features of the code. You could use the “Discussions” field in Github to keep track of personal support requests/advice of how to do something in the code, so that other users/devs can also benefit from the answer. In YADE, we keep a historical, open-source record of answers, and this is very helpful to us as devs and to the users who might have a question now that was answered already years ago.
- [x] Are you considering using the Sphinx LaTeΧ builder to create a pdf version of the documentation? This is also very handy for easy navigation of all docs in one place.
- [x] Section 5.3 Interaction / Contact - Contact model for polyhedra: Do you use GJK for contact detection and EPA for contact force calculations or something in-house? I suggest documenting how the contact point / contact normal / contact force are calculated for polyhedra, and if needed point to the relevant Rockable docs.
- [x] Have you checked if your polyhedra conserve energy (always a tricky point for non-spherical particles in the DEM).
- [x] Section 7.1. Contact Force Laws - Regarding Contact Laws, you might want to look into more advanced contact models with time, such as Hertz-Mindlin (1949), Hertz-Mindlin and Deresiewicz (1953) and into more advanced viscous damping models, where the coefficient of restitution is velocity-dependent.
- [x] Section 8. Numerical Scheme - Regarding time integration, you might want to look into more advanced time integration schemes, such as SPIRAL (del Valle et al, Comput. Phys. Commun. 2024) for non-spherical particles.

2. Some suggestions for improvement within this review:

- [x] The file CONTRIBUTING.md provides the very helpful `ctest --test-dir example` for unit tests, but IMO this should be added also in the docs, right after installation, so that a new/any user can quickly check the validity of their installation of exaDEM.
- [x] Do these tests only check that the script can start running, or is there any unit testing also based on results, e.g. compared to analytical/expected results?
- [x] To use exaDEM more easily, I found handy defining an alias in `~/.bashrc`, so that might be helpful advice for new users in the docs.
- [x] Example 3.1.1: The docs suggest two scripts for the rotating drum simulation (`/rotating-drum-full.msp` and `/rotating-no-periodic.msp`), but there are additional scripts in the folder (e.g. `/rotating-drum.msp`). I would suggest to either document all scripts or remove any redundant ones. This also happens in other examples, where only a few scripts are documented in the docs, and other variants of the scripts exist in the same folder, with no information of what why they are there - this can be confusing to a new user.
- [x] Regarding vtk outputs:
    - [x] #101
    - [x] You could also save stress tensors (static and/or dynamic) per particle, later on.
    - [x] I suggest documenting how the user can visualise the shape of polyhedra. I only see the points/positions in the vtp files.
    - [x] I suggest documenting how the user can visualise interparticle forces / force chains, if this is available.
    - [x] The output of all runs is saved in a folder “ExaDEMOutputDir”, and every run overwrites that files in that folder. For users running many examples from the same directory, it might be beneficial to add the choice to name the output folder in the input file, or show how to do this if this is already available.
- [x] With the exception of the first example, all other examples take a fairly long time to run / write vtk files, which could put off a new user. You could reduce the computation time by reducing the number of particles and/or modify the contact stiffness/density to accelerate these examples.
- [x] Examples 3.1.1 to 3.1.3, and 3.3.1: you use URLs as `exaDEM/examples/spheres/...`, but for Examples 3.1.4 to 3.2.5 you use truncated URLs as `examples/spheres/..`. This is a detail, but I would suggest using the same formatting across all cases.
- [x] Example 3.1.7: You should add the missing URL to the files.
- [x] Example 3.3.1: You are noting the old URL, as this example is now in `examples/spheres/cylinder_stl/cylinder_stl.msp`
- [x] Example 3.3.2: If this example is missing, I would suggest commenting it out, and add it when you are ready to release working example. This is a suggestion to help keep the docs clean.

Hope these comments help improve your already high-quality docs. Vasileios
