### 2024-12-19 #7484 @vsangelidakis
Hi @rprat-pro, I have added a few comments regarding documentation as an issue under https://github.com/Collab4exaNBody/exaDEM/issues/100. Great docs, I enjoyed going through them! Do let me know if you have any questions.
