### 2024-12-19 Collab4exaNBody/exaDEM#100 @rprat-pro
Thank you @vsangelidakis for your suggestions, they will indeed greatly improve the documentation. First, I'll try to answer and add your suggestions quickly (especially part 2). Then with @jmvanson and @lamarsid we'll think about improvements on the physics side. ExaDEM is still a young code that's flexible enough to integrate different contact modeling or time integration schemes so it's a good way to stress its design by trying new physiques =)
