### 2024-12-19 Collab4exaNBody/exaDEM#100 @vsangelidakis
That's great @rprat-pro. Please do not overthink the answers to the points I raise, you can fix the few broken links/straightforward changes, and let me know so that I mark the review on JOSS as completed.

You can keep my more open-ended suggestions as ideas for improvement outside of the review process (i.e. do not feel rushed to implement everything now).
