### 2024-12-22 ab829eb
add barycenter operator
