+++ CONTRIBUTING.md
@@ -12,7 +12,7 @@ Using the issue tracker
 
 The issue tracker is the preferred channel for bug reports, features requests and submitting pull requests, but please respect the following restrictions:
 
-- Please do not use the issue tracker for personal support requests (contact directly the authors raphael.prat@cea.fr
+- Please do not use the issue tracker for personal support requests, open a "discussion" so that others can benefit from the exchanges and answers (or contact directly the authors raphael.prat@cea.fr for very specific personal support requests).
 
 Bug reports
 ===========
