### 2024-12-22 Collab4exaNBody/exaDEM#100 @vsangelidakis
One more thing: it would be helpful to report in the docs and in the JOSS paper if your spheropolyhedra have to be convex or if they can have any arbitrary shape.
