### 2024-12-22 Collab4exaNBody/exaDEM#100 @rprat-pro
Yes, this approach enables us to handle non-convex shapes such as hexapods.
