+++ docs/paper/joss/paper.md
@@ -55,7 +55,7 @@ A comprehensive description of the behavior of granular media remains a signific
 
 # DEM Background
 
-The `DEM` method, employed to study granular media, falls within the scope of so-called N-body methods. It involves numerically replicating the kinetics and interactions of a collection of rigid particles over time. Time is divided into discrete steps, where at each step n, Newton's equation **f**=m**a** is solved to determine the translation and rotation acceleration for each particle. Subsequently, the corresponding velocity and momentum are computed, leading to the calculation of the new position at step n+1. The force **f** is evaluated from particle interactions, including contact interactions (normal and tangential force) and external forces such as gravity. A commonly employed numerical scheme is the Velocity Verlet, while the Hooke law is often used to model contact interactions. The versatility of the `DEM` method allows for simulating rigid bodies with various shapes, ranging from spherical to polyhedral particles. Note that in `ExaDEM`, complex particle shapes are handled using a sphero-polyhedral approach, facilitating simplified contact detection and representing intricate shapes [@alonso_marroquin_2008]. For clarity, the term "polyhedron" is used to refer to a sphero-polyhedral particle.
+The `DEM` method, employed to study granular media, falls within the scope of so-called N-body methods. It involves numerically replicating the kinetics and interactions of a collection of rigid particles over time. Time is divided into discrete steps, where at each step n, Newton's equation **f**=m**a** is solved to determine the translation and rotation acceleration for each particle. Subsequently, the corresponding velocity and momentum are computed, leading to the calculation of the new position at step n+1. The force **f** is evaluated from particle interactions, including contact interactions (normal and tangential force) and external forces such as gravity. A commonly employed numerical scheme is the Velocity Verlet, while the Hooke law is often used to model contact interactions. The versatility of the `DEM` method allows for simulating rigid bodies with various shapes, ranging from spherical to polyhedral particles. Note that in `ExaDEM`, complex particle shapes are handled using a sphero-polyhedral approach, facilitating simplified contact detection and representing intricate shapes [@alonso_marroquin_2008]. For clarity, the term "polyhedron" is used to refer to a sphero-polyhedral particle. Note that this approach can be used to simulate non-convex polyhedra such as hexapods.
 
 Efficient neighbor particle detection is a critical aspect of `DEM` simulation codes. Typically, this is achieved by combining the linked cells method [@ciccotti1987simulation] and Verlet lists [@verlet1967computer], which optimize neighbor searches using a cartesian grid of cells (with a complexity of N, N is the number of particles) while limiting the refresh rate of neighbor lists.  
   
