+++ example/polyhedra/balls/CMakeLists.txt
@@ -1,2 +1,2 @@
 file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/alpha3.shp DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
-add_omp_mpi_test("HexapodsBalls" "balls.msp")
+add_omp_mpi_test("HexapodsBalls" "balls_mini.msp")
