### 2024-12-23 Collab4exaNBody/exaDEM#100 @rprat-pro
Hi @vsangelidakis, here are some answers to your questions/tasks:

```
Are you considering using the Sphinx LaTeΧ builder to create a pdf version of the documentation? This is also very handy for easy navigation of all docs in one place.
```

We had added the commands to generate the pdf in the github `doc_exaDEM` but for the moment we hadn't planned to add a pdf as the site seemed sufficient. I think I'll add a documentation version to the next release if I can improve the rendering (which I think is pretty ugly).

```
    Section 5.3 Interaction / Contact - Contact model for polyhedra: Do you use GJK for contact detection and EPA for contact force calculations or something in-house? I suggest documenting how the contact point / contact normal / contact force are calculated for polyhedra, and if needed point to the relevant Rockable docs.
```

@lamarsid :   GJK (Gilbert-Johnson-Keerthi) and EPA (Expanding Polytope Algorithm) are commonly used in DEM and can be implemented. However, we use an approach based on a spheropolyhedral representation  (with a Minkowski radius) of shapes, with detection relying on bounding boxes and cell-linking. In this approach, we identify four types of contacts: vertex-to-vertex, vertex-to-edge, edge-to-edge, and     vertex-to-face. The remaining types of contacts are combinations of these.

In addition, we're currently working with Vincent Richefeu (creator of Rockable) on a paper that explains all these aspects and will make a very good additional resource.

```
    Have you checked if your polyhedra conserve energy (always a tricky point for non-spherical particles in the DEM).
```

@lamarsid :  Spheropolyhedra generate simultaneous multiple contacts (edge-edge and vertex-face at the same time for example), which can complicate the dynamics and potentially introduce energy losses if the forces are not calculated accurately. But as you say, event for other approaches it is a tricky point.

```
    Section 7.1. Contact Force Laws - Regarding Contact Laws, you might want to look into more advanced contact models with time, such as Hertz-Mindlin (1949), Hertz-Mindlin and Deresiewicz (1953) and into more advanced viscous damping models, where the coefficient of restitution is velocity-dependent.
```

@lamarsid :  For now, we are using linear viscoelastic contact laws, but nothing prevents us from using models such as Hertz-Mindlin. thank you for your advising.

```
    Section 8. Numerical Scheme - Regarding time integration, you might want to look into more advanced time integration schemes, such as SPIRAL (del Valle et al, Comput. Phys. Commun. 2024) for non-spherical particles.
```

@lamarsid : exaDEM is a part of project for n-body interaction involing also molecular dynamics and SPH, for time integration, we use velocity Verlet witch is common scheme to these methods. for the future, we also want to add  Runge-Kutta and predictor-corrector.

In addition,  ExaDEM makes it easy to change the integration scheme by modifying / adding operators in ``numerical_scheme: numerical_scheme_verlet`` with currently :

```
numerical_scheme_verlet: numerical_scheme_verlet_macro(compute_force)

numerical_scheme_verlet_macro(compute_force_op):
  name: scheme
  body:
#    - push_f_v_r: { dt_scale: 1.0 }
#    - push_f_v: { dt_scale: 0.5 }
#    - push_to_quaternion: { dt_scale: 1.0 }
    - combined_compute_prolog
    - push_av_to_quat_driver
    - push_v_to_r_driver
    - mpi_barrier
    - check_and_update_particles
    - reset_force_moment
    - mpi_barrier
    - compute_force_op
    - mpi_barrier
    - force_to_accel
#    - update_angular_acceleration
#    - update_angular_velocity: { dt_scale: 1 }
#    - push_f_v: { dt_scale: 0.5 }
    - combined_compute_epilog
```

```
 Do these tests only check that the script can start running, or is there any unit testing also based on results, e.g. compared to analytical/expected results?
 ```

 Indeed, we don't test that the physical results remain exactly the same. We had initially planned to compare particle positions, but this turned out to be more complicated than expected, as depending on the hardware architecture, compiler and type of parallelization, we didn't obtain exactly the same result. We had thought of adding a relative tolerance, but finally we moved on to other problems... On the other hand, as soon as the code starts to go wrong, we generally miss a contact and the code crashes very quickly with an error or calculation times increase enormously. In the end, it's still an open question, perhaps by adding post-processing ...

```
 You could also save stress tensors (static and/or dynamic) per particle, later on.
 ```

 Issue #44 . Yes, it's planned, it's just a question of time and priority.

 For other points:

 - I added the operator `io_config:` to the examples to specify the output directory,
 - I've renamed the examples so that the minimal cases contain “mini” in their names,
 - I've added some time slots to give the reader an idea of the simulations before running them,
 - I've provided a few links to post-processing with Paraview at the beginning of the Test cases section,
 - Two examples propose to display the contact network,
 - This example is "ready" 3.3.2 but I need to write the description in the following days with the input files.

![pale-huge-slide](https://github.com/user-attachments/assets/5a38d8fd-355b-4cf4-84ac-9ab9a4e5344e)
![pale-huge-profile](https://github.com/user-attachments/assets/bdf9e9de-d7db-4ca4-91c6-23a7e7953a7e)

I hope I've answered all your questions.
Thanks again for your great comments.
