+++ .github/workflows/cmake.yml
@@ -15,7 +15,7 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4.2.2
       with:
         persist-credentials: false
     - uses: actions/setup-node@v4
