### 2024-12-23 Collab4exaNBody/exaDEM#100 @vsangelidakis
Thank you for the detailed answers - Happy to approve the review at JOSS.

Regarding regression tests and checks to verify physical results, I would suggest implementing these sooner rather than later, to make sure that every new commit does not break the reproducibility of your code. There are a few available two-particle problems with analytical solutions (e.g. see Chung and Ooi, 2011, Granul. Matter), as well as bulk-scale benchmarks you can look at. For these you might want to introduce a tolerance as you say, as indeed parallelisation and system architecture can influence the sequence of operations and induce small differences that do not allow for perfectly deterministic results.

Well done writing this code and releasing it in an open-source manner.

All the best,
Vasileios
