### 2024-12-23 #7484 @vsangelidakis
@jedbrown I have finished my review. This is a great software to simulate particulate materials efficiently, the code and examples are well documented, and the authors have put significant effort in developing this computational tool. I believe this contribution meets the criteria to be published in JOSS.
