### 2024-12-23 Collab4exaNBody/exaDEM#100 @rprat-pro
Thank you again for your review.

I've just had a look at the article “Benchmark tests for verifying discrete element modelling codes at particle impact level” and it seems easy enough to add it for spherical particles. Thank you.
