### 2024-12-23 5952631
change cell list to traversal (real or all)
