+++ src/io/simulation_state.cpp
@@ -35,7 +35,7 @@ under the License.
 
 #include <exaDEM/simulation_state.h>
 #include <exaDEM/dem_simulation_state.h>
-#include <exaDEM/cell_list_wrapper.hpp>
+#include <exaDEM/traversal.hpp>
 #include <exaDEM/interaction/classifier.hpp>
 #include <exaDEM/itools/itools.hpp>
 
@@ -52,7 +52,7 @@ namespace exaDEM
     ADD_SLOT(MPI_Comm, mpi, INPUT, MPI_COMM_WORLD);
     ADD_SLOT(GridT, grid, INPUT, REQUIRED);
     ADD_SLOT(Domain, domain, INPUT, REQUIRED);
-    ADD_SLOT(CellListWrapper, cell_list, INPUT, DocString{"list of non empty cells within the current grid"});
+    ADD_SLOT(Traversal, traversal_real, INPUT, DocString{"list of non empty cells within the current grid"});
     ADD_SLOT(SimulationState, simulation_state, OUTPUT);
 
     // DEM data
@@ -106,7 +106,7 @@ namespace exaDEM
       double mass = 0.;
       uint64_t total_particles = 0;
 
-      auto [cell_ptr, cell_size] = cell_list->info();
+      auto [cell_ptr, cell_size] = traversal_real->info();
       exaDEM::simulation_state_variables sim{}; // kinetic_energy, rotation_energy, mass, potential_energy, total_particles};
       ReduceSimulationStateFunctor func = {};
       reduce_cell_particles(*grid, false, func, sim, reduce_field_set, parallel_execution_context(), {}, cell_ptr, cell_size);
