### 2024-12-23 76cb8ba
adapt the exmaple with the barycenter operator with the documentation
