### 2024-12-23 42bae61
98 compute barycenter region and type (#103)
* add type (field) during the ghost update process

* add barycenter operator

* add example with the calculation of the barycenters

* adapt the exmaple with the barycenter operator with the documentation

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
