### 2024-12-24 214693b
add stress stensor / reset stress / avg stress stensor. Compilation ok
