### 2024-12-25 545902c
fix issue when computing stress tensors, add stress tensor to the exadem workflow, fix examples
