+++ src/interaction/stress_tensor.cpp
@@ -49,9 +49,9 @@ namespace exaDEM
             uint64_t idx,
 						Interaction& I, 
 						TMPLC *const __restrict__ cells, 
-						Vec3d *const __restrict__ fnp, 
-						Vec3d *const __restrict__ ftp, 
-						Vec3d *const __restrict__ cpp) const
+						const Vec3d *const __restrict__ fnp, 
+						const Vec3d *const __restrict__ ftp, 
+						const Vec3d *const __restrict__ cpp) const
 				{
 					assert( type == I.type );
 					// get fij and cij
@@ -67,7 +67,7 @@ namespace exaDEM
 						Vec3d fji = -fij;
 						Vec3d pos_j = {cellj[field::rx][I.p_j], cellj[field::ry][I.p_j], cellj[field::rz][I.p_j]};
 						Vec3d cji = cpp[idx] - pos_j;
-						exanb::atomic_add_contribution(cell[field::stress][I.p_j], exanb::tensor(fji, cji));
+						exanb::atomic_add_contribution(cellj[field::stress][I.p_j], exanb::tensor(fji, cji));
 					}
 				}
 		};
@@ -81,14 +81,17 @@ namespace exaDEM
 				void iteration(Classifier<InteractionSOA>& classifier, TMPLC *const __restrict__ cells)
 				{
 					static_assert(Type >= 0 && Type < NTypes);
-					ParallelForOptions opts;
-					opts.omp_scheduling = OMP_SCHED_STATIC;
-					auto [Ip, size] = classifier.get_info(Type);           // get interactions
-					auto [dnp, cpp, fnp, ftp] = classifier.buffer_p(Type); // get forces (fn, ft) and contact positions (cp) computed into the contact force operators.
-					InteractionWrapper<InteractionSOA> interactions(Ip);
-					compute_stress_tensor<Type, Sym> func;
-					WrapperForAll wrapper(interactions, func , cells, fnp, ftp, cpp);
-					parallel_for(size, wrapper, oper->parallel_execution_context(), opts);
+					auto [Ip, size] = classifier.get_info(Type); // get interactions
+          if (size > 0 )
+          {
+					  ParallelForOptions opts;
+					  opts.omp_scheduling = OMP_SCHED_STATIC;
+					  const auto [dnp, cpp, fnp, ftp] = classifier.buffer_p(Type); // get parameters: get forces (fn, ft) and contact positions (cp) computed into the contact force operators.
+					  InteractionWrapper<InteractionSOA> interactions(Ip);         // get data: interaction
+					  compute_stress_tensor<Type, Sym> func;                       // get kernel
+					  WrapperForAll wrapper(interactions, func , cells, fnp, ftp, cpp); // pack data, kernel, and interaction in a wrapper
+					  parallel_for(size, wrapper, oper->parallel_execution_context(), opts); // launch kernel
+          }
 				}
 
 			template<int Type, typename... Args> void loop(Args... args)
@@ -105,18 +108,12 @@ namespace exaDEM
 		};
 
 
-	template <typename GridT, class = AssertGridHasFields<GridT, field::_rx, field::_ry, field::_rz>> class StressTensor : public OperatorNode
+	template <typename GridT, class = AssertGridHasFields<GridT, field::_stress>> class StressTensor : public OperatorNode
 	{
 		// attributes processed during computation
-		using ComputeFields = FieldSet<field::_vrot, field::_arot>;
-		static constexpr ComputeFields compute_field_set{};
-
 		ADD_SLOT(MPI_Comm, mpi, INPUT, MPI_COMM_WORLD);
 		ADD_SLOT(GridT, grid, INPUT_OUTPUT, REQUIRED);
[…]
