+++ src/interaction/include/exaDEM/interaction/classifier.hpp
@@ -24,352 +24,352 @@ namespace exaDEM
   template <typename T> struct InteractionWrapper;
 
   template <> 
-		struct InteractionWrapper<InteractionSOA>
-		{
-			// forces
-			double * __restrict__ ft_x;
-			double * __restrict__ ft_y;
-			double * __restrict__ ft_z;
-			// moment
-			double * __restrict__ mom_x;
-			double * __restrict__ mom_y;
-			double * __restrict__ mom_z;
-			// particle id
-			uint64_t * __restrict__ id_i;
-			uint64_t * __restrict__ id_j;
-			// cell id
-			uint32_t * __restrict__ cell_i;
-			uint32_t * __restrict__ cell_j;
-			// position into the cell
-			uint16_t * __restrict__ p_i;
-			uint16_t * __restrict__ p_j;
-			// sub id
-			uint16_t * __restrict__ sub_i;
-			uint16_t * __restrict__ sub_j;
-			uint16_t m_type;
+    struct InteractionWrapper<InteractionSOA>
+    {
+      // forces
+      double * __restrict__ ft_x;
+      double * __restrict__ ft_y;
+      double * __restrict__ ft_z;
+      // moment
+      double * __restrict__ mom_x;
+      double * __restrict__ mom_y;
+      double * __restrict__ mom_z;
+      // particle id
+      uint64_t * __restrict__ id_i;
+      uint64_t * __restrict__ id_j;
+      // cell id
+      uint32_t * __restrict__ cell_i;
+      uint32_t * __restrict__ cell_j;
+      // position into the cell
+      uint16_t * __restrict__ p_i;
+      uint16_t * __restrict__ p_j;
+      // sub id
+      uint16_t * __restrict__ sub_i;
+      uint16_t * __restrict__ sub_j;
+      uint16_t m_type;
 
-			InteractionWrapper(InteractionSOA &data)
-			{
-				ft_x = onika::cuda::vector_data(data.ft_x);
-				ft_y = onika::cuda::vector_data(data.ft_y);
-				ft_z = onika::cuda::vector_data(data.ft_z);
+      InteractionWrapper(InteractionSOA &data)
+      {
+        ft_x = onika::cuda::vector_data(data.ft_x);
+        ft_y = onika::cuda::vector_data(data.ft_y);
+        ft_z = onika::cuda::vector_data(data.ft_z);
 
-				mom_x = onika::cuda::vector_data(data.mom_x);
-				mom_y = onika::cuda::vector_data(data.mom_y);
-				mom_z = onika::cuda::vector_data(data.mom_z);
+        mom_x = onika::cuda::vector_data(data.mom_x);
+        mom_y = onika::cuda::vector_data(data.mom_y);
+        mom_z = onika::cuda::vector_data(data.mom_z);
 
-				id_i = onika::cuda::vector_data(data.id_i);
-				id_j = onika::cuda::vector_data(data.id_j);
+        id_i = onika::cuda::vector_data(data.id_i);
+        id_j = onika::cuda::vector_data(data.id_j);
 
-				cell_i = onika::cuda::vector_data(data.cell_i);
-				cell_j = onika::cuda::vector_data(data.cell_j);
+        cell_i = onika::cuda::vector_data(data.cell_i);
+        cell_j = onika::cuda::vector_data(data.cell_j);
 
-				p_i = onika::cuda::vector_data(data.p_i);
-				p_j = onika::cuda::vector_data(data.p_j);
+        p_i = onika::cuda::vector_data(data.p_i);
+        p_j = onika::cuda::vector_data(data.p_j);
 
-				sub_i = onika::cuda::vector_data(data.sub_i);
[…]
