+++ src/interaction/stress_tensor.cpp
@@ -1,13 +1,13 @@
 /*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
 
-  http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing,
 software distributed under the License is distributed on an
@@ -15,7 +15,7 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
-*/
+ */
 #include <exanb/core/operator.h>
 #include <exanb/core/operator_slot.h>
 #include <exanb/core/operator_factory.h>
@@ -37,105 +37,105 @@ under the License.
 
 namespace exaDEM
 {
-	using namespace exanb;
+  using namespace exanb;
   using namespace onika::parallel;
 
-	template <int type, bool sym>
-		struct compute_stress_tensor
-		{
-			template <typename TMPLC> 
-      ONIKA_HOST_DEVICE_FUNC 
-      inline void operator()(
+  template <int type, bool sym>
+    struct compute_stress_tensor
+    {
+      template <typename TMPLC> 
+        ONIKA_HOST_DEVICE_FUNC 
+        inline void operator()(
             uint64_t idx,
-						Interaction& I, 
-						TMPLC *const __restrict__ cells, 
-						const Vec3d *const __restrict__ fnp, 
-						const Vec3d *const __restrict__ ftp, 
-						const Vec3d *const __restrict__ cpp) const
-				{
-					assert( type == I.type );
-					// get fij and cij
-					auto &cell = cells[I.cell_i];
-					Vec3d fij = fnp[idx] + ftp[idx];
-					Vec3d pos_i = {cell[field::rx][I.p_i], cell[field::ry][I.p_i], cell[field::rz][I.p_i]};
-					Vec3d cij = cpp[idx] - pos_i;
-					exanb::atomic_add_contribution(cell[field::stress][I.p_i], exanb::tensor(fij, cij));
+            Interaction& I, 
+            TMPLC *const __restrict__ cells, 
+            const Vec3d *const __restrict__ fnp, 
+            const Vec3d *const __restrict__ ftp, 
+            const Vec3d *const __restrict__ cpp) const
+        {
+          assert( type == I.type );
+          // get fij and cij
+          auto &cell = cells[I.cell_i];
+          Vec3d fij = fnp[idx] + ftp[idx];
+          Vec3d pos_i = {cell[field::rx][I.p_i], cell[field::ry][I.p_i], cell[field::rz][I.p_i]};
+          Vec3d cij = cpp[idx] - pos_i;
+          exanb::atomic_add_contribution(cell[field::stress][I.p_i], exanb::tensor(fij, cij));
 
-					if constexpr ( type <= 3 && sym == true) // polyhedron - polyhedron || sphere - sphere
-					{
-						auto &cellj = cells[I.cell_j];
-						Vec3d fji = -fij;
[…]
