### 2024-12-25 6c0151a
it works on parallel (stress tensor)
