### 2024-12-25 6dea9e1
Add stress stensor (#104)
* add type (field) during the ghost update process

* add barycenter operator

* add example with the calculation of the barycenters

* adapt the exmaple with the barycenter operator with the documentation

* add stress stensor / reset stress / avg stress stensor. Compilation ok

* fix issue when computing stress tensors, add stress tensor to the exadem workflow, fix examples

* indent

* it works on parallel (stress tensor)

* add example/polyhedra/analyses/octahedron.shp

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
