+++ include/exanb/field_sets.h
@@ -26,7 +26,7 @@ namespace exanb
   // DEM model field set
   using DEMFieldSet = FieldSet<
       // rx, ry and rz are added implicitly
-      field::_vx, field::_vy, field::_vz, field::_fx, field::_fy, field::_fz, field::_mass, field::_homothety, field::_radius, field::_orient, field::_mom, field::_vrot, field::_arot, field::_inertia, field::_id, field::_type, field::_vertices>;
+      field::_vx, field::_vy, field::_vz, field::_fx, field::_fy, field::_fz, field::_mass, field::_homothety, field::_radius, field::_orient, field::_mom, field::_vrot, field::_arot, field::_inertia, field::_id, field::_type, field::_vertices, field::_stress>;
 
   // the standard set of FieldSet
   // use FieldSetsWith<fields...> (at the bottom of this file) to select a subset depending on required fields
