+++ src/interaction/avg_stress_tensor.cpp
@@ -1,0 +1,119 @@
+/*
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing,
+software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+KIND, either express or implied.  See the License for the
+specific language governing permissions and limitations
+under the License.
+ */
+#include <exanb/core/operator.h>
+#include <exanb/core/operator_slot.h>
+#include <exanb/core/operator_factory.h>
+#include <exanb/core/make_grid_variant_operator.h>
+#include <exanb/core/parallel_grid_algorithm.h>
+#include <exanb/core/concurent_add_contributions.h>
+#include <exanb/compute/reduce_cell_particles.h>
+
+#include <exanb/core/grid.h>
+#include <exaDEM/traversal.hpp>
+#include <exaDEM/type/add_contribution_mat3d.hpp>
+
+#include <memory>
+#include <mpi.h>
+
+namespace exaDEM
+{
+  using namespace exanb;
+
+  struct ReduceStressTensorFunctor
+  { 
+    ONIKA_HOST_DEVICE_FUNC inline void operator()(Mat3d &local, const Mat3d& stress, reduce_thread_local_t = {}) const
+    {
+      local += stress;
+    }
+
+    ONIKA_HOST_DEVICE_FUNC inline void operator()(Mat3d &global, Mat3d& local, reduce_thread_block_t) const
+    {
+      exanb::mat3d_atomic_add_block_contribution(global, local);
+    }
+
+    ONIKA_HOST_DEVICE_FUNC inline void operator()(Mat3d &global, Mat3d& local, reduce_global_t) const
+    {
+      exanb::mat3d_atomic_add_contribution(global, local);
+    }
+  };
+
+}; // namespace exaDEM
+
+
+namespace exanb
+{
+  template <> struct ReduceCellParticlesTraits<exaDEM::ReduceStressTensorFunctor>
+  {
+    static inline constexpr bool RequiresBlockSynchronousCall = false;
+    static inline constexpr bool RequiresCellParticleIndex = false;
+    static inline constexpr bool CudaCompatible = true;
+  };
+}; // namespace exanb
+
+
+
+namespace exaDEM
+{
+  using namespace exanb;
+
+  template <typename GridT, class = AssertGridHasFields<GridT, field::_stress>> class AverageStressTensor : public OperatorNode
+  {
+    // attributes processed during computation
+    using ReduceFields = FieldSet<field::_stress>;
+    static constexpr ReduceFields reduce_field_set{};
+
+    ADD_SLOT(MPI_Comm, mpi, INPUT, MPI_COMM_WORLD);
+    ADD_SLOT(GridT, grid, INPUT_OUTPUT, REQUIRED);
+    ADD_SLOT(Traversal, traversal_real, INPUT, DocString{"list of non empty cells within the current grid"});
+    ADD_SLOT(double, volume, INPUT, REQUIRED, DocString{"Volume of the domain simulation. >0 "});
+    ADD_SLOT(Mat3d, stress_tensor, OUTPUT, DocString{"Write an Output file containing stress tensors."});
+
+    public:
[…]
