### 2024-12-25 651fd8f
add particle map type (#105)
Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
