+++ src/interaction/include/exaDEM/classifier/classifier.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include <exaDEM/interaction/interaction.hpp>
-#include <exaDEM/interaction/interactionAOS.hpp>
-#include <exaDEM/interaction/interactionSOA.hpp>
+#include <exaDEM/classifier/interactionAOS.hpp>
+#include <exaDEM/classifier/interactionSOA.hpp>
 #include <exaDEM/interaction/grid_cell_interaction.hpp>
 #include <exaDEM/itools/buffer.hpp>
 
