### 2024-12-26 e179b9e
reduce the dependencies of the plugins driver and shape
