+++ src/polyhedron/vertex_displ_over.cpp
@@ -33,7 +33,7 @@ under the License.
 
 #include <exanb/compute/reduce_cell_particles.h>
 #include <exanb/mpi/particle_displ_over_async_request.h>
-#include <exaDEM/shape/shapes.hpp>
+#include <exaDEM/shapes.hpp>
 #include <exaDEM/backup_dem.h>
 #include <exaDEM/traversal.hpp>
 
