+++ src/interaction/include/exaDEM/network.hpp
@@ -19,8 +19,8 @@ under the License.
 #pragma once
 
 #include <exaDEM/interaction/interaction.hpp>
-#include <exaDEM/shape/shapes.hpp>
-#include <exaDEM/shape/shape_detection.hpp>
+#include <exaDEM/shapes.hpp>
+#include <exaDEM/shape_detection.hpp>
 #include <exaDEM/contact_force_parameters.h>
 #include <exaDEM/compute_contact_force.h>
 #include <vector>
