+++ src/driver/include/exaDEM/driver_params.hpp
@@ -1,0 +1,248 @@
+/*
+	 Licensed to the Apache Software Foundation (ASF) under one
+	 or more contributor license agreements.  See the NOTICE file
+	 distributed with this work for additional information
+	 regarding copyright ownership.  The ASF licenses this file
+	 to you under the Apache License, Version 2.0 (the
+	 "License"); you may not use this file except in compliance
+	 with the License.  You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing,
+software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+KIND, either express or implied.  See the License for the
+specific language governing permissions and limitations
+under the License.
+ */
+#pragma once
+
+#include <exaDEM/normalize.hpp>
+#include <climits>
+#include <chrono>
+#include <thread>
+
+
+namespace exaDEM
+{
+	using namespace exanb;
+
+
+	enum MotionType
+	{
+		STATIONARY,                /**< Stationary state, with no motion. */
+		LINEAR_MOTION,             /**< Linear movement type, straight-line motion. */
+		COMPRESSIVE_FORCE,         /**< Movement influenced by compressive forces. */
+		LINEAR_FORCE_MOTION,       /**< Linear motion type influenced by applied forces. */
+		FORCE_MOTION,              /**< General movement caused by applied forces. */
+		LINEAR_COMPRESSIVE_MOTION, /**< Linear movement combined with compressive forces. */
+    UNKNOWN
+	};
+
+
+	inline std::string motion_type_to_string(MotionType motion_type)
+	{
+		switch(motion_type)
+		{
+			case STATIONARY: return "STATIONARY";
+			case LINEAR_MOTION: return "LINEAR_MOTION";
+			case COMPRESSIVE_FORCE: return "COMPRESSIVE_FORCE";
+			case LINEAR_FORCE_MOTION: return "LINEAR_FORCE_MOTION";
+			case FORCE_MOTION: return "FORCE_MOTION";
+			case LINEAR_COMPRESSIVE_MOTION: return "LINEAR_COMPRESSIVE_MOTION";
+			default: return "UNKNOWN";
+		}
+	}
+
+	inline MotionType string_to_motion_type(const std::string& str)
+	{
+		if (str == "STATIONARY") return STATIONARY;
+		if (str == "LINEAR_MOTION") return LINEAR_MOTION;
+		if (str == "COMPRESSIVE_FORCE") return COMPRESSIVE_FORCE;
+		if (str == "LINEAR_FORCE_MOTION") return LINEAR_FORCE_MOTION;
+		if (str == "FORCE_MOTION") return FORCE_MOTION;
+		if (str == "LINEAR_COMPRESSIVE_MOTION") return LINEAR_COMPRESSIVE_MOTION;
+
+		// If the string doesn't match any valid MotionType, return a default value
+		return UNKNOWN;  // Or some other default action like throwing an exception or logging
+	}
+
+	struct Driver_params
+	{
+		// motio, part
+		MotionType motion_type = STATIONARY;
+		Vec3d motion_vector = {0,0,0};
+		double motion_start_threshold = 0;
+		double motion_end_threshold = std::numeric_limits<double>::max();
+		double const_vel = 0;
+
+		inline bool is_stationary() { return motion_type == STATIONARY; }
+
+		void set_params(Driver_params& in)
+		{ 
+			(*this) = in;
+		}
+
+		inline bool is_linear()
+		{
+			return (motion_type == LINEAR_MOTION || motion_type == LINEAR_FORCE_MOTION || motion_type == LINEAR_COMPRESSIVE_MOTION);
+		}
+
[…]
