+++ src/driver/include/exaDEM/ball.h
@@ -1,13 +1,13 @@
 /*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
 
-  http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing,
 software distributed under the License is distributed on an
@@ -15,125 +15,151 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
-*/
+ */
 #pragma once
 #include <exanb/core/basic_types.h>
 #include <exaDEM/driver_base.h>
+#include <exaDEM/driver_params.hpp>
 
 namespace exaDEM
 {
   using namespace exanb;
 
-  /**
-   * @brief Struct representing a ball in the exaDEM simulation.
-   */
-  struct Ball
+  const std::vector<MotionType> ball_valid_motion_types = {STATIONARY, LINEAR_MOTION}; //, COMPRESSIVE_FORCE, FORCE_MOTION};
+
+/**
+ * @brief Struct representing a ball in the exaDEM simulation.
+ */
+struct Ball : public Driver_params
+{
+  double radius;       /**< Radius of the ball. */
+  exanb::Vec3d center; /**< Center position of the ball. */
+  exanb::Vec3d vel;    /**< Velocity of the ball. */
+  exanb::Vec3d vrot;   /**< Angular velocity of the ball. */
+  double c;            /**< compression value */
+
+  Ball(double r, exanb::Vec3d& c, exanb::Vec3d& v, exanb::Vec3d& vr) : Driver_params(), radius(r), center(c), vel(v), vrot(vr) {}  
+
+  void set_params(Driver_params& in) 
   {
-    double radius;       /**< Radius of the ball. */
-    exanb::Vec3d center; /**< Center position of the ball. */
-    exanb::Vec3d vel;    /**< Velocity of the ball. */
-    exanb::Vec3d vrot;   /**< Angular velocity of the ball. */
+    Driver_params::set_params(in);
+  }
 
-    /**
-     * @brief Get the type of the driver (in this case, BALL).
-     * @return The type of the driver.
-     */
-    constexpr DRIVER_TYPE get_type() { return DRIVER_TYPE::BALL; }
+  /**
+   * @brief Get the type of the driver (in this case, BALL).
+   * @return The type of the driver.
+   */
+  constexpr DRIVER_TYPE get_type() { return DRIVER_TYPE::BALL; }
 
-    /**
-     * @brief Print information about the ball.
-     */
-    void print()
+  /**
+   * @brief Print information about the ball.
+   */
+  void print()
+  {
+    lout << "Driver Type: Ball" << std::endl;
+    lout << "Radius: " << radius << std::endl;
+    lout << "Center: " << center << std::endl;
+    lout << "Vel   : " << vel << std::endl;
+    lout << "AngVel: " << vrot << std::endl;
+    if ( is_compressive() ) lout << "C: "<< c << std::endl;
[…]

[…]
