### 2024-12-29 a7ae8fc
add compressive motion mode for the driver type ball
