### 2024-12-29 c1504e7
reset force and update force to accel for drivers
