+++ src/driver/include/exaDEM/driver_params.hpp
@@ -1,11 +1,11 @@
 /*
-	 Licensed to the Apache Software Foundation (ASF) under one
-	 or more contributor license agreements.  See the NOTICE file
-	 distributed with this work for additional information
-	 regarding copyright ownership.  The ASF licenses this file
-	 to you under the Apache License, Version 2.0 (the
-	 "License"); you may not use this file except in compliance
-	 with the License.  You may obtain a copy of the License at
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
 
 http://www.apache.org/licenses/LICENSE-2.0
 
@@ -26,84 +26,84 @@ under the License.
 
 namespace exaDEM
 {
-	using namespace exanb;
+  using namespace exanb;
 
 
-	enum MotionType
-	{
-		STATIONARY,                /**< Stationary state, with no motion. */
-		LINEAR_MOTION,             /**< Linear movement type, straight-line motion. */
-		COMPRESSIVE_FORCE,         /**< Movement influenced by compressive forces. */
-		LINEAR_FORCE_MOTION,       /**< Linear motion type influenced by applied forces. */
-		FORCE_MOTION,              /**< General movement caused by applied forces. */
-		LINEAR_COMPRESSIVE_MOTION, /**< Linear movement combined with compressive forces. */
+  enum MotionType
+  {
+    STATIONARY,                /**< Stationary state, with no motion. */
+    LINEAR_MOTION,             /**< Linear movement type, straight-line motion. */
+    COMPRESSIVE_FORCE,         /**< Movement influenced by compressive forces. */
+    LINEAR_FORCE_MOTION,       /**< Linear motion type influenced by applied forces. */
+    FORCE_MOTION,              /**< General movement caused by applied forces. */
+    LINEAR_COMPRESSIVE_MOTION, /**< Linear movement combined with compressive forces. */
     UNKNOWN
-	};
+  };
 
 
-	inline std::string motion_type_to_string(MotionType motion_type)
-	{
-		switch(motion_type)
-		{
-			case STATIONARY: return "STATIONARY";
-			case LINEAR_MOTION: return "LINEAR_MOTION";
-			case COMPRESSIVE_FORCE: return "COMPRESSIVE_FORCE";
-			case LINEAR_FORCE_MOTION: return "LINEAR_FORCE_MOTION";
-			case FORCE_MOTION: return "FORCE_MOTION";
-			case LINEAR_COMPRESSIVE_MOTION: return "LINEAR_COMPRESSIVE_MOTION";
-			default: return "UNKNOWN";
-		}
-	}
+  inline std::string motion_type_to_string(MotionType motion_type)
+  {
+    switch(motion_type)
+    {
+      case STATIONARY: return "STATIONARY";
+      case LINEAR_MOTION: return "LINEAR_MOTION";
+      case COMPRESSIVE_FORCE: return "COMPRESSIVE_FORCE";
+      case LINEAR_FORCE_MOTION: return "LINEAR_FORCE_MOTION";
+      case FORCE_MOTION: return "FORCE_MOTION";
+      case LINEAR_COMPRESSIVE_MOTION: return "LINEAR_COMPRESSIVE_MOTION";
+      default: return "UNKNOWN";
+    }
+  }
 
-	inline MotionType string_to_motion_type(const std::string& str)
-	{
-		if (str == "STATIONARY") return STATIONARY;
-		if (str == "LINEAR_MOTION") return LINEAR_MOTION;
[…]
