+++ src/driver/force_to_accel_driver.cpp
@@ -1,13 +1,13 @@
 /*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
 
-  http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing,
 software distributed under the License is distributed on an
@@ -15,11 +15,12 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
-*/
+ */
 #include <exanb/core/operator.h>
 #include <exanb/core/operator_slot.h>
 #include <exanb/core/operator_factory.h>
 #include <exaDEM/drivers.h>
+#include <mpi.h>
 
 namespace exaDEM
 {
@@ -38,15 +39,31 @@ namespace exaDEM
     void operator()(auto&& arg) { arg.force_to_accel(); }
   };
 
+  struct tmp_reduce
+  {
+    std::tuple<bool, Vec3d> operator()(Ball& arg)
+    {
+      if( arg.is_compressive() || arg.is_force_motion() )
+      {
+        return {true, arg.forces};
+      }
+      else 
+      { 
+        return {false, {0,0,0}};
+      }
+    }
+    std::tuple<bool, Vec3d> operator()(auto && arg) { return {false, {0,0,0}};}
+  };
+
 
   class ForceToAccelDriverFunctor : public OperatorNode
   {
+    ADD_SLOT(MPI_Comm, mpi, INPUT, MPI_COMM_WORLD, DocString{"MPI communicator for parallel processing."});
     ADD_SLOT(Drivers, drivers, INPUT_OUTPUT, REQUIRED, DocString{"List of Drivers"});
     ADD_SLOT(double, dt, INPUT, DocString{"dt is the time increment of the timeloop"});
-    ADD_SLOT(double, system_mass, INPUT, REQUIRED);
+    ADD_SLOT(double, system_mass, INPUT, REQUIRED); 
 
-
-  public:
+    public:
     inline std::string documentation() const override final
     {
       return R"EOF(
@@ -56,6 +73,35 @@ namespace exaDEM
 
     inline void execute() override final
     {
+      // we need to update forces if required
+      std::vector <int> ids; // id, forces
+      std::vector <double> pack; // id, forces
+      tmp_reduce reduce;
+      for (size_t id = 0; id < drivers->get_size(); id++)
+      {
+        auto &driver = drivers->data(id);
+        auto [update , forces] = std::visit(reduce, driver);
+        if( update )
+        {
+          ids.push_back(id);
+          pack.push_back(forces.x);
+          pack.push_back(forces.y);
+          pack.push_back(forces.z);
+        }
+      }
+
+      if(pack.size() > 0)
+      {
+        std::vector <double> unpack(pack.size());
+        MPI_Allreduce(pack.data(), unpack.data(), pack.size(), MPI_DOUBLE, MPI_SUM, *mpi);
[…]
