+++ src/driver/include/exaDEM/ball.h
@@ -56,25 +56,6 @@ namespace YAML
   using exanb::Quantity;
   using exanb::UnityConverterHelper;
 
-  inline bool check_error(const Node &node, std::string item)
-  {
-    if (!node[item])
-    {
-      exanb::lout << "\033[31m" << item << " is missing\033[0m\n";
-      return false;
-    }
-    return true;
-  }
-
-  inline bool check(const Node &node, std::string item)
-  {
-    if (!node[item])
-    {
-      return false;
-    }
-    return true;
-  }
-
   template <> struct convert<Ball_params>
   {
     static bool decode(const Node &node, Ball_params &v)
@@ -148,7 +129,7 @@ namespace exaDEM
      */
     void dump_driver(int id, std::stringstream &stream)
     {
-      stream << "  - add_ball:" << std::endl;
+      stream << "  - register_ball:" << std::endl;
       stream << "     id: " << id << std::endl;
       stream << "     state: { radius:" << this->radius;
       stream << ",center: [" << this->center << "]";
