+++ src/driver/include/exaDEM/cylinder.h
@@ -22,126 +22,180 @@ under the License.
 
 namespace exaDEM
 {
-  using namespace exanb;
+	using namespace exanb;
 
-  /**
-   * @brief Struct representing a cylinder in the exaDEM simulation.
-   */
-  struct Cylinder
+	using namespace exanb;
+	struct Cylinder_params
+	{
+		double radius = -1;       /**< Radius of the cylinder. */
+		exanb::Vec3d axis = {1,0,1};   /**< Axis direction of the cylinder. */
+		exanb::Vec3d center = {0,0,0}; /**< Center position of the cylinder. */
+		exanb::Vec3d vel = {0,0,0};    /**< Velocity of the cylinder. */
+		exanb::Vec3d vrot = {0,0,0};   /**< Angular velocity of the cylinder. */
+	};
+}
+
+namespace YAML
+{
+  using exaDEM::Cylinder_params;
+  using exaDEM::MotionType;
+  using exanb::lerr;
+  using exanb::Quantity;
+  using exanb::UnityConverterHelper;
+
+  template <> struct convert<Cylinder_params>
   {
-    double radius;       /**< Radius of the cylinder. */
-    exanb::Vec3d axis;   /**< Axis direction of the cylinder. */
-    exanb::Vec3d center; /**< Center position of the cylinder. */
-    exanb::Vec3d vel;    /**< Velocity of the cylinder. */
-    exanb::Vec3d vrot;   /**< Angular velocity of the cylinder. */
-    exanb::Vec3d forces;
-    
-    /**
-     * @brief Get the type of the driver (in this case, CYLINDER).
-     * @return The type of the driver.
-     */
-    constexpr DRIVER_TYPE get_type() { return DRIVER_TYPE::CYLINDER; }
-
-    /**
-     * @brief Print information about the cylinder.
-     */
-    void print()
+    static bool decode(const Node &node, Cylinder_params &v)
     {
-      lout << "Driver Type: Cylinder" << std::endl;
-      lout << "Radius: " << radius << std::endl;
-      lout << "Axis  : " << axis << std::endl;
-      lout << "Center: " << center << std::endl;
-      lout << "Vel   : " << vel << std::endl;
-      lout << "AngVel: " << vrot << std::endl;
-    }
-
-    /**
-     * @brief Write cylinder information into a stream.
-     */
-    void dump_driver(int id, std::stringstream &stream)
-    {
-      stream << "  - add_cylinder:" << std::endl;
-      stream << "     id: " << id << std::endl;
-      stream << "     axis: [" << this->axis << "]" << std::endl;
-      stream << "     radius: " << this->radius << std::endl;
-      stream << "     center: [" << this->center << "]" << std::endl;
-      stream << "     velocity: [" << this->vel << "]" << std::endl;
-      stream << "     angular_velocity: [" << this->vrot << "]" << std::endl;
-    }
-
-    /**
-     * @brief return driver velocity
-     */
-    ONIKA_HOST_DEVICE_FUNC inline Vec3d &get_vel() { return vel; }
-
-    ONIKA_HOST_DEVICE_FUNC inline void force_to_accel() {}
-  	ONIKA_HOST_DEVICE_FUNC inline void push_f_v(const double dt) {}
-  	ONIKA_HOST_DEVICE_FUNC inline void push_f_v_r(const double dt) { center = center + dt * vel; }
-
-    /**
-     * @brief Filter function to check if a point is within a certain radius of the cylinder.
-     * @param rcut The cut-off radius. Note: rcut = rverlet + r shape
-     * @param vi The vector representing the point to check.
-     * @return True if the point is within the cut-off radius of the cylinder, false otherwise.
-     */
-    ONIKA_HOST_DEVICE_FUNC inline bool filter(const double rcut, const Vec3d &vi)
-    {
[…]
