+++ src/driver/include/exaDEM/ball.h
@@ -164,32 +164,6 @@ namespace exaDEM
     }
 
     /**
-     * @brief return driver velocity
-     */
-    ONIKA_HOST_DEVICE_FUNC inline Vec3d &get_vel() { return vel; }
-
-    /**
-     * @brief Update the position of the ball.
-     * @param dt The time step.
-     */
-    ONIKA_HOST_DEVICE_FUNC inline void push_f_v_r(const double dt) 
-    {
-      if( motion_type == LINEAR_MOTION )
-      {
-        vel = motion_vector * const_vel; // I prefere reset it 
-      }
-      if( is_compressive() )
-      {
-        if( motion_type == COMPRESSIVE_FORCE )
-        {
-          vel = {0,0,0};
-        }      
-        push_ra_rv_to_rad(dt);
-      }
-      center = center + dt * vel; 
-    }
-
-    /**
      * @brief Update the position of the ball.
      * @param dt The time step.
      */
@@ -207,122 +181,151 @@ namespace exaDEM
     }
 
     /**
+     * @brief return driver velocity
+     */
+    ONIKA_HOST_DEVICE_FUNC inline Vec3d &get_vel() { return vel; }
+
+    /**
+     * @brief Update the position of the ball.
+     * @param dt The time step.
+     */
+    ONIKA_HOST_DEVICE_FUNC inline void push_f_v_r(const double dt) 
+    {
+      if( !is_stationary() )
+      {
+        if( is_compressive() )
+        {
+          push_ra_rv_to_rad(dt);
+        }
+        center = center + dt * vel; 
+      }
+    }
+
+    /**
      * @brief Update the position of the ball.
      * @param dt The time step.
      */
     ONIKA_HOST_DEVICE_FUNC inline void push_f_v(const double dt)
     {
-      if( is_force_motion() )
-      {
-        vel = acc * dt;
-      }
+			if( is_force_motion() )
+			{
+				vel = acc * dt;
+			}
 
-      // update ra
-      if( is_compressive() )
-      {
-        push_ra_to_rv(dt);
-      }
-    }
-    /**
-     * @brief Update the "velocity raduis" of the ball.
-     * @param t The time step.
-     */
-    ONIKA_HOST_DEVICE_FUNC inline void push_ra_to_rv(const double dt) 
-    {
-      if( is_compressive() )
-      {
-        if( sigma != 0 ) rv += 0.5 * dt * ra;
-      }
-    }
+			if( motion_type == LINEAR_MOTION )
+			{
+				vel = motion_vector * const_vel; // I prefere reset it 
+			}
 
-    /**
-     * @brief Update the "velocity raduis" of the ball.
-     * @param t The time step.
-     */
-    ONIKA_HOST_DEVICE_FUNC inline void push_ra_rv_to_rad(const double dt) 
-    {
-      if( is_compressive() )
-      {
-        radius += dt * rv + 0.5 * dt * dt * ra;
-      }
-    }
+			if( is_compressive() )
+			{
+				if( motion_type == COMPRESSIVE_FORCE )
+				{
+					vel = {0,0,0};
+				}
+				push_ra_to_rv(dt);
+			}
+		}
+		/**
+		 * @brief Update the "velocity raduis" of the ball.
+		 * @param t The time step.
+		 */
+		ONIKA_HOST_DEVICE_FUNC inline void push_ra_to_rv(const double dt) 
+		{
+			if( is_compressive() )
+			{
+				if( sigma != 0 ) rv += 0.5 * dt * ra;
+			}
+		}
 
-    /**
-     * @brief Compute the surface.
-     */
-    ONIKA_HOST_DEVICE_FUNC inline 
-      double surface()
-      {
[…]
