+++ src/driver/register_surface.cpp
@@ -21,39 +21,40 @@ under the License.
 #include <exanb/core/operator_factory.h>
 #include <exaDEM/driver_base.h>
 #include <exaDEM/drivers.h>
-#include <exaDEM/cylinder.h>
+#include <exaDEM/surface.h>
 
 namespace exaDEM
 {
 
   using namespace exanb;
 
-  class RegisterCylinder : public OperatorNode
+  class RegisterSurface : public OperatorNode
   {
+
     static constexpr Driver_params default_params = Driver_params();
 
     ADD_SLOT(Drivers, drivers, INPUT_OUTPUT, REQUIRED, DocString{"List of Drivers"});
-    ADD_SLOT(int, id, INPUT, REQUIRED, DocString{"Driver index"});
-    ADD_SLOT(Cylinder_params, state, INPUT, REQUIRED, DocString{"Current cylinder state, default is {radius: REQUIRED, axis: REQUIRED, center: REQUIRED, vel: [0,0,0], vrot: [0,0,0], rv: 0, ra: 0}. You need to specify the radius and center"});
+    ADD_SLOT(size_t, id, INPUT, REQUIRED, DocString{"Driver index"});
+    ADD_SLOT(Surface_params, state, INPUT, REQUIRED, DocString{"Current Cylinder state, default is {offset: REQUIRED, normal: REQUIRED, vel: 0, vrot: [0,0,0], suface: -1, center: [0,0,0]}. You need to specify the offset, and the normal vector. You need to specify the center if vrot != [0,0,0]"});
     ADD_SLOT(Driver_params, params, INPUT, default_params, DocString{"List of params, motion type, motion vectors .... Default is { motion_type: STATIONARY}."});
+
 
     public:
     inline std::string documentation() const override final
     {
       return R"EOF(
-        This operator add a cylinder to the drivers list.
+        This operator add a surface to the drivers list.
         )EOF";
     }
 
     inline void execute() override final
     {
-      // proj center over axis
-      exaDEM::Cylinder driver{*state, *params};
-      driver.initialize();
+      exaDEM::Surface driver = {*state, *params}; //
+      driver.initialize(); // initialize some values from input parameters such as the projected center of the surface (normal line)
       drivers->add_driver(*id, driver);
     }
   };
 
   // === register factories ===
-  CONSTRUCTOR_FUNCTION { OperatorNodeFactory::instance()->register_factory("register_cylinder", make_simple_operator<RegisterCylinder>); }
+  CONSTRUCTOR_FUNCTION { OperatorNodeFactory::instance()->register_factory("register_surface", make_simple_operator<RegisterSurface>); }
 } // namespace exaDEM
