+++ src/driver/include/exaDEM/paraview_driver.hpp
@@ -1,13 +1,13 @@
 /*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
 
-  http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing,
 software distributed under the License is distributed on an
@@ -15,7 +15,7 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
-*/
+ */
 #pragma once
 #include <exanb/core/basic_types.h>
 #include <filesystem>
@@ -26,11 +26,177 @@ namespace exaDEM
   using namespace exanb;
 
   struct info_ball { int id; Vec3d center; double radius; Vec3d vel; };
+  struct info_surface { int id; Vec3d normal; double offset; double vel; };
 
-  void write_balls_paraview(std::vector<info_ball> balls, std::filesystem::path path, std::string filename)
+  std::tuple<bool, bool, Vec3d> intersect(Vec3d& n, double offset, Vec3d& corner1, Vec3d& corner2)
   {
-    std::string full_path = path.string() + "/" + filename;
-    if( !std::filesystem::exists(full_path))
+    Vec3d c1 = corner1 - offset * n;
+    Vec3d c2 = corner2 - offset * n;
+    Vec3d dist = c2 - c1;
+    double norm_proj_dist = exanb::dot(n, dist);
+    if ( norm_proj_dist == 0) 
+    {
+      return {false, false, Vec3d{0,0,0}};
+      //return {false, true, Vec3d{0,0,0}};
+    }
+    double t = -(exanb::dot(n, c1)) / norm_proj_dist;
+    if (t < 0 || t > 1) 
+    {
+      return {false, false,Vec3d{0,0,0}};
+    }
+    Vec3d point = c1 + t * dist + offset * n;
+    return {true, false, point};
+  }
+
+  std::vector<Vec3d> compute_intersections(Domain& domain, Vec3d& normal, double offset)
+  {
+    std::vector<Vec3d> points;
+    auto [inf, sup] = domain.bounds();
+    std::array<Vec3d, 8> corners = {
+      Vec3d{inf.x, inf.y, inf.z}, // Corner 0
+      Vec3d{sup.x, inf.y, inf.z}, // Corner 1
+      Vec3d{sup.x, sup.y, inf.z}, // Corner 2
+      Vec3d{inf.x, sup.y, inf.z}, // Corner 3
+      Vec3d{inf.x, inf.y, sup.z}, // Corner 4
+      Vec3d{sup.x, inf.y, sup.z}, // Corner 5
+      Vec3d{sup.x, sup.y, sup.z},  // Corner 6
+      Vec3d{inf.x, sup.y, sup.z} // Corner 7
+    };
+
+    std::vector<std::pair<int, int>> edges = {
[…]
