### 2024-12-31 73b3e9a
update stl mesh
