+++ src/driver/include/exaDEM/stl_mesh.h
@@ -18,6 +18,8 @@ under the License.
  */
 #pragma once
 #include <exanb/core/basic_types.h>
+#include <exanb/core/quaternion.h>
+#include <exanb/core/quaternion_yaml.h>
 #include <exaDEM/driver_base.h>
 #include <exaDEM/shape.hpp>
 #include <exaDEM/shape_reader.hpp>
@@ -42,20 +44,60 @@ namespace exaDEM
     void clean() { vertices.clear(); edges.clear(); faces.clear(); }
   };
 
+  struct Stl_params
+  {
+    exanb::Vec3d center = Vec3d{0,0,0}; /**< Center position of the STL mesh. */
+    exanb::Vec3d vel = Vec3d{0,0,0};    /**< Velocity of the STL mesh. */
+    exanb::Vec3d vrot = Vec3d{0,0,0};   /**< Angular velocity of the STL mesh. */
+    exanb::Quaternion quat = {1,0,0,0};             /**< Quaternion of the STL mesh. */
+    double mass = std::numeric_limits<double>::max() / 4; /**< Mass of the ball */
+  };
+}
+
+namespace YAML
+{
+  using exaDEM::Stl_params;
+  using exaDEM::MotionType;
+  using exanb::lerr;
+  using exanb::Quantity;
+  using exanb::UnityConverterHelper;
+
+  template <> struct convert<Stl_params>
+  {
+    static bool decode(const Node &node, Stl_params &v)
+    {
+      if (!node.IsMap())
+      {
+        return false;
+      }
+      if( check(node, "center") ) { v.vel = node["center"].as<Vec3d>(); }
+      if( check(node, "vel") )    { v.vel = node["vel"].as<Vec3d>(); }
+      if( check(node, "vrot") )   { v.vrot = node["vrot"].as<Vec3d>(); }
+      if( check(node, "mass") ) { v.mass = node["mass"].as<double>(); }
+      if( check(node, "quat") ) { v.quat = node["quat"].as<exanb::Quaternion>(); }
+      return true;
+    }
+  };
+}
+
+
+
+namespace exaDEM
+{
+  const std::vector<MotionType> stl_valid_motion_types = {STATIONARY, LINEAR_MOTION};
+
+  using namespace exanb;
   /**
    * @brief Struct representing a STL mesh in the exaDEM simulation.
    */
-  struct Stl_mesh
+  struct Stl_mesh : public Stl_params, Driver_params
   {
-    exanb::Vec3d center;    /**< Center position of the STL mesh. */
-    exanb::Vec3d vel;       /**< Velocity of the STL mesh. */
-    exanb::Vec3d vrot;      /**< Angular velocity of the STL mesh. */
-    exanb::Quaternion quat; /**< Quaternion of the STL mesh. */
     shape shp;              /**< Shape of the STL mesh. */
     vector_t<Vec3d> vertices;      /**< Collection of vertices (computed from shp, quat and center). */
     std::vector<list_of_elements> grid_indexes; /**< Grid indices of the STL mesh. */
     std::vector<omp_lock_t> grid_mutexes; /**< Grid indices of the STL mesh. */
-    exanb::Vec3d forces;
+    /** We don't need to save these values */
+    exanb::Vec3d acc = {0,0,0};       /**< Acceleration of the mesh */
 
     /**
      * @brief Get the type of the driver (in this case, STL_MESH).
@@ -63,7 +105,10 @@ namespace exaDEM
      */
     constexpr DRIVER_TYPE get_type() { return DRIVER_TYPE::STL_MESH; }
 
-    /**
+
+    void set_shape(shape& s) {shp = s;}
+
+   /**
      * @brief Print information about the STL mesh.
      */
     void print()
@@ -104,96 +149,137 @@ namespace exaDEM
       // remove relative paths
[…]
