### 2024-12-31 7b7dd85
update examples with register stl_mesh
