### 2024-12-31 2015327
add motion types, paraview output files for balls and apply motion ty… (#113)
* add motion types, paraview output files for balls and apply motion types to ball drivers

* add compressive motion mode for the driver type ball

* reset force and update force to accel for drivers

* fix balls test case

* update cylinder with Cylinder params and Driver params

* update examples

* add surface, it looks ok for linear_motion

* add surface / paraview

* update stl mesh

* update examples with register stl_mesh

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
