+++ src/driver/include/exaDEM/face.h
@@ -19,6 +19,7 @@ under the License.
 #pragma once
 
 #include <exanb/core/basic_types.h>
+#include <exaDEM/normalize.hpp>
 
 namespace exaDEM
 {
@@ -42,25 +43,6 @@ namespace exaDEM
     };
   };
 
-  /**
-   * @brief Calculate the length of a 3D vector.
-   * @param v The input vector.
-   * @return The length of the vector.
-   */
-  inline double length(Vec3d &v) { return std::sqrt(v.x * v.x + v.y * v.y + v.z * v.z); }
-
-  /**
-   * @brief Calculate the length of a const 3D vector.
-   * @param v The input vector.
-   * @return The length of the vector.
-   */
-  inline double length(const Vec3d &v) { return std::sqrt(v.x * v.x + v.y * v.y + v.z * v.z); }
-
-  /**
-   * @brief Normalize a 3D vector.
-   * @param v The input vector to be normalized.
-   */
-  inline void _normalize(Vec3d &v) { v = v / exanb::norm(v); }
 
   /**
    * @brief Struct representing a 3D face.
