+++ src/driver/push_f_v_driver.cpp
@@ -19,41 +19,60 @@ under the License.
 #include <exanb/core/operator.h>
 #include <exanb/core/operator_slot.h>
 #include <exanb/core/operator_factory.h>
-#include <exaDEM/driver_base.h>
 #include <exaDEM/drivers.h>
-#include <exaDEM/surface.h>
 
 namespace exaDEM
 {
-
   using namespace exanb;
 
-  class AddSurface : public OperatorNode
+  struct push_f_v_r
+  {
+    const double dt;
+    void operator()(Ball& arg)
+    {
+      arg.push_f_v(dt);
+    }
+
+    void operator()(Surface& arg)
+    {
+      arg.push_f_v(dt);
+    }
+
+    void operator()(Stl_mesh& arg)
+    {
+      arg.push_f_v(dt);
+    }
+    void operator()(auto&& arg)
+    {
+      /** nothing */
+    }
+  };
+
+  class PushAccelToVelocityDriver : public OperatorNode
   {
     ADD_SLOT(Drivers, drivers, INPUT_OUTPUT, REQUIRED, DocString{"List of Drivers"});
-    ADD_SLOT(size_t, id, INPUT, REQUIRED, DocString{"Driver index"});
-    ADD_SLOT(double, offset, INPUT, 0.0, DocString{"Offset from the origin (0,0,0) of the rigid surface"});
-    ADD_SLOT(double, velocity, INPUT, 0.0, DocString{"Surface velocity"});
-    ADD_SLOT(Vec3d, center, INPUT, Vec3d{0.0, 0.0, 0.0}, DocString{"Normal vector of the rigid surface"});
-    ADD_SLOT(Vec3d, normal, INPUT, Vec3d{0.0, 0.0, 1.0}, DocString{"Normal vector of the rigid surface"});
-    ADD_SLOT(Vec3d, vrot, INPUT, Vec3d{0.0, 0.0, 0.0}, DocString{"Angular velocity of the surface, default is 0 m.s-"});
+    ADD_SLOT(double, dt, INPUT, DocString{"dt is the time increment of the timeloop"});
 
   public:
     inline std::string documentation() const override final
     {
       return R"EOF(
-        This operator add a surface to the drivers list.
+          This operator updates driver centers using their velocities. Not that accelerations are not used.
         )EOF";
     }
 
     inline void execute() override final
     {
-      exaDEM::Surface driver = {*offset, *normal, *center, *velocity, *vrot}; //
-      driver.initialize();                                                    // initialize some values from input parameters such as the projected center of the surface (normal line)
-      drivers->add_driver(*id, driver);
+      double t = *dt;
+      push_f_v_r func = {t};
+      for (size_t id = 0; id < drivers->get_size(); id++)
+      {
+        auto &driver = drivers->data(id);
+        std::visit(func, driver);
+      }
     }
   };
 
   // === register factories ===
-  CONSTRUCTOR_FUNCTION { OperatorNodeFactory::instance()->register_factory("add_surface", make_simple_operator<AddSurface>); }
+  CONSTRUCTOR_FUNCTION { OperatorNodeFactory::instance()->register_factory("push_f_v_driver", make_simple_operator<PushAccelToVelocityDriver>); }
 } // namespace exaDEM
