### 2025-01-01 d913a11
add binary mode for stl mesh reader
