### 2025-01-01 4ab6778
add a description for the binary mode of the stl mesh reader
