### 2025-01-01 407f3ca
update binary_mode to binary [SLOT / register_stl_mesh]
