### 2025-01-01 a13ef31
106 read binary stl files (#114)
* add binary mode for stl mesh reader

* add a description for the binary mode of the stl mesh reader

* update binary_mode to binary [SLOT / register_stl_mesh]

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
