### 2025-01-03 921eebc
Update register_stl_mesh.cpp
