### 2025-01-04 acac1b6
add restart_template.py
