### 2025-01-05 e503f28
update restart template script
