### 2025-01-06 100fa4f
Resolution compilation problems on the GPU
