### 2025-01-06 7e7be65
fix gpu compilation
