### 2025-01-07 bf03ce8
fix performance issue on GPU when updating the driver forces
