### 2025-01-07 35e36c4
modify reduce stl mesh (avoid useless atomicAdd)
