### 2025-01-08 f4b2981
remove useless print into the stl reader
