+++ src/interaction/unclassify_interactions.cpp
@@ -91,6 +91,8 @@ namespace exaDEM
   {
   	int idx = threadIdx.x + blockIdx.x * blockDim.x;
   	
+  	//printf("RAZ\n");
+  	
   	int incr = blocks_incr[blockIdx.x];
   	
   	__shared__ int s[256];
@@ -242,13 +244,13 @@ namespace exaDEM
       
       //getchar();
       
-      if (grid->number_of_cells() == 0)
+      /*if (grid->number_of_cells() == 0)
       {
         return;
       }
       if (!ic.has_value())
         return;
-      ic->unclassify(*ges);
+      ic->unclassify(*ges);*/
       
       //printf("UNCLASSIFY END\n");
       
