### 2025-01-14 073099d
add traversal all for compute_vertices
