### 2025-01-16 b09bf09
add minimal set_fields operator
