+++ src/field/set_fields.cpp
@@ -1,13 +1,13 @@
 /*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
+	 Licensed to the Apache Software Foundation (ASF) under one
+	 or more contributor license agreements.  See the NOTICE file
+	 distributed with this work for additional information
+	 regarding copyright ownership.  The ASF licenses this file
+	 to you under the Apache License, Version 2.0 (the
+	 "License"); you may not use this file except in compliance
+	 with the License.  You may obtain a copy of the License at
 
-  http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing,
 software distributed under the License is distributed on an
@@ -15,7 +15,7 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
-*/
+ */
 //#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
 
 #include <exanb/core/operator.h>
@@ -24,227 +24,170 @@ under the License.
 #include <exanb/core/make_grid_variant_operator.h>
 #include <exanb/core/parallel_grid_algorithm.h>
 #include <exanb/core/grid.h>
+#include <exanb/core/particle_type_id.h>
 #include <memory>
+#include <exaDEM/shapes.hpp>
 #include <exaDEM/set_fields.h>
 
 namespace exaDEM
 {
-  using namespace exanb;
 
-  template <typename GridT, class = AssertGridHasFields<GridT, field::_type>> class SetFields : public OperatorNode
-  {
-    using ComputeFieldsType = FieldSet<field::_type>;
-    using ComputeRegionFieldsType = FieldSet<field::_rx, field::_ry, field::_rz, field::_id, field::_type>;
-    static constexpr ComputeFieldsType compute_field_set_type{};
-    static constexpr ComputeRegionFieldsType compute_region_field_set_type{};
-    ADD_SLOT(GridT, grid, INPUT_OUTPUT);
-    // vector version
-    ADD_SLOT(std::vector<uint32_t>, types, INPUT, OPTIONAL, DocString{""});
-    ADD_SLOT(std::vector<double>, densities, INPUT, OPTIONAL, DocString{""});
-    ADD_SLOT(std::vector<double>, radii, INPUT, OPTIONAL, DocString{""});
-    ADD_SLOT(std::vector<Vec3d>, velocities, INPUT, OPTIONAL, DocString{""});
-    ADD_SLOT(std::vector<Vec3d>, random_velocities, INPUT, OPTIONAL, DocString{""});
-    ADD_SLOT(std::vector<Vec3d>, angular_velocities, INPUT, OPTIONAL, DocString{""});
-    ADD_SLOT(std::vector<Vec3d>, random_angular_velocities, INPUT, OPTIONAL, DocString{""});
-    ADD_SLOT(std::vector<Quaternion>, quaternions, INPUT, OPTIONAL, DocString{""});
-    ADD_SLOT(std::vector<bool>, random_quaternions, INPUT, OPTIONAL, DocString{""});
-    // scalar version
-    ADD_SLOT(uint32_t, type, INPUT, OPTIONAL, DocString{""});
-    ADD_SLOT(double, density, INPUT, OPTIONAL, DocString{""});
[…]
