### 2025-01-19 c04c0d1
minor change, update rsa mpi spack repo
