### 2025-01-20 6d75792
add n_max particles criterion for rsa algo
