### 2025-01-20 fee9fcb
116 rsa algo random radius (#117)
* add operator rsa_rnd_rad

* update rsa rnd rad operator, it works

* add n_max particles criterion for rsa algo

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
