+++ src/rsa/rsa_rnd_rad.cpp
@@ -1,0 +1,207 @@
+#include <chrono>
+#include <ctime>
+#include <mpi.h>
+#include <string>
+#include <numeric>
+
+#include <exanb/core/basic_types_yaml.h>
+#include <exanb/core/operator.h>
+#include <exanb/core/operator_slot.h>
+#include <exanb/core/operator_factory.h>
+#include <exanb/core/make_grid_variant_operator.h>
+#include <exanb/core/grid.h>
+#include <exanb/core/basic_types_stream.h>
+#include <exanb/core/log.h>
+//#include "ustamp/vector_utils.h"
+#include <exanb/core/file_utils.h>
+#include <exanb/core/domain.h>
+#include <exanb/core/check_particles_inside_cell.h>
+#include <exanb/core/simple_block_rcb.h>
+#include <exanb/grid_cell_particles/particle_region.h>
+
+
+// rsa_mpi stuff
+#include <rsa_data_storage.hxx>
+#include <rsa_random.hxx>
+#include <rsa_domain.hxx>
+#include <rsa_decoration.hxx>
+#include <operator_algorithm.hxx>
+#include <radius_generator.hxx>
+
+namespace exaDEM
+{
+	using namespace exanb;
+
+	struct ParticleType
+	{
+		static inline constexpr size_t MAX_STR_LEN = 16;
+
+		double m_mass = 1.0;
+		double m_radius = 1.0;
+		char m_name[MAX_STR_LEN] = {'\0'};
+
+		inline void set_name(const std::string &s)
+		{
+			if (s.length() >= MAX_STR_LEN)
+			{
+				std::cerr << "Particle name too long : length=" << s.length() << ", max=" << (MAX_STR_LEN - 1) << "\n";
+				std::abort();
+			}
+			std::strncpy(m_name, s.c_str(), MAX_STR_LEN);
+			m_name[MAX_STR_LEN - 1] = '\0';
+		}
+		inline std::string name() const { return m_name; }
+	};
+
+	using ParticleTypes = onika::memory::CudaMMVector<ParticleType>;
+
+	template <typename GridT> class RSARndRad : public OperatorNode
+	{
+		ADD_SLOT(MPI_Comm, mpi, INPUT, MPI_COMM_WORLD);
+		ADD_SLOT(Domain, domain, INPUT_OUTPUT);
+		ADD_SLOT(GridT, grid, INPUT_OUTPUT);
+		ADD_SLOT(double, enlarge_bounds, INPUT, 0.0);
+		ADD_SLOT(ReadBoundsSelectionMode, bounds_mode, INPUT, ReadBoundsSelectionMode::FILE_BOUNDS);
+		ADD_SLOT(std::vector<bool>, periodicity, INPUT, OPTIONAL, DocString{"if set, overrides domain's periodicity stored in file with this value"});
+		ADD_SLOT(bool, expandable, INPUT, OPTIONAL, DocString{"if set, override domain expandability stored in file"});
+		ADD_SLOT(AABB, bounds, INPUT, REQUIRED, DocString{"if set, override domain's bounds, filtering out particle outside of overriden bounds"});
+		ADD_SLOT(int, type, INPUT, 0);
+		ADD_SLOT(bool, pbc_adjust_xform, INPUT, true);
+
+		ADD_SLOT(double, r_min, INPUT, REQUIRED, DocString{"Value of the smallest radius possible"}); 
+		ADD_SLOT(double, r_max, INPUT, REQUIRED, DocString{"Value of the biggest radius possible"}); 
+		ADD_SLOT(uint64_t, n_max, INPUT, 1e16, DocString{"Maximal number of particles. Default is 1e16."}); 
+
[…]
