### 2025-01-20 5376ce4
Xdm custom lattice (#78)
* Added legacy_lattice operator with internal init_rcb_grid through a boolean slot

* Added an example using legacy_lattice operator (internal init_rcb_grid) + custom lattice definition

* minor change, update rsa mpi spack repo

* fix legacy example with impoce velocity

* rm py cache

---------

Co-authored-by: Paul Lafourcade <paul.lafourcade@cea.fr>
Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
