+++ spack_repo/packages/exadem/package.py
@@ -18,6 +18,7 @@ class Exadem(CMakePackage):
     depends_on("yaml-cpp@0.6.3")
     depends_on("openmpi")
     depends_on("exanbody")
+    depends_on("rsampi", when="+rsampi")
     depends_on("cuda", when="+cuda")
     depends_on("rsampi", when="+rsampi")
 
