+++ src/polyhedron/include/exaDEM/compute_vertices.hpp
@@ -26,16 +26,19 @@ namespace exaDEM
   struct PolyhedraComputeVerticesFunctor
   {
     const shape *shps;
+    const shapes_GPU shps2;
     ONIKA_HOST_DEVICE_FUNC inline void operator()(const uint32_t type, const double rx, const double ry, const double rz, const double h, const exanb::Quaternion &orient, ::onika::oarray_t<::exanb::Vec3d, EXADEM_MAX_VERTICES> &vertices) const
     {
       // h will be used in a next development
-      const auto &shp = shps[type];
-      const unsigned int nv = shp.get_number_of_vertices();
+      //const auto &shp = shps[type];
+      //const unsigned int nv = shp.get_number_of_vertices();
+      const unsigned int nv = shps2.get_number_of_vertices(type);
       const exanb::Vec3d position = {rx, ry, rz};
       assert(nv < EXADEM_MAX_VERTICES);
       for (size_t i = 0; i < nv; i++)
       {
-        vertices[i] = shp.get_vertex(i, position, orient);
+        //vertices[i] = shp.get_vertex(i, position, orient);
+        vertices[i] = shps2.get_vertex(type, i, position, orient);
       }
     }
   };
