### 2025-01-22 df7e8ef
Binary search
