+++ src/interaction/include/exaDEM/interaction/interaction_manager.hpp
@@ -46,7 +46,7 @@ namespace exaDEM
       {
         info[p].offset = offset;
         if constexpr (use_history)
-          //update_friction_moment(list[p], hist);
+          update_friction_moment(list[p], hist);
         std::copy(list[p].begin(), list[p].end(), data.data() + offset);
         info[p].size = list[p].size();
         offset += list[p].size();

[…]
