### 2025-01-22 f7ecb55
stl_mesh
