+++ src/interaction/unclassify_interactions.cpp
@@ -46,6 +46,39 @@ namespace exaDEM
 {
   using namespace exanb;
   
+  enum ReduceType {
+    MIN,
+    MAX
+  };
+  
+  void findMinMax(uint16_t* d_in, int size, uint16_t* result, ReduceType type) {
+    // Temporary storage
+    void* d_temp_storage = nullptr;
+    size_t temp_storage_bytes = 0;
+
+    if (type == MAX) {
+        // Calculate temporary storage size for max
+        cub::DeviceReduce::Max(d_temp_storage, temp_storage_bytes, d_in, result, size);
+    } else {
+        // Calculate temporary storage size for min
+        cub::DeviceReduce::Min(d_temp_storage, temp_storage_bytes, d_in, result, size);
+    }
+
+    // Allocate temporary storage
+    cudaMalloc(&d_temp_storage, temp_storage_bytes);
+
+    if (type == MAX) {
+        // Launch reduction to find the max
+        cub::DeviceReduce::Max(d_temp_storage, temp_storage_bytes, d_in, result, size);
+    } else {
+        // Launch reduction to find the min
+        cub::DeviceReduce::Min(d_temp_storage, temp_storage_bytes, d_in, result, size);
+    }
+
+    // Free temporary storage
+    cudaFree(d_temp_storage);
+  } 
+  
   __global__ void filtre_un( double* ft_x,
   			double* ft_y,
   			double* ft_z,
@@ -70,8 +103,10 @@ namespace exaDEM
   
   __global__ void filtre_deux( uint64_t* id_i,
   				uint64_t* id_j,
+  				uint16_t* sub_j,
   				uint64_t* id_i_res,
   				uint64_t* id_j_res,
+  				uint16_t* sub_j_res,
   				double* ft_x,
   				double* ft_y,
   				double* ft_z,
@@ -119,14 +154,15 @@ namespace exaDEM
   				if( s[i] == 1 ) index++;
   			}
   			
-  			id_i_res[index + incr] = id_i[threadIdx.x];
-  			id_j_res[index + incr] = id_j[threadIdx.x];
-  			ft_x_res[index + incr] = ft_x[threadIdx.x];
-  			ft_y_res[index + incr] = ft_y[threadIdx.x];
-  			ft_z_res[index + incr] = ft_z[threadIdx.x];
-  			mom_x_res[index + incr] = mom_x[threadIdx.x];
-  			mom_y_res[index + incr] = mom_y[threadIdx.x];
-  			mom_z_res[index + incr] = mom_z[threadIdx.x];
+  			id_i_res[index + incr] = id_i[idx];
+  			id_j_res[index + incr] = id_j[idx];
+  			sub_j_res[index + incr] = sub_j[idx];
+  			ft_x_res[index + incr] = ft_x[idx];
+  			ft_y_res[index + incr] = ft_y[idx];
+  			ft_z_res[index + incr] = ft_z[idx];
+  			mom_x_res[index + incr] = mom_x[idx];
+  			mom_y_res[index + incr] = mom_y[idx];
+  			mom_z_res[index + incr] = mom_z[idx];
 
   			indices[incr + index] = incr + index;
   		}
@@ -147,6 +183,31 @@ namespace exaDEM
   	{
   		int range = max - min + 1;
   		keys[idx] = (id_i[idx] - min) * range + (id_j[idx] - min);
+  	}
+  }
+  
+  __global__ void generateKeys2( uint64_t* keys,
+  				 const uint64_t* id_i,
+  				 const uint64_t* id_j,
+  				 const uint16_t* sub_j,
+  				 int min, int max,
+  				 int min2, int max2,
+  				 int size)
+  {
[…]
