### 2025-01-24 305b65e
small changes into stress tensor
