### 2025-01-24 3c7b195
add a paraview field containing the mpi rank into write_paraview_polyhedra
