+++ src/io/write_paraview_polyhedra.cpp
@@ -48,9 +48,12 @@ namespace exaDEM
     ADD_SLOT(MPI_Comm, mpi, INPUT, MPI_COMM_WORLD);
     ADD_SLOT(GridT, grid, INPUT_OUTPUT);
     ADD_SLOT(Domain, domain, INPUT, REQUIRED);
-    ADD_SLOT( std::string , filename     , INPUT , "output");
+    ADD_SLOT(std::string , filename, INPUT , "output");
     ADD_SLOT(long, timestep, INPUT, DocString{"Iteration number"});
     ADD_SLOT(shapes, shapes_collection, INPUT_OUTPUT, DocString{"Collection of shapes"});
+
+    // optionnal
+    ADD_SLOT(bool, mpi_rank, INPUT, false, DocString{"Add a field containing the mpi rank."});
 
   public:
     inline std::string documentation() const override final
@@ -76,7 +79,7 @@ namespace exaDEM
       auto &shps = *shapes_collection;
       const auto cells = grid->cells();
       const size_t n_cells = grid->number_of_cells();
-      par_poly_helper buffers; // it conatins streams 
+      par_poly_helper buffers = {*mpi_rank}; // it conatins streams 
 
       // fill string buffers
       for (size_t cell_a = 0; cell_a < n_cells; cell_a++)
@@ -103,8 +106,17 @@ namespace exaDEM
 
       if (rank == 0)
       {
-        exaDEM::write_pvtp_polyhedron(*filename, size);
+        exaDEM::write_pvtp_polyhedron(*filename, size, buffers);
       }
+
+      if(buffers.mpi_rank) // add ranks 
+      {
+        for(int i = 0 ; i < buffers.n_vertices ; i++)
+        {
+          buffers.ranks << rank << " ";
+        }
+      }
+
       std::string file = *filename + "/%06d.vtp";
       file = format_string(file,  rank);
       exaDEM::write_vtp_polyhedron(file, buffers);
