### 2025-01-24 13cc916
Add mpi fields (#118)
* small changes into stress tensor

* add a paraview field containing the mpi rank into write_paraview_polyhedra

---------

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
