### 2025-01-28 7a52f08
First version on GPU
