+++ src/grid/chunk_neighbors_contact.cpp
@@ -17,6 +17,8 @@ specific language governing permissions and limitations
 under the License.
 */
 
+//#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
+
 #include <exanb/core/operator.h>
 #include <exanb/core/operator_slot.h>
 #include <exanb/core/operator_factory.h>
@@ -41,9 +43,193 @@ under the License.
 
 #include <exanb/particle_neighbors/chunk_neighbors_execute.h>
 
+#include <cub/cub.cuh>
+
+#include <exaDEM/classifier/interactionSOA.hpp>
+
+#include <exaDEM/drivers.h>
+#include <exaDEM/shape_detection_driver.hpp>
+
+#include <exaDEM/traversal.hpp>
+
 namespace exaDEM
 {
   using namespace exanb;
+  
+    template< class GridT > ONIKA_HOST_DEVICE_FUNC bool nbh_filter_GPU(GridT* cells,
+  						double rcut_inc,
+  						double d2,
+  						double rcut2,
+  						size_t cell_a,
+  						size_t p_a,
+  						size_t cell_b,
+  						size_t p_b)
+  {
+      const double r_a = cells[cell_a][field::radius][p_a];
+      const double r_b = cells[cell_b][field::radius][p_b];
+      const double rmax = r_a + r_b + rcut_inc;
+      const double rmax2 = rmax * rmax;
+      return d2 > 0.0 && d2 < rmax2 && d2 < rcut2 ;
+  }
+  
+  template< class GridT > bool filter( GridT* cells, double rcut_inc, double d2, double rcut2, int cell_a, int p_a, int cell_b, int p_b)
+  {
+      const double r_a = cells[cell_a][field::radius][p_a];
+      const double r_b = cells[cell_b][field::radius][p_b];
+      const double rmax = r_a + r_b + rcut_inc;
+      const double rmax2 = rmax * rmax;
+      return d2 > 0.0 && d2 < rmax2 && d2 < rcut2 ;  	
+  }
+  
+  bool is_in(int index, size_t* cells, size_t size)
+  {
+  	//bool b = true;
+  	int i = 0;
+
+  	while( i < size)
+  	{
+  		if(index == cells[i]) return true;
+  		
+  		//if(index > cells[i]) b = false;
+  		
+  		i++;
+  	}
+
+  	return false;
+  }
+  
+  template< class GridT > __global__ void kernelUN(GridT* cells,
+  							int* cell_id,
+  							int* nb_particles,
+  							int* cell_neighbors_ids,
+  							int* cell_neighbors_size,
+  							int* cell_start,
+  							int* cell_end,
+  							int* cell_particle_start,
+  							const double dist_lab,
+  							Mat3d xform,
+  							double rcut_inc,
+  							uint64_t* id_i,
+  							uint64_t* id_j,
+  							uint32_t* cell_i,
+  							uint32_t* cell_j,
+  							uint16_t* p_i,
+  							uint16_t* p_j,
+  							int* nb_nbh,
+  							int* id,
+  							int* celli,
+  							int* pi,
+  							int* total_interactions,
+  							int* total_interactions_driver,
+  							int nombre_voisins_potentiels,
+  							Cylinder driver,
+  							int* interaction_driver )
+  {
+  	
+  	int cell = cell_id[blockIdx.x];
+	int nb = nb_particles[blockIdx.x];
+	int start = cell_start[blockIdx.x];
+	int end = cell_end[blockIdx.x];
[…]
