+++ src/sphere/nbh_sphere.cpp
@@ -234,52 +234,7 @@ namespace exaDEM
 
           item.type = 0; // === Vertex - Vertex
 
-          if (sym)
-          {
-            // Second, we add interactions between two spheres.
-            apply_cell_particle_neighbors(*grid, *chunk_neighbors, cell_a, loc_a, std::false_type() /* not symetric */,
-                [&g, &manager, &cells, cell_a, &item, id_a](int p_a, size_t cell_b, unsigned int p_b, size_t p_nbh_index)
-                {
-                // default value of the interaction studied (A or i -> B or j)
-                const uint64_t id_nbh = cells[cell_b][field::id][p_b];
-                if (id_a[p_a] >= id_nbh)
-                {
-                if (!g.is_ghost_cell(cell_b))
-                return;
-                }
 
-                // Add interactions
-                item.id_i = id_a[p_a];
-                item.p_i = p_a;
-                item.id_j = id_nbh;
-                item.p_j = p_b;
-                item.cell_j = cell_b;
-                manager.add_item(p_a, item);
-                });
-          }
-          else
-          {
-            // Second, we add interactions between two spheres.
-            apply_cell_particle_neighbors(*grid, *chunk_neighbors, cell_a, loc_a, std::false_type() /* not symetric */,
-                [&g, &manager, &cells, cell_a, &item, id_a](int p_a, size_t cell_b, unsigned int p_b, size_t p_nbh_index)
-                {
-                // default value of the interaction studied (A or i -> B or j)
-                const uint64_t id_nbh = cells[cell_b][field::id][p_b];
-                // Add interactions
-                item.id_i = id_a[p_a];
-                item.p_i = p_a;
-                item.id_j = id_nbh;
-                item.p_j = p_b;
-                item.cell_j = cell_b;
-                manager.add_item(p_a, item);
-                });
-          }
-
-          manager.update_extra_storage<true>(storage);
-
-          assert(interaction_test::check_extra_interaction_storage_consistency(storage.number_of_particles(), storage.m_info.data(), storage.m_data.data()));
-
-          assert(migration_test::check_info_value(storage.m_info.data(), storage.m_info.size(), 1e6));
         } //    GRID_OMP_FOR_END
       }
     }
