### 2025-01-30 e6aa133
divide the stress tensor with the volume (#119)
Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
