+++ src/interaction/unclassify_interactions.cpp
@@ -445,15 +445,15 @@ namespace exaDEM
       
       }
       }
-      /*
-      if (grid->number_of_cells() == 0)
+      
+      /*if (grid->number_of_cells() == 0)
       {
         return;
       }
       if (!ic.has_value())
         return;
-      ic->unclassify(*ges);
-      */
+      ic->unclassify(*ges);*/
+      
      printf("UNCLASSIFY_END\n");
       
     }
