+++ src/grid/chunk_neighbors_contact.cpp
@@ -99,6 +99,8 @@ namespace exaDEM
   }
   
   template< class GridT > __global__ void kernelUN(GridT* cells,
+  							//GridT g,
+  							int* ghost_cell,
   							int* cell_id,
   							int* nb_particles,
   							int* cell_neighbors_ids,
@@ -165,8 +167,10 @@ namespace exaDEM
   				const Vec3d dr = { cells[cell][field::rx][p_a] - cells[cell_b][field::rx][p_b] , cells[cell][field::ry][p_a] - cells[cell_b][field::ry][p_b] , cells[cell][field::rz][p_a] - cells[cell_b][field::rz][p_b] };
                 		double d2 = norm2( xform * dr );
                 	
-  				if(nbh_filter_GPU(cells, rcut_inc, d2, rcut2, cell, p_a, cell_b, p_b) && (cells[cell][field::id][p_a] < cells[cell_b][field::id][p_b]) && (cell!=cell_b || p_a!=p_b)) 
-  				{ 
+  				if(nbh_filter_GPU(cells, rcut_inc, d2, rcut2, cell, p_a, cell_b, p_b)) //&& (cells[cell][field::id][p_a] < cells[cell_b][field::id][p_b]) && ) 
+  				{
+  					if(cells[cell][field::id][p_a] < cells[cell_b][field::id][p_b])
+  					{
   					id_i[incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels + nb_interactions] = cells[cell][field::id][p_a];
   					id_j[incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels + nb_interactions] = cells[cell_b][field::id][p_b];
   					cell_i[incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels + nb_interactions] = cell;
@@ -175,8 +179,19 @@ namespace exaDEM
   					p_j[incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels + nb_interactions] = p_b;
   					
   					nb_interactions++;
+  					}
+  					else if( ghost_cell[i]==1 )
+  					{
+   					id_i[incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels + nb_interactions] = cells[cell][field::id][p_a];
+  					id_j[incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels + nb_interactions] = cells[cell_b][field::id][p_b];
+  					cell_i[incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels + nb_interactions] = cell;
+  					cell_j[incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels + nb_interactions] = cell_b;
+  					p_i[incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels + nb_interactions] = p_a;
+  					p_j[incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels + nb_interactions] = p_b;
   					
-  					atomicAdd(&total_interactions[0], 1);
+  					nb_interactions++;
+  					}
+  					//atomicAdd(&total_interactions[0], 1);
   				}				
 			}
 		}
@@ -185,6 +200,7 @@ namespace exaDEM
 		id[incr + p_a] = cells[cell][field::id][p_a];
 		celli[incr + p_a] = cell;
 		pi[incr + p_a] = p_a;
+		atomicAdd(&total_interactions[0], nb_interactions); 
 	}
    }
    
@@ -307,8 +323,8 @@ namespace exaDEM
       auto cells = grid->cells();
[…]
