### 2025-02-02 11d03dc
add linear force motion for stl meshes
