+++ src/grid/chunk_neighbors_contact.cpp
@@ -125,8 +125,7 @@ namespace exaDEM
   							int* total_interactions_driver,
   							int nombre_voisins_potentiels,
   							Cylinder driver,
-  							int* interaction_driver,
-  							int* anas )
+  							int* interaction_driver)
   {
   	
   	int cell = cell_id[blockIdx.x];
@@ -191,8 +190,6 @@ namespace exaDEM
   						p_j[incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels + nb_interactions] = p_b;
   					
   						nb_interactions++;
-  						
-  						atomicAdd(&anas[0], 1);
   						
   					}
   					//atomicAdd(&total_interactions[0], 1);
@@ -328,7 +325,7 @@ namespace exaDEM
       ContactNeighborFilterFunc<decltype(cells)> nbh_filter{cells, *rcut_inc};
       static constexpr std::false_type no_z_order = {};
       
-      if (!domain->xform_is_identity())
+      /*if (!domain->xform_is_identity())
       {
         LinearXForm xform = {domain->xform()};
         chunk_neighbors_execute(ldbg, *chunk_neighbors, *grid, *amr, *amr_grid_pairs, *config, *chunk_neighbors_scratch, cs, cs_log2, *nbh_dist_lab, xform, gpu_enabled, no_z_order, nbh_filter);
@@ -337,7 +334,7 @@ namespace exaDEM
       {
         NullXForm xform = {};
         chunk_neighbors_execute(ldbg, *chunk_neighbors, *grid, *amr, *amr_grid_pairs, *config, *chunk_neighbors_scratch, cs, cs_log2, *nbh_dist_lab, xform, gpu_enabled, no_z_order, nbh_filter);
-      }
+      }*/
       
       
       //NOUVELLE MÉTHODE
@@ -345,9 +342,9 @@ namespace exaDEM
       auto number_of_particles = grid->number_of_particles();
        
       onika::memory::CudaMMVector<int> nb_particles_cell;
-      //onika::memory::CudaMMVector<onika::memory::CudaMMVector<int>> cell_particles_neighbors;
+
       onika::memory::CudaMMVector<int> cell_particles_neighbors;
-      //onika::memory::CudaMMVector<onika::memory::CudaMMVector<int>> cell_particles_neighbors_size;
+
       onika::memory::CudaMMVector<int> cell_particles_neighbors_size;
       
       onika::memory::CudaMMVector<int> cell_particles_number_of_neighbors_cells;
@@ -376,14 +373,7 @@ namespace exaDEM
       	cell_particles_neighbors_size.resize( n_cells*27 );
       	
       	cell_particles_number_of_neighbors_cells.resize(n_cells);
-      	
-      	/*#pragma omp parallel for
-      	for(int i = 0; i < n_cells; ++i)
-      	{
-      		cell_particles_neighbors[i].resize(27);
-      		cell_particles_neighbors_size[i].resize(27);
-      	}*/
- 
+
       	unsigned int max_threads = omp_get_max_threads();
 
 #	pragma omp parallel
@@ -400,8 +390,6 @@ namespace exaDEM
 			ssize_t sgstart_a = sub_grid_start[cell_a];
 			ssize_t sgsize_a = sub_grid_start[cell_a+1] - sgstart_a;
 			ssize_t n_sub_cells_a = sgsize_a+1;
-			ssize_t sgside_a = icbrt64( n_sub_cells_a );
-			assert( sgside_a <= static_cast<ssize_t>(GRID_CHUNK_NBH_MAX_AMR_RES) );
 			
[…]
