+++ src/grid/chunk_neighbors_contact.cpp
@@ -121,33 +121,44 @@ namespace exaDEM
   							int* id,
   							int* celli,
   							int* pi,
-  							int* total_interactions,
-  							int* total_interactions_driver,
+  							//int* total_interactions,
+  							//int* total_interactions_driver,
   							int nombre_voisins_potentiels,
   							Cylinder driver,
   							int* interaction_driver)
   {
   	
-  	int cell = cell_id[blockIdx.x];
 	int nb = nb_particles[blockIdx.x];
-	int start = cell_start[blockIdx.x];
-	int end = cell_end[blockIdx.x];
-	int incr = cell_particle_start[blockIdx.x];
 	
 	if(threadIdx.x < nb)
 	{
+		int cell = cell_id[blockIdx.x];
+		int start = cell_start[blockIdx.x];
+		int end = cell_end[blockIdx.x];
+		int incr = cell_particle_start[blockIdx.x];
+		
 		int p_a = threadIdx.x;
+				
+		int incr2 = incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels;
+		
 		int nb_interactions = 0;
 		
 		const double rVerletMax = cells[cell][field::radius][p_a] + rcut_inc;
 		
-		const Vec3d r = {cells[cell][field::rx][p_a], cells[cell][field::ry][p_a], cells[cell][field::rz][p_a]};
+		auto &rx_a = cells[cell][field::rx][p_a];
+		auto &ry_a = cells[cell][field::ry][p_a];
+		auto &rz_a = cells[cell][field::rz][p_a];
+		auto &id_a = cells[cell][field::id][p_a];
+		
+		const Vec3d r = { rx_a , ry_a , rz_a };
 		
 		if (driver.filter(rVerletMax, r))
 		{
-			atomicAdd(&total_interactions_driver[0], 1);
+			//atomicAdd(&total_interactions_driver[0], 1);
 			interaction_driver[incr + p_a] = 1;
-			
+			id[incr + p_a] = id_a;
+			pi[incr + p_a] = p_a;
+			celli[incr + p_a] = cell;
 		}
 		else
 		{
@@ -161,62 +172,65 @@ namespace exaDEM
 			for(int j = 0; j < cell_neighbors_size[i]; j++)
 			{
 				int p_b = j;
+				
+				auto &id_b = cells[cell_b][field::id][p_b];
 	  			
 	  			double rcut2 = dist_lab * dist_lab;
   		
-  				const Vec3d dr = { cells[cell][field::rx][p_a] - cells[cell_b][field::rx][p_b] , cells[cell][field::ry][p_a] - cells[cell_b][field::ry][p_b] , cells[cell][field::rz][p_a] - cells[cell_b][field::rz][p_b] };
+  				const Vec3d dr = { rx_a - cells[cell_b][field::rx][p_b] , ry_a - cells[cell_b][field::ry][p_b] , rz_a - cells[cell_b][field::rz][p_b] };
                 		double d2 = norm2( xform * dr );
                 	
   				if(nbh_filter_GPU(cells, rcut_inc, d2, rcut2, cell, p_a, cell_b, p_b)) //&& (cells[cell][field::id][p_a] < cells[cell_b][field::id][p_b]) && ) 
   				{
-  					if(cells[cell][field::id][p_a] < cells[cell_b][field::id][p_b])
+  					if( id_a < id_b)
   					{
-  						id_i[incr*nombre_voisins_potentiels + p_a*nombre_voisins_potentiels + nb_interactions] = cells[cell][field::id][p_a];
[…]
