### 2025-02-04 2e1b52f
add linear compressive force for stl meshes
