+++ src/interaction/unclassify_interactions.cpp
@@ -55,8 +55,7 @@ namespace exaDEM
   			double* mom_y,
   			double* mom_z,
   			size_t size,
-  			int* blocks,
-  			int* total)
+  			int* filtre)
   {
   	int idx = threadIdx.x + blockIdx.x * blockDim.x;
   	
@@ -64,8 +63,7 @@ namespace exaDEM
   	{
   		if(ft_x[idx]!=0 || ft_y[idx]!=0 || ft_z[idx]!=0 || mom_x[idx]!=0 || mom_y[idx]!=0 || mom_z[idx]!=0)
   		{
-  			atomicAdd(&blocks[blockIdx.x], 1);
-  			atomicAdd(&total[0], 1);
+  			filtre[idx] = 1;
   		}
   	}
   }
@@ -75,7 +73,6 @@ namespace exaDEM
   				uint16_t* sub_j,
   				uint64_t* id_i_res,
   				uint64_t* id_j_res,
-  				uint16_t* sub_j_res,
   				double* ft_x,
   				double* ft_y,
   				double* ft_z,
@@ -88,56 +85,29 @@ namespace exaDEM
   				double* mom_x_res,
   				double* mom_y_res,
   				double* mom_z_res,
-  				int* blocks_incr,
+  				int* filtre_incr,
   				int* indices,
-  				uint64_t* keys,
   				int size)
   {
   	int idx = threadIdx.x + blockIdx.x * blockDim.x;
-  	
-  	//printf("RAZ\n");
-  	
-  	int incr = blocks_incr[blockIdx.x];
-  	
-  	__shared__ int s[256];
-  	
-  	s[threadIdx.x] = 0;
-  	
-  	__syncthreads();
-  	
-  	
+
   	if(idx < size)
   	{
   		if(ft_x[idx]!=0 || ft_y[idx]!=0 || ft_z[idx]!=0 || mom_x[idx]!=0 || mom_y[idx]!=0 || mom_z[idx]!=0)
   		{
-  			s[threadIdx.x] = 1;
-  		}
-  		
-  		__syncthreads();
-  		
-  		int index = 0;
-  		
-  		if(s[threadIdx.x] == 1)
-  		{
-  			for(int i = 0; i < threadIdx.x; i++)
-  			{
-  				if( s[i] == 1 ) index++;
-  			}
+  			int &incr = filtre_incr[idx];
   			
-  			id_i_res[index + incr] = id_i[idx];
-  			id_j_res[index + incr] = id_j[idx];
-  			sub_j_res[index + incr] = sub_j[idx];
-  			ft_x_res[index + incr] = ft_x[idx];
-  			ft_y_res[index + incr] = ft_y[idx];
-  			ft_z_res[index + incr] = ft_z[idx];
-  			mom_x_res[index + incr] = mom_x[idx];
-  			mom_y_res[index + incr] = mom_y[idx];
-  			mom_z_res[index + incr] = mom_z[idx];
-
-  			indices[incr + index] = incr + index;
+  			id_i_res[incr] = id_i[idx];
+  			id_j_res[incr] = id_j[idx];
+  			ft_x_res[incr] = ft_x[idx];
+  			ft_y_res[incr] = ft_y[idx];
+  			ft_z_res[incr] = ft_z[idx];
+  			mom_x_res[incr] = mom_x[idx];
+  			mom_y_res[incr] = mom_y[idx];
+  			mom_z_res[incr] = mom_z[idx];
   			
+  			indices[incr] = incr;
   		}
-  		
   	}
   }
   
@@ -251,52 +221,60 @@ namespace exaDEM
       int numBlocks = ( size + blockSize - 1 ) / blockSize;
       
       
-      onika::memory::CudaMMVector<int> blocks;
[…]
