+++ src/grid/chunk_neighbors_contact.cpp
@@ -52,6 +52,8 @@ under the License.
 
 #include <exaDEM/traversal.hpp>
 
+#include <exaDEM/classifier/classifier.hpp>
+
 namespace exaDEM
 {
   using namespace exanb;
@@ -248,6 +250,128 @@ namespace exaDEM
   		}
   	}
   }
+  
+  __global__ void filtre_un( double* ft_x,
+  			double* ft_y,
+  			double* ft_z,
+  			double* mom_x,
+  			double* mom_y,
+  			double* mom_z,
+  			size_t size,
+  			int* filtre)
+  {
+  	int idx = threadIdx.x + blockIdx.x * blockDim.x;
+  	
+  	if(idx < size)
+  	{
+  		if(ft_x[idx]!=0 || ft_y[idx]!=0 || ft_z[idx]!=0 || mom_x[idx]!=0 || mom_y[idx]!=0 || mom_z[idx]!=0)
+  		{
+  			filtre[idx] = 1;
+  		}
+  	}
+  }
+  
+  __global__ void filtre_deux( uint64_t* id_i,
+  				uint64_t* id_j,
+  				uint16_t* sub_j,
+  				uint64_t* id_i_res,
+  				uint64_t* id_j_res,
+  				double* ft_x,
+  				double* ft_y,
+  				double* ft_z,
+  				double* ft_x_res,
+  				double* ft_y_res,
+  				double* ft_z_res,
+  				double* mom_x,
+  				double* mom_y,
+  				double* mom_z,
+  				double* mom_x_res,
+  				double* mom_y_res,
+  				double* mom_z_res,
+  				int* filtre_incr,
+  				int* indices,
+  				int size)
+  {
+  	int idx = threadIdx.x + blockIdx.x * blockDim.x;
+
+  	if(idx < size)
+  	{
+  		if(ft_x[idx]!=0 || ft_y[idx]!=0 || ft_z[idx]!=0 || mom_x[idx]!=0 || mom_y[idx]!=0 || mom_z[idx]!=0)
+  		{
+  			int &incr = filtre_incr[idx];
+  			
+  			id_i_res[incr] = id_i[idx];
+  			id_j_res[incr] = id_j[idx];
+  			ft_x_res[incr] = ft_x[idx];
+  			ft_y_res[incr] = ft_y[idx];
+  			ft_z_res[incr] = ft_z[idx];
+  			mom_x_res[incr] = mom_x[idx];
+  			mom_y_res[incr] = mom_y[idx];
+  			mom_z_res[incr] = mom_z[idx];
+  			
+  			indices[incr] = incr;
+  		}
+  	}
+  }
+  
+  __global__ void generateKeys( uint64_t* keys, 
+  				const uint64_t* id_i, 
+  				const uint64_t* id_j,
+  				int min,
+  				int max,
+  				int type,
+  				int size)
+  {
+  	int idx = threadIdx.x + blockIdx.x * blockDim.x;
+  	
+  	if(idx < size)
+  	{
+  		int range = max - min + 1;
+  		if(type == 0)keys[idx] = (id_i[idx] - min) * range + (id_j[idx] - min);
+  		if(type == 4)keys[idx] = id_i[idx];
+  	}
+  }
+  
+  
+  void exclusive_sum( int* d_in, int* d_out, int size )
+  {
+  	void* d_temp_storage = nullptr;
+  	size_t temp_storage_bytes = 0;
+  	
+  	cub::DeviceScan::ExclusiveSum(d_temp_storage, temp_storage_bytes, d_in, d_out, size);
+  	
+  	cudaMalloc(&d_temp_storage, temp_storage_bytes);
+  	
[…]
