### 2025-02-06 #7484 @jedbrown
Thanks, and sorry this took me a while to get to. Once you've made the edits suggested below, I'll be able to proceed.

- [ ] `LIGGGHTS` is misspelled (multiple occurrences) and the Kloss DOI resolves to 404 (that looks like a publisher failure, but perhaps you can find a new DOI)
- [ ] line 97: "guided by [the] exaNBody framework"
- [ ] figure 1 caption: "700 thousand octahedra" (no plural "thousands")
- [ ] capitalize "MPI" in figure 1 and 2 captions
- [ ] please use `\autoref` when referring to figures (see [example paper](https://joss.readthedocs.io/en/latest/example_paper.html))
- [ ] line 113: "MPIIO" -> "MPI-IO"
- [ ] You can remove the link to the repository and documentation in the text (it always appears in the sidebar and on the JOSS website for the paper). The "Future of ExaDEM" is speculative and can be removed.
- [ ] Please add a `# References` anchor at the bottom of your document (Acknowledgements currently flows into references)
- [ ] "poisson" in the bib probably needs to protected capitalization `{Poisson}`
- [ ] `richefeu_20` should not protect capitalization (not proper names)
- [ ] "lennard-jones" needs protection
