### 2025-02-06 e8128f0
Update paper.md
LIGGGTHS -> LIGGGHTS
