+++ docs/paper/joss/paper.md
@@ -59,7 +59,7 @@ The `DEM` method, employed to study granular media, falls within the scope of so
 
 Efficient neighbor particle detection is a critical aspect of `DEM` simulation codes. Typically, this is achieved by combining the linked cells method [@ciccotti1987simulation] and Verlet lists [@verlet1967computer], which optimize neighbor searches using a cartesian grid of cells (with a complexity of N, N is the number of particles) while limiting the refresh rate of neighbor lists.  
   
-Several `DEM` software packages have emerged in recent years, offering `HPC` capabilities. Examples include `LIGGGTHS` [@kloss2012models], which is based on `LAMMPS` [@thompson2022lammps] data structures for Molecular Dynamics simulations with spherical particles (`MPI`), and `Blaze-DEM` [@govender2018study] which employs spheres and polyhedra on `GPU` via `CUDA`. `ExaDEM` aims to establish itself as a software solution that combines `MPI` parallelization, `OpenMP` thread parallelization for both polyhedral and spherical particles, and `CUDA` `GPU` parallelization for spherical particles. Similar to `LIGGGTHS` with `LAMMPS`, `ExaDEM` benefits from various `HPC` features developed in the Molecular Dynamics code `ExaSTAMP` [@cieren2014exastamp] that have been mutualized in the `exaNBody` framework, such as `AMR` data structures [@prat2018combining] with load balancing [@prat2020amr] and in-situ analysis [@dirand2018tins]. `ExaDEM` aims to incorporate the physics of interest from the `Rockable` [@richefeu_2016, @gitrockable] (open source), originally created at CNRS for polyhedra on an `HPC` framework. To the best of our knowledge, this is a non-exhaustive list of other well-known `DEM` codes: `EDEM` (not open source), `Rocky DEM` (not open source), `MercuryDPM` [@weinhart2020fast] (open source), and `Yade` [@smilauer2023yade] (open source).
+Several `DEM` software packages have emerged in recent years, offering `HPC` capabilities. Examples include `LIGGGHTS` [@kloss2012models], which is based on `LAMMPS` [@thompson2022lammps] data structures for Molecular Dynamics simulations with spherical particles (`MPI`), and `Blaze-DEM` [@govender2018study] which employs spheres and polyhedra on `GPU` via `CUDA`. `ExaDEM` aims to establish itself as a software solution that combines `MPI` parallelization, `OpenMP` thread parallelization for both polyhedral and spherical particles, and `CUDA` `GPU` parallelization for spherical particles. Similar to `LIGGGHTS` with `LAMMPS`, `ExaDEM` benefits from various `HPC` features developed in the Molecular Dynamics code `ExaSTAMP` [@cieren2014exastamp] that have been mutualized in the `exaNBody` framework, such as `AMR` data structures [@prat2018combining] with load balancing [@prat2020amr] and in-situ analysis [@dirand2018tins]. `ExaDEM` aims to incorporate the physics of interest from the `Rockable` [@richefeu_2016, @gitrockable] (open source), originally created at CNRS for polyhedra on an `HPC` framework. To the best of our knowledge, this is a non-exhaustive list of other well-known `DEM` codes: `EDEM` (not open source), `Rocky DEM` (not open source), `MercuryDPM` [@weinhart2020fast] (open source), and `Yade` [@smilauer2023yade] (open source).
 
 # Implementation Details
 
