+++ docs/paper/joss/paper.md
@@ -70,34 +70,32 @@ Regarding the intra-`MPI` parallelization, we distinguish two main differences b
 - For spherical particles, `OpenMP` parallelization involves iterating over cells. In `GPU` parallelization, a block of `GPU` threads is assigned to a cell, with each `GPU` thread processing a particle.
 - For polyhedral particles, another parallel level is chosen for thread parallelization, the interaction. Unlike spherical particles, polyhedra can have multiple contacts of different types (vertex-vertex, vertex-edge, vertex-face, edge-edge). Hence, it is more effective to consider interactions rather than cells to achieve thread parallelization. However, this strategy introduces synchronizations, such as the usage of mutexes. The `GPU` parallelization of polyhedra is still an upcoming development.
 
-In conclusion, the design of `ExaDEM`, guided by the `exaNBody` framework, facilitates the addition or removal of individual operators or features without disrupting other functionalities, provided these operators are independent. For instance, removing the `gravity_force` from the `ExaDEM` repository is feasible while ensuring the preservation of the `contact_neighbor` operator (building neighbor lists for every particle), which is crucial for running the `hooke_force` operator. Significant efforts have been made to minimize interactions between operators, simplifying the process of adding or removing new modules/operators. 
+In conclusion, the design of `ExaDEM`, guided by the the `exaNBody` framework, facilitates the addition or removal of individual operators or features without disrupting other functionalities, provided these operators are independent. For instance, removing the `gravity_force` from the `ExaDEM` repository is feasible while ensuring the preservation of the `contact_neighbor` operator (building neighbor lists for every particle), which is crucial for running the `hooke_force` operator. Significant efforts have been made to minimize interactions between operators, simplifying the process of adding or removing new modules/operators. 
 
 # Main Features
 
-![Simulation of 700 thousands octahedra in a rotating drum running on 128 mpi processes with 8 OpenMP threads per mpi process (processor: AMD EPYC Milan 7763). \label{fig:rotating-drum}](./rotating-drum.png "test"){width=85%}
+![Simulation of 700 thousands octahedra in a rotating drum running on 128 MPI processes with 8 OpenMP threads per MPI process (processor: AMD EPYC Milan 7763). \label{fig:rotating-drum}](./rotating-drum.png "test"){width=85%}
 
-![Simulation of 20 million spherical particles falling into a funnel. This simulation runs on 512 mpi processes with 8 OpenMP threads per mpi process (processor: AMD EPYC Milan 7763).  \label{fig:funnel}](./funnel.png "test"){width=70%}
+![Simulation of 20 million spherical particles falling into a funnel. This simulation runs on 512 MPI processes with 8 OpenMP threads per MPI process (processor: AMD EPYC Milan 7763).  \label{fig:funnel}](./funnel.png "test"){width=70%}
 
-`ExaDEM` aims to meet scientific expectations, particularly for nuclear fuel simulations involving scenarios such as rotating drums (see figure \ref{fig:rotating-drum}) or compaction processes. To achieve this goal, `ExaDEM` provides the following key features:
+`ExaDEM` aims to meet scientific expectations, particularly for nuclear fuel simulations involving scenarios such as rotating drums (see figure \autoref{fig:rotating-drum}) or compaction processes. To achieve this goal, `ExaDEM` provides the following key features:
 
 - Handling of different particle types: spherical and polyhedral,
 - Hybrid parallelisation `MPI` + X,
 	- X = `OpenMP` or `CUDA` for spherical particles,
 	- X = `OpenMP` for polyhedral particles,
 	- The Recursive Coordinate Bisection method is used for the load balancing,
[…]
