+++ doc/paper/joss/paper.bib
@@ -9,7 +9,7 @@
 }
 
 @article{JOSIEN2024109354,
-title = {Parallel and bias-free RSA algorithm for maximal Poisson-sphere sampling},
+title = {Parallel and bias-free RSA algorithm for maximal {Poisson}-sphere sampling},
 journal = {Computer Physics Communications},
 volume = {305},
 pages = {109354},
@@ -64,7 +64,7 @@ keywords = {RSA, HPC, MPI},
 
 
 @article{verlet1967computer,
-  title={Computer" experiments" on classical fluids. I. Thermodynamical properties of Lennard-Jones molecules},
+  title={Computer" experiments" on classical fluids. I. Thermodynamical properties of {Lennard-Jones} molecules},
   author={Verlet, Loup},
   journal={Physical review},
   volume={159},
@@ -83,7 +83,7 @@ keywords = {RSA, HPC, MPI},
   number={2-3},
   pages={140--152},
   year={2012},
-  doi={10.1504/pcfd.2012.047457},
+  doi={10.1504/PCFD.2012.047457},
   publisher={Inderscience Publishers}
 }
 
