+++ doc/paper/joss/paper.bib
@@ -15,7 +15,7 @@ volume = {305},
 pages = {109354},
 year = {2024},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2024.109354},
+doi = {10.1016/j.cpc.2024.109354},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465524002777},
 author = {Marc Josien and Raphaël Prat},
 keywords = {RSA, HPC, MPI},
