### 2025-02-06 2a6c8e9
add linear compressive force for stl meshes (#121)
Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
