+++ src/driver/include/exaDEM/stl_mesh.h
@@ -49,7 +49,8 @@ namespace exaDEM
     exanb::Vec3d center = Vec3d{0,0,0}; /**< Center position of the STL mesh. */
     exanb::Vec3d vel = Vec3d{0,0,0};    /**< Velocity of the STL mesh. */
     exanb::Vec3d vrot = Vec3d{0,0,0};   /**< Angular velocity of the STL mesh. */
-    exanb::Quaternion quat = {1,0,0,0};             /**< Quaternion of the STL mesh. */
+    exanb::Quaternion quat = {1,0,0,0}; /**< Quaternion of the STL mesh. */
+    double surface = -1;                /**< Surface, used with linear_compression_motion. */
     double mass = std::numeric_limits<double>::max() / 4; /**< Mass of the ball */
   };
 }
@@ -75,16 +76,15 @@ namespace YAML
       if( check(node, "vrot") )   { v.vrot = node["vrot"].as<Vec3d>(); }
       if( check(node, "mass") ) { v.mass = node["mass"].as<double>(); }
       if( check(node, "quat") ) { v.quat = node["quat"].as<exanb::Quaternion>(); }
+      if( check(node, "surface") ) { v.surface = node["surface"].as<double>(); }
       return true;
     }
   };
 }
 
-
-
 namespace exaDEM
 {
-  const std::vector<MotionType> stl_valid_motion_types = {STATIONARY, LINEAR_MOTION, LINEAR_FORCE_MOTION};
+  const std::vector<MotionType> stl_valid_motion_types = {STATIONARY, LINEAR_MOTION, LINEAR_FORCE_MOTION, LINEAR_COMPRESSIVE_MOTION};
 
   using namespace exanb;
   /**
@@ -114,11 +114,12 @@ namespace exaDEM
     void print()
     {
       lout << "Driver Type: MESH STL" << std::endl;
-      lout << "Name   : " << shp.m_name << std::endl;
-      lout << "Center : " << center << std::endl;
-      lout << "Vel    : " << vel << std::endl;
-      lout << "AngVel : " << vrot << std::endl;
-      lout << "Quat   : " << quat.w << " " << quat.x << " " << quat.y << " " << quat.z << std::endl;
+      lout << "Name               : " << shp.m_name << std::endl;
+      lout << "Center             : " << center << std::endl;
+      lout << "Velocity           : " << vel << std::endl;
+      lout << "Angular Velocity   : " << vrot << std::endl;
+      lout << "Orientation        : " << quat.w << " " << quat.x << " " << quat.y << " " << quat.z << std::endl;
+      if(surface > 0.0) lout << "Surface            : " << surface << std::endl;
       lout << "Number of faces    : " << shp.get_number_of_faces() << std::endl;
       lout << "Number of edges    : " << shp.get_number_of_edges() << std::endl;
       lout << "Number of vertices : " << shp.get_number_of_vertices() << std::endl;
@@ -158,11 +159,39 @@ namespace exaDEM
         lout << "Please, define a positive mass." << std::endl;
         std::exit(EXIT_FAILURE);
       }
+
+      if( is_compressive() )
+      {
+        double s = shp.compute_surface();
+        if( surface <= 0 )
+        {
+          lout << "\033[31m[STL Error]: The surface value must be positive for LINEAR_COMPRESSIVE_FORCE. You need to specify surface: XX in the 'state' slot.\033[0m" << std::endl;
+          lout << "\033[33m[STL Error]: The computed surface of all faces is: " << s << "\033[0m" << std::endl;
+          std::exit(EXIT_FAILURE);
[…]
