+++ doc/paper/joss/paper.bib
@@ -31,7 +31,7 @@ keywords = {RSA, HPC, MPI},
 
 @misc{gitrockable,
   author = {Richefeu, V.},
-  title = {Rocakble},
+  title = {Rockable},
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
