### 2025-02-09 3d60ac6
add an example for stl meshes (piston) with linear force motion and linear compression motion
