+++ src/interaction/classify_interactions.cpp
@@ -154,15 +154,15 @@ namespace exaDEM
       {
         return;
       }
-      /*auto [cell_ptr, cell_size] = traversal_real->info();
+      auto [cell_ptr, cell_size] = traversal_real->info();
       if (!ic.has_value())
         ic->initialize();
       ic->classify(*ges, cell_ptr, cell_size);
       ic->prefetch_memory_on_gpu(); // GPU only*/
 
-     auto &olds = *ic_olds;
+     //auto &olds = *ic_olds;
             
-      auto& c = *ic;
+      //auto& c = *ic;
       
       /*auto [data, size] = c.get_info(0);
       
@@ -200,7 +200,7 @@ namespace exaDEM
       waves[4].clear();
       waves[4] = *interaction_type4;*/
       
-      for(int type = 0; type < 13; type++)
+      /*for(int type = 0; type < 13; type++)
       {
       
       auto [data, size] = c.get_info(type);
@@ -239,7 +239,7 @@ namespace exaDEM
       find_common_elements<<<numBlocks, blockSize>>>( keys.data(), old.keys, size, old.size, interactions.ft_x, interactions.ft_y, interactions.ft_z, old.ft_x, old.ft_y, old.ft_z, interactions.mom_x, interactions.mom_y, interactions.mom_z, old.mom_x, old.mom_y, old.mom_z, indices_sorted.data(), old.indices);
       
       }
-      }
+      }*/
       
      //printf("END_CLASSIFY\n");  
     }
