### 2025-02-21 c9891ad
first refactoring pass
