+++ src/sphere/CMakeLists.txt
@@ -15,4 +15,4 @@
 # specific language governing permissions and limitations
 # under the license.
 set(exadem_sphere_LINK_LIBRARIES exanbDefBox exanbCompute exanbIO exanbMPI exanbParticleNeighbors exanbGridCellParticles exanbExtraStorage exadem_driver exadem_shape exadem_force_field exadem_grid exadem_interaction)
-xstamp_add_plugin(exadem_sphere ${CMAKE_CURRENT_SOURCE_DIR})
+onika_add_plugin(exadem_sphere ${CMAKE_CURRENT_SOURCE_DIR})
