+++ src/field/radius_from_shape.cpp
@@ -17,9 +17,9 @@ specific language governing permissions and limitations
 under the License.
 */
 //#pragma xstamp_cuda_enable // DO NOT REMOVE THIS LINE
-#include <exanb/core/operator.h>
-#include <exanb/core/operator_slot.h>
-#include <exanb/core/operator_factory.h>
+#include <onika/scg/operator.h>
+#include <onika/scg/operator_slot.h>
+#include <onika/scg/operator_factory.h>
 #include <exanb/core/make_grid_variant_operator.h>
 #include <exanb/core/parallel_grid_algorithm.h>
 #include <exanb/core/grid.h>
