### 2025-02-21 be29c4c
suppressed all #pragma xstamp_*
