+++ src/driver/driver_displ_over.cu
@@ -16,7 +16,6 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
  */
-//#pragma xstamp_cuda_enable
 
 #include <mpi.h>
 #include <onika/scg/operator.h>
