+++ src/polyhedron/vertex_displ_over.cu
@@ -16,9 +16,7 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
  */
-#pragma xstamp_cuda_enable
 
-#pragma xstamp_grid_variant
 
 #include <onika/scg/operator.h>
 #include <onika/scg/operator_slot.h>
