+++ src/io/paraview_file_name.cpp
@@ -19,7 +19,7 @@ under the License.
 #include <onika/scg/operator.h>
 #include <onika/scg/operator_factory.h>
 #include <onika/scg/operator_slot.h>
-#include <exanb/core/log.h>
+#include <onika/log.h>
 #include <exanb/core/string_utils.h>
 
 #include <memory>
