+++ src/io/print_simulation_state.cpp
@@ -19,7 +19,7 @@ under the License.
 #include <onika/scg/operator.h>
 #include <onika/scg/operator_slot.h>
 #include <onika/scg/operator_factory.h>
-#include <exanb/core/log.h>
+#include <onika/log.h>
 #include <exanb/core/string_utils.h>
 #include <exanb/core/physics_constants.h>
 
