+++ src/field/include/exaDEM/random_quaternion.h
@@ -21,7 +21,7 @@ under the License.
 #include <onika/math/basic_types.h>
 #include <exanb/compute/compute_cell_particles.h>
 #include <exanb/grid_cell_particles/particle_region.h>
-#include <exanb/core/quaternion_operators.h>
+#include <onika/math/quaternion_operators.h>
 
 namespace exaDEM
 {
