### 2025-02-21 a4c8168
renamed CONSTRUCTOR_FUNCTION to ONIKA_AUTORUN_INIT(xxx)
