+++ src/interaction/classify_interactions.cpp
@@ -69,5 +69,5 @@ namespace exaDEM
   template <class GridT> using ClassifyInteractionsTmpl = ClassifyInteractions<GridT>;
 
   // === register factories ===
-  CONSTRUCTOR_FUNCTION { OperatorNodeFactory::instance()->register_factory("classify_interactions", make_grid_variant_operator<ClassifyInteractionsTmpl>); }
+  ONIKA_AUTORUN_INIT(classify_interactions) { OperatorNodeFactory::instance()->register_factory("classify_interactions", make_grid_variant_operator<ClassifyInteractionsTmpl>); }
 } // namespace exaDEM
