### 2025-02-21 575b0d9
2 more plugins compile
