+++ src/CMakeLists.txt
@@ -19,9 +19,9 @@
 # === exaDEM application package  ===
 # ===================================
 
-#add_subdirectory(grid)
+add_subdirectory(grid)
 add_subdirectory(shape)
-#add_subdirectory(field)
+add_subdirectory(field)
 #add_subdirectory(numericalscheme)
 #add_subdirectory(forcefield)
 #add_subdirectory(io)
