+++ src/field/CMakeLists.txt
@@ -14,5 +14,5 @@
 # kind, either express or implied.  see the license for the
 # specific language governing permissions and limitations
 # under the license.
-set(exadem_field_LINK_LIBRARIES exanbCore exanbDefBox exanbCompute exanbIO exanbMPI exanbParticleNeighbors exanbGridCellParticles exadem_shape exadem_interaction)
+set(exadem_field_LINK_LIBRARIES exanbCore exanbDefBox exanbCompute exanbIO exanbMPI exanbParticleNeighbors exanbGridCellParticles exadem_shape)
 onika_add_plugin(exadem_field ${CMAKE_CURRENT_SOURCE_DIR})
