### 2025-02-21 2547b4d
compiles, but only with gcc-11
