+++ src/driver/include/exaDEM/stl_mesh.h
@@ -16,6 +16,7 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
  */
+ 
 #pragma once
 #include <onika/math/basic_types.h>
 #include <onika/math/quaternion.h>
@@ -25,6 +26,7 @@ under the License.
 #include <exaDEM/shape_reader.hpp>
 //#include <exaDEM/interaction/interaction.hpp>
 #include <filesystem>
+#include <onika/physics/units.h>
 
 namespace exaDEM
 {
@@ -60,8 +62,7 @@ namespace YAML
   using exaDEM::Stl_params;
   using exaDEM::MotionType;
   using exanb::lerr;
-  using exanb::Quantity;
-  using exanb::UnityConverterHelper;
+  using onika::physics::Quantity;
 
   template <> struct convert<Stl_params>
   {
