+++ src/forcefield/include/exaDEM/contact_parameters.h
@@ -16,8 +16,10 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
+
 #pragma once
 #include <onika/math/basic_types.h>
+#include <onika/physics/units.h>
 
 namespace exaDEM
 {
@@ -53,8 +55,7 @@ namespace YAML
 {
   using exaDEM::ContactParams;
   using exanb::lerr;
-  using exanb::Quantity;
-  using exanb::UnityConverterHelper;
+  using onika::physics::Quantity;
 
   template <> struct convert<ContactParams>
   {
