+++ src/driver/include/exaDEM/surface.h
@@ -16,9 +16,11 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
  */
+ 
 #pragma once
 #include <onika/math/basic_types.h>
 #include <exaDEM/driver_base.h>
+#include <onika/physics/units.h>
 
 namespace exaDEM
 {
@@ -45,8 +47,7 @@ namespace YAML
   using exaDEM::Surface_params;
   using exaDEM::MotionType;
   using exanb::lerr;
-  using exanb::Quantity;
-  using exanb::UnityConverterHelper;
+  using onika::physics::Quantity;
 
   template <> struct convert<Surface_params>
   {
