+++ src/rsa/init_rsa.cpp
@@ -13,7 +13,7 @@
 #include <onika/math/basic_types_stream.h>
 #include <onika/log.h>
 //#include "ustamp/vector_utils.h"
-#include <exanb/core/file_utils.h>
+#include <onika/file_utils.h>
 #include <exanb/core/domain.h>
 #include <exanb/core/check_particles_inside_cell.h>
 #include <exanb/core/simple_block_rcb.h>
