+++ src/io/read_xyz.cpp
@@ -31,7 +31,7 @@ under the License.
 #include <onika/math/basic_types_stream.h>
 #include <onika/log.h>
 //#include "ustamp/vector_utils.h"
-#include <exanb/core/file_utils.h>
+#include <onika/file_utils.h>
 #include <exanb/core/domain.h>
 #include <exanb/core/check_particles_inside_cell.h>
 
