+++ src/CMakeLists.txt
@@ -24,9 +24,9 @@ add_subdirectory(shape)
 add_subdirectory(field)
 add_subdirectory(numericalscheme)
 add_subdirectory(driver)
-#add_subdirectory(forcefield)
+add_subdirectory(forcefield)
+add_subdirectory(interaction)
 #add_subdirectory(io)
-#add_subdirectory(interaction)
 #add_subdirectory(sphere)
 #add_subdirectory(polyhedron)
 #add_subdirectory(analysis)
