+++ src/io/include/exaDEM/simulation_state.h
@@ -16,6 +16,7 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
+
 #pragma once
 
 #include <onika/math/basic_types.h>
