+++ src/analysis/include/exaDEM/AnalysisManager.hpp
@@ -30,7 +30,7 @@ namespace exaDEM
       void add_element(std::string name, T& new_element, std::string format)
       {
         header << name << " ";
-        std::string element = exanb::format_string(format, new_element);
+        std::string element = onika::format_string(format, new_element);
         line << element << " "; 
       }
 
