+++ src/CMakeLists.txt
@@ -26,10 +26,10 @@ add_subdirectory(numericalscheme)
 add_subdirectory(driver)
 add_subdirectory(forcefield)
 add_subdirectory(interaction)
-#add_subdirectory(io)
+add_subdirectory(analysis)
+add_subdirectory(io)
 #add_subdirectory(sphere)
 #add_subdirectory(polyhedron)
-#add_subdirectory(analysis)
 
 
 #if(USE_RSA)
