+++ src/io/print_simulation_state.cpp
@@ -121,7 +121,7 @@ namespace exaDEM
         }
         else
         {
-          lb_value = format_string("%.1e", lb_inbalance);
+          lb_value = onika::format_string("%.1e", lb_inbalance);
         }
       }
 
@@ -146,7 +146,7 @@ namespace exaDEM
       double avg_act_I = double(active_interactions) / sim_info.particle_count();
       double avg_I = double(total_interactions) / sim_info.particle_count();
 
-      std::string line = format_string("%9ld % .6e %13ld  %c %c %8s %.3e    %9.3f  %9.3f % .3e % .3e % .3e % .3e % .4e",
+      std::string line = onika::format_string("%9ld % .6e %13ld  %c %c %8s %.3e    %9.3f  %9.3f % .3e % .3e % .3e % .3e % .4e",
                                        *timestep,                           // %9ld
                                        *physical_time,                      // %.6e
                                        sim_info.particle_count(),           // %13ld
