### 2025-02-24 26302cc
updated build instrcutions to get proper particle type integer definition
