+++ docs/build-examples.txt
@@ -1,5 +1,8 @@
+
+
+
 # Tested with 
-# Ubuntu 24.04 , g++-13.4.0 and Cuda 12.8
+# Ubuntu 24.04 , g++-11.4.0 and Cuda 12.8
 #
 # Note: cofigure exaNBody with EXANB_BUILD_CONTRIB_MD=ON to enable contributed MD potentials
 # in exaStamp (i.e ML SNAP)
@@ -20,4 +23,13 @@ source ./exaDEM
 
 # run executable, with or without profiling tools
 /home/carrardt/local/onika/bin/onika-exec mytest.msp
+
+
+# Additional notes :
+# follow instructions in doc/build-exemples.txt in both onika and exaNBody
+# use branch main for onika and release-2.0 for exaNBody
+# IMPORTANT !!
+# to have 32bits integer particle types in exaDEM, you must set the following cmake
+# variable when configuring exaNBody :
+# XNB_PARTICLE_TYPE_INT=uint32_t
 
