### 2025-02-24 d5aa003
renamed main cnofig file to main-config.msp
