+++ src/grid/grid_flavor_dem.cpp
@@ -54,6 +54,6 @@ namespace exaDEM
   };
 
   // === register factories ===
-  ONIKA_AUTORUN_INIT(grid_flavor) { OperatorNodeFactory::instance()->register_factory("grid_flavor_dem", make_compatible_operator<InitGridFlavorNode<GridFromFieldSet<DEMFieldSet>>>); }
+  ONIKA_AUTORUN_INIT(grid_flavor) { OperatorNodeFactory::instance()->register_factory("grid_flavor_dem", make_compatible_operator< InitGridFlavorNode< GridFromFieldSet<DEMFieldSet> > >); }
 
 } // namespace exaDEM
