+++ src/driver/include/exaDEM/drivers.h
@@ -24,7 +24,7 @@ under the License.
 #include <exaDEM/ball.h>
 #include <exaDEM/stl_mesh.h>
 #include <exaDEM/undefined_driver.h>
-#include <variant>
+#include <onika/flat_tuple.h>
 
 namespace exaDEM
 {
@@ -36,17 +36,77 @@ namespace exaDEM
      * @brief Alias template for a CUDA memory managed vector.
      * @tparam T The type of elements in the vector.
      */
-    template <typename T> using vector_t = onika::memory::CudaMMVector<T>;
-    using data_t = std::variant<exaDEM::Cylinder, exaDEM::Surface, exaDEM::Ball, exaDEM::Stl_mesh, exaDEM::UndefinedDriver>;
-    using driver_t = DRIVER_TYPE;
-    vector_t<driver_t> m_type; /**< Vector storing the types of drivers. */
-    vector_t<data_t> m_data;   /**< Vector storing the data of drivers. */
+    template <typename T> using vector_t = onika::memory::CudaMMVector<T>;    
+
+    struct DriverTypeAndIndex
+    {
+      DRIVER_TYPE m_type = DRIVER_TYPE::UNDEFINED;
+      int m_index = -1;
+    };
+    
+    vector_t<DriverTypeAndIndex> m_type_index; /**< Vector storing the types of drivers. */
+    
+    //vector_t<data_t> m_data;   /**< Vector storing the data of drivers. */
+    onika::FlatTuple< vector_t<Cylinder> , vector_t<Surface> , vector_t<Ball> , vector_t<Stl_mesh> > m_data;
 
     /**
      * @brief Get the size of the Drivers collection.
      * @return The size of the Drivers collection.
      */
-    inline size_t get_size() { return m_type.size(); }
+    inline size_t get_size() const { return m_type_index.size(); }
+
+    template<size_t driver_type>
+    inline const auto & get_driver_vec() const
+    {
+      static_assert( driver_type != DRIVER_TYPE::UNDEFINED );
+      return m_data.get_nth_const< driver_type >();
+    }
+
+    template<size_t driver_type>
+    inline auto & get_driver_vec()
+    {
+      static_assert( driver_type != DRIVER_TYPE::UNDEFINED );
+      return m_data.get_nth< driver_type >();
+    }
+
+    template<class T>
+    inline const T& get_typed_driver(const int idx) const
+    {
+      constexpr DRIVER_TYPE t = get_type<T>();
+      static_assert( t != DRIVER_TYPE::UNDEFINED );
+      const auto & driver_vec = m_data.get_nth_const<t>();
+      assert( idx>=0 && idx<m_type_index.size() );
+      assert( m_type_index[idx].m_type == t );
+      assert( m_type_index[idx].m_index >= 0 && m_type_index[idx].m_index < driver_vec.size() );
+      return driver_vec[m_type_index[idx].m_index];
+    }
+
+    template<class T>
+    inline T& get_typed_driver(const int idx)
+    {
+      constexpr DRIVER_TYPE t = get_type<T>();
+      static_assert( t != DRIVER_TYPE::UNDEFINED );
+      auto & driver_vec = m_data.get_nth<t>();
+      assert( idx>=0 && idx<m_type_index.size() );
+      assert( m_type_index[idx].m_type == t );
+      assert( m_type_index[idx].m_index >= 0 && m_type_index[idx].m_index < driver_vec.size() );
+      return driver_vec[m_type_index[idx].m_index];
+    }
+
+    template<class FuncT>
+    inline auto apply(const int idx , const FuncT& func)
+    {
+      assert( idx>=0 && idx<m_type_index.size() );
+      DRIVER_TYPE t = m_type_index[idx].m_type;
[…]
