+++ src/driver/include/exaDEM/cylinder.h
@@ -26,7 +26,6 @@ namespace exaDEM
 {
   using namespace exanb;
 
-  using namespace exanb;
   struct Cylinder_params
   {
     double radius = -1;       /**< Radius of the cylinder. */
@@ -107,7 +106,7 @@ namespace exaDEM
     /**
      * @brief Print information about the cylinder.
      */
-    void print()
+    inline void print() const
     {
       lout << "Driver Type: Cylinder" << std::endl;
       lout << "Radius: " << radius << std::endl;
@@ -208,3 +207,22 @@ namespace exaDEM
     }
   };
 } // namespace exaDEM
+
+
+namespace onika { namespace memory
+{
+
+  template<>
+  struct MemoryUsage< exaDEM::Cylinder >
+  {
+    static inline size_t memory_bytes(const exaDEM::Cylinder& obj)
+    {
+      const exaDEM::Cylinder_params * cparms = &obj;
+      const exaDEM::Driver_params * dparms = &obj;
+      return onika::memory::memory_bytes( *cparms , *dparms );
+    }
+  };
+
+} }
+
+
