### 2025-02-25 d4fe87d
version without std::variant compiling with GCC-11, still issues with GCC-13
