+++ src/interaction/dump_contact_network.cpp
@@ -94,7 +94,7 @@ namespace exaDEM
 			}
 
 			std::string file = *filename + "/%06d.vtp";
-			file = format_string(file,  rank);
+			file = onika::format_string(file,  rank);
 			auto ids = manager.create_indirection_array();
 			manager.fill_position(ids);
 			manager.fill_connect_and_value(ids);
