### 2025-02-26 #7484 @jedbrown
Two small edits:
- [ ] This line needs to remove case protection. https://github.com/Collab4exaNBody/exaDEM/blob/main/docs/paper/joss/paper.bib#L136
- [ ] This line needs protection for `{Blaze-DEM}`

It looks good otherwise, so after doing that, you may proceed with:
