+++ docs/paper/joss/paper.bib
@@ -79,7 +79,7 @@
 }
 
 @article{govender2018study,
-  title={A study of shape non-uniformity and poly-dispersity in hopper discharge of spherical and polyhedral particle systems using the Blaze-DEM GPU code},
+  title={A study of shape non-uniformity and poly-dispersity in hopper discharge of spherical and polyhedral particle systems using the {Blaze-DEM} GPU code},
   author={Govender, Nicolin and Wilke, Daniel N and Pizette, Patrick and Abriak, Nor-Edine},
   journal={Applied Mathematics and Computation},
   volume={319},
