+++ docs/paper/joss/paper.bib
@@ -133,7 +133,7 @@
 }
 
 @book{richefeu_2016,
-  title = {Modeling {{Gravity Hazards}} from {{Rockfalls}} to {{Landslides}}},
+  title = {Modeling Gravity Hazards from Rockfalls to Landslides},
   author = {Richefeu, Vincent and Villard, Pascal},
   editor = {Richefeu, Vincent and Villard, Pascal},
   year = {2016},
