+++ doc/paper/joss/paper.bib
@@ -98,7 +98,7 @@ keywords = {RSA, HPC, MPI},
 }
 
 @article{govender2018study,
-  title={A study of shape non-uniformity and poly-dispersity in hopper discharge of spherical and polyhedral particle systems using the Blaze-DEM GPU code},
+  title={A study of shape non-uniformity and poly-dispersity in hopper discharge of spherical and polyhedral particle systems using the {Blaze-DEM} GPU code},
   author={Govender, Nicolin and Wilke, Daniel N and Pizette, Patrick and Abriak, Nor-Edine},
   journal={Applied Mathematics and Computation},
   volume={319},
@@ -158,7 +158,7 @@ keywords = {RSA, HPC, MPI},
 }
 
 @book{richefeu_2016,
-  title = {Modeling {{Gravity Hazards}} from {{Rockfalls}} to {{Landslides}}},
+  title = {Modeling Gravity Hazards from Rockfalls to Landslides},
   author = {Richefeu, Vincent and Villard, Pascal},
   editor = {Richefeu, Vincent and Villard, Pascal},
   year = {2016},
