### 2025-02-26 #7484 @rprat-pro
@jedbrown I have updated the bib file.

The lastest tagged version is `1.0.2`, I uploaded it on Zenodo at: https://zenodo.org/records/14929748 .
The DOI is: https://doi.org/10.5281/zenodo.14929748 .
