### 2025-02-27 #7484 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
