+++ doc/paper/joss/paper.bib
@@ -8,16 +8,30 @@
   organization={Springer}
 }
 
+@article{JOSIEN2024109354,
+title = {Parallel and bias-free RSA algorithm for maximal {Poisson}-sphere sampling},
+journal = {Computer Physics Communications},
+volume = {305},
+pages = {109354},
+year = {2024},
+issn = {0010-4655},
+doi = {10.1016/j.cpc.2024.109354},
+url = {https://www.sciencedirect.com/science/article/pii/S0010465524002777},
+author = {Marc Josien and Raphaël Prat},
+keywords = {RSA, HPC, MPI},
+}
+
 @article{ciccotti1987simulation,
   title={Simulation of liquids and solids},
   author={Ciccotti, Giovanni and Frenkel, Daan and Mc Donald, Ian R},
   year={1987},
+  doi = {},
   publisher={Elsevier Science Pub. Co. Inc., New York, NY}
 }
 
 @misc{gitrockable,
   author = {Richefeu, V.},
-  title = {Rocakble},
+  title = {Rockable},
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -32,6 +46,7 @@
   volume={271},
   pages={108171},
   year={2022},
+  doi = {10.1016/j.cpc.2021.108171},
   publisher={Elsevier}
 }
 
@@ -43,18 +58,20 @@
   number={1},
   pages={1--19},
   year={1995},
+  doi = {10.2172/10176421},
   publisher={Elsevier}
 }
 
 
 @article{verlet1967computer,
-  title={Computer" experiments" on classical fluids. I. Thermodynamical properties of Lennard-Jones molecules},
+  title={Computer" experiments" on classical fluids. I. Thermodynamical properties of {Lennard-Jones} molecules},
   author={Verlet, Loup},
   journal={Physical review},
   volume={159},
   number={1},
   pages={98},
   year={1967},
+  doi={10.1103/physrev.159.98 },
   publisher={APS}
 }
 
@@ -66,6 +83,7 @@
   number={2-3},
   pages={140--152},
   year={2012},
+  doi={10.1504/PCFD.2012.047457},
   publisher={Inderscience Publishers}
 }
 
@@ -75,16 +93,18 @@
   booktitle={Supercomputing Frontiers: 4th Asian Conference, SCFA 2018, Singapore, March 26-29, 2018, Proceedings 4},
   pages={159--178},
   year={2018},
+  doi={10.1007/978-3-319-69953-0_10},
   organization={Springer}
 }
 
 @article{govender2018study,
-  title={A study of shape non-uniformity and poly-dispersity in hopper discharge of spherical and polyhedral particle systems using the Blaze-DEM GPU code},
+  title={A study of shape non-uniformity and poly-dispersity in hopper discharge of spherical and polyhedral particle systems using the {Blaze-DEM} GPU code},
   author={Govender, Nicolin and Wilke, Daniel N and Pizette, Patrick and Abriak, Nor-Edine},
   journal={Applied Mathematics and Computation},
   volume={319},
   pages={318--336},
   year={2018},
+  doi={10.1016/j.amc.2017.03.037},
   publisher={Elsevier}
 }
 
@@ -94,6 +114,7 @@
   booktitle={Euro-Par 2014: Parallel Processing Workshops: Euro-Par 2014 International Workshops, Porto, Portugal, August 25-26, 2014, Revised Selected Papers, Part II 20},
   pages={121--132},
   year={2014},
+  doi={10.1007/978-3-319-14313-2_11},
   organization={Springer}
 }
 
@@ -102,6 +123,7 @@
[…]
