### 2025-02-27 7494764
update joss paper from the branch joss_paper (#125)
Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>
