### 2025-02-27 #7484 @rprat-pro
There was a minor problem updating the paper in the `v1.0.2` branch. Everything's fine for me now.
