+++ doc/paper/joss/paper.bib
@@ -29,14 +29,20 @@ keywords = {RSA, HPC, MPI},
   publisher={Elsevier Science Pub. Co. Inc., New York, NY}
 }
 
-@misc{gitrockable,
-  author = {Richefeu, V.},
-  title = {Rockable},
-  year = {2023},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/richefeu/rockable.git}},
+@softwareversion{gitrockable,
+  TITLE = {{Rockable}},
+  AUTHOR = {Richefeu, Vincent and Combe, Ga{\"e}l and Villard, Pascal and Delenne, Jean-Yves and Amarsid, Lhassan and Nezamabadi, Saeid and Radjai, Farhang and Vanson, Jean-Mathieu and Prat, Raphael and Mutabaruka, Patrick},
+  URL = {https://hal.univ-grenoble-alpes.fr/hal-04933604},
+  NOTE = {},
+  YEAR = {2025},
+  MONTH = Feb,
+  SWHID = {swh:1:dir:e981bb9fae3c064abb9ddff3bbd19ad1a3b698e7;origin=https://hal.archives-ouvertes.fr/hal-04933604;visit=swh:1:snp:3a5c01138e6e944de9ff6ec37770d7dd7744c5cc;anchor=swh:1:rel:bf7c270991242bd75e4f44c0d525cb4493d903ba;path=/},
+  REPOSITORY = {https://github.com/richefeu/rockable},
+  LICENSE = {CeCILL-B FREE SOFTWARE LICENSE AGREEMENT},
+  HAL_ID = {hal-04933604},
+  HAL_VERSION = {v1},
 }
+
 
  
 @article{thompson2022lammps,
