### 2025-02-28 #7484 @kyleniemeyer
@rprat-pro a few minor things to address before publishing:
- the Zenodo archive has some random line breaks, perhaps left over from a copy-paste
- in the paper, on page 4, the in-text reference of Figure 1 has a repeated "figure": `see figure Figure 1`. I believe the figure reference command automatically includes "Figure".
- The text does not reference Figure 2, could you add a mention somewhere or remove the figure?
- The Ciccotti et al. reference only has a title and year. If that is a book, please add publisher info & ISBN.
