+++ doc/paper/joss/paper.bib
@@ -22,11 +22,12 @@ keywords = {RSA, HPC, MPI},
 }
 
 @article{ciccotti1987simulation,
-  title={Simulation of liquids and solids},
-  author={Ciccotti, Giovanni and Frenkel, Daan and Mc Donald, Ian R},
-  year={1987},
+  title = {Simulation of liquids and solids},
+  author = {Ciccotti, Giovanni and Frenkel, Daan and Mc Donald, Ian R},
+  year = {1987},
   doi = {},
-  publisher={Elsevier Science Pub. Co. Inc., New York, NY}
+  publisher = {Elsevier Science Pub. Co. Inc., New York, NY},
+  isbn = {978-0-44487-062-9}
 }
 
 @softwareversion{gitrockable,
