### 2025-02-28 3e74c2b
V1.0.2 (#126)
* fix bug when we use the GPU parallelization with the operator set_rand_velocity

* Update README.md

* add mirror example (#34)

Co-authored-by: rprat-pro <rp269144@pleiades1217.intra.cea.fr>

* Integration first dev carlo gpu (#35)

* update data structure for gpu + add gpu parallelization with classifier for spherical particles

* refactoring sym/no sym stuff for the contact law

* update hooke force with a classifier (block parallel for)

* fix issue in classifier sphere gpu (// exec context)

* remove size from wrapper

* fix the computation of the cylinder center

* move f
[…]
