+++ src/interaction/include/exaDEM/interaction/classifier_for_all.hpp
@@ -1,0 +1,193 @@
+#pragma once
+#include <onika/parallel/parallel_execution_context.h>
+#include <onika/parallel/block_parallel_for.h>
+#include <onika/parallel/parallel_for.h>
+
+namespace exaDEM
+{
+  using namespace onika::parallel;
+
+  /**
+   * @brief Namespace for utilities related to tuple manipulation.
+   */
+  namespace tuple_helper
+  {
+    template <size_t... Is> struct index
+    {
+    };
+
+    template <size_t N, size_t... Is> struct gen_seq : gen_seq<N - 1, N - 1, Is...>
+    {
+    };
+
+    template <size_t... Is> struct gen_seq<0, Is...> : index<Is...>
+    {
+    };
+  } // namespace tuple_helper
+
+  struct AnalysisDataPackerNull
+  {
+    template <typename... Args> ONIKA_HOST_DEVICE_FUNC inline void operator()(Args &&...args) const
+    { /* do nothing */
+    }
+  };
+
+  /**
+   * @struct AnalysisDataPacker
+   * @brief Packs analysis data for interactions, including overlap (dn), contact point, and forces.
+   *
+   * This structure stores interaction-related data such as overlap (dn), contact point,
+   * normal force (fn), and tangential force (ft) into their respective buffers.
+   */
+  struct AnalysisDataPacker
+  {
+    double * __restrict__ dnp; /**< Pointer to the buffer storing overlap (dn) values between particles. */
+    Vec3d * __restrict__ cpp;  /**< Pointer to the buffer storing contact point positions. */
+    Vec3d * __restrict__ fnp;  /**< Pointer to the buffer storing normal force vectors (fn). */
+    Vec3d * __restrict__ ftp;  /**< Pointer to the buffer storing tangential force vectors (ft). */
+
+    /**
+     * @brief Constructor that initializes the data packer with buffers from a classifier.
+     *
+     * @param ic The classifier used to access the buffers for a given interaction type.
+     * @param type The interaction type identifier to retrieve the appropriate buffers.
+     *
+     * The constructor retrieves the buffers for overlap (dn), contact points, normal forces,
+     * and tangential forces from the classifier based on the interaction type.
+     */
+    template <typename T> AnalysisDataPacker(Classifier<T> &ic, int type)
+    {
+      auto [_dnp, _cpp, _fnp, _ftp] = ic.buffer_p(type);
+      dnp = _dnp;
+      cpp = _cpp;
+      fnp = _fnp;
+      ftp = _ftp;
+    }
+
+    /**
+     * @brief Stores the interaction data at a specific index in the corresponding buffers.
+     *
+     * @param idx The index of the interaction.
+     * @param dn The overlap between interacting particles.
+     * @param contact The contact point vector.
+     * @param fn The normal force vector at the contact point.
+     * @param ft The tangential force vector at the contact point.
+     *
+     * This operator writes the given interaction data to the respective buffers
+     * for overlap, contact point, normal force, and tangential force.
+     */
+    ONIKA_HOST_DEVICE_FUNC inline void operator()(const uint64_t idx, const double dn, const Vec3d &contact, const Vec3d &fn, const Vec3d &ft) const
+    {
+      dnp[idx] = dn;
+      cpp[idx] = contact;
+      fnp[idx] = fn;
+      ftp[idx] = ft;
+    }
+  };
+
+  /**
+   * @brief Wrapper for applying a kernel function to elements of an array in parallel.
+   *
+   * This struct provides a mechanism to apply a kernel function to each element of
+   * an array in parallel. It stores a pointer to the array (`ptr`),
+   * the kernel function (`kernel`), and a tuple of parameters (`params`) to be passed
+   * to the kernel function along with each element.
+   *
[…]
