+++ doc/paper/joss/paper.bib
@@ -25,8 +25,8 @@ keywords = {RSA, HPC, MPI},
   title = {Simulation of liquids and solids},
   author = {Ciccotti, Giovanni and Frenkel, Daan and Mc Donald, Ian R},
   year = {1987},
+  publisher={Elsevier},
   doi = {},
-  publisher = {Elsevier Science Pub. Co. Inc., New York, NY},
   isbn = {978-0-44487-062-9}
 }
 
