### 2025-02-28 #7484 @rprat-pro
Hello @kyleniemeyer and thank you :

-  `the Zenodo archive has some random line breaks, perhaps left over from a copy-paste.` I found one in the description.
- `in the paper, on page 4, the in-text reference of Figure 1 has a repeated "figure": see figure Figure 1. I believe the figure reference command automatically includes "Figure". ` Indeed, it's been fixed.
- `the text does not reference Figure 2, could you add a mention somewhere or remove the figure?` The mention is in the line 116.
- `The Ciccotti et al. reference only has a title and year. If that is a book, please add publisher info & ISBN.`  I've added the ISBN number for this quote. On the other hand, the “publisher” is in the bibtex but it's a mystery why it's not displayed. On the other hand, I noticed that the ISBN is not displayed for the citation “Modeling gravity hazards from rockfalls to landslides”, whereas the publisher is visible. A real mystery ...
- I've updated the citation for the github with Rockable with a HAL repository referring to the latest version of Rockable (v1).
