### 2025-02-28 #7484 @kyleniemeyer
Ah, I think the issue is that the type needs to be `book` for Ciccotti et al., it is currently `article`.

I imagine the "Modeling gravity..." reference (which is a `book` type) does not include the ISBN since there is a DOI.
