+++ doc/paper/joss/paper.bib
@@ -21,7 +21,7 @@ author = {Marc Josien and Raphaël Prat},
 keywords = {RSA, HPC, MPI},
 }
 
-@article{ciccotti1987simulation,
+@book{ciccotti1987simulation,
   title = {Simulation of liquids and solids},
   author = {Ciccotti, Giovanni and Frenkel, Daan and Mc Donald, Ian R},
   year = {1987},
